% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{Change Door}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay}{A value indicating whether the contestant stays with their original selectoin.}

\item{opened.door}{A number between 1 and 3 representing the door opened by the host.}

\item{a.pick}{A number between 1 and 3 representing the contestant's original door selection.}
}
\value{
The function returns a number between 1 and 3 indicating the contestant's final door selection.
}
\description{
Determines the contestant's final door selection based on whether they stay or switch.
}
\details{
If the contestant stays, the function returns their original door selection.
If the contestant switches, the function returns the remaining unopened door.
}
\examples{
game <- create_game()
pick <- select_door()
opened <- open_goat_door(game, pick)
change_door(stay = FALSE, opened.door = opened, a.pick = pick)

}
