% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_door.R
\name{change_door_orig}
\alias{change_door_orig}
\title{Gives the player the opion to switch to a different door or to stay with the original door.}
\usage{
change_door_orig(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay}{= T True means that the player would stay with the original door, False would mean to switch to a different door}

\item{opened.door}{The door that has been opened with a goat behind it.}

\item{a.pick}{The original door selected by the player}

\item{door.choices}{The total number of doors in the game.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
This function allows the player to choose if they would like to stay on their original door or switch their choice to a different door given the results of the opened door.
}
\description{
Gives the player the opion to switch to a different door or to stay with the original door.
}
\examples{
## Player switches the door to the only remaining door.
change_door_orig(F, 2, 1)

## Randomly chooses a remaining door from those that remain.
change_door(F, opened.ddor = 3, a.pick = 2, door.choices = 8)

}
