% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_status.R
\name{check_status}
\alias{check_status}
\title{Check status}
\usage{
check_status(status, prize = NULL, chosen_door = NULL)
}
\arguments{
\item{status}{A logical vector of length \code{n_doors} (at least three) with the
status of the doors (\code{TRUE} for closed doors, \code{FALSE} for open doors): can
use element \code{status} of the output of \code{\link[=setup_MontyHall]{setup_MontyHall()}}.}

\item{prize}{A vector with \link[checkinput:is_natural]{natural} numbers
giving the indices of the doors behind which a prize is hidden: can use
element \code{prize} of the output of \code{\link[=setup_MontyHall]{setup_MontyHall()}}.}

\item{chosen_door}{A single \link[checkinput:is_natural]{natural} number with
the index of the door chosen by the contestant: can use the output of
\code{\link[=choose_door]{choose_door()}}.}
}
\value{
\link{NULL}, returned \link[=invisible]{invisibly}.
}
\description{
Helper function to check the content of object \code{status}.
}
\examples{
check_status(status = c(TRUE, TRUE, TRUE))
check_status(status = c(TRUE, TRUE, TRUE), prize = 1L, chosen_door = 1L)
try(check_status(status = c(TRUE, TRUE)))
try(check_status(status = c(TRUE, FALSE, FALSE)))
try(check_status(status = c(TRUE, TRUE, FALSE), prize = 3L))
try(check_status(status = c(TRUE, TRUE, TRUE), prize = 1L))
try(check_status(status = c(TRUE, TRUE, FALSE), prize = 1L, chosen_door = 3))

}
