% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_door.R
\name{choose_door}
\alias{choose_door}
\title{The contestant chooses a closed door}
\usage{
choose_door(status, chosen_door = NULL)
}
\arguments{
\item{status}{A logical vector of length \code{n_doors} (at least three) with the
status of the doors (\code{TRUE} for closed doors, \code{FALSE} for open doors): can
use element \code{status} of the output of \code{\link[=setup_MontyHall]{setup_MontyHall()}}.}

\item{chosen_door}{A single \link[checkinput:is_natural]{natural} number with
the index of the door chosen by the contestant: can use the output of
\code{\link[=choose_door]{choose_door()}}.}
}
\value{
A single \link{integer} with the index of the door chosen by the contestant.
}
\description{
The contestant chooses a closed door behind which the price might be
}
\details{
At least two doors should be closed. \code{chosen_door} should be provided if the
participant changes to another door.
}
\examples{
set.seed(314)
(setup <- setup_MontyHall(n_doors = 3, n_prizes = 1))
(chosen_door <- choose_door(status = setup$status))

}
\seealso{
\link{play_MontyHall} to play the game once and \link{play_MontyHall_repeated} to play
the game repeatedly.
}
