% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{create_game_5}
\alias{create_game_5}
\title{Create a 5-door Monty Hall game (3 goats, 2 cars)}
\usage{
create_game_5()
}
\value{
A character vector of length 5.
}
\description{
Generates a random arrangement of 3 goats and 2 cars behind five doors.
}
\examples{
create_game_5()

}
