% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine the winner}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{The contestants final door choice.}

\item{game}{The game with two goats and one car.}
}
\value{
Returns WIN or LOSE.
}
\description{
Checks if the final door has the car or a goat.
}
\details{
If the final door has the car the result is WIN.
If the final door has a goat the result is LOSE.
}
\examples{
game <- c("goat","car","goat")
determine_winner(2, game)
determine_winner(1, game)

}
