% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HolfordMeetsMonty.R
\docType{data}
\name{game}
\alias{game}
\title{\strong{TEST}}
\format{
An object of class \code{character} of length 3.

An object of class \code{character} of length 3.

An object of class \code{character} of length 3.
}
\usage{
game

game

game
}
\description{
\strong{TEST}

\strong{TEST}

\strong{TEST OF FUNCTION ONLY}
}
\keyword{datasets}
