% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{getfw}
\alias{getfw}
\title{Compute game win frequencies across game simulations of the Monty Hall Problem.}
\usage{
getfw(
  nsimulations = 5,
  niterations = 2,
  ndoors = 3,
  prize.index = NULL,
  doorswitch = 1,
  verbose.results = FALSE
)
}
\arguments{
\item{nsimulations}{Number of simulations to run.}

\item{niterations}{Number of iterations/games per simulation run.}

\item{ndoors}{Quantity of doors per game.}

\item{prize.index}{Door index for prize in each game.}

\item{doorswitch}{Frequency (0 - 100%) with which player switches doors when given the option.}
}
\value{
Vector of win fractions across simulations run, reflecting the numnber of iterations won over the total for that simulation.
}
\description{
Wrapper to compute win frequencies across simulations.
}
