% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{getggdat}
\alias{getggdat}
\title{Get results data formatted for ggplot2 functions.}
\usage{
getggdat(ld)
}
\arguments{
\item{ld}{List of win fractions.}
}
\value{
Formatted dataset.
}
\description{
Takes a list of win fractions and results the data in ggplot2 function-ready format.
}
