% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{getlineplot}
\alias{getlineplot}
\title{Generate a ggplot2 line plot.}
\usage{
getlineplot(
  ld,
  ptitle = "Plot title",
  ribbontype = "sd",
  xlim = NULL,
  ylim = NULL,
  xlab = "ndoors"
)
}
\arguments{
\item{ld}{List of win fractions.}

\item{ptitle}{Plot main title from `ggtitle()`.}

\item{xlim}{Vector of 2 x-axis coordinates (min and max), or NULL.}

\item{ylim}{Vector of 2 y-axis coordinates (min and max) or NULL.}
}
\value{
Line plot object.
}
\description{
Generates ggplot2 line plot object from the list of win fraction results.
}
