% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{getprettygif}
\alias{getprettygif}
\title{Generate a gif composite of violin and line plot animations}
\usage{
getprettygif(
  ld,
  gifname = "mh_ndoors.gif",
  plottype = c("composite_ndoors", "lineplots_winfractions")
)
}
\arguments{
\item{ld}{List of win fractions.}

\item{gifname}{Name of gif file to store.}

\item{plottype}{Type of plots for gif animation (either "composite_ndoors" or "lineplots_winfractions").}
}
\value{
Saves a gif file of the composite plots.
}
\description{
Generates a gif of composite plot animations for violin and line plots. Currently supports results from the `ndoors` experiment.
}
