% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{getprettyplots}
\alias{getprettyplots}
\title{Generate a composite of 3 ggplot2 plots.}
\usage{
getprettyplots(ld, topmain = "Top Title")
}
\arguments{
\item{ld}{List of win fractions.}

\item{topmain}{Top title of composite plot.}
}
\value{
Composite image of 3 ggplot2 plots.
}
\description{
Generates a composite of 3 ggplot2 plots from list of win fraction results.
}
