% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{mhsim}
\alias{mhsim}
\title{Run simulations of the Monty Hall problem}
\usage{
mhsim(
  niter = 100,
  seed = 1,
  ndoors = 3,
  nprize = 1,
  ndec1 = 1,
  ndec2 = 1,
  nrevealdif = 1,
  prize.index = NULL,
  selectdec1 = "random",
  doorswitch = "always",
  montyselect = "random",
  verbose.results = FALSE
)
}
\arguments{
\item{niter}{Number of simulation iterations or games.}

\item{seed}{Seed integer to set in `set.seed()`.}

\item{ndoors}{Total quantity of doors for game simulations.}

\item{nprize}{Total number of prizes for game simulations.}

\item{ndec1}{Door index to *always* choose in decision 1.}

\item{ndec2}{Door index to *always* choose in decision 2.}

\item{nrevealdif}{Number of doors Monty does not reveal between player decisions 1 and 2.}

\item{prize.index}{Door index for prize location in each game.}

\item{selectdec1}{Either "random" or a door index to choose for decision 1.}

\item{doorswitch}{Whether the player switches their door choice. Either "always" (100%) or some number between 0-1.}

\item{montyselect}{Indices of doors Monty reveals between decisions 1 and 2.}

\item{verbose.results}{Whether to return iteration/game details alongside results.}
}
\value{
A vector of game results, with simulation details if verbose.results = TRUE.
}
\description{
Run simulations of the Monty Hall problem. Default settings are canonical/classic game parameters, with arguments to change various parameters.
}
