% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MontyHall-package.R
\docType{package}
\name{MontyHall-package}
\alias{MontyHall}
\alias{MontyHall-package}
\title{MontyHall: Simulate the Monty Hall game}
\description{
Simulate the Monty Hall game as illustration of simulation studies.
}
\author{
\strong{Maintainer}: Jesse Alderliesten \email{jessealderliesten@hotmail.com} (\href{https://orcid.org/0000-0003-1132-4310}{ORCID})

}
\keyword{internal}
