% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monty Hall.R
\name{montyhall}
\alias{montyhall}
\title{몬티홀 딜레마함수 montyhall(k)}
\usage{
montyhall(k)
}
\description{
함수 설명
몬티홀문제를 3가지 경우로 나누어 어떠한 선택에도 결론(선택을 바꾸는 것이 유리하다.)을 도출할 수 있다는 것을
증명한 함수입니다.
각 케이스의 첫째 행은 쇼의 주인공이 모르는 문 뒤의 상황입니다. (k는 주인공이 선택한 문의 번호입니다.k=1,2,3)
두번째 행은 주인공의 선택(Choice)과 사회자가 열어서 보여준 염소(Open)의 위치입니다.
}
