% Generated by roxygen2 (4.0.1): do not edit by hand
\name{MontyHallStrat}
\alias{MontyHallStrat}
\title{Simulation of Let's Make a Deal game}
\usage{
MontyHallStrat(games, doors = 3, strategy = "switch", set.seed = 1,
  plot = F, output = F)
}
\arguments{
\item{games}{integer number of games in simulation}

\item{doors}{integer number of doors in each game}

\item{strategy}{character argument: 'switch', 'random', 'stay'}

\item{set.seed}{integer number for pseudorandom number generator}

\item{plot}{logical to plot cumulative avg plot, default FALSE}

\item{output}{logical return sim data to var or print to console}
}
\value{
list 1. mean: numeric success probability
2. cumu.avg: vector of cumulative avg success probability
3. data.frame: df of simulation values
}
\description{
This function simulates and computes long-run averages of success
by strategy for Monty Hall problem.
}
\details{
This function runs a simulation for on of three strategies in the
Monty Hall problem and returns a list of the relevant simulation results.
Plotting the cumulative average success probability is an optional logical
argument, default FALSE.
}
\author{
Stephen Downing
}

