% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Open a Goat Door}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{Represents the game, with two goats and one car.}

\item{a.pick}{A number between 1 and 3 representing the contestant's initial door selection.}
}
\value{
The function returns a number between 1 and 3 indicating the door opened by the host.
}
\description{
Host selects a door that contains a goat.
}
\details{
If the contestant selected the car, the function randomly selects one of the two goat doors.
If the contestant selected a goat, the function selects the other goat door.
}
\examples{
game <- create_game()
pick <- select_door()
open_goat_door(game, pick)

}
