% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_goat_door.R
\name{open_goat_door_orig}
\alias{open_goat_door_orig}
\title{Randomly chooses one of the doors with a goat that is not the selected door}
\usage{
open_goat_door_orig(game, a.pick)
}
\arguments{
\item{game}{Random assortment of goats and cars}

\item{a.pick}{The player selected door}

\item{door.choices}{Number of total possible doors}

\item{...}{further arguments passed to or from other methods.}
}
\value{
This function will open a random door that is not the door the player has selected as well as one that has a goat hidden behind.
}
\description{
Randomly chooses one of the doors with a goat that is not the selected door
}
\examples{
## Opens a single door between 1 and 3 from the original problem
open_goat_door_orig(create_game_orig(), 2)

## Randomly selects 1 door out of any remaining doors that fit the above condition.
open_goat_door(create_game(5, 3), 2, 8) 

}
