% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem (1).R
\name{opened.door}
\alias{opened.door}
\title{Host Opens Goat Door}
\usage{
opened.door(game, a.pick)
}
\arguments{
\item{…}{if the car door is selected then a sample size of 1 can be taken
from the remaining doors. If the goat door is selected, then the game has to
eliminate the door with the car behind it as an option. The remaining door to
be chosen will be a goat door.}
}
\value{
The return should be the selected goat door by the game.
}
\description{
The game randomly selects one of the goat doors to open.
}
\details{
The goat door function consists of the game and the contestant’s door
selection. The door selected by the contestant can not be chosen as one of
the doors so it must be accounted for in the function. As a result, the
contestant’s choice will affect the remaining doors. If the contestant pick
has a car behind the door than the two remaining doors have a goat behind
them; therefore, it does not matter which door is selected and a sample can
be chosen. If the contestant chooses a goat door, there will be both a car
and goat behind the remaining doors. The game as to make sure to not pick
door with the car behind it and to select the door with the goat behind it.
}
\examples{
If (game[a.pick] = “goat”)
  opened.door <- function(game, a.pick)

}
