% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome.R
\name{outcome}
\alias{outcome}
\title{Get the outcome of the Monty Hall game}
\usage{
outcome(status, prize, chosen_door)
}
\arguments{
\item{status}{A logical vector of length \code{n_doors} (at least three) with the
status of the doors (\code{TRUE} for closed doors, \code{FALSE} for open doors): can
use element \code{status} of the output of \code{\link[=setup_MontyHall]{setup_MontyHall()}}.}

\item{prize}{A vector with \link[checkinput:is_natural]{natural} numbers
giving the indices of the doors behind which a prize is hidden: can use
element \code{prize} of the output of \code{\link[=setup_MontyHall]{setup_MontyHall()}}.}

\item{chosen_door}{A single \link[checkinput:is_natural]{natural} number with
the index of the door chosen by the contestant: can use the output of
\code{\link[=choose_door]{choose_door()}}.}
}
\value{
A list of length two, each containing a vector with the same length as the
number of games that were played, containing \code{TRUE} or \code{FALSE} indicating if
the contestant chose a door with a prize behind it:
\itemize{
\item element \code{stick}: the contestant sticks to the initial choice made before
the host revealed what was behind one of the closed doors
\item element \code{change}: the contestant changes to another door after the host
revealed what was behind one of the closed doors
}
}
\description{
Get the outcome of the Monty Hall game if the contestant sticks to the
initial choice or changes to another door.
}
\details{
\code{status} here is the status after the host has revealed what is behind one of
the doors.
}
\examples{
set.seed(314)
(setup <- setup_MontyHall(n_doors = 3, n_prizes = 1))
(chosen_door <- choose_door(status = setup$status))
(new_status <- reveal_door(status = setup$status, prize = setup$prize,
                           chosen_door = chosen_door))
(outcome(status = new_status, prize = setup$prize,
         chosen_door = chosen_door))

}
\seealso{
\link{play_MontyHall} to play the game once and \link{play_MontyHall_repeated} to play
the game repeatedly.
}
