% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_game}
\alias{play_game}
\title{Play one game}
\usage{
play_game()
}
\value{
Returns a data frame with the strategy and outcome.
}
\description{
Plays one full Monty Hall game.
}
\details{
The function creates a game, picks a door, opens a goat
door and checks the results for staying and switching.
}
\examples{
play_game()

}
