% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg_Lab02.R
\name{play_game_5_doors}
\alias{play_game_5_doors}
\title{Plays an entire game.}
\usage{
play_game_5_doors()
}
\arguments{
\item{...}{no parameters}
}
\value{
The function returns a dataframe with two columns, strategy and outcome.
The strategy column has stay and switch, and the outcome is whether or not the
strategy led to a win or loss.

#'   new.game <- create_game_5_doors()
a.pick <- select_door_5_doors()
opened.door <- open_goat_door_5_doors(new.game,a.pick)
final.pick <- change_door_5_doors(stay=T,opened.door,a.pick)
determine_winner(final.pick,new.game)
}
\description{
\code{play_game_5_doors()} Wraps up all the functions prior into one.
}
\details{
The game setup replicates the game on the TV show "Let's
Make a Deal" where there are five doors for a contestant
to choose from, one of which has a car behind it and four
have goats. The contestant selects a door, then the host
opens a door to reveal a goat, and then the contestant is
given an opportunity to stay with their original selection
or switch to the other unopened doors. There was a famous
debate about whether it was optimal to stay or switch when
given the option to switch, so this simulation was created
to test both strategies.
}
\examples{
  play_game_5_doors()
}
