% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_game_orig.R
\name{play_game_orig}
\alias{play_game_orig}
\title{Original Monty Hall game with 2 goats and 1 car}
\usage{
play_game_orig()
}
\arguments{
\item{...}{further arguments passed to or from other methods.}
}
\value{
The number of goats added to the number of cars makes the total number of doors to choose from.
}
\description{
Original Monty Hall game with 2 goats and 1 car
}
\examples{
## Single simulation of the original Monty Hall Game
play_game()

}
