% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_MontyHall.R
\name{play_MontyHall}
\alias{play_MontyHall}
\alias{play_MontyHall_repeated}
\title{Play the Monty Hall game}
\usage{
play_MontyHall(n_doors = 3, n_prizes = 1)

play_MontyHall_repeated(n_doors = 3, n_prizes = 1, n_games = 100)
}
\arguments{
\item{n_doors}{A \link[checkinput:is_natural]{natural} number giving the used
number of doors. Should be at least three.}

\item{n_prizes}{A \link[checkinput:is_natural]{natural} number giving the
number of prizes to be hidden. \code{n_prizes} should be at least one and at least
two smaller than \code{n_doors}.}

\item{n_games}{A \link[checkinput:is_natural]{natural} number giving the
number of games to play.}
}
\value{
A list of length two, each containing a vector with the same length as the
number of games that were played, containing \code{TRUE} or \code{FALSE} indicating if
the contestant chose a door with a prize behind it:
\itemize{
\item element \code{stick}: the contestant sticks to the initial choice made before
the host revealed what was behind one of the closed doors
\item element \code{change}: the contestant changes to another door after the host
revealed what was behind one of the closed doors
}
}
\description{
Play the Monty Hall game
}
\section{Programming notes}{

Rewrite using \link{apply} instead of a \link[=Control]{for}-loop, then adjust to use
parallel processing, e.g., using package \code{parallel}, see
\code{optim_bruteforce_knip.R}.
}

\examples{
#### Play typical game once
set.seed(314)
play_MontyHall(n_doors = 3, n_prizes = 1)

# Same as:
set.seed(314)
(setup <- setup_MontyHall(n_doors = 3, n_prizes = 1))
(chosen_door <- choose_door(status = setup$status))
(new_status <- reveal_door(status = setup$status, prize = setup$prize,
                           chosen_door = chosen_door))
(outcome(status = new_status, prize = setup$prize,
         chosen_door = chosen_door))

#### Play game with multiple doors and multiple prizes once
set.seed(314)
play_MontyHall(n_doors = 10, n_prizes = 3)

# Same as:
set.seed(314)
(setup <- setup_MontyHall(n_doors = 10, n_prizes = 3))
(chosen_door <- choose_door(status = setup$status))
(new_status <- reveal_door(status = setup$status, prize = setup$prize,
                           chosen_door = chosen_door))
(outcome(status = new_status, prize = setup$prize,
         chosen_door = chosen_door))

#### Play typical game a hundred times
set.seed(314)
many_games <- play_MontyHall_repeated(n_doors = 3, n_prizes = 1, n_games = 100)
(frac_won_stick <- length(which(many_games$stick)) / length(many_games$stick))
(frac_won_change <- length(which(many_games$change)) / length(many_games$change))

}
