% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_n_games}
\alias{play_n_games}
\title{Play multiple games}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n}{Number of games to play. Default is 100.}
}
\value{
Returns a data frame with all of the game results.
}
\description{
Plays the Monty Hall game multiple times.
}
\details{
Repeats the game n times and saves the results for the
stay and switch strategies.
}
\examples{
play_n_games(10)

}
