% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg_Lab02.R
\name{play_n_games_5_doors}
\alias{play_n_games_5_doors}
\title{Plays N number of games and returns a dataframe with the results of all N
games.}
\usage{
play_n_games_5_doors(n = 100)
}
\arguments{
\item{n}{The number of games to simulate.}
}
\value{
The function returns a dataframe results.df
}
\description{
\code{play_n_games_5_doors()} loops play_game n times and appends the returned dataframe
to a list results.list. After all N games have been played, the list is binded
into a dataframe results.df via bind_rows. A table is produced to show the
percentages of wins and losses for each strategy.
}
\details{
The game setup replicates the game on the TV show "Let's
Make a Deal" where there are five doors for a contestant
to choose from, one of which has a car behind it and four
have goats. The contestant selects a door, then the host
opens a door to reveal a goat, and then the contestant is
given an opportunity to stay with their original selection
or switch to the other unopened doors. There was a famous
debate about whether it was optimal to stay or switch when
given the option to switch, so this simulation was created
to test both strategies.
}
\examples{
  play_n_games_5_doors(100)
}
