% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MontyHall_repeated.R
\name{plot_MontyHall_repeated}
\alias{plot_MontyHall_repeated}
\title{Plot outcome of repeated Monty Hall games}
\usage{
plot_MontyHall_repeated(
  n_doors = 3,
  n_prizes = 1,
  n_games = 100,
  n_trials = 100,
  sort_values = TRUE,
  show_ranges = TRUE,
  print_summary = TRUE
)
}
\arguments{
\item{n_doors}{A \link[checkinput:is_natural]{natural} number giving the used
number of doors. Should be at least three.}

\item{n_prizes}{A \link[checkinput:is_natural]{natural} number giving the
number of prizes to be hidden. \code{n_prizes} should be at least one and at least
two smaller than \code{n_doors}.}

\item{n_games}{A \link[checkinput:is_natural]{natural} number giving the
number of games to play.}

\item{n_trials}{A single \link[checkinput:is_natural]{natural} number with
the number of trials, e.g., how many times to run \link{play_MontyHall_repeated}.}

\item{sort_values}{\code{TRUE} or \code{FALSE}: sort values in the plot?}

\item{show_ranges}{\code{TRUE} or \code{FALSE}: show the ranges as dotted vertical
lines in the plot?}

\item{print_summary}{\code{TRUE} or \code{FALSE}: print a summary of the fraction of
games won for the different tactics?}
}
\value{
A list of length two, returned \link[=invisible]{invisibly}. Each element contains
a numeric vector with a length equal to \code{n_trials}, containing the fraction
of games the contestant won by choosing a door with a prize behind it:
\itemize{
\item element \code{frac_won_stick}: the contestant sticks to the initial choice made
before the host revealed what was behind one of the closed doors
\item element \code{frac_won_change}: the contestant changes to another door after the
host revealed what was behind one of the closed doors
}
}
\description{
Plot outcome of repeated Monty Hall games
}
\details{
The plot shows the fraction of games won in each trial. Different colours
indicate different tactics. Solid vertical lines indicate the median value of
all trials and, if \code{show_ranges} is \code{TRUE}, dotted vertical lines show the
ranges.

It is easier to compare the overall chances of winning for the different
tactics if \code{sort_values} is \code{TRUE}, but then the linkage between the tactics
is broken. In contrast, if \code{sort_values} is \code{FALSE}, the
}
\section{Side effects}{

A summary of the outcomes is plotted.
}

\examples{
plot_MontyHall_repeated(n_doors = 3, n_prizes = 1, n_games = 100, n_trials = 50,
                        sort_values = TRUE, show_ranges = TRUE, print_summary = TRUE)
plot_MontyHall_repeated(n_doors = 10, n_prizes = 3, n_games = 100, n_trials = 50,
                        sort_values = FALSE, show_ranges = FALSE, print_summary = FALSE)

}
