% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reveal_door.R
\name{reveal_door}
\alias{reveal_door}
\title{Reveal a door}
\usage{
reveal_door(status, prize, chosen_door)
}
\arguments{
\item{status}{A logical vector of length \code{n_doors} (at least three) with the
status of the doors (\code{TRUE} for closed doors, \code{FALSE} for open doors): can
use element \code{status} of the output of \code{\link[=setup_MontyHall]{setup_MontyHall()}}.}

\item{prize}{A vector with \link[checkinput:is_natural]{natural} numbers
giving the indices of the doors behind which a prize is hidden: can use
element \code{prize} of the output of \code{\link[=setup_MontyHall]{setup_MontyHall()}}.}

\item{chosen_door}{A single \link[checkinput:is_natural]{natural} number with
the index of the door chosen by the contestant: can use the output of
\code{\link[=choose_door]{choose_door()}}.}
}
\value{
\code{status} with one more open door that was revealed by the host: a logical
vector of length \code{n_doors} with the status of the doors: \code{TRUE} for closed
doors, \code{FALSE} for open doors.
}
\description{
The host opens a closed door to reveal what is behind it. It is \emph{not} the
door chosen by the contestant and \emph{not} a door with a price behind it.
}
\details{
At least two doors should be closed.
}
\section{Programming notes}{

For games with more than three doors, you can reveal more doors.
}

\examples{
set.seed(314)
(setup <- setup_MontyHall(n_doors = 3, n_prizes = 1))
(chosen_door <- choose_door(status = setup$status))
(new_status <- reveal_door(status = setup$status, prize = setup$prize,
                           chosen_door = chosen_door))

}
\seealso{
\link{play_MontyHall} to play the game once and \link{play_MontyHall_repeated} to play
the game repeatedly.
}
