% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_safe.R
\name{sample_safe}
\alias{sample_safe}
\title{Safe drop-in replacement for sample}
\usage{
sample_safe(x, size, replace = FALSE, prob = NULL)
}
\arguments{
\item{x}{A vector of one or more elements from which to choose.}

\item{size}{a non-negative integer giving the number of items to choose.}

\item{replace}{should sampling be with replacement?}

\item{prob}{a vector of probability weights for obtaining the elements
    of the vector being sampled.}
}
\value{
A vector of length \code{size} with elements drawn from \code{x}.
}
\description{
Safe drop-in replacement for sample
}
\details{
The base-\R function \code{\link[=sample]{sample()}} is \emph{unsafe} if \code{x} can vary in length, as
as noted in the section \code{Details} of its documentation:
\itemize{
\item if \code{x} is a \link{length}-one \link[=is.numeric]{numeric} with a
\link[=is.finite]{finite} value \link[=Comparison]{equal to or larger than} one,
samples are taken from \code{1:x}, e.g., \code{sample(3)} samples from \code{c(1, 2, 3)}.
\item otherwise, samples are taken from \code{x}, e.g., \code{sample(c(3, 4))} samples from
\code{c(3, 4)}.
}

In contrast, function \code{sample_safe()} \emph{always} samples from \code{x}.
}
\section{Programming notes}{

\code{sample_safe()} is similar to \code{resample()} defined in the \code{Examples} of
\code{\link[=sample]{sample()}} as safer alternative to \code{\link[=sample]{sample()}}, but uses named arguments when
passing them to \code{\link[=sample.int]{sample.int()}} instead of relying on positional matching and
determines \code{size} from the length of \code{x} if \code{size} is missing.
}

\examples{
# Taken from the examples in help(sample)

x <- 1:10
sample_safe(x = x) # a random permutation
sample_safe(x = c(0, 1), size = 5, replace = TRUE) # five Bernoulli trials

# Comparing sample() and sample_safe()
sample(x = x[x > 8]) # length two: c(9, 10) or c(10, 9)
sample(x = x[x > 9]) # length ten (!): random permutation of 1:10
sample(x = x[x > 10]) # length zero: integer(0)

sample_safe(x = x[x > 8]) # length two: c(9, 10) or c(10, 9)
sample_safe(x = x[x > 9]) # length one: 10
sample_safe(x = x[x > 10]) # length zero: integer(0)

}
\seealso{
\code{\link[=sample]{sample()}}; package \href{https://github.com/hadley/strict}{strict} which gives an
error if \code{\link[=sample]{sample()}} is used with length-one input; \link{Distributions} to draw
random numbers from distributions.
}
