% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Select a Door}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a number between 1 and 3 indicating the selected door.
}
\description{
Randomly selects one of the the three doors in a Monty Hall Problem game.
}
\details{
The function randomly selects one door from the three available doors.
}
\examples{
select_door()

}
