% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg_Lab02.R
\name{select_door_5_doors}
\alias{select_door_5_doors}
\title{Make an initial selection as contestant in the 5 door Monty Hall game.}
\usage{
select_door_5_doors()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a single door value, 1-5
}
\description{
\code{select_door_5_doors()} selects a door in the 5 door Monty Hall game
as the contestant
}
\details{
The game setup replicates the game on the TV show "Let's
Make a Deal" where there are five doors for a contestant
to choose from, one of which has a car behind it and four
have goats. The contestant selects a door, then the host
opens a door to reveal a goat, and then the contestant is
given an opportunity to stay with their original selection
or switch to the other unopened doors. There was a famous
debate about whether it was optimal to stay or switch when
given the option to switch, so this simulation was created
to test both strategies.
}
\examples{
  select_door_5_doors()

}
