% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_door.R
\name{select_door_orig}
\alias{select_door_orig}
\title{Randomly select a door based on the total number of goats and cars.}
\usage{
select_door_orig()
}
\arguments{
\item{door.choices}{Total number of doors to choose from.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
door.choices should equal goats + cars from other functions.
}
\description{
Randomly select a door based on the total number of goats and cars.
}
\examples{
## Randomly selects doors 1, 2, or 3
select_door_orig()

## Randomly chooses one of the total number of doors.
select_door(15)

}
