% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_MontyHall.R
\name{setup_MontyHall}
\alias{setup_MontyHall}
\title{Setup Monty Hall game}
\usage{
setup_MontyHall(n_doors = 3, n_prizes = 1)
}
\arguments{
\item{n_doors}{A \link[checkinput:is_natural]{natural} number giving the used
number of doors. Should be at least three.}

\item{n_prizes}{A \link[checkinput:is_natural]{natural} number giving the
number of prizes to be hidden. \code{n_prizes} should be at least one and at least
two smaller than \code{n_doors}.}
}
\value{
A \link{list} of length two containing two elements:
\itemize{
\item \code{status} with a logical vector of length \code{n_doors} with the status of the
doors: \code{TRUE} for closed doors, \code{FALSE} for open doors
\item \code{prize} with an \link{integer} vector of length \code{n_prizes} with the indices of
the doors behind which a prize is hidden.
}
}
\description{
Setup Monty Hall game
}
\section{Programming notes}{

Check \code{\link[=rep.int]{rep.int()}}.
}

\examples{
# Typical game
set.seed(314)
setup_MontyHall(n_doors = 3, n_prizes = 1)

# Game with many doors and multiple prizes
set.seed(314)
setup_MontyHall(n_doors = 10, n_prizes = 3)

}
\seealso{
\link{play_MontyHall} to play the game once and \link{play_MontyHall_repeated} to play
the game repeatedly.
}
