% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_game.R
\name{simulate_game}
\alias{simulate_game}
\title{Monty Hall Game Simulation}
\usage{
simulate_game(game.size = 3, num.cars = 1, rounds = 10000)
}
\arguments{
\item{game.size}{An integer.}

\item{num.cars}{An integer greater than zero and not larger than \code{game.size}-2.}

\item{rounds}{An integer.}
}
\value{
The function returns a table showing the number of times where stay
and switch strategies lead to winning or losing.
}
\description{
\code{simulate_game(game.size=3, num.cars=1, rounds=10000)} simulates a generalized
version of the Monty Hall game.
}
\details{
This function first generates and plays random Monty Hall games where in all of the games
there are \code{game.size} doors and behind \code{num.cars} doors there are cars and behind the rest
we have goats. The game is played for \code{rounds} number of times. The function reports the
number of times where stay and switch strategies lead to winning or losing.

Read more about the game in \url{https://en.wikipedia.org/wiki/Monty_Hall_problem}.
}
\examples{
simulate_game(4, 2, 20000)
simulate_game()

\dontrun{
simulate_game(3, 2, 10000)
}

}
