% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{simulate_games}
\alias{simulate_games}
\title{Play multiple Monty Hall games}
\usage{
simulate_games(n = 1000, switch = TRUE)
}
\arguments{
\item{n}{Integer, number of games to play (> 0).}

\item{switch}{Logical, \code{TRUE} to switch, \code{FALSE} to stay.}
}
\value{
Numeric proportion of wins in \verb{[0, 1]}.
}
\description{
Simulate many games to estimate the winning probability with stay/switch strategy.
}
\examples{
\donttest{
  simulate_games(100, switch = TRUE)
  simulate_games(100, switch = FALSE)
}
}
