% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{simulation}
\alias{simulation}
\title{Simulation Wrapper}
\usage{
simulation(num.cars, num.goats)
}
\arguments{
\item{...}{The function takes two positional arguments, number of cars and number of goats}
}
\value{
returns a table of results
}
\description{
Runs 10,000 simulations of the game
}
\details{
Outputs table to report results
}
