% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{sum}
\alias{sum}
\title{Sum of vector elements.}
\usage{
sum(..., na.rm = TRUE)
}
\arguments{
\item{na.rm}{A logical scalar. Should missing values (including \code{NaN})
be removed?}

\item{x}{Numeric, complex, or logical vectors.}
}
\value{
If all inputs are integer and logical, then the output
will be an integer. Otherwise it will be a length-one numeric or
complex vector.

Zero-length vectors have sum 0 by definition. See
\url{http://en.wikipedia.org/wiki/Empty_sum} for more details.
}
\description{
\code{sum(x)} returns the sum of all the values present in its arguments.
}
\details{
This is a generic function: methods can be defined for it directly
or via the \link{Summary} group generic. For this to work properly,
the arguments \code{...} should be unnamed, and dispatch is on the
first argument.
}
\examples{
sum(1:10)
sum(1:5, 6:10)
sum(F, F, F, T, T)

sum(.Machine$integer.max, 1L)
sum(.Machine$integer.max, 1)

\dontrun{
sum("a")
}
}
