% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HolfordMeetsMonty.R
\docType{data}
\name{this.game}
\alias{this.game}
\title{\strong{TEST OF ALL FUNCTIONS IN GAME}}
\format{
An object of class \code{character} of length 3.
}
\usage{
this.game
}
\description{
\strong{TEST OF ALL FUNCTIONS IN GAME}
}
\keyword{datasets}
