% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem (1).R
\name{change_door}
\alias{change_door}
\title{The contestant can either stay with their initial selection,
or switch to the remaining unopened door.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{...}{Arguments: \code{stay} (its value is TRUE if the contestant is staying
with their initial choice, or FALSE if they choose to switch to the other door.),
\code{opened.door} (its value is the door which the host opened in open_goat_door),
and \code{a.pick} (its value is the contestant's initially selected door).}
}
\value{
The function returns an integer representing the contestant's final selection.
This is either the contestant's intitial pick, or the remaining unopened door.
}
\description{
\code{change_door()} Type: Integer -> Generates an integer of the contestant's final
selection door.
}
\details{
This allows the contestant to employ their chosen game strategy of either
staying with their first choice, or switching to the other unopened door.
}
\examples{
  change_door(stay = T, opened.door = openedDoor,
  a.pick = firstDoor)
}
