% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{Change the selected door.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{...}{the arguments are stay (whether the contestant chooses to stay), the door that
the host opened, and the door the contestant chose.}
}
\value{
The function returns a number the door the contestant chooses.
}
\description{
\code{change_door()} generates a number corresponding with the door the contestant selects.
}
\details{
The contestant is given the option to change their door selection.
}
\examples{
  change_door()

}
