% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{CHANGE DOORS}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{if}{the contestant stays they will have their first pick, if changes will need to choose the opposing ! argument 
for their choice}
}
\value{
Which door that the contestant chooses
}
\description{
change_door() will allow the contestant to switch doors or stay
}
\details{
The contestant will be able to switch doors or stay with their first selection. If they stay, they will 
keep their original pick. If the contestant switches, they will chose the other door, the one the host
did not open.
}
\examples{
return( final.pick )
}
