% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{change_door}
\alias{change_door}
\title{Change or not change selected door.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{...}{no arguments are used by the function}
}
\value{
Return a final picked value of the door.
}
\description{
The host offers participants a chance to change the selected
door. Participants can choose change or not change.
}
\details{
Input a random number of stay value and return the changed
door or the same door.
}
