% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{Change doors}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{If}{contestant stays (stay=T)
If contestant does not stay (stay=F)}
}
\value{
The function returns a length of 1 numeric vector indicating
that a one door has been selected that is different from
the door that the host opened.
}
\description{
The contestant decides if they want to change doors.
}
\details{
After seeing the host open one of the goat doors,
the contestant will decide if they want to stay with
the door they first chose or switch to the remaining
door that has not been chosen yet.
}
\examples{
  change_door(stay=T, opened.door, a.pick)
  change_door(stay=F, opened.door, a.pick)
}
