% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{Determine the final door choice based on strategy.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay}{A logical value (\code{TRUE} or \code{FALSE}) indicating whether the
contestant stays with their original selection. Defaults to \code{TRUE}.}

\item{opened.door}{A single integer (1, 2, or 3) representing the index
of the door containing a goat opened by the host.}

\item{a.pick}{A single integer (1, 2, or 3) representing the contestant's
initial door selection.}
}
\value{
The function returns a single integer (1, 2, or 3) representing
the index of the contestant's final door choice.
}
\description{
\code{change_door()} simulates the contestant's decision to either stay with
their initial door choice or switch to the remaining unopened door.
}
\details{
After the host opens a door to reveal a goat, the contestant is faced
with a choice. If the contestant decides to stay (\code{stay = TRUE}), their
final choice remains their initial pick. If they decide to switch
(\code{stay = FALSE}), their final choice becomes the one remaining closed
door that was neither their initial pick nor the door opened by the host.
}
\examples{
  this_game <- create_game()
  my_pick <- select_door()
  goat_door <- open_goat_door( game = this_game, a.pick = my_pick )

  # Strategy: Stay
  change_door( stay = TRUE, opened.door = goat_door, a.pick = my_pick )

  # Strategy: Switch
  change_door( stay = FALSE, opened.door = goat_door, a.pick = my_pick )

}
