% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{Stay With or Switch Doors}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay}{A logical value indicating whether the contestant stays
with the original selection. TRUE means stay and FALSE means switch.}

\item{opened.door}{A numeric value indicating the door opened by
the host. The value should be 1, 2, or 3.}

\item{a.pick}{A numeric value indicating the contestant's initial
door selection. The value should be 1, 2, or 3.}
}
\value{
A numeric value indicating the contestant's final door
selection. The returned value will be 1, 2, or 3.
}
\description{
\code{change_door()} determines the contestant's final door selection
based on whether the contestant chooses to stay or switch.
}
\details{
If \code{stay} is TRUE, the contestant keeps the original door
selection. If \code{stay} is FALSE, the function selects the remaining
unopened door that was neither the contestant's initial choice
nor the door opened by the host.
}
\examples{
  change_door(stay = TRUE, opened.door = 2, a.pick = 1)
  change_door(stay = FALSE, opened.door = 2, a.pick = 1)

}
