% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{Change Doors?}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay=T}{or stay=F: a character signifying TRUE or FALSE
to the contestant's choice of keeping their current
door (stay=T), or switching doors with the host (stay=F)\preformatted{ open.door- an integer (1:3) returned from open_goat_door
 that signifies the goat door opened and removed from
 gameplay

 a.pick- an integer (1:3) returned as the contestant's
 initial pick
}}
}
\value{
\preformatted{ stores one of the following in final.pick:
      [1] 1
      [1] 2
      [1] 3
}
}
\description{
\preformatted{ change_door allows the contestant the opportunity to
 switch the initially assigned door with the one
 remaining door opened
}
}
\details{
\preformatted{ the function receives stay=T or stay=F

 if stay=T the function stores the integer stored in
 a.pick in final.pick

 if stay=F the function stores the door that is NOT
 the door determined by open_goat_door (goat.door)
 or the door that determined by select_door (a.pick)
 and stores the remaining integer (door) in final.pick
}
}
\examples{
     change_door
}
