% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{Contestant Chooses to Switch or Stay}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{...}{This function has 3 arguments: stay = True or False,
opened.door which is inputted by the open_goat_door function,
and a.pick which is inputted by the select_door function.}
}
\value{
This function returns a length of 1 numeric vector
that represents the contestant's final pick.
}
\description{
\code{change_door} allows contestant to switch their initial pick to
the remaining unopened door or allows them to keep their
initial pick.
}
\details{
This function replicates the part of the game where the contestant
is allowed to stay with their initial pick or switch their pick to
the unopened door.
}
\examples{
 change_door(stay = F, opened.door, a.pick)
 
 This function have values put in for the defaults or
 it must be used with the outputs of open_goat_door
 and select_door assigned to objects.  
 
 You can test this function with the following examples.
 
 Test example: 
 
 change_door(stay = T, opened.door = 1, a.pick = 2)
 This should result in 2 since stay is TRUE.
 
 change_door(stay = F, opened.door = 1, a.pick = 2)
 This should result in 3 since stay = FALSE.
 
 Test example:
 
 object_select_door <- select_door()
 object_open_goat_door <- open_goat_door()
 change_door(stay = T, object_open_goat_door, object_select_door)
 
}
