% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{Option to switch doors.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay}{A logical value representing if the contestant decides to stay with
the first pick (TRUE) or switch to a different door (FALSE). The default value
is TRUE.}

\item{opened.door}{An integer representing the opened goat door (a value between
1 and 3).}

\item{a.pick}{An integer representing the contestant's initial door pick (a value
between 1 and 3).}
}
\value{
The function returns a number between 1 and 3 that is either the first
pick or the last remaining unopened door.
}
\description{
\code{change_door()} randomly assigns if the contestant chooses to stay with the
original door or to switch to the last remaining unopened door.
}
\details{
If the contestant stays, the final pick will be the same as the first pick.
If the contestant switches, the final pick will be the last remaining unopened
door.
}
\examples{
  change_door(opened.door, a.pick = 2)

}
