% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{Make final door selection}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay:}{TRUE or FALSE. TRUE returns the results from the select_door()
function. FALSE returns the other unopened door. Defaults to TRUE.
opened.door: the numerical value from the open_goat_door() function. No default.
a.pick: the vector from the create_game() function. No default.}
}
\value{
The function returns a number between 1 and 3: either the value from the select_door() function
(if stay = TRUE) or the number that is neither the opened_goat_door() value nor the
select_door() value (if stay = FALSE)
}
\description{
Having seen a goat door, the participant has the otpion to change doors or switch doors.
The function selects the final door depending on the participant's choice of stay or switch.
}
\details{
Now that a goat door has been revealed, the participant has two doors to choosee from.
One of those doors has already been selected, and the participant has the option of
staying with that door or switching to the other unopened door. If the participant decides
to stay with the original door selection, the function returns the originally selected door
from the select_door() function. If the participant decides to switch doors, the function
returns the number of the other unopened door.
}
\examples{
  change_door( T, 1, 3 )
  change_door( F, 1, 3 )

}
