% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.r
\name{change_door}
\alias{change_door}
\title{Change doors}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{No}{arguments are used by the function
stay: TRUE or FALSE. TRUE returns the results from the select_door()
function. FALSE returns the other unopened door. Defaults to TRUE.
opened.door: the numerical value from the open_goat_door() function. No default.
a.pick: the vector from the create_game() function. No default.}
}
\value{
The function returns a number between 1 and 3: either the value from the select_door() function
(if stay = TRUE) or the number that is neither the opened_goat_door() value nor the
select_door() value (if stay = FALSE)
}
\description{
\code{change_door()} function returns a final pick.
}
\details{
If the contestant decides to stay with the original pick, the function returns
the originally selected door. If the contestant decides to switch doors, then
the function reveals the number of the other unopened door.
}
\examples{
  change_door( T, 1, 3 )
  change_door( F, 1, 3 )

}
