% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem (4).R
\name{change_door}
\alias{change_door}
\title{Change Doors}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{...}{Based on the game playing strategy stay=TRUE or stay=False will be
assigned.}
}
\value{
The function returns a length 3 character vector
indicating a number between 1 and 3, this will indicate a final pick made
by the contestant.
}
\description{
\code{change_door( stay=T, opened.door, a.pick )} generates the option given to
the contestant to change from their initial selection to the other door
that still remains closed.
}
\details{
The contestant is now granted a chance to change from his/her initial
selection to the door that still remains closed. Contestant must now decide
if they wish to stick with original selection, or change to the remaining
unopened door.
}
\examples{
  change_door( stay=T, opened.door, a.pick )

}
