% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{Change Door Selection}
\usage{
change_door(stay = TRUE, opened.door, a.pick)
}
\arguments{
\item{stay}{A logical value. If \code{TRUE}, the contestant stays with their original choice. If \code{FALSE}, the contestant switches to the other unopened door.}

\item{opened.door}{An integer representing the door that the host has opened to reveal a goat.}

\item{a.pick}{An integer representing the contestant's initial door selection.}
}
\value{
The function returns an integer (1, 2, or 3) representing the contestant's final door choice.
}
\description{
This function allows the contestant to either stay with their initial choice or switch to the other unopened door.
}
\details{
If the contestant decides to stay, their final choice will be the same as their initial pick. If they switch, they will choose the other unopened door.
}
\examples{
opened.door <- open_goat_door(game = create_game(), a.pick = select_door())
change_door(stay = TRUE, opened.door, a.pick = 1)

}
