% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{Does the contestant stay with their initial pick or switch doors?}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns the door number of the initial pick if the contestant chose to stay or the number of the door the contestant switched to.
}
\description{
\code{change_door()} provides the conditions if the contestant decides to stay with their initial pick or change doors.
}
\details{
If the contestant chooses to stay with their initial pick, the function retains their initial pick. If the contestant chooses to switch doors, the function selects a door that is not their initial pick or a door opened by the host.
}
\examples{
  change_door()

}
