% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{Change or stay with the initial door.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay}{Logical, indicating whether to stay (`TRUE`) or switch (`FALSE`).}

\item{opened.door}{An integer (1, 2, or 3) indicating the door opened by the host.}

\item{a.pick}{An integer (1, 2, or 3) representing the contestant’s initial choice.}
}
\value{
An integer between 1 and 3 corresponding to the contestant’s
   final choice.
}
\description{
`change_door()` determines the contestant’s final choice:
   stay or switch.
}
\details{
If `stay = TRUE`, the contestant keeps their original door.
   If `stay = FALSE`, the contestant switches to the other unopened
   door that is not the initial pick. The host has already revealed the 
   goat behind the door not already mentioned.
}
\examples{
  gg <- create_game()
  choice <- select_door()
  o <- open_goat_door(gg, choice)
  change_door(stay = FALSE, opened.door = o, a.pick = choice)

}
