% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem(bh).R
\name{change_door}
\alias{change_door}
\title{Contestant chooses game strategy: stay or switch.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{Logical}{arguments "T" or "F" are used to indicate
whether contestant stays at the a.pick door or switches
to the last remaining door.}
}
\value{
The function returns a numeric vector (1-3) that is not
the same as the numeric vector as 'opened.door'.
}
\description{
The 'change_door()' function uses the argument "stay" if the
contestant chooses to stay at the door they initially chose
or the argument "switch" if the contestant choose to instead
select the last remaining door.
}
\details{
If the contestant decides to stay (stay=T), the numeric vector
'a.pick' does not change. If the contestant decides to switch
(stay=F), the numeric vector that is not 'a.pick' or
'opened.door' is returned.
}
\examples{
  change_door(stay=T, opened.door, a.pick)
  change_door(stay=F, opened.door, a.pick)

}
