% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{Allows user to select strategy for game: "stay" or "switch"}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay}{logical indicating strategy, TRUE = stay, FALSE
= switch}

\item{opened.door}{positive integer value of the door opened by the host}

\item{a.pick}{postiive integer value of the first door chosen by
the contestant}
}
\value{
a positive integer value between 1 and 3 indicating
the final door selection for the contestant.
}
\description{
After the host reveals one door with a goat behind it, the
contestant has two choices, stay with the first choice
or switch to the other door.
}
\details{
\code{change_door()} requires the contestant to choose their
strategy for the game and therefore their final choice
of doors. The function requires a logical argument for
strategy "stay" (T/F), the integer value of the host opened
door and the door initially picked by the contestant.
}
\examples{
  change_door( stay=F, 2, 1)
  change_door( stay=T, 3, 2)

}
