% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{Change Door Selection or Stay.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{Uses}{3 arguments: stay = TRUE, opened.door object, and a.pick object.}
}
\value{
Returns a value between 1 and 3.
}
\description{
Simulate the decision to either stay or change their pick.
}
\details{
Creates a vector for 3 doors. The function then checks the value of the stay
argument, either TRUE or FALSE. If the player stays, final.pick is set to a.pick.
If the player chooses not to stay, the function checks remaining options that
was not opened.
}
\examples{
 final.pick.stay <- change_door(stay = TRUE, opened.door = opened.door, a.pick = my.initial.pick)
 final.pick.switch <- change_door(stay = FALSE, opened.door = opened.door, a.pick = my.initial.pick)
}
