% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{Contestant can decide to change doors}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay}{A boolean.}

\item{opened.door}{A number.}

\item{a.pick}{A number.}
}
\value{
A number representing the final door number to open in the game
}
\description{
\code{change_door()} contestant decides if they want to change
their selected door to open
}
\details{
The contestant can choose to change doors to the other unopened door,
or keep the existing door they originally chose
}
\examples{
  change_door( stay=T, opened.door, a.pick )

}
