% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{Contestant decides whether to 'stay' with or 'switch' the door they
originally selected in the Monty Hall Problem game.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{The}{arguments 'stay=T', 'opened.door', and 'a.pick' are used by the
function.}
}
\value{
The function returns a number indicating the door that the contestant
ultimately chooses. What is behind the door is not indicated, although the
person running the game can determine that it is a goat from the character
vector returned by the first function.
}
\description{
\code{change_door} allows the contestant continue on through the game using the
door they originally selected, or to switch their door to the remaining
unopened door.
}
\details{
This is the portion of the game where the contestant decides whether to
continue by using the door they originally selected, or to change their
door choice to the remaining unopened door, knowing that the host has
opened a door with a goat behind it, so one of the remaining two doors has
the other goat behind it, while the other has a car behind it.
}
\examples{
  change_door(stay=T, opened.door, a.pick)
  
}
