% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{Contestant given option to change door.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{no}{params}
}
\value{
Returns the number that corresponds to the unopened door that the contestant
did not select in select_door(). Changing the door prompts the next step where
the host reveals whether the contestant's final choice reveals a car or goat.
}
\description{
\code{change_door()} The host gives the contestant the option to change the original
selection to the other remaining unselected door.
}
\details{
The contestant has an option to stay with the original choice in select_door()
or to change the selection to the last remaining unselected door. Depending on
the game's paramaters, it is generally a better strategy to switch doors.
}
\examples{
Select_door()=1; open_goat_door()=2; change_door()=3
}
