% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{Changing Door Function.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{...}{If the argument is \code{TRUE} it means that the contestant stayed
and \code{FALSE} means that the contestant switched doors.}
}
\value{
\preformatted{This function returns a number representing the number of the final door choice for the contestant.
}
}
\description{
\preformatted{`change_door()` the contestant is asked if they would like to change their 
selection to a different door or if they would like to keep their first choice.
}
}
\details{
\preformatted{a door is generated that represents the door that the 
contestant chose, whether they switched or stayed.
}
}
\examples{
	change_door(stay = F, opened.door = 2, a.pick = 1)

}
