% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{Change doors in a Monty Hall game}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay}{A logical indicating whther the constestant will stay with their original door selection or change to a different door}

\item{opened.door}{An interger indicating the number of the door that was opened}

\item{a.pick}{An interger indicating the intial door selected by the player}
}
\value{
The function returns the number of the final door selected by the player
}
\description{
This function lets the constestant change their door selection
}
