% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{Change Doors}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay}{logical vector that is either T or F}

\item{opened.door}{number between 1 and 3, reflecting the goat door that the host opened}

\item{a.pick}{integer (1,2,3) that aligns with the door that the contestant chose first}
}
\value{
number between 1 and 3
}
\description{
\code{change_door()} simulates the contestant either switching to a new door or staying with the door they originally picked
}
\details{
starts with \code{stay=T} meaning the contestant stays with their first pick, or \code{stay=F} meaning the contestant decides to switch to one of the unopened doors. the function also pulls from earlier arguments: \code{opened.door} and \code{a.pick}. next it creates a \code{doors} vector with 3 doors, then starts the following logical argument: if \code{stay=T} then \code{final.pick} becomes their \code{a.pick} value or their first choice. if \code{stay=F}, then \code{final.pick} becomes the unopened, unpicked door. lastly, this function will return the \code{final.pick} value which is a number between 1 and 3.
}
\examples{
  change_door( stay=T, opened.door=1, a.pick=3 ) # should be 3
  change_door( stay=F, opened.door=1, a.pick=3 ) # should be 2
  
}
