% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{Contestant makes the final pick}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay:}{logical flag to determine if you intend to stay
or switch;
opened.door: a number (between 1 to 3) for the door revealed
by host;
a.pick:      a number (between 1 to 3) for the door opened by
the contestant at the first guess.}
}
\value{
The function returns a number between 1 and 3 indicating
the contestant's final pick.
}
\description{
\code{change_door(stay=T, opened.door, a.pick )} makes the final pick depending upon
if you intend to stay or switch.
}
\details{
The function provides the contestant an opportunity
to stay with their original selection or switch to
the other unopened door.
}
\examples{
opened.door <- open_goat_door( this.game, my.initial.pick )
change_door( stay=T,
            opened.door=opened.door,
            a.pick=my.initial.pick )
change_door( stay=F,
            opened.door=opened.door,
            a.pick=my.initial.pick )
my.final.pick <- change_door( stay=F,
                             opened.door=opened.door,
                             a.pick=my.initial.pick )
}
