% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{Selects final door based on strategy chosen.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay}{Boolean describing strategy of stay or switch. Switch = False}

\item{opened.door}{Door with goat opened by the host.}

\item{a.pick}{Door opened by the contestant.}
}
\value{
The function returns a single door value, 1-3
}
\description{
\code{change_door()} selects a door based on the strategy in the
params. If the strategy is stay, the contestant keeps the pick
from select_door. If not, they choose a new door that is neither
the door from select_door nor open_goat_door.
}
\details{
The game setup replicates the game on the TV show "Let's
Make a Deal" where there are three doors for a contestant
to choose from, one of which has a car behind it and two
have goats. The contestant selects a door, then the host
opens a door to reveal a goat, and then the contestant is
given an opportunity to stay with their original selection
or switch to the other unopened door. There was a famous
debate about whether it was optimal to stay or switch when
given the option to switch, so this simulation was created
to test both strategies.
}
\examples{
  change_door(stay=T,opened.door=2,a.pick=1)
}
