% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{Change door}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{`stay`}{is to choose whether the play will switches doors or keep the same door. `opened.door` is the door opened with a goat behind it. `a.pick` is the first door selected by the player.}
}
\value{
The function returns the final decision the player has made. Either to stay with the original selection, or switch doors after seeing the first goat.
}
\description{
This is the door the player has switched to after seeing one of the other doors a goat was behind.
}
\details{
The game setup replicates the game on the TV show "Let's
  Make a Deal" where there are three doors for a contestant
  to choose from, one of which has a car behind it and two
  have goats. The contestant selects a door, then the host
  opens a door to reveal a goat, and then the contestant is
  given an opportunity to stay with their original selection
  or switch to the other unopened door. There was a famous
  debate about whether it was optimal to stay or switch when
  given the option to switch, so this simulation was created
  to test both strategies.
}
\examples{
 `change_door(stay=T, opened.door, a.pick)`
}
