% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{Contestant Keeps or Switches Door}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{logical}{vector of stay
numeric vectors of previously opened host door and original contestant choice}
}
\value{
Returns final choice that will be opened to determine game outcome - numeric
vector between 1 and 3
}
\description{
In this function, the contestant decides whether they will keep the original door they chose
or if they will switch to the remaining unopened door
}
\details{
Determines final door based on decision to stay or switch
}
\examples{
change_door(stay = F, opened.door, a.pick)
}
