% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{Contestant choses a final selection: Stay with or Switch door selection.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{Parameter}{stay is logical: stay=T indicating the contestant not switching and
staying with their original choice of door and stay=F for switching to the
third mystery door neither disclosed or chosen. Parameter opened.door is the
number of the known goat door, disclosed by the Host. Parameter a.pick is the
integer representing the door originally selected by the contestant of the
current game.}
}
\value{
Returns length 1 vector or 1, 2, or 3 representing the position of the stay
or switch door selection, the final player door selection.
}
\description{
\code{change_door()} selects one of the initial door selected from \code{select_door()}
or the third remaining door, but not the door known to have a goat from
\code{open_goat_door()}.
}
\details{
The Current game Contestant has selected a mysterious door which may have a
car or goat behind it, a known goat door has been revealed, and there is a
third door that is also mysterious goat or car. The Contestant can choose to
stay with their original door selection a.pick, or select the remaining
mysterious door, given the knowledge of a known goat door that was revealed
by the Host.
}
\examples{
change_door()

}
