% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{Contestant stay or switch doors.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{...}{A numeric value}
}
\value{
The function returns a length 1 numeric vector
indicating the position of the door they chose.
}
\description{
\code{change_door()} randomly generates the contestant's choice
to either switch to the remaining door or
remain with their initial selection.
}
\details{
The Contestant now chooses whether they would like to
switch to the remaining unopened door or to stay
with their initial selection. Their are aiming to
open the door with a car behind it.
}
\examples{
`change_door ()`
[1]

`change_door ()``
[2]

`change_door ()`
[3]

}
