% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{Change door.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay}{A logical value indicating whether the player decides to stay with the initial door selection (TRUE) or switch to the other door (FALSE).}

\item{opened.door}{A single integer between 1 and 3 representing the door number that was opened.}

\item{a.pick}{A single integer between 1 and 3 representing the player's initial door selection.}
}
\value{
A single integer between 1 and 3 representing the final door selection of the player.
}
\description{
change_door simulates the player's decision to change or stay with the initial door selection
}
\details{
The function returns the final door selection of the player based on whether they choose to stay with the initial selection or switch to the other door.
}
\examples{
  change_door(FALSE, 2, 1) # returns 3

}
