% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{change _door}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{if}{contestant selected car,
randomly select one of two goats}
}
\value{
one door, car or goat
}
\description{
The contestant changes a door
}
\details{
the contestant is given an opportunity to stay with their original selection
or switch to the other unopened door. There was a famous
debate about whether it was optimal to stay or switch when
given the option to switch, so this simulation was created
to test both strategies.
}
\examples{

}
