% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{Step 04: Contestant has opportunity to change door selection.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{Numeric}{}
}
\value{
If the contestant decides to stay (stay=T), the initial pick number is returned. If the contestant decides to switch (stay=F), the final pick is returned.
}
\description{
'change_door()' allows the contestant to adpot a stratgy to either 'Stay' or 'Switch' to increase chances of winning.
}
\details{
This function simulates a contestant making a final pick, choosing either to 'Stay' with the initial door selection, or
choosing to 'Switch and forfeit the initial selection in favor of the third, remaining door in hopes of selecting the
door with the car behind it.
}
\examples{
change_door()
##CODE CHUNK 4 UNIT TEST

opened.door <- open_goat_door( this.game, my.initial.pick)

change_door( stay=T,
             opened.door=opened.door,
             a.pick=my.initial.pick )
change_door( stay=F,
             opened.door=opened.door,
             a.pick=my.initial.pick )

my.final.pick <- change_door( stay=F,
                              opened.door=opened.door,
                              a.pick=my.initial.pick)

this.game
my.initial.pick
my.final.pick
}
