% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{Contestant changes doors.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{opened.door, }{a.pick ... these are default arguments from the previous
function `open_goat_door` to run the function correctly.

stay=TRUE or FALSE ... this argument represents the game-playing
                strategy to stay or change the door.}
}
\value{
The function returns a number (numeric vector) between 1 and 3 or
logical vector [TRUE (the same initial position)] indicating the position
of the finally picked door by the contestant.
}
\description{
`change_door()` this function is to give the chance for
the contestant to change his/her first selection.
}
\details{
After opening one of the goats' doors, now this function gives
the host contestant the option to change to the other door that
is still closed or stay with his/her first selection.
}
\examples{
change_door(stay=T, opened.door, a.pick = T)
change_door(stay=F, opened.door, a.pick = T)

}
