% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monty Hall Script.R
\name{change_door}
\alias{change_door}
\title{Switch or stay with initial door}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay}{This argument's default vaule is TRUE but can be made FALSE.}

\item{opened.door}{This argument is the numeric vector of length one equal to the opened door.}

\item{a.pick}{This is a numeric vector of length one equal to the player's initial door.}
}
\value{
This function returns the final pick of the player based on whether they chose to stay or switch.
}
\description{
If the player wishes to stay with their initial door, then their initial pick is made their final pick.
If the player wishes to switch to the remaining unopen door, then their final pick is made the unopened door.
}
\details{
This function determines whether or not the player wishes to stay through a logical argument. If true,
then the final pick is returned as the initial pick. If false, then the position of the door that was neither the
player's first choice or opened door is determined. This value is then made the player's final pick.
}
\examples{
change_door(stay = T, opened.door = 3, a.pick = 1)

change_door(stay = F, opened.door = 3, a.pick = 1)

}
