% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{Monty Hall Door Switch Opportunity}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{...}{The default stay value is true. It also requires the return
from open_goat_door(game, a.pick) and select_door().}
}
\value{
This function returns a value between 1 and 3 representing the
position of the participant's final choice.
}
\description{
This function represents the point in the show where the game show host
offers the participant an opportunity to change their choice to the one
remaining door.
}
\details{
It gives the user a choice to switch doors.  If stay is false, the
return is the only remaining door, if true, the return is the initial
chosen door.
}
\examples{

game <- create_game()
a.pick ,_ select_door()
opened.door <- open_goat_door(game, a.pick)
change_door( TRUE, opened.door, a.pick ) 
change_door( FALSE, opened.door, a.pick )
change_door( opened.door = opened.door, a.pick = a.pick )

}
