% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{Simulates whether or not the contestant will change their selection after
the host reveals a goat door.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay}{A logical vector}

\item{opened.door}{An integer between 1 and 3}

\item{a.pick}{An integer between 1 and 3}
}
\value{
the function returns an integer between 1 and 3
}
\description{
\code{change_door} determines the contestants final choice of door depending
on their behavior is set to Stay or Switch.
}
\details{
The game setup replicates the game on the TV show "Let's
Make a Deal" where there are three doors for a contestant
to choose from, one of which has a car behind it and two
have goats. The contestant selects a door, then the host
opens a door to reveal a goat, and then the contestant is
given an opportunity to stay with their original selection
or switch to the other unopened door. There was a famous
debate about whether it was optimal to stay or switch when
given the option to switch, so this simulation was created
to test both strategies.
}
\examples{
  change_door(stay=t, opened.door = 2, a.pick = 1)
  change_door(FALSE, 1, 2)
  
}
