% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPackageProject_Roach.R
\name{change_door}
\alias{change_door}
\title{Determine door strategy.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{Requires}{three arguments, including strategy, initial selection, and the opened door selection.}
}
\value{
The function returns a length of 1 numeric vector indicating the final door selected.
}
\description{
'change_door' selects a final door selection.
}
\details{
Selects the final door choice. Contestants can either choose to stay
with their original pick or switch to the other unopened door.
}
\examples{
change_door(stay = TRUE, opened.door, a.pick)
change_door(stay = FALSE, opened.door, a.pick)
}
