% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{Change Door}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{User}{opened.door function, and a.pick which was the first door picked}
}
\value{
Return the door number
}
\description{
Does the contestant want to change their pick
}
\details{
If the user wants to stick with their pick, we will call it the final pick.
}
\examples{
object <- function     return (number)
}
