% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{Contestant stays or switches.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay}{Logical argument to equal "T" or "F" to decide if the
contestant stays at the a.pick door or switches.}
}
\value{
The function returns a numeric vector (1-3) that cannot
match 'opened.door'.
}
\description{
'change_door()' uses the argument "stay" for the
contestant deciding to stay at the door they first picked or
switch to the remaining door.
}
\details{
If the contestant decides to stay (stay=T), the numeric vector
'a.pick' does not change. If the contestant decides to switch
(stay=F), the numeric vector that is not 'a.pick' or
'opened.door' is returned.
}
\examples{
  change_door(stay=T, opened.door, a.pick)
  change_door(stay=F, opened.door, a.pick)

}
