% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhallproblem.R
\name{change_door}
\alias{change_door}
\title{Allow the contestant to change doors in the Monty Hall Problem.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay}{is a logical vector that indicates whether the contestant chose
to stick with their selection (TRUE) or change their selection (FALSE)}

\item{opened.door}{is the numeric value of the door opened in the
'open_goat_door' function.}

\item{a.pick}{is the numeric value of the door selected by the contestant in
the game.}

\item{Doors}{is a numeric vector. Refer to the function 'select_door'}
}
\value{
The output is the numeric value of the contestant's final
door selection.
}
\description{
\code{change_door()} is a function that gives the contestant the option
of selecting the other remaining door or sticking with the door
they originally selected.
}
\details{
This function returns the number of the contestant's final choice.
The function works by eliminating the option of the door opened by the host.
If the contestant chooses to stay with their selection (stay = T), the
final door will be equal to the original selection. If the contestant
chooses to switch doors (stay = F), the function also eliminates the
option of the door previously selected by the contestant.
}
\examples{
  opened.door <- open_goat_door( this.game, my.initial.pick )
  change_door( stay=T, 
  opened.door=opened.door, 
  a.pick=my.initial.pick )
  change_door( stay=F, 
  opened.door=opened.door, 
  a.pick=my.initial.pick )
  
}
