% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{Changing doors option}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{Input}{numeric vector}

\item{If}{statement, if contestant chooses to not stay. Final pick is the door that
is not the open_goat_door and not a.pick in previous step.}
}
\value{
The function returns a number between 1 and 3 indicating which door the
contestant has decided to switch their selection to.
}
\description{
\code{open_goat_door()} This function was created to simulate the contestant's
game-playing strategy if they decided to choose another door that is still
closed.
}
\details{
First, a vector of three doors numbered 1,2,3 are created. Then two control
"If" statements return the contestant's final door selected based on whether
the contestant chooses to stay with their original selected door or choose
the other door that the host did not select.
}
\examples{
change_door()
}
