% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{Stay with or change the selected door.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay}{A logical value indicating whether the contestant stays with their original selection. The default is TRUE.}

\item{opened.door}{A numeric value between 1 and 3 indicating the door opened to reveal a goat.}

\item{a.pick}{A numeric value between 1 and 3 indicating the contestant's initial door selection.}
}
\value{
The function returns a numeric value between 1 and 3 indicating the contestant's final door selection.
}
\description{
'change_door()' determines the contestant's final door selection based on whether they choose to stay or switch.
}
\details{
If stay is TRUE, the contestant keeps their original door selection. If stay is FALSE, the contestant switches to the remaining unopened door.
}
\examples{
change_door(stay=TRUE, opened.door=3, a.pick=1)
change_door(stay=FALSE, opened.door=3, a.pick=1)
}
