% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{change_door}
\alias{change_door}
\title{Contestant decides to stay or switch.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay}{a true or false value}

\item{opened.door}{a number between 1 and 3}

\item{a.pick}{a number between 1 and 3}
}
\value{
The function returns a value between 1 and 3
}
\description{
\code{change_door()} reveals the final door choice depending if the contestant
stays or switches doors.
}
\details{
This function chooses the initial pick as the final pick if the contestant
decides to stay and switches to the remaining door if they decide to switch.
}
\examples{
change_door()

}
