% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhallfunctions.R
\name{change_door}
\alias{change_door}
\title{Contest chooses to stay or switch.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{"T"}{or "F", representing TRUE and FALSE, respectively,
decide whether the contest stays or remains.}
}
\value{
The function returns a number of 1-3 that does not match
'opened.door'.
}
\description{
In this function, 'change_door()', the argument "stay" represents the
contestant keeping their original choice or switching to the other
remaining door.
}
\details{
Again, a binary choice emerges. If the contestant stays with
their original choice (stay=T), then 'a.pick' remains the same.
However, if the contestant chooses to switch to the other remaining
door, then the vector that is not represented by 'a.pick' or
'opened.door' is selected.
}
\examples{
  change_door(stay=T, opened.door, a.pick)
  change_door(stay=F, opened.door, a.pick)

}
