% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{Make a Final Door Selection.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay}{A logical value (TRUE or FALSE) indicating whether the contestant decides to stay with their initial pick (default is TRUE).
opened.door A numeric value representing the door number (1, 2, or 3) that the host opened to reveal a goat.
a.pick A numeric value representing the contestant's initial door choice (1, 2, or 3).}
}
\value{
A numeric value representing the contestant's final door selection (1, 2, or 3).
}
\description{
This function determines the contestant's final door choice in the Monty Hall problem based on whether they choose to stay with their initial pick or switch to the remaining unopened door.
}
\details{
The contestant has the option to either stay with their original door selection or switch to the other unopened door. If they stay, the initial door choice is returned. If they switch, the function returns the other unopened door.
}
\examples{
Contestant chooses to stay
change_door(stay = TRUE, opened.door = 3, a.pick = 1)
Contestant chooses to switch
change_door(stay = FALSE, opened.door = 2, a.pick = 1)
}
