% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{\itemize{
\item this is called the change_door function
}}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{-}{The arguments here are if statements which account for whether or not a contestant decides to stay with their original pick, which would then make their final pick their initial pick represented by a.pick. If they deicde to switch or not stay, then their final pick will represent a choice from the doors that are not an opened door, and no a,pick.}
}
\value{
we are returned a final.pick, or the contestant's final selection
}
\description{
\itemize{
\item The contestant is given the option to change from their initial selection to the other door that is still closed. The function will represent the game-playing strategy as the argument stay=TRUE or stay=FALSE.
}
}
\details{
\itemize{
\item in the change_door obeject, we initiate a function which by default has stay = TRUE, which indicates
that the contestant has stayed. We utilize our opened.door and a.pick objects, as well as bring in our doors argument. we then use if statements to come to our final pick.
}
}
\examples{
- # test it
opened.door <- open_goat_door( this.game, my.initial.pick )

change_door( stay=T, 
           opened.door=opened.door, 
            a.pick=my.initial.pick )
change_door( stay=F, 
           opened.door=opened.door, 
           a.pick=my.initial.pick )

my.final.pick <- change_door( stay=F, 
                             opened.door=opened.door, 
                             a.pick=my.initial.pick )

this.game
my.initial.pick
my.final.pick 
}
