% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{Select final door.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{The}{paramaters for this function are stay, opened.door, and a.pick.}
}
\value{
The function returns a number from 1-3 that is the value of the final picked door.
}
\description{
\code{change_door()} allows contestants to change their original door selection.
}
\details{
This function determines whether stay is true or false.
}
\examples{
 change_door()

}
