% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{Change Door}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay}{Logical; \code{TRUE} if the contestant stays with their choice, \code{FALSE} if they switch.}

\item{opened.door}{A numeric value representing the door opened by the host.}

\item{a.pick}{A numeric value representing the contestant's current door choice.}
}
\value{
A numeric value representing the contestant's final door choice (1, 2, or 3).
}
\description{
Changes the contestant's door choice based on their decision to stay or switch.
}
\details{
If the contestant decides to stay, their original choice is returned. If they switch, the unopened door is returned.
}
\examples{
change_door(stay = TRUE, opened.door = 2, a.pick = 1)
}
