% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem (1).R
\name{change_door}
\alias{change_door}
\title{Change Doors}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{...arguments}{are (stay=T, opened.door, a.pick)
stay=T: True if contestant decides to proceed with the initial selection
stay=F: False if contestant switches the selection to a new door
opened.door: door opened by host
a.pick: initial selection of contestant}
}
\value{
The function returns a character vector of length 1
indicating the final pick of the contestant
}
\description{
Contestant chooses to either stay on the initial selection
or switch to a different one
}
\details{
The contestant has the option to proceed with the first selection of the door
or choose a different one that it is still closed. After the host make a selection
which is neither similar to the contestant's choice nor a car door, the
contestant is given the chance to choose.
}
\examples{
change_door (stay=T,opened.door,a.pick)
}
