% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{Change door or stay}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay}{indicates if the contestant keeps their initially
chosen door}

\item{opened.door}{The goat door opened by the host}

\item{a.pick}{An integer that represents the initially chosen door,
a number between 1 and 3.}
}
\value{
The function returns a length 1 numeric vector of a
number between 1 and 3 indicating the final opened door of
the contestant. It will either be the same door or a
different door than the initial pick.
}
\description{
\code{change_door()} generates a door number based on if the
player decides to keep their initial door pick or switch.
}
\details{
After the host opens the goat door, the contestant has the
option to stay with their initially chosen door, or "switch"
to choose the unopened door.
}
\examples{
  change_door( stay = T, opened.door = 2, a.pick=3)
    # 2 
  
}
