% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{Change the chosen door.}
\usage{
change_door(stay, opened.door, a.pick)
}
\arguments{
\item{stay}{Boolean indicating whether to stay with the original choice.}

\item{opened.door}{The door that was opened to reveal a goat.}

\item{a.pick}{The original choice.}
}
\value{
A number between 1 and 3 representing the final chosen door.
}
\description{
Change the chosen door.
}
\examples{
change_door(stay = TRUE, opened.door = 2, a.pick = 1)

}
