% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{Stay at original pick or change the door.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{...}{stay=T or stay = TRUE refers to staying at the
original pick. stay=F or stay = FALSE refers to switching.
opened.door is the previously opened goat door by the host
and a.pick is the original pick by the participant.
opened.door and a.pick are necessary for the logical
statements wrapped in this function to ensure that we aren't
choosing the already chosen or opened door from before if
the contestant decides to switch. opened.door and a.pick
is numeric and requires numeric input. stay needs either
TRUE or FALSE input.}
}
\value{
... Returns a number(door)/numeric vector.
}
\description{
This function represents the choice of the participant
to stay or to change doors (stay = T means to stay,
stay = F means to switch). Staying simply means remaining
at the original pick, while switching means not picking
the door that was originally chosen and not picking the
door that was opened by the host.
}
\details{
This function (change_door()) uses the arguments stay = T,
opened.door and a.pick. stay = T represents staying, versus
stay = F represents switching. opened.door is used to ensure
if the participant switches, that the participant does not
choose the door already opened by the host. a.pick is used to
represent the choice of staying, but also to ensure that if
the participant switches, that they don't choose the door
they already chose. If a participant stays, then they simply
remain with their original pick (a.pick), but if they switch
then they cannot choose the door that was opened by the host
(goat door, as represented by opened.door) and they can't
choose their original pick (a.pick) and thus the if
statement will return the only door that meets those two
requirements (not opened.door and definitely not a.pick).
}
\examples{

change_door( stay=T, opened.door=1, a.pick=3 )
change_door( stay=F, opened.door=1, a.pick=3 )

}
