% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{change_door}
\alias{change_door}
\title{Decide if you want to change the selected door or not.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{This}{function passes the \code{opened.door} (the opened goat door),
the \code{a.pick} (the originally selected door), and the new
parameter: \code{stay = T} (stay with \code{a.pick}) or \code{stay = F} (switch
to the other unopened door). Creating an object for the opened
door would be useful and easy to pass for the game's sake:
opened.doors <- \code{open_goat_door( this.game, my.initial,pick)}}
}
\value{
This function returns the final door (a number between 1 and 3)
that the contestant picks. This is dependent on the stay or
switch that is decided in the parameter.
}
\description{
\code{change_door()} tells the game whether or not a person switches
the door or stays on the originally selected door.
}
\details{
This function takes all the previous information from the game:
the order of the car and goats, the originally selected door,
and the open door that shows one of the unchosen doors that has
a goat behind it. This function then goes further to allow the
contestant to decide if they want to stay with the selected
door or switch to the only other unopened door.
}
\examples{
 `change_door( stay = T, opened.doors, my.initial.pick )`

 `change_door( stay = F, opened.doors, my.initial.pick )`

}
