% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_door.R
\name{change_door}
\alias{change_door}
\title{Monty Hall Game Door Change}
\usage{
change_door(stay = T, game, opened.door, a.pick)
}
\arguments{
\item{stay}{Logical. Stay on the same door or not?}

\item{game}{A vector of any class and length.}

\item{opened.door}{An integer that is not larger than the length of the \code{game} vector.}

\item{a.pick}{An integer that is not larger than the length of the \code{game} vector.}
}
\value{
The output is an integer number between, and including, 1 and \code{length(game)}
which equals \code{a.pick} if \code{stay = TRUE}. Otherwise, it returns a number different from
\code{a.pick} and \code{opened.door}.
}
\description{
\code{change_door ( stay = TRUE, game, opened.door, a.pick )} runs the third step of
a Monty Hall game where the contestant decides either to stay on the same door or to change
their decision.
}
\details{
Given a game vector, a door number that was the contestant's first selection, the
goat door revealed by the host, and a decision by the contestant whether to stay or
switch doors, this function makes the final door selection. If the contestant stays,
the function returns the door number of the initial selection, otherwise, it randomly
chooses one of the doors that were neither selected by the contestant or revealed by
the host.
}
\examples{
change_door( stay=T, c("goat", "goat", "car" ), 2, 1 )
change_door( stay=F, create_game(5), 4, 2 )

\dontrun{
change_door( stay=T, c("goat", "car", "car" ), 1, 5 )
}
}
