% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg (2).R
\name{change_door}
\alias{change_door}
\title{change door}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{#no}{arguments are used by the function.}
}
\value{
The function returns a number between 1 and 3
}
\description{
change_door() allows contestants to switch door.
}
\details{
then the contestant is given an opportunity to stay with their original selection
or switch to the other unopened door.
}
\examples{
change_door
}
