% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{Change the selected door.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay}{A logical value indicating whether the contestant
stays with the original door.}

\item{opened.door}{A number between 1 and 3 indicating the
door that was opened.}

\item{a.pick}{A number between 1 and 3 indicating the
contestant's original selected door.}
}
\value{
The function returns a number between 1 and 3
indicating the contestant's final selected door.
}
\description{
\code{change_door()} determines the contestant's final door
based on whether they stay or switch.
}
\details{
The function keeps the contestant's original selection
when staying and selects the other unopened door when
switching.
}
\examples{
  game <- create_game()
  a.pick <- select_door()
  opened.door <- open_goat_door( game, a.pick )
  change_door( stay=TRUE, opened.door, a.pick )

}
