% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall-pkg.R
\name{change_door}
\alias{change_door}
\title{Contestant changes their selection.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{The}{arguments for this function are "stay=T" which is if the contestant
chooses to stay or switch. opened.door is the numeric value of the opened
goat door in the previous step "open_goat_door". And a.pick is the
numeric value for the contestants original pick.}
}
\value{
The function returns a numeric value representing the
final door selection.
}
\description{
After the host opens a goat door, the change_door() is the contestant
deciding whether to stay with their current selection or switch to a
new door using the argument "stay".
}
\details{
If the contestant chooses to stay (stay = T), then the numeric value
"a.pick" is returned. If the contestant chooses to switch (stay = F),
than the numeric value that is not "a.pick" or "opened.door" is returned.
}
\examples{
   change_door( stay=T, opened.door, a.pick )
   change_door( stay=F, opened.door, a.pick )

}
