% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{TITLE}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{details}{}
}
\value{
details
}
\description{
Change door
}
\details{
details
}
\examples{
details
}
