% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{Change or Stay}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{if(}{stay ), final.pick <- a.pick
if( ! stay ), final.pick <- doors[ doors != opened.door & doors != a.pick}
}
\value{
A door between 1 and 3 is returned.
}
\description{
The player can choose to stay with the door they selected,
or change to one of the two remaining doors. The final
door is whichever door has not alreadybeen selected by
the player or the host.
}
\details{
If the play stays, the final pick is the first door selected,
a.pick. If the player changes, the remaining door is final.pick,
any door that is not the opened doors.
}
\examples{
 1, 2 or 3
}
