% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{Contestant chooses to stay with original choice or change doors.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{`stay`}{\itemize{
\item This tells the function if the contestant decides to stay
with their initial pick or not. This is a TRUE/FALSE value.
\code{opened.door} - This tells the function which goat door was opened.
\code{a.pick} - This tells the function what the contestant's
initial door pick was.
}}
}
\value{
The functions returns a number between 1 and 3 that represents
the contestants final door pick.
}
\description{
\code{change_door()} generates the contestant's final choice.
They can either stay with their original door chosen, or switch
to the last unopened door.
}
\details{
This function uses if statements that determine if the
contestant stayed or changed doors. If the contestant decides to stay
then the contestant's original door choice is returned as the final pick.
If the contestant decides not to stay, then the final
unopened door is returned as the final pick instead.
}
\examples{
  change_door( stay=T, opened.door, a.pick )
  change_door( stay=F, 3, 1)

}
