% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_door.R
\name{change_door}
\alias{change_door}
\title{Change Door}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{...}{stay=T, opened.door, a.pick}
}
\value{
returns to stay and a different door
}
\description{
\code{change_door} switches your inital door
}
\details{
This function switches your door to the sny of the random doors not already
selected or the opened goat door
}
\examples{
  change_door()
}
