% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{Change doors}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay}{logical value indicating if contestant stays or switches their initial pick.
`opened.door` represents the host's door
`a.pick` represents the contestants initial pick}
}
\value{
The function returns the contestant's final pick.
}
\description{
The contestant chooses to switch or keep their initial pick.
}
\details{
If the contestant chooses to keep their initial pick, the function returns it
as their final pick. If the contestant chooses to switch, the function picks the
remaining unopened door.
}
\examples{
change_door(stay=T, opened.door, a.pick=2)
}
