% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{Change Doors}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{The}{parameters to this function include the amount of doors in the game and whether
or not the contestant chooses to keep their current door choice.}
}
\value{
The function returns a length 1 character vector indicating the door choice.
}
\description{
'change_door' function allows the contestant the option to change their choice
in doors when they have already opened one of the losing doors.
}
\details{
This step in the game is when the contestant has to decide to stay or switch to
a new door choice before the final prize reveal.
}
\examples{
change_door( stay=T, opened.door=1, a.pick=3 )

}
