% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{Switching doors}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{Numeric}{}
}
\value{
The function will be a numeric value that indicates a door.
}
\description{
'change_door()' reflects the choices when contestant wants
the change their choice.
}
\details{
If the contestant stays with their choice, the final pick is the same.
If the contestant switches, the choice will be one of the remaining door.
}
\examples{
 change_door()
}
