% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{Contestant makes final door choice.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay}{Logical. TRUE/FALSE, representing the choice
to stay=TRUE, or switch stay=FALSE (the first argument).}

\item{opened.door}{A number, representing the goat door opened
by host (the second argument).}

\item{a.pick}{A number, representing the first door pick of the
contestant (the third argument).}
}
\value{
The function returns a numeric value indicating the
contestant's final pick, after staying or switching.
}
\description{
\code{change_door()} generates a number that represents the
door the contestant picks after host reveals the goat door.
}
\details{
From the TV show, "Let's Make a Deal," the contestant has made
their first pick and the host has to opened a door to reveal a
goat. The contestant is now given the opportunity to stay with
their original selection or switch to the other unopened door.
}
\examples{
  change_door(stay=TRUE, opened.door=2, a.pick = 3)

}
