% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{Stay or switch doors}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay}{TRUE to stay and FALSE to switch.}

\item{opened.door}{The door opened by the host.}

\item{a.pick}{The contestants first door choice.}
}
\value{
Returns the final door number.
}
\description{
Decides the final door based on if the contestant stays
or switches.
}
\details{
If stay is TRUE the contestant keeps their first door.
If stay is FALSE they switch to the other unopened door.
}
\examples{
change_door(stay=T, opened.door=2, a.pick=1)
change_door(stay=F, opened.door=2, a.pick=1)

}
