% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{Change the Chosen Door}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay}{Logical, indicating whether the contestant chooses to stay with their original choice.}

\item{opened.door}{The door number that was opened to reveal a goat.}

\item{a.pick}{The contestant's original choice of door number.}
}
\value{
The final chosen door number (1, 2, or 3).
}
\description{
Simulates the contestant either staying with their original choice or switching to another unopened door.
}
\details{
The \code{change_door()} function simulates the decision-making process of the contestant in the Monty Hall Problem.
When the \code{stay} parameter is set to \code{TRUE}, the contestant retains their original choice.
However, if \code{stay} is set to \code{FALSE}, the contestant opts to switch to the remaining unopened door.
This function provides the final selected door number as its output.
}
\examples{
  change_door(stay = TRUE, opened.door = 2, a.pick = 1)
  change_door(stay = FALSE, opened.door = 2, a.pick = 1)

}
