% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem functions Hauer.R
\name{change_door}
\alias{change_door}
\title{Contestant is Given the Option to Change Door Selection.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay}{A logical value indicating whether the contestant wants to stay
with their original door selection (TRUE) or switch to another door (FALSE).
The default value is TRUE.}

\item{opened.door}{An integer representing the door that was opened to reveal
a goat (a value between 1 and 3).}

\item{a.pick}{An integer representing the door that the contestant has
selected (a value between 1 and 3).}
}
\value{
The final selected door (a door number between 1-3) is returned.
}
\description{
The contestant makes a final pick on their door by either
staying with their original door selection or switching.
}
\details{
The contestant can choose to change from their initial selection to the other
door that is still closed. The function will represent
the game-playing strategy as the argument stay=TRUE or stay=FALSE.
}
\examples{
change_door()

}
