% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{change selected door}
\usage{
change_door(nDoor, stay = T, opened.door, a.pick)
}
\arguments{
\item{nDoor}{number of doors in the game}

\item{stay}{true of false depicting whether to stay with the picked door or not}

\item{opened.door}{the door opened by the host}

\item{a.pick}{the door picked by the player}
}
\value{
the final door selected
}
\description{
'change_door()' returns final choice to whether stay or change the door
}
\details{
if the stay parameter is true, then stay with the picked door
     if the stay parameter is false, then pick any other door from
     the game array except the one player picked and the one host opened
}
\examples{
     change_door(5, T, 2, 1)
}
