% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{The chance to change or stick with choice.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{...}{the function takes in parameters: stay (logical value),
opened.door (integer value), and a.pick (integer value).}
}
\value{
It outputs an integer value of the final pick if the user decided to stick
with their original pick or switch picks.
}
\description{
\code{change_door(stay, opened.door, a.pick)}The chance to change choice to the
remaining door available after revealing a goat door or sticking with the
original choice.
}
\details{
The function allows the user to have the option to change their choice to
the last door choice or stick with their choice after the host has revealed
a goat door.
}
\examples{
  When the user stays -> change_door(stay = T, opened.door, a.pick)
  When the user switches -> change_door(stay = F, opened.door, a.pick)

}
