% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{Contestant chooses to change or not change door}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{parameters}{are set around the contestant's choice to stay or switch}
}
\value{
return(final.pick) -> returns the door number of the contestant's final pick
}
\description{
At this point in the game, the contestant is now given the option
to keep their first door selection or switch their choice
to the last remaining door
}
\details{
'change_door" function is created building on the 'opened.door' and 'a.pick' functions.
if, else statements are again created, this time based on the choice of the contestant.
if the contestant chooses to stay,
their final pick 'final.pick' is equal to their initial pick 'a.pick'
if the contestant chooses to switch,
their final pick 'final.pick' is not equal to their initial pick
and also not equal to the host's pick 'opened.door'
}
\examples{
change_door -> 1
change_door -> 2
}
