% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{The other door opportunity.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a length 3 character vector
  indicating the door the door number that is not the original selection, nor the 
  goat door that was shown to them in the previous segement.
}
\description{
Monty always gives each contestant the ability to switch doors.
}
\details{
If the contestant wishes to keep their first selection stay will = T, if they choose
to switch doors the selection will be stay = F.
}
\examples{
change_door

}
