% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{Contestant chooses final door in Monty Hall Problem game.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{`stay`}{is a logical value indicating whether a contestant wants to
stay with their initial selection (\code{TRUE}) or switch to the  remaining
door (\code{FALSE}). \code{opened.door} is an integer (1, 2, or 3) representing the
door the host opened to reveal a goat in \code{open_goat_door}. \code{a.pick} is an
integer (1, 2, or 3) representing the contestant's initial door selection
from \code{select_door()}.}
}
\value{
An integer (1, 2, or 3) representing the door number that the
contestant ends up with after deciding to keep or switch.
}
\description{
\code{change_door()} simulates what happens when a contestant is given the
choice to keep their initial selection or switch to the remaining closed
door after the host eliminates a door with a goat behind it.
}
\details{
In the Monty Hall Problem game, the contestant chooses 1 of 3 doors, then
the host opens one of the remaining doors to reveal a goat. This leaves
2 closed doors--the contestant's initial selection and a door they did
not initially select. The contestant decides whether to keep their
initial selection or to switch doors. This function simulates that will
happen for both choices.
}
\examples{
  change_door()
}
