% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{Contestant chooses their strategy.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{...}{This function takes three arguments. The first is a boolean (\code{stay}),
which determines whether to stick with the initial door picked by contestant
(\code{a.pick}) or to pick the remaining door, which is represented by the integer
not occupied by \code{opened.door} or \code{a.pick}. For the former, \code{stay} should be T,
for True and for the latter, \code{stay} should be F, or False, i.e. the strategy
adopted by the contestant is "switch". \code{opened.door} is the door with a goat
behind it. \code{a.pick} is the initial door picked by contestant. Both are integers
between 1 and 3 inclusive.}
}
\value{
This function returns an integer value between 1 and 3 inclusive. \code{a.pick}
if \code{stay}=T, and another value which is not \code{opened.door} or \code{a.pick} if \code{stay}=F.
This value is stored in \code{final.pick}.
}
\description{
\code{change_door()} function will return the final door number picked
based on whether a "stay" or "switch" strategy was opted.
}
\details{
When the door with the goat behind it is opened, the contestant
has to now decide whether to stay with their initial pick or to
switch to the other door before the outcome can be revealed.
}
\examples{
change_door(stay=F,opened.door=2,a.pick=1)
}
