% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{Option to Change Doors}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{The}{function is using the whether the contestant chose to stay, the host door,
and the first pick as the parameters.}
}
\value{
The return will either have the same door as the first initial pick or a new
door number.
}
\description{
'change_door()' gives the contestant the option to either stay with their
first choice or switch doors.
}
\details{
If the contestant chooses to stay (stay=T), then their final pick will be the
same as their first picl; otherwise, the contestant chooses to switch doors,
their final decision will be a different door number.
}
\examples{
If a goat door = 1 and I chose to 'stay' - the return value would be '1'.
}
