% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{Switch or stay}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay}{opened.door
a.pick}
}
\value{
final choice numeric number
}
\description{
Does the contestant keep their initial choice or switch?
}
\details{
Switch or stay choice
}
\examples{
  opened.door <- open_goat_door(create_game(), select_door())
  a.pick <- select_door()
  change_door(TRUE, opened.door, a.pick)
}
