% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{change_door}
\alias{change_door}
\title{The contestant can stay or change doors.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns the contestant's final choice.
}
\description{
`change_door()` selects either stay or change door based
  on the contestant's choice.
}
\details{
The contestant is given the option to change doors or
  to stay with their original choice. Depending on the
  contestant's choice, the function will open the opposite door.
}
\examples{
  change_door()

}
