% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{change_door}
\alias{change_door}
\title{Contestant choses to switch doors or stick with initial pick.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay}{The strategy used (T = stay, F= switch)}

\item{opened.door}{Door opened by the host}

\item{a.pick}{initial pick of the contestant}
}
\value{
The final pick of the contestant
}
\description{
Contestant gets an option to switch his initial selection after host reveals a door with a goat behind it.
}
\details{
If the contestant chooses to stay with with the initial pick, the final pick is the same as the initial. If the contestant chooses to switch his pick, the function selects the only door that is left after eliminating the initial pick and the door opened by the host.
}
\examples{
change_door(stay = T, 2, 1)
}
