% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{Final door selection}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{`stay`}{A logical value (T or F) indicating whether the contestant chooses to stay with their original door selection \code{TRUE} or switch to the other unopened door \code{FALSE}}

\item{`opened.door`}{An integer (1, 2, or 3) representing the door that was opened by the host to reveal a goat.}

\item{`a.pick`}{An integer (1, 2, or 3) representing the door initially selected by the contestant.}
}
\value{
The function returns an integer (1, 2, or 3) representing the contestant's final door selection.
}
\description{
\code{change_door()} determines the contestant's final door selection based on whether they choose to stay with their original pick or switch to the other unopened door.
}
\details{
After the host opens a door to reveal a goat, the contestant has the option to either stay with their inital door selection or switch to the other unopened door. This function updates the contestant's final pick based on their decision.
}
\examples{
`change_door()`
}
