% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{Change Doors}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The return will be the remaining door selected by the contestant
}
\description{
'if(stay)' will be used in conjunction with
'if( ! stay)' for the contestant to choose
whether or not to switch doors
}
\details{
The contestant will have the option to switch doors, or keep
their previously selected door. A if function will be used to
create an option if the contestant chooses STAY, or if the
contestant chooses SWAP
}
\examples{
change_door
}
