% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty_hall_pkg1.R
\name{change_door}
\alias{change_door}
\title{Change doors.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay}{a logical vector indicating whether the contestant stay with
their initial selection or switch to a different door. This argument's
default value is TRUE.}

\item{opened.door}{a number between 1 and 3 indicating the host's selection.}

\item{a.pick}{a number between 1 and 3 indicating the contestant's selection.}
}
\value{
The function returns a number between 1 and 3 indicating the
contestant's final selection.
}
\description{
\code{change_door()} give the contestant the option to stay with
their initial selection or switch to the other door that is still closed.
}
\details{
\code{change_door()} assigns the contestant' inital selection as their
final decision if the contestant does not want to change; otherwise,
\code{change_door()} selects the remaining closed door that was not
previously selected by either the contestant or the host.
}
\examples{
  game <- create_game()
  first.pick <- select_door()
  host.door <- open_goat_door(game, first.pick)
  change_door(stay=T, host.door, first.pick )

}
