% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{Contestant stays or switches}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{'doors'}{\itemize{
\item the numbers 1, 2, 3 that correspond to the the three doors in
the game
'a.pick' - one door numbered either 1, 2, or 3, which is the contestant's
initially selected door
'stay' - the contestant's initial pick is their final pick
'switch' - the contestant's final pick is the door that they did  not
initially pick that the host has not opened
}}
}
\value{
The function returns 'final.pick', which is a number between 1 and 3 that
corresponds to the door the contestant ended up with at the end of the
game depening on if they chose to stay or to switch.
}
\description{
The 'change_door()' function generates a number from 1 to 3 that
determines if the contestant stays with the original door or switches to
a different door.
}
\details{
After the host opens a goat door, the contestant is given the choice to
stay with their original pick of door or to switch to the other
unopened door. This decision determines which door is the contestant's
final pick.
}
\examples{
    change_door()

}
