% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHallKRB.R
\name{change_door}
\alias{change_door}
\title{Returns final pick of the guest, door 1, 2, or 3.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay}{= T or F, opened.door, a.pick}
}
\value{
Returns a vector of length one and class numeric, which represents the guest's final choice of door.
}
\description{
If guest chooses to stay with their first pick of door, then a.pick is equal to their final.pick.
If the guest chooses to change doors, then their final pick is neither a.pick (guest's first pick)
nor the opened.goat.door (or the door opened by the host)
}
\details{
Because the host opened a goat door, the final.pic of the guest cannot be this door. The guest's
final pick of door can either be their first pick (a.pick) or the remaining door (not opened.door
or a.pick)
}
\examples{
final.pick
}
