% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg-Wickerham.R
\name{change_door}
\alias{change_door}
\title{The contestant's final choice.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{The}{argument for this function is stay=TRUE or stay=FALSE, and it also
contains opened.door (the door number opened by the host) and a.pick (the
door originally selected by the contestant) to return the final pick.}
}
\value{
This function returns a numeric value between 1 and 3. If stay=TRUE,
it will return the value of the door the contestant chose. If stay=FALSE,
it will return the remaining door.
}
\description{
'change_door()' will return the final selection of the contestant based
off whether the contestant chooses to keep their original door (stay)
or change it (switch).
}
\details{
The contestant now makes their final choice to either stay with their
original door or switch to the lat one remaining.
}
\examples{
    change_door( stay=T, opened.door=1, a.pick=3 )
}
