% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{Determines whether a contestant will decide to choose to keep the same door
or choose to "switch" doors.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{`opened.door`}{Numeric integer
\code{a.pick} Numeric integer
\code{doors} Character vector
\code{final.pick} Numeric integer}
}
\value{
The function will return a character string of either "car" or "goat"
}
\description{
\code{change_door} will filter out the original contestant's door choice and the
door selected by the game host and leaves the remaining door, and will
provide the contestant the option to keep their door choice or switch.
}
\details{
This is the step in the game that will determine whether the contestant
wins or loses the game.
}
\examples{
change_door(1:3, 2, 1:2, "SWITCH")
}
