% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{Change Door}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{The}{`change_door()` function take three arguments: stay, opened.door
and a.pick. Stay is a True or False boolean. opened.door and a.pick
are both integers between 1 and 3.}
}
\value{
The value returned is a integer between 1 and three that represents
  the final door picked by a contestant.
}
\description{
A contestant can elect to choose a door different from the one they
  initially chose. This function handles that choice.
}
\details{
If the contestant choosees to stay, then the final.pick variable
  takes on th value of a.pick. If a contestant does not choose to stay,
  the final.pick variable will take on the value of the third,
  not-yet-revealed door.
}
\examples{
  change_door( stay = T, opened.door, a.pick )

}
