% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{Contestant Switches or Stays with Door}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{...}{This function has three arguments: stay = T or stay = F,
opened.door, and a.pick: stay is the argument that will
reflect if the door selection is switched ( stay = F ) or
stays as the initial selection ( stay = T ). The argument
opened.door is the argument that indicates the door that
has been opened by the host. The argument a.pick is the
argument that indicates the door that was initially selected
by the contestant}
}
\value{
The function returns a length of one numeric vector
that is either \code{a.pick}, or of the remaining unopened
door that is not \code{opened_door}. This will be saved
as \code{final.pick}.
}
\description{
\code{change_door()} generates a selection of a door by
the contestant through a numeric vector of 1, 2, or 3,
depending on if the strategy "stay" is true (stay = T)
or false (stay = F).
}
\details{
The contestant will be given an opportunity to stay
with the same door initially selected or to switch to
the other unopened door. If the contestant chooses to
stay with the same door that was initially selected
( stay = T ), this will generate \code{a.pick}, the initial
door selection. If the contestant decides to switch doors,
where ( stay = F ), the door that remains closed and
is not \code{a.pick} will be selected.
}
\examples{
  change_door ( stay=T, opened.door, a.pick )
  change_door ( stay=F, opened.door, a.pick )

}
