% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{Change the door selection.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay}{A logical value indicating whether the contestant chooses to stay
with their initial door (\code{T}) or switch to the other door (\code{F}).}

\item{opened.door}{An integer that represents the door number that was opened
to reveal a goat.}

\item{a.pick}{An integer that represents the initial door pick by the player.}
}
\value{
An integer that represents the contestant's final door choice (1, 2, or 3).
}
\description{
\code{change_door()} allows the contestant to either stay with their original
door or switch to the remaining unopened door in the Monty Hall game.
}
\details{
This function determines the final door choice of the contestant based on
whether they choose to stay with their initial selection or switch to the
other unopened door. If the contestant chooses to stay, the function
returns their original pick. If they choose to switch, the function returns
the other available door.
}
\examples{
  opened.door <- 2
  a.pick <- 1
  change_door(stay = T, opened.door, a.pick)
}
