% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{Change or stay with chosen door.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay:}{Logical; TRUE if the contestant stays with their first pick,
FALSE if they switch.
opened.door: A numeric value (1, 2, or 3) representing the door opened
by the host to reveal a goat.
a.pick: A numeric value (1, 2, or 3) indicating the contestant's
original chosen door.}
}
\value{
A single numeric value (1, 2, or 3) representing the door opened
by the host.
}
\description{
change_door() simulates the contestant’s decision to either stay with
their original door or switch to the remaining unopened door.
}
\details{
If stay = TRUE, the contestant keeps their original pick.
If stay = FALSE, the contestant switches to the unopened door that is
neither the host’s opened door nor their original pick.
}
\examples{
#' @examples
game <- create_game()
first.pick <- select_door()
opened.door <- open_goat_door(game, first.pick)
change_door(stay = FALSE, opened.door = opened.door, a.pick = first.pick)


}
