% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{CHANGE DOORS}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{'stay'}{indicates whether the contestant should stay with their initial pick (TRUE)
or switch to the other remaining door (FALSE).
'opened.door' is the door the host opened to reveal a goat (1,2 or 3)
'a.pick' is the initial door chosen by the contestant (1,2, or 3)}
}
\value{
The final door number that is chosen by the contestant (1,2 or 3)
}
\description{
The contestant is given the option to change their initial selection
to the remaining closed door.
}
\details{
This function decides the contestant's final door choice. STAY = TRUE
if the contestant keeps their initial choice. STAY = FALSE if the contestant
switches to the remaining unopened door.
}
\examples{
 change_door(stay = TRUE, opened.door = 2, a.pick = 1)
}
