% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Montyhall.R
\name{change_door}
\alias{change_door}
\title{Change doors option}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{The}{stay argument is a logical vector, as it contains "True" and "false" value.}
}
\value{
The function returns a length of one numeric vector indicating the final pick made by the contestant after
      being offered by the host the option to change doors.
}
\description{
The contestant is given the option of changing doors from his first selection and change to the other door
           that is still closed.
}
\details{
The change_door()uses the game-playing strategy of sticking to the contestant's first selection or not
       by using the arguments stay = True or stay = False. If the contestant decided to stay (i.e. stay = T), the
       returned result or the final pick will be the contestant's initial selection (a.pick). If the contestant
       decided not to stay(i.e. stay = F or ! stay), the returned result or the final pick will be the other
       closed door (i.e the door that was not initially selected by the contestant or the goat door opened by the
       host)
}
\examples{
        change_door( stay=T, opened.door=1, a.pick=3 )
        change_door( stay=F, opened.door=1, a.pick=3 )

}
