% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_monty-hall-problem (2).R
\name{change_door}
\alias{change_door}
\title{Option to change door selected.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{In}{this, staying is represented as True or False}
}
\value{
Function will return final pick for the contestant. They either stay with
initial pick (Stay= T), or change to closed door (Stay=F).
}
\description{
the 'change_door'  function allows the contestant to change their initial
selection to the other door that is still closed.
}
\details{
This function represents the playing strategy as the argument, staying or
switching. With this, the contestant is offered an opportunity to switch
to the closed door or stay with original selection.
}
\examples{
change_door()

}
