% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{Change doors}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{...}{arguments are ( stay=T, opened.door, a.pick )}
}
\value{
The function returns a length 1 numeric vector.
}
\description{
Contestant has the choice to switch or stay with original choice
of the door.
}
\details{
The contestant is given the option to change from their initial selection
to the other door that is still closed.
The function will represent the game-playing strategy
as the argument stay=TRUE or stay=FALSE.
After host opens one of the doors that is not what the contestant picked nor the car door
the contestant has the choice of switching to the other door.
}
\examples{
change_door( stay=T, opened.door, a.pick )


}
