% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem-brentlinger.R
\name{change_door}
\alias{change_door}
\title{Player chooses to stay or switch.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{arguments}{used by the function are 'stay' in which stay=T means the
player is keeping their original pick, or stay=F the player will switch doors.
The default is stay=T, 'opened.door' which is the door that the host opened
in the 'open_door()' function, and 'a.pick' which is the player's original
selection in the 'select_door()' function.}
}
\value{
The function returns a length 1 number vector, indicating the final player
selection.
}
\description{
After the host has opened a remaining goat door, there are now two doors
left: the player's original selection and one more door.  The player now
must choose whether they want to keep their original door, or
change to the remaining unopened door.
}
\details{
The player can either stay with their original pick, in which case that
becomes their final selection.  Or they can choose to switch doors,
in which case their final selection becomes the remaining unopened door
(not the original pick and not the opened goat door).
}
\examples{
change_door(stay=T, opened.door, a.pick)

}
