% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{Stay with or change the selected door}
\usage{
change_door(stay = TRUE, opened.door, a.pick)
}
\arguments{
\item{stay}{Logical value indicating whether the contestant stays with
the original choice.}

\item{opened.door}{The door opened by the host.}

\item{a.pick}{The contestant's original door selection.}
}
\value{
An integer indicating the contestant's final door choice.
}
\description{
Determines the contestant's final door choice based on strategy.
}
\details{
If \code{stay = TRUE}, the contestant keeps the original door. If
\code{stay = FALSE}, the contestant switches to the remaining unopened door.
}
\examples{
change_door(TRUE, opened.door = 2, a.pick = 1)
change_door(FALSE, opened.door = 2, a.pick = 1)

}
