% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{\preformatted{Stay with Initial Door or Switch?
}}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a numeric vector 1-3.
}
\description{
\preformatted{`change_door()` return a numeric vector in regards to 
one of the two remaining doors.   
}
}
\details{
\preformatted{In the original "Let's Make a Deal" game, after one of
the goat doors was revealed, contestants have the option
to stay with the door that was initially picked, or to 
switch to the other remaining door. As mentioned above, 
this is the part of the debate about whether it was 
optimal to stay or switch when given the option to switch.
}
}
\examples{
    change_door()

}
