% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{The decision of a lifetime}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{Lets}{break down the arguements piece by piece, shall we? first arguement
again, is the doors function that has an assigned value of 1,2,3, which as
discussed from the previous question. The purpose for assining the numbers
1,2,3 is to be associated with the criteria for the gameplay which has doors
1 to 3. Now to the if statements, the first if statement is simply stating
that if user has decided to stay then the "final.pick" for the user shall
be the same as the initial pick, therefor no change is needed. On the
contrary, "! stay" if the user choose not to stay then the final pick
cannot be either the initial pick or cannot be the door that has already
been opened. Hence why this is a conditional statement for the final pick
for the user that wishes not to stay.}
}
\value{
The return function is simply calling the statement final pick which has
been defined based on weather the user decides to stay or switch doors
one the user decides to stay or switch then the argument that is associated
with the final.pick , based on the decision of the user, is then executed.
This results will then populate one the change door function is called due
the return function being called at the end for 'final.pick'
}
\description{
After the reveal of the goat door occurs the user is given a choice as to
weather they shall stay with their current selection for the door vs being
able to switch to the alternative door. This gives the false perception to
the user that they posess a 50/50 change of opening the door with the car
prize, When i reality this isn't entirely the case. After the user has
picked their door the goat door is then revealed which then only gives the
user two doors to choose from. The decision is then left to the user creating
a sense of uncertainty and tension within the audience.
}
\details{
Details regarding the gameplay is simple and especially in the code. This
game is ulitmatly a pyschological game which plays on the inner conflict of
the individual, thus assuming that what isn't obtainable is often the answer
within our inner psyche. This montey hall problem was actually solved by
Marilyan vos Savant who was able to answer this question on a TV show which
qas about her being able to answer the most question, thus earning her title
as the "smartest person alive" She had recieved a lot of backlash for her
decision and rationale but ultimatley was proven right. This problem has
baffled mathematical professionals thus earning its reputation.
}
\examples{
change_door()
}
