% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{Determines whether the contestant keeps their initial door choice or switches
to the remaining door choice.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay}{Logical integer}

\item{opened.door}{Numeric integer}

\item{a.pick}{Numeric integer}
}
\value{
Numeric integer representing the door that was opened.
}
\description{
'change_door()' uses 'if' to determine whether the contestant keeps the initial
door choice of switches to the remaining door.
}
\details{
This function is the third step of the game and represents the contestant's
final door choice.
}
\examples{
change_door(stay=T, opened.door = 1, a.pick = 2)
}
