% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problemb.R
\name{change_door}
\alias{change_door}
\title{Contestant changes or uses same door.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{...}{three arguments are used by the function.}
}
\value{
The function returns a length 1 vector
indicating the position of the choice of the constant after stay or switch.
}
\description{
\code{change_door()} generates a options for contenstant to change door
choice or stay with door choice.
}
\details{
The function allows a random decision of a contestant to choose
to stay with the choice made, or change the door selection.
This decision impacts the play_n_game() function.
}
\examples{
  change_door()
}
