% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen comments.R
\name{change_door}
\alias{change_door}
\title{Strategy stay or switch}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{The}{parameter stay is a logical vector which can be true or false. It represents
whether the player keeps their original pick or switches to the other door.
The parameter opened.door is the number of the door that the host opened to
reveal the goat.
The parameter a.pick is an integer that represents the door number that
the player picked first.}
}
\value{
This function returns a number between 1 and 3 which coincides with
a stay or switch selection.
}
\description{
\code{change_door()} picks either the initial door identified during \code{select_door()}
or the third remaining door.
}
\details{
After the game player makes the initial door pick, and the game host opens a
second door revealing a goat, the player must decide whether to stay with their
original pick or choose the other remaining door.
}
\examples{
change_door()

}
