% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{change_door}
\alias{change_door}
\title{Allows the user to switch their previously selected door.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{This}{function utilizes the arguements 'opened.door' from
the 'open_goat_door' function and 'a.pick' from the
'select_door' function.}
}
\value{
This function will return an integer variable that
corresponds to the door number that the user has
selected after deciding if they will switch doors.
Doors can be switched by utilizing 'stay=FALSE'
}
\description{
'change_door' allows the user to determine if they
would like to switch their selected door to the
remaining door.
}
\details{
This function allows the user to switch their selected
door to the one not chosen by the 'select_door' and
'open_goat_door' functions. Users have the ability to
keep their original choice or to switch doors.
}
\examples{
 final.pick <- change_door( stay=F,
 opened.door=opened.door,
 a.pick=my.initial.pick )

 final.pick <- change_door( stay=T,
 opened.door=opened.door,
 a.pick=my.initial.pick )

}
