% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOCUMENTATION.R
\name{change_door}
\alias{change_door}
\title{Change door}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{If}{the contestant decides to stay with their initial selection,
then that door is the final selected door and not the other unopened door.
If the contestant decides to choose the other unopened door,
then that door is selected and made the final selected door in the game.}
}
\value{
The function returns five vectors. The first and fourth are a length 1 numerical
vector indicating the initial door selected by the contestant.
The second is a length 1 numerical vector indicating the
the final unopened door should the contestant choose to exchange it
for their initial selection.The third is a length 3
character vector indicating the positions of goats and the car.
The fifth is a length 1 numerical vector indicating the final selected door
in the game.
}
\description{
\code{change_door()} selects the final unopened door in the game.
}
\details{
The contestant is given the option to exchange their initial
selection for the remaining unopened door.
}
\examples{
  opened.door <- open_goat_door( this.game, my.initial.pick )
  change_door( stay=T, opened.door=opened.door, a.pick=my.initial.pick )
  change_door( stay=F, opened.door=opened.door, a.pick=my.initial.pick )
  my.final.pick <- change_door( stay=F, opened.door=opened.door, a.pick=my.initial.pick )
  this.game
  my.initial.pick
  my.final.pick 

}
