% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem-cable.R
\name{change_door}
\alias{change_door}
\title{The contestant chooses to either keep their original door or switch to
the other unopened door.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{This}{function uses three arguments: Whether the contestant stays or not
(which defaults to True), the goat door that the host opened, and the
original door that the contestant opened.}
}
\value{
The function returns a length 1 numeric vector indicating the final door
that the contestant has committed to.
}
\description{
\code{change_door()} sets three doors labeled "1", "2", and "3", then
determines what the final choice of the contestant will be.
If the contestant stays with the door they originally chose, then
their final pick will be the same door that they chose in the beginning.
If the contestant switches doors, then their final pick will be
the door that isn't the door the host opened and also isn't the door
they picked originally. It then returns the value of the final door
selected.
}
\details{
Now that the host has revealed a goat door, the contestant has two options.
They can stay with their original door choice and hope that they
picked the car door the first time, or they can switch to the unopened
door and hope that that door hides the car.
}
\examples{
  change_door()
  change_door( stay=F, opened.door, a.pick )
  
}
