% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.r
\name{change_door}
\alias{change_door}
\title{Change or keep the initially chosen door}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay}{Logic Operator. If TRUE, the contestant keeps their initial choice.}

\item{opened.door}{Intiger. Intiger 1, 2, or 3. Represents the goat door opened by the host.}

\item{a.pick}{Intiger 1, 2, or 3. Represents the contestant's initial door choice.}
}
\value{
The function returns an integer 1, 2, or 3. This integer
represents the contestant's final door choice.
}
\description{
Updates the contestant's choice depending on whether to stay with
their initial selection or switch to the other door.
}
\details{
If "stay = TRUE", the contestant remains with their initial selection
If "stay = FALSE", the contestant will switch doors to the only remaining door.
}
\examples{
g <- create_game()
p <- select_door()
open_goat_door(g, p)
change_door <- (stay = TRUE, opened.door = 0, a.pick = p)
change_door <- (stay = FALSE, opened.door = 0, a.pick = p)
}
