% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{\preformatted{Want to stay or switch doors?
}}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
the function returns a number between 1 and 3 to represent one of the three doors
}
\description{
\preformatted{The contestant will select their final pick
}
}
\details{
\preformatted{The contestant can either choose to stick with 
their original pick or switch to the other remaining
closed door. 
}
}
\examples{
	change_door()

}
