% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{Will the contestant choose a different door?}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay}{A logical variable (T, F) with a default STAY=TRUE.}

\item{opened.door}{The opened door; integer between 1 and 3.}

\item{a.pick}{The initial door chosen; integer between 1 and 3.}
}
\value{
The function returns an integer between 1 and 3, representing
the contestant's final choice of door to open.
}
\description{
\code{change_door()} allows the contestant to stay with their
initial pick or select the other unopened door.
}
\details{
The function returns the initial pick as the final pick if
the contestant chooses to STAY.  If the contestant chooses
not to STAY, then the function returns the other unopened door.
}
\examples{
  change_door( stay=T, 1, 2 )

}
