% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{Contestant has opportunity to change door selection.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{If}{contestant chooses to switch doors, the selected door
cannot be the opened door or the original selection.}
}
\value{
Returns a numeric value indicating final door selection.
}
\description{
\code{change_door()} gives the option to stay on first choice
or switch to the remaining unopened door.
}
\details{
The game setup replicates the game on the TV show "Let's
Make a Deal" where there are three doors for a contestant
to choose from, one of which has a car behind it and two
have goats. The contestant selects a door, then the host
opens a door to reveal a goat, and then the contestant is
given an opportunity to stay with their original selection
or switch to the other unopened door. There was a famous
debate about whether it was optimal to stay or switch when
given the option to switch, so this simulation was created
to test both strategies.
}
\examples{
  change_door()
  
}
