% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{Simulate contestant staying with original door
or switching doors after goat door is revealed.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{opened.door}{is a numeric integer.}

\item{a.pick}{is a numeric integer.}

\item{stay=T}{is a logical value meaning stay equals true.}
}
\value{
This function returns a numeric vector.
}
\description{
\code{change_door()} uses the logical statement stay=T and if funtion to determine which
door will be the final door depending if the contestant chooses to stay or switch.
}
\details{
This function performs the third step in the game where the contestant chooses
to stay with their original door or switch doors, now that a goat door has been revealed.
}
\examples{
change_door()

}
