% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem-R-script.R
\name{change_door}
\alias{change_door}
\title{The contestant is given the opportunity to switch to a
different door that neither they nor the host have
already selected/opened or they can stay at their
initial door.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{...}{The arguments are \code{stay} (its value is TRUE if
contestant stays at their door or FALSE if contestant
chooses to switch to the other door), \code{opened.door}
(its value is the number of the door that the host
opened), and \code{a.pick} (its value is the number of
the initially chosen door).}
}
\value{
The function returns an integer representing the number
of the door that the contestant switched to or stayed at.
Returns a value which can be stored in a variable \code{finalDoor}.
Example output: 3
}
\description{
\code{change_door()} Type: Integer -> generates a door number
that represents the door that the participant switched
to or stayed at.
}
\details{
This function is provided in order to allow the
contestant to switch their door if they wish to.
This function helps to understand if staying or
switching leads to greater chances of winning overall.
}
\examples{
  change_door(stay = T, opened.door = openedDoor,
  a.pick = firstDoor)
  change_door(stay = F, opened.door = 2, a.pick = 1)

}
