% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{Indicates whether a contestant would like to switch doors or keep their
original selection}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay}{= T/F logical, indicates whether the contestant will keep their
original selection}

\item{opened.door}{ numeric value between 1 and 3 representing the door opened
by the host (see select_door())}

\item{a.pick}{ numeric value between 1 and 3 representing the door originally
selected by the contestant (see open_goat_door())}
}
\value{
Generates a numeric value between 1 and 3
}
\description{
After a goat door is revealed by the host, the contestant can choose to
switch their door selection to the other unopened door, or keep their
original selection.
}
\details{
If the contestant elects to switch doors, the function will return the
number of the remaining unopened door. If the contestant elects to keep
their selection, the function will return the numeric value associated
with the original selection
}
\examples{
change_door(stay=T, opened.door = 1, a.pick = 2) #should deliver output 2
change_door(stay = F, opened.door = 1, a.pick = 2) #should deliver output 3
change_door(stay = T, opened.door = 2, a.pick = 1) #should deliver output 1
change_door(stay = F, opened.door = 2, a.pick = 1) #should deliver output 3
}
