% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg-3.R
\name{change_door}
\alias{change_door}
\title{\if{html}{\out{<div class="sourceCode">}}\preformatted{The host offers the player the option to swap their door 
}\if{html}{\out{</div>}}}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{This}{function takes in a 'stay' variable, the opened.door variable, and the first a.pick variable.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{This function returns a final.pick variable that is an integer between 1 and 3
}\if{html}{\out{</div>}}
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{change_door() represents the host offering the contestant the option to swap their door selection to the only unopened remaining door. 
}\if{html}{\out{</div>}}
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{This function takes in a value for a 'stay' variable which represents the player sticking to their initial choice or swapping doors. It outputs the final door selection. 
}\if{html}{\out{</div>}}
}
\examples{
	change_door(stay=F, opened.door, this.pick)
}
