% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{The contestant receives an option to select a different door.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{...}{three arguments ( stay which is assinged to T, opened.door. a.pick) are used.}
}
\value{
The function returns 'final.pick' which is either 
   a.pick (initial selected door) or the remained door.
}
\description{
Change_door creates a vector of doors numbered 1, 2, and 3.
   Then, based on the contestant's choice to either stay with their 
   initial pick or switch, determine the final selection. 
   Finally, it returns the chosen door.
}
\details{
Once the host has revealed a remaining goat door, only two options remain for the contestant.
    Those two options are their initial pick and one more remaining door. At this point,
    the player must decide whether to stay with their initial choice or switch to the 
    remaining unopened door.
}
\examples{
   this.game
   my.initial.pick
   my.final.pick <- change_door( stay=T, opened.door=opened.door, a.pick=my.initial.pick )
   determine_winner( final.pick=my.final.pick, game=this.game )
   my.final.pick <- change_door( stay=F, opened.door=opened.door, a.pick=my.initial.pick )
   determine_winner( final.pick=my.final.pick, game=this.game )
}
