% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg copy 2.R
\name{change_door}
\alias{change_door}
\title{The host offers the player the option to swap their door}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{This}{function takes in a \code{stay} variable, the \code{opened.door} variable, and
the initial \code{a.pick} variable.}
}
\value{
This function returns \code{final.pick}, an integer between 1 and 3.
}
\description{
change_door() represents the host offering the contestant the option to
swap their door selection to the only unopened remaining door.
}
\details{
This function takes a value for a \code{stay} variable, which represents the
player either sticking with their initial choice or switching doors. It
outputs the final door selection.
}
\examples{
  goat.door <- open_goat_door(this.game, this.pick)
  change_door(stay=F, goat.door, this.pick)

}
