% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem (1).R
\name{change_door}
\alias{change_door}
\title{Change Doors}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{…}{if the contestant stays then there is no changes that need to occur
and all choices would be True. If the contestant decides to switch choices,
the final pick would have to eliminate the original choice and the opened
goat door by using the (!).}
}
\value{
The final pick based on the contestant’s choice to stay or switch.
}
\description{
The contestant has the opportunity to keep the door they originally picked or
they can switch to the other door.
}
\details{
The function consists of whether they chose to stay the opened goat door and
the original choice. If they decide to stay than the stay option would be
true which means that the pick and the opened goat door would remain the same.
The final pick would be the originally selected door. If the contestant
choses to switch doors than the opened goat door would remain the same, but
the contestant’s choice would change to the other remaining door; therefore,
the opened goat door and the original choice would have to be ignored as a
choice.
}
\examples{
final.pick <- doors[doors != opened.door & doors != a.pick]

}
