% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{Choose to stay with the initial door or change to the final door}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{This}{function requires the arguments of "stay" (either TRUE or
FALSE), "opened.door", and "a.pick". The two later arguments
come from previous functions, the "stay" argument is input
by the user.}
}
\value{
the function returns a single numeric, either 1, 2, or 3.
}
\description{
\code{change_door()} shows the result of if the contestant wants to
keep the door they initially selected, or switch to the other
unopened door.
}
\details{
This function represents the third part of the game, once the
host has revealed a door with a goat behind it. The participant
can either keep the door they initially selected, or change to
the other unopened door.
}
\examples{
  change_door(stay = TRUE, 3, 1)
}
