% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{change_door}
\alias{change_door}
\title{Stay or switch to the other unopened door}
\usage{
change_door(stay = TRUE, opened.door, a.pick)
}
\arguments{
\item{stay}{Logical. TRUE = stay; FALSE = switch.}

\item{opened.door}{Integer (1:3) opened by the host.}

\item{a.pick}{Integer (1:3) initial pick.}
}
\value{
Integer (1:3) final chosen door.
}
\description{
Computes the final pick given the strategy.
}
\examples{
  change_door(stay=TRUE,  opened.door=2, a.pick=1)
  change_door(stay=FALSE, opened.door=2, a.pick=1)
}
