% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{Give a chance to change the door}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{Stay}{a Boolean that can take wither true if the user will change the door or false when
the user will keep the original door

opened.door a vector that will indicate which door that was previously opened

a pick a vector that will indicate the door that was first selected}
}
\value{
the function returns a length 1 vector indicating the final pick of the participant
}
\description{
\code{change_door} Asks the user to change or keep the previously selected door
}
\details{
The game will give a last chance to the user to either keep his door or change it and with
this game we will try to get the results of the two scenarios
}
\examples{
function( stay=T, opened.door, a.pick )

}
