% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall_script_rj.R
\name{change_door}
\alias{change_door}
\title{:
Choose to change doors or to stay with the original choice.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{:}{The door number returned is a numeric value.}
}
\value{
:
Depending on whethey stay was set to TRUE or FALSE, the final door number will be returned as final.pick.
}
\description{
:
The change_door function chooses the final door selection based on whether the contestant stays with the original choice or changes.
}
\details{
:
The function is based on whether stay is set to TRUE or FALSE, as well as the door opened by the host "opened.door" and the initial pick "a.pick." Is stay is set to TRUE, the initial pick is used as the final selection. If stay is set to FALSE, the contestant switches doors and the door that is not the inital pick and is not the door revelead by the host is set as the final pick.
}
\examples{
change_door( stay=T, opened.door=1, a.pick )
}
