% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty_hall_scripts.R
\name{change_door}
\alias{change_door}
\title{Contestant stays or switches.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay}{A boolean object representing TRUE if the contestant stays with their initial door pick or
FALSE if the contestant switches to the unopened door.}

\item{opened.door}{The position of the door opened by the host with a goat behind it, a number between 1 and 3.}

\item{a.pick}{The position of the contestant's initial door select, a number between 1 and 3.}
}
\value{
This function returns the position of the contestant's final choice after either staying
or switching: 1, 2, or 3.
}
\description{
\code{change_door()} simulates the contestant choosing to either stay with their initial door
selection or switching.
}
\details{
In the Monty Hall Problem, once the contestant has opened a door and the host has opened
one of the door with a goat behind it, the contestant must choose to either stay with their
original door selection or switch to the unopened door.
}
\examples{
game <- create_game()
a.pick <- select_door()
opened.door <- open_goat_door(game, a.pick)
change_door( stay = T, opened.door, a.pick)

}
