% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{Switch or Stay}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{This}{functions takes on three inputs: whether the contestant stays with their initial pick (stay=T) or switches (stay=F), which door was
opened in the previous step, and which door was the contestant's initial pick.}
}
\value{
The function will return either 1, 2, or 3 corresponding to the contestant's final pick
}
\description{
This function will either let the contestant either stay with their initial pick, or switched to the other unopened door.
}
\details{
This function takes on three inputs, which will allow the contestant to stick with their initial door, or change their decision to the
as of yet unopened door. The output will be the final decision of the contestant as to which door has the car behind it.
}
\examples{
change_door(stay=T, opened.door, a.pick )
}
