% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{Determine the contestant's final door selection.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay}{A logical value indicating whether the contestant stays
with the original door selection. TRUE means stay and FALSE means
switch.}

\item{opened.door}{A numeric value representing the door opened by
the host.}

\item{a.pick}{A numeric value representing the contestant's original
door selection.}
}
\value{
A numeric value representing the contestant's final door
selection. The value will be 1, 2, or 3.
}
\description{
\code{change_door()} determines the contestant's final door selection
based on whether the contestant stays with the original choice
or switches to the other unopened door.
}
\details{
If \code{stay} is TRUE, the function returns the contestant's original
door selection. If \code{stay} is FALSE, the function returns the
remaining unopened door.
}
\examples{
  game <- create_game()
  a.pick <- select_door()
  opened.door <- open_goat_door(game, a.pick)
  change_door(TRUE, opened.door, a.pick)
  change_door(FALSE, opened.door, a.pick)

}
