% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{Change door selection in the Monty Hall game}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay}{A logical value. If TRUE, the contestant stays with their initial pick.
If FALSE, they switch to the other unopened door.}

\item{opened.door}{An integer between 1 and 3 representing the door opened by the host,
as returned by the `open_goat_door()` function.}

\item{a.pick}{An integer between 1 and 3 representing the contestant's initial choice,
as returned by the `select_door()` function.}
}
\value{
An integer between 1 and 3, representing the final door selection.
}
\description{
`change_door()` determines the final door selection based on the contestant's decision to stay or switch.
}
\details{
This function simulates the contestant's final decision in the Monty Hall Problem.
After the host opens a door revealing a goat, the contestant can either stick with
their original choice (stay) or switch to the other unopened door. This function
determines which door is the final selection based on this decision.
}
\examples{
# Simulate staying with the initial choice
change_door(stay = TRUE, opened.door = 2, a.pick = 1)

# Simulate switching to the other unopened door
change_door(stay = FALSE, opened.door = 2, a.pick = 1)

}
