% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{Change the contestant's door choice in the Monty Hall Problem.}
\usage{
change_door(stay = TRUE, opened.door, a.pick)
}
\arguments{
\item{switch}{A logical value indicating whether the contestant decides to
switch doors (TRUE) or stay with their original choice (FALSE).}

\item{opened}{An integer (1, 2, or 3) indicating which door the host opened
to reveal a goat.}

\item{original_choice}{An integer (1, 2, or 3) representing the contestant's
original door choice.}
}
\value{
The function returns an integer (1, 2, or 3) representing the contestant's
final door choice.
}
\description{
\code{change_door()} determines the contestant's final door choice based on
whether they decide to switch, which door was opened by the host, and
their original choice.
}
\details{
In the Monty Hall Problem, after the contestant's initial choice and
the host reveals a goat behind one of the unchosen doors, the contestant
is given the option to switch their choice to the remaining unopened door.
This function simulates that decision process, taking into account whether
the contestant chooses to switch, which door the host opened, and the
contestant's original choice. It then returns the contestant's final door
selection.
}
\examples{
  change_door(switch = TRUE, opened = 2, original_choice = 1)
  change_door(switch = FALSE, opened = 3, original_choice = 2)

}
