% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{Contestant has the option to stay or switch doors after a goat is revealed}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay}{A logical value indicating whether the player will stay with their initial choice.}

\item{opened.door}{The revealed door not contains the prize.}

\item{a.pick}{The door initially chosen by the player.}
}
\value{
The numeric vector index for the selected door.  If the player stays, the initial selection
is returned.  If the player switches, the remaining unopened door is returned.
}
\description{
The contestant is given an option to change doors after a Zonk (goat) has been revealed.
}
\examples{
change_door(1, 3)
 Output: 2

change_door(2, 2)
 Output: 1

}
