% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{change_door}
\alias{change_door}
\title{Change Door Function}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{This}{function contains two basic arguments.  The first is that if the
player decides to stay, then their initial choice \code{a.pick} becomes
their \code{final.pick}. If the player decides to switch, then the host
cannot select \code{opened.door} or \code{a.pick}.  To perform this, the function
uses the argument that final.pick is not equal to \code{opened.door} or
\code{a.pick}.}
}
\value{
This function returns a length one character vector that is a number
between 1 and 3.  That number is assigned to variable \code{final.pick}.
}
\description{
The \code{change_door} function is where the player makes the decision to
stay with their initial choice, \code{a.pick}, or change to the other
unopened door.  They do not know what is behind either of the remaining
closed doors.
}
\details{
if the player decides to stay, then the host will move forward with
opening the door associated with \code{a.pick}. If the player decides to
switch doors, the host will open the third door.  The third door is
not \code{a.pick} or \code{opened.door}.  The final choice of the player is
then assigned to variable \code{final.pick}.
}
\examples{
  change_door()

}
