% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{change_door}
\alias{change_door}
\title{CHANGE DOORS.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{if}{arguments are used by the function.}
}
\value{
The function returns the final pick of the door whether  to Stay
or Switch.
}
\description{
\code{change_door ()} CHANGE DOORS.
}
\details{
The contestant is given the option to change from their
initial selection to the other door that is still closed.
The function will represent the game-playing strategy as
the argument stay=TRUE or stay=FALSE.
}
\examples{
change_door ()

}
