% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{Changing doors - Function}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{The}{function "if" will help to run different commands depending on the
contestant choice. If the contestant selects to "stay" then his previous
pick will be remain the same (and will be a final pick). If the contestant
chooses to change the door then the function "if" will run the code to
choose a different door from the remaining ones, and this will be his final pick.}
}
\value{
The function will be returning the final pick of the contestant, (either if it is still their first choice o a new choice).
}
\description{
This function will allow the contestant to change their first pick if
they wanted to.
}
\details{
This function will be using the variables "stay", "opened door", and
the "a.pick", this variable was previously used in the previous function.
The contestant will decide if they want to stay or get another pick.
}
\examples{
change_door( )
}
