% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{Contestant's Final Decision}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{Stay=T}{is used in this function as part of an if/else statement. If it is
true the contestant stayed with their original decision compared to if they
chose to switch doors}
}
\value{
final.pick
}
\description{
Contestant decides whether they want to stick to their first pick(a.pick)
or switch to a different door
}
\details{
This function states that if the contest chooses to stay with their first
pick (a.pick), then that is their final pick. If the contestant chooses
to switch doors, then they must choose the remaining door as the other two
were previously chosen (a.pick) and already opened by the host (opened.door)
}
\examples{
change_door(stay=T, opened.door, a.pick)
change_door(stay=F, opened.door, a.pick)
}
