% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_game.R
\name{change_door}
\alias{change_door}
\title{Change or keep the selected door}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay}{Logical value indicating whether the contestant
stays with their original selection. `TRUE` means stay and
`FALSE` means switch.}

\item{opened.door}{The number of the door opened by the host.}

\item{a.pick}{The contestant's original door selection.}
}
\value{
The number of the contestant's final door selection.
}
\description{
`change_door()` determines the contestant's final door
selection based on whether they stay with their original
choice or switch to the remaining unopened door.
}
\examples{
game <- create_game()
pick <- select_door()
opened <- open_goat_door(game, pick)
change_door(TRUE, opened, pick)

}
