% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{Change Doors}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{The}{change_door function calls upon the arguments stay, opened.door and a.pick.
The door function takes in the logical vectors 1-3.
If statements evaluate the stay argument and assigns
the output to function final.pick.}
}
\value{
The return would be a logical vector, an integer, between 1 and 3.
}
\description{
The game continue with adding more code to allow the contestant the option of selecting
another door that is still closed with a change door function.
}
\details{
The contestant is given the option to change from their initial selection to the other
door that is still closed. The function will represent the game-playing strategy as the
argument stay=TRUE or stay=FALSE.
}
\examples{
  1
  2
  3

}
