% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg-finished.R
\name{change_door}
\alias{change_door}
\title{Create a change door function.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a number between 1 and 3 that corresponds with a door
}
\description{
\code{change_door()} generates a new function that allows the contestant to either
keep their initial selected door, or change to the remaining unopened door.
}
\details{
After one losing door is revealed by the host, the contestant is then given
the option to either keep their initial door, or change their selection to the
final remaining door.
}
\examples{
 change_door()


}
