% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{Contestant decides to switch or stay with his/her original choice
 of the door.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{The}{function takes three arguments.
1. option- We are considering that constants picks to stay
with original choice as defualt.
2. opened.door - the goat door which is opened by the host
(numberic vector of length one containing number between 1-3).
3. a.pick- Contestent's original pick of the door
(numberic vector of length one containing number between 1-3).}
}
\value{
Function returns a numeric vector of length one conating a number
 between 1-3.
}
\description{
The contestant has a choice to stay to his/her original choice or
 switch doors to one of the two availble doors at this point.
}
\details{
The contestant decides to stay with his/her original pick or s/he has
 an option to switch.If the contestant decide to switch s/he can pick
 between one of the two available doors.The function return either the
 original pick or one of the two availble doors depending upon
 contestant's choice.
}
\examples{

 change_door(stay=T, opened.door, a.pick)

}
