% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{The contestant chooses to retain his initial selection or switch to the last available door}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{This}{function uses the arguments stay=, opened.door, and a.pick which represent the choice to STAY = T (retain initial
door selection), STAY = F (switch doors), opened.door (which door the host opened), a.pick (the door the contestant
initially chose).}
}
\value{
The function returns the door the contestant selected - a single numeric representing which door was chosen.
}
\description{
After the host reveals a goat door, the contestant is able to choose between switching his selection to the only
other door remaining, or to retain his initial selection.
}
\details{
The contestant must decide at this point to stay or switch doors. There are only two doors left at this point: one
with a goat, and one with a car. The contestant chooses to switch to the last remaining door or keep his initial
selection. The available doors are determined by previous functions and used as arguments. The initially chosen door
by the contestant, the opened door by the host revealing a goat, provide the final 2 doors available to pick from.
}
\examples{
game <- create_game()
a.pick <- select_door()
opened.door <- open_goat_door(game.a.pick)
change_door(stay = TRUE, opened.door, a.pick)
}
