% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{Player switch or stay.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{`stay=T`}{Logical. Controls whether or not the player switches.}

\item{`opened.door`}{Integer. The goat door opened by Monty.}

\item{`a.pick`}{Integer. The player's initial pick.}
}
\value{
Integer. The number of the chosen door.
}
\description{
This function determines whether or not the player switches doors.
}
\details{
After Monty opens a goat door, the player may either double down
on their original pick or switch to the last remaining door.
}
\examples{
    change_door()

}
