% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monty Hall.package.R
\name{change_door}
\alias{change_door}
\title{Switch or stay with the chosen door}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay}{Logical, whether the player stays with the original pick(\code{True})}

\item{opened.door}{An integer (1-3), the door opened by the host.}

\item{a.pick}{An integer (1-3), the contestant's initial pick.}
}
\value{
An integer (1-3) indicating the contestant's final door choice.
}
\description{
`change_door() determines the contestant's final choice based on their strategy
}
\details{
If the player stays, they keep their first pick.
If the player switches, they pick the unopened door that is not their first pick and not the opened goat door.
}
\examples{
g <- create_game()
p <- select_door()
o <- open_goat_door(g, p)
change_door(stay =FALSE, opened.door =o, a,pick = p)

}
