% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{Change Doors}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay}{TRUE if the player stays and FALSE if the player switches.}

\item{opened.door}{The goat door that was opened.}

\item{a.pick}{The player's original door choice.}
}
\value{
The number of the player's final door choice.
}
\description{
Determines the player's final door choice.
}
\details{
If the player stays, the original door is kept. If the player
switches, the other unopened door is selected.
}
\examples{
change_door(stay=TRUE, opened.door=2, a.pick=1)
change_door(stay=FALSE, opened.door=2, a.pick=1)

}
