% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{#'   The contestant chooses to stay with initial pick of door or
change to the other door not revealed.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{...}{the stay argument takes a logical vector for staying or
changing doors. The opened.door argument is from the
open_goat_door function and the a.pick is from the select_door
function.}
}
\value{
The function returns a length 1 numerical vector
indicating the position of the contestant choice.
}
\description{
#'   \code{change_door()} works from a stay argument that indicates
whether contestant will stick with initial pick or go with
the other option.
}
\details{
#'   At this point in the Monty Hall game, the guest has been shown
one door with a goat behind it. They are either already paired
with a winning door or the other door has a car behind it. The guest
now has the option to stay with their initial pick or change doors.
This code takes in which door has been opened, and which doors
have a goat or a car which is not known to the contestant. The
change leads into determining a winner or loser in the final.
}
\examples{
  change_door(stay=T,1,1)
  
}
