% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{Decide whether to switch doors or stay with the first pick}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay}{\itemize{
\item logical, T or F
opened.door - the output of open_goat_door
a.pick - output of select_door()
}}
}
\value{
return a number between 1-3
}
\description{
This step is the opportunity to switch doors or stay with the first pick
}
\details{
No details
}
\examples{
 change_door(stay=T,  open_goat_door( create_game(), select_door()), select_door())
}
