% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem edited.R
\name{change_door}
\alias{change_door}
\title{Contestant Chooses Whether to Change their Selection}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{This}{function has three arguments, \verb{stay=T, opened.door, a.pick}.
\code{stay=T} is a logical value. \code{stay=T} is the default and will keep the contestant's choice unchanged.
\code{stay=F} will randomly choose one of the two unopened doors as the contestant's new choice.
\code{opened.door} is a numeric value, usually stored in an object, that is the return of 'opened_goat_door\code{. }a.pick\verb{is the return of}select_door`, a numeric value stored as an object}
}
\value{
The return of this function is a numeric value between 1 and 3.
}
\description{
During this step, the contestant is given the opportunity to
either keep their original door selection, or switch to the
other unopened door. They do not know what is behind the two
doors, just that one is a goat and one is a car.
}
\details{
This function creates an \code{if} scenario regarding whether \code{stay=T} and the
contestant's final choice is the same as the initial choice, or if \code{stay=F}
and the contestant's final choice then changes to the other unopened door.
}
\examples{
  change_door(stay=T, opened.goat.door, initial.pick)
  change_door(stay=F, montys.goat.door, first.pick)

}
