% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{Change door or Stay}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{`stay`}{\itemize{
\item A Boolean representing whether the contestant switches their initial pick. The default value is "T"
\code{opened.door} - A number representing the hosts door.
\code{a.pick} - A number representing the contestants initial pick.
}}
}
\value{
A number representing the contestants final pick.
}
\description{
\code{change_door} will give you the choice to either change your pick or stay with you initial pick.
}
\details{
In the classic game show, Lets Make a Deal, after the host opens a non winning door, they then give you the option to change or stay with your original door.
}
\examples{
change_door(stay=T, host.door, my.initial.pick)
}
