% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{Test}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{...}{no arguments are Test.}
}
\value{
The function returns a Test
}
\description{
Test
}
\details{
Test
  Test
  Test
}
\examples{
  Test()

}
