% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{Change Door?}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{Stay=T}{means the contestant has stayed with their original pick, when stay=F, they switch their pick to the remaining door}
}
\value{
return a number 1, 2, or 3.
}
\description{
Here the contestant has the opportunity to change their pick to the only remaining closed door, or stick with their current and first pick.
}
\details{
If the contestant stays, that door is revealed, if they switch, the door they switch to will be revealed in the next step
}
\examples{
: 1
}
