% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{change_door}
\alias{change_door}
\title{Select the whether the contestant stays or switches doors}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{opened.door}{A number between 1 and 3}

\item{a.pick}{A number between 1 and 3}

\item{Stay=T}{A logical vector}
}
\value{
The function returns a number between 1 and 3
}
\description{
\code{change_door()} selects the final door for the contestant
}
\details{
The function produces the final door selection based on the initial
door choice and whether the contestant stays with that choice or
switches to the other remaining door.
}
\examples{
  change_door()

}
