% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\name{change_door}
\alias{change_door}
\title{two conditions for the player}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{The}{first door picked by played, the opened door by host, and a boolean option}
}
\value{
The final picked door by player
}
\description{
The player will get to decide if they will switch door or stay
}
\details{
Depending on the player choice, the function will return either
their first door selction or their new pick
}
\examples{
  change_door(stay=F, opened.door, a.pick)
  
}
