% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.r
\name{change_door}
\alias{change_door}
\title{Change Door}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{Stay}{is a Boolean value for whether the contestant decides
to stay or not. Open.door is a numeric value between 1 and 3,
which represents which goat door was opened. a.pick is a single
numeric value between 1 and 3 which represents the initial
contestant pick.}
}
\value{
The function returns final.pick which is nua numeric value
 between 1 and 3 representing the final choice of the
 contestant.
}
\description{
`change_door()` allows the contestant to change their pick
  from their initial pick to a different unopened door if they
  choose.
}
\details{
This finalizes the door choice of the contestant.
}
\examples{
  opened.door <- open_goat_door( this.game, my.initial.pick )
  change_door(stay=T,opened.door=opened.door,a.pick=my.initial.pick )
  change_door(stay=F,opened.door=opened.door,a.pick=my.initial.pick)

}
