% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{Player decides whether to change chosen door.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{This}{function uses the same doors vector of
previous functions, which is a numeric vector that
is three characters long. There are also if, then
statements to indicate that if the player stays
with their initial choice, then the game should
refer back to the 'select_door()' result. If the
the player chooses to not stay with their initial
choice, then the game will select the door that was not
initially picked and also not revealed by the
'open_goat_door()' function.}
}
\value{
the function returns one number between 1 and 3.
}
\description{
'change_door()' allows the player to either
stay with their original choice or choose
the other door that was not revealed.
}
\details{
The player needs to make a choice whether they
keep their initial chosen door or if they will
switch to the door that the game did not open.
Whichever door they select in this part of the
game is the door that will be revealed to either
be a car that they won, or a goat indicating loss.
}
\examples{
change_door()
}
