% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{Contestant chooses to stay or change doors.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{`stay`}{- This tells the function if the contestant stayed
with their initial pick or not.
`opened.door` - This tells the function which door was opened.
`a.pick` - This tells the function what the contestant's
inital door pick was.}
}
\value{
The functions returns a number between 1 and 3 that represents
  the contestants final door pick.
}
\description{
`change_door()` generates the contestant's final pick.
  They can iether stay with their first pick, or change
  with the last unopened door.
}
\details{
This function uses if statements that determine if the
  contestant stayed or changed doors. If the contested
  stayed, then it turns the contestants inital door pick into
  a final pick that is returned. If the contestant did not
  stay, then it passes the final unopened door into the
  final pick and returns that door instead.
}
\examples{
  change_door( stay=T, opened.door, first.pick )
  change_door( stay=F, 3, 1)

}
