% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-script.R
\name{change_door}
\alias{change_door}
\title{Change Door}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{None}{}
}
\value{
Contest will automatically either switch or stay with their initial pick.
}
\description{
Contestant has an option to change their door choice.
}
\details{
Since the host exposed which of the 2 doors with a goat by opening
one of the doors, the contestant now has an option to change their
initial pick.
}
\examples{
change_door()

}
