% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{Step 4: Change doors}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{...}{two positional arguments and one keyword argument}
}
\value{
an integer representing the final choice of door selected by
player
}
\description{
Gives the option for the player to change choice
of door.this function receives two positional arguments
with a default value of True (stay=True)
}
\details{
opened_door is the return value of the open_goat_door call
a.pick is the return value of the select_door call
}
