% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{change_door}
\alias{change_door}
\title{The contestant decided to change a door when given chance to play.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{There}{are arguments used by the function.}
}
\value{
The function returns a numeric vector
}
\description{
\code{change_door()} The contestant was given the change to either stay or
switch.
}
\details{
As planned in the game the contestant has the opportunity to decided
whether to continue with the game or not. In this case the decided to
switch. the functions the final picks of the contestant.
}
\examples{
  change_door()

}
