% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{Determine final door choice}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{stay}{Logical value. TRUE if player stays with the original pick; FALSE to switch}

\item{opened.door}{An integer of the door opened by the host}

\item{a.pick}{An integer of the initial door picked by the player}
}
\value{
An integer of the final door picked by the player
}
\description{
Decides whether the player stays with the initial choice or switches
}
\details{
If stays is TRUE, returns the initial pick; otherwise returns the other unopened door
}
\examples{
change_door(TRUE, 3, 1)
}
