% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{Change Doors (Stay or Switch)}
\usage{
change_door(stay = TRUE, opened.door, a.pick)
}
\arguments{
\item{stay}{A logical indicating whether to stay (\code{TRUE})
or switch (\code{FALSE}).}

\item{opened.door}{An integer 1, 2, or 3 that the host opened
(revealing a goat).}

\item{a.pick}{An integer for the door the contestant originally picked.}
}
\value{
An integer \code{1}, \code{2}, or \code{3} representing the final door choice.
}
\description{
\code{change_door()} gives the contestant's final choice after
deciding whether to stay with the initial pick or switch to
the unopened door.
}
\details{
If \code{stay=TRUE}, the function returns the original door (\code{a.pick}).
If \code{stay=FALSE}, it returns the other unopened door (neither \code{a.pick}
nor \code{opened.door}).
}
\examples{
  g <- create_game()
  pick <- select_door()
  opened <- open_goat_door(g, pick)
  change_door(TRUE, opened, pick)
  change_door(FALSE, opened, pick)

}
