% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{create_game}
\alias{create_game}
\title{Create a new Monty Hall Problem game}
\usage{
create_game()
}
\value{
A character vector of length 3 with entries \code{"goat"} or \code{"car"}.
}
\description{
\code{create_game()} generates a new game with two goats and one car.
}
\details{
The game mirrors "Let's Make a Deal": three doors, one with a car and two with goats.
}
\examples{
create_game()
}
