% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{create_game}
\alias{create_game}
\title{Step 1: Monty Hall Game Initialization.}
\usage{
create_game()
}
\arguments{
\item{...}{\code{create_game()} uses no arguments.}
}
\value{
\code{create_game()} returns a 3x1 vector with the elements
"goat", "goat", and "car" in random order
}
\description{
\code{create_game()} produces a 3x1 vector with the elements
"goat", "goat", and "car" in random order
}
\details{
The game setup replicates the game on the TV show "Let's
Make a Deal". There are three closed doors. Behind one of the doors
there is a car, and behind the other two there are goats.
The contestant wins the car if they select the corresponding door.
}
\examples{
  game <-create_game()
  game
  [1] "car"  "goat" "goat"

}
