% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{create_game}
\alias{create_game}
\title{\preformatted{Create a New Game.
}}
\usage{
create_game()
}
\arguments{
\item{x}{Numeric vector}

\item{values}{must be whole integers greater than zero}
}
\value{
\preformatted{Returns a length 3 charater vector showing the positions of the goats and the car 
}
}
\description{
\preformatted{`create_game()` randomly assigns a specified number of both goats and cars to the
}

number
}
\details{
\preformatted{This funtion creates a new game for the monty hall problem. There are three doors: 
two goats and one car. this simulation gives you the opportunity to test the 
probabilities!
}
}
\examples{
	create_game()


}
