% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monty Hall Script.R
\name{create_game}
\alias{create_game}
\title{Creates vector of user inputs}
\usage{
create_game()
}
\arguments{
\item{...}{This function does not take any arguments}
}
\value{
This function returns a character vector, a.game.
}
\description{
This function creates a character vector of length three that consists of two goats and one car.
The goats and car are randomly assembled to create unique combimations.
}
\details{
This function is the basis of game and is referenced many times throughout the game.
}
\examples{
create_game()

}
