% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{create_game}
\alias{create_game}
\title{Create a 3-door Monty Hall game}
\usage{
create_game()
}
\value{
A character vector of length 3 containing two \code{"goat"} values and one \code{"car"}.
}
\description{
Generates a random arrangement of two goats and one car behind three doors.
}
\examples{
create_game()

}
