% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{create_game}
\alias{create_game}
\title{Create a new Monty Hall Problem game.}
\usage{
create_game()
}
\value{
A length 3 character vector indicating the positions of goats and the car.
}
\description{
Generates a new game that consists of two doors with goats behind them,
and one with a car.
}
\examples{
create_game()

}
