% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_game.R
\name{create_game}
\alias{create_game}
\title{Monty Hall Game Generator}
\usage{
create_game(game.size = 3, num.cars = 1)
}
\arguments{
\item{game.size}{An integer.}

\item{num.cars}{An integer greater than zero and not larger than \code{game.size}-2.}
}
\value{
The output is a character vector of length \code{game.size}. Each element is
either "car" or "goat".
}
\description{
\code{create_game(game.size, num.cars)} creates a randomly generated
Monty Hall game, generalized for any number of doors in the game.
}
\details{
This function generates a vector of length \code{game.size} the elements of which represent
the doors where either a goat or a car is hidden behind each. \code{num.cars} is the number of
cars in the game.
}
\examples{
create_game(3,1)
create_game(10,3)

\dontrun{
create_game(5,4)
}

}
