% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{create_game}
\alias{create_game}
\title{Create a new Monty Hall Problem game.}
\usage{
create_game(nCars, nGoats)
}
\arguments{
\item{nCars}{number of cars}

\item{nGoats}{number of goats}
}
\value{
The function returns a length n character vector
  indicating the positions of goats and the car.
}
\description{
`create_game()` generates a new game that consists of n doors
  with goats and cars behind them.
}
\details{
The game setup replicates the game on the TV show "Let's
  Make a Deal" where there are n doors for a contestant
  to choose from. Then given constraints in questions are applied
  and results are noted. There was a famous
  debate about whether it was optimal to stay or switch when
  given the option to switch, so this simulation was created
  to test both strategies.
}
\examples{
  create_game(1,2)

}
