% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{create_game}
\alias{create_game}
\title{Create a New Monty Hall Game}
\usage{
create_game()
}
\value{
A character vector of length three containing two
goats and one car.
}
\description{
Creates a new Monty Hall game by randomly assigning
two goats and one car behind three doors.
}
\details{
The position of the car is randomized each time the
function is called.
}
\examples{
create_game()

}
