% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{create_game}
\alias{create_game}
\title{Create a Monty Hall Game}
\usage{
create_game()
}
\value{
A character vector representing the game setup (e.g., c("car", "goat", "goat")).
}
\description{
Randomly generates a Monty Hall game setup with one car and two goats.
}
\examples{
create_game()
}
