% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{montyhall}
\alias{montyhall}
\title{Create a new Monty Hall Problem game.}
\usage{}
\arguments{}
\value{}
\description{
\code{} 
}
\details{
The game setup replicates the game on the TV show "Let's
Make a Deal" where there are three doors for a contestant
to choose from, one of which has a car behind it and two
have goats. The contestant selects a door, then the host
opens a door to reveal a goat, and then the contestant is
given an opportunity to stay with their original selection
or switch to the other unopened door. There was a famous
debate about whether it was optimal to stay or switch when
given the option to switch, so this simulation was created
to test both strategies.

The game setup replicates the game on the TV show "Let's
Make a Deal" where there are three doors for a contestant
to choose from, one of which has a car behind it and two
have goats. The contestant selects a door, then the host
opens a door to reveal a goat, and then the contestant is
given an opportunity to stay with their original selection
or switch to the other unopened door. There was a famous
debate about whether it was optimal to stay or switch when
given the option to switch, so this simulation was created
to test both strategies.
}
\examples{
  play_game()

}
\name{create_game}
\alias{create_game}
\title{}
\usage{
create_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a length 3 character vector
indicating the positions of goats and the car.
}
\description{
\code{create_game()} generates a new game that consists of two doors
with goats behind them, and one with a car.
}
\details{

}
\examples{
  create_game()

}

}
\name{select_door}
\alias{select_door}
\title{}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a number between 1 and 3.
}
\description{
\code{select_door()} randomly one of the doors for the contestant.
}
\details{

}
\examples{
  select_door()

}
}
\name{open_goat_door}
\alias{open_goat_door}
\title{}
\usage{
open_goat_door()
}
\arguments{
\item{This function takes two arguments:}
\item{game : the game with two doors with goats behind them and one with a car.}\item{a.pick: the door selected by the contestant.}
}
\value{
returns a number between 1 and 3 corresponding to a door with a goat.
}
\description{
\code{open_goat_door()} selects one of the remaining two doors with a goat behind it.
}
\details{

}
\examples{
  open_goat_door()

}
}
\name{change_door}
\alias{change_door}
\title{}
\usage{
change_door()
}
\arguments{
\item{This function takes three arguments:}
\item{stay : Logical where `TRUE` returns contestants original choice and `FALSE` switches contestants choice to last remaining door.}\item{opened.door: Number between 1-3 corresponding to goat door opened by host.}
\item{a.pick:  Number between 1-3 corresponding to the original door selected by the contestant.}
}
\value{
returns number between 1 and 3.
}
\description{
\code{open_goat_door()} determines the last remaining door that was not selected by the contestant or opened by the host to reveal a goat, and allows the contestant's choice to be changed if desired. 
}
\details{

}
\examples{
  change_door()

}