% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{create_game}
\alias{create_game}
\title{Create a new Monty Hall Problem game}
\usage{
create_game()
}
\value{
A character vector of length three indicating the placement of
two goats and one car.
}
\description{
\code{create_game()} generates a new game that consists of two doors
with goats behind them and one door with a car.
}
\details{
The game setup replicates the TV show "Let's Make a Deal". One car
and two goats are randomly assigned to three doors. The contestant
initially selects a door, after which the host reveals a goat behind
one of the remaining doors.
}
\examples{
create_game()

}
