% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{create_game}
\alias{create_game}
\title{Create a Monty Hall game}
\usage{
create_game()
}
\value{
A character vector of length 3 with values "goat"/"car".
}
\description{
Randomly assigns 2 goats and 1 car behind three doors.
}
\examples{
set.seed(1); create_game()
}
