% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen comments.R
\name{create_game}
\alias{create_game}
\title{Create a new Monty Hall Problem game.}
\usage{
create_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a length 3 character vector
indicating the positions of goats and the car.
}
\description{
\code{create_game()} generates a new game that consists of two doors
with goats behind them, and one with a car.
}
\details{
There are three doors for a contestant
to choose from, one of which has a car behind it and two
have goats. The contestant selects a door, then the host
opens a door to reveal a goat. The contestant is
given an opportunity to stay with their original selection
or switch to the other unopened door. This simulation will
test which strategy is better to stay or switch.
}
\examples{
  create_game()

}
