% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_game.R
\name{create_game}
\alias{create_game}
\title{Create the orginization of goats and cars.}
\usage{
create_game(goats = 2, cars = 1)
}
\arguments{
\item{goats}{Numer of goats that will be hidden behind doors.}

\item{cars}{Number of cars that will be hidden behind doors.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
create_game() will create a sample game of two goats and one car by default,
but also will allow the user to define the number of goats and cars in the simulation.
}
\description{
Create the orginization of goats and cars.
}
\examples{
## Randomly sorts 2 goats and 1 car
create_game()

## Randomly sorts 4 goats and 3 cars.
create_game(4, 3)

}
