% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg copy 2.R
\name{create_game}
\alias{create_game}
\title{The Monty Hall Problem}
\usage{
create_game()
}
\arguments{
\item{...}{no arguments are used by the function create_game().}
}
\value{
The function returns a length 3 character vector
indicating the positions of goats and the car.
}
\description{
There are three doors, behind one there is a car. Behind the other two there are goats.
Can you find the car and win the game?
}
\details{
The game setup replicates the game on the TV show "Let's
Make a Deal" where there are three doors for a contestant
to choose from, one of which has a car behind it and two
have goats.
Once the contestant has selected a door, the host
opens one of the remaining two doors to reveal a goat. and then the contestant is
The contestant at this point must now make another choice:
Should they keep their selection the same, or choose the remaining door?
Fortunately, with statistics and an R simulation we can figure out which option
gives them the highest chance of finding the car!
}
\examples{
  create_game()

}
