% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{create_game}
\alias{create_game}
\title{Create a new Monty Hall Problem game.}
\usage{
create_game()
}
\value{
A length-3 character vector indicating the positions of goats and the car.
}
\description{
\code{create_game()} generates a new game that consists of two doors
with goats behind them, and one with a car.
}
\details{
The game setup replicates the game on the TV show "Let's
Make a Deal" where there are three doors for a contestant
to choose from, one of which has a car behind it and two
have goats. The contestant selects a door, then the host
opens a door to reveal a goat, and then the contestant is
given an opportunity to stay with their original selection
or switch to the other unopened door. This simulation tests both strategies.
}
\examples{
  create_game()
}
