% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{create_game}
\alias{create_game}
\title{Create a new Monty Hall Problem game.}
\usage{
create_game()
}
\value{
A character vector of length 3, where each element represents the contents
behind each door. The vector will contain two "goat" elements and one "car" element,
randomly ordered.
}
\description{
`create_game()` generates a new game setup for the Monty Hall Problem,
consisting of three doors with two goats and one car behind them.
}
\details{
This function simulates the initial setup of the Monty Hall Problem game.
It randomly assigns a car to one of the three doors and goats to the other two.
This setup is used as the basis for playing through the game scenario.

The Monty Hall Problem is based on a game show scenario where a contestant
must choose between three doors, behind one of which is a car (the prize)
and behind the other two are goats. After the contestant chooses a door,
the host (who knows what's behind each door) opens another door to reveal
a goat, and then offers the contestant the chance to switch their choice
to the remaining unopened door.
}
\examples{
# Generate a new game setup
create_game()

# Generate multiple game setups
replicate(5, create_game())

}
