% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_game.R
\name{create_game}
\alias{create_game}
\title{Create a new Monty Hall Problem game}
\usage{
create_game()
}
\value{
A length-three character vector indicating the
positions of the goats and the car.
}
\description{
`create_game()` generates a new game consisting of three doors,
with two goats and one car placed randomly behind the doors.
}
\details{
The game setup replicates the game on the TV show "Let's
Make a Deal," where there are three doors for a contestant
to choose from. One door has a car behind it and two have
goats. The contestant selects a door, then the host opens
a door to reveal a goat. The contestant is then given an
opportunity to stay with their original selection or switch
to the other unopened door.
}
\examples{
create_game()

}
