% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{create_game}
\alias{create_game}
\title{Create a new Monty Hall Problem game.}
\usage{
create_game(num.cars, num.goats)
}
\arguments{
\item{...}{the function takes 2 positional arguments,
number of cars and number of goats.}
}
\value{
The function returns a character vector
indicating the positions of goats and the cars.
}
\description{
Step 1: Create a new game
\code{create_game()} generates a new game that consists of doors
with goats behind them, and cars behind the other doors.
}
\details{
The game setup replicates the game on the TV show "Let's
Make a Deal" where there are three doors for a contestant
to choose from, one of which has a car behind it and two
have goats. The contestant selects a door, then the host
opens a door to reveal a goat, and then the contestant is
given an opportunity to stay with their original selection
or switch to the other unopened door. There was a famous
debate about whether it was optimal to stay or switch when
given the option to switch, so this simulation was created
to test both strategies.
}
\examples{
  create_game(3,7)
}
