% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{create_game}
\alias{create_game}
\title{Create a Monty Hall game}
\usage{
create_game()
}
\value{
A character vector of length 3, with one "car" and two "goat" (order randomized).
}
\description{
Initialize a new Monty Hall game with one car and two goats hidden behind three doors.
}
\examples{
g <- create_game()
length(g) == 3
}
