% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{create_game}
\alias{create_game}
\title{Create a new Monty Hall Problem game (3 doors).}
\usage{
create_game()
}
\arguments{
\item{...}{No arguments are used by this function.}
}
\value{
A length-3 character vector, e.g. \code{c("goat","goat","car")},
indicating the random positions of goats and the car.
}
\description{
\code{create_game()} generates a new Monty Hall game with 3 doors:
2 goats and 1 car placed in a random order.
}
\details{
This setup replicates the TV show "Let's Make a Deal,"
where there are three doors for a contestant to choose from:
behind one door is a car, and behind the other two are goats.
The contestant picks a door, the host opens a different door
to reveal a goat, and the contestant then has the option
to stay with the original door or switch to the other unopened door.
}
\examples{
  create_game()

}
