% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem (1).R
\name{determine_winner}
\alias{determine_winner}
\title{Determine if the contestant has won or lost the game.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{...}{Arguments: \code{final.pick} (its value is the number of the final selection door)
and \code{game} (the name of the game)}
}
\value{
Returns a character string of either "WIN" or "LOSE" depending if the
contestant's final selection was a car door.
}
\description{
\code{determine_winner()} Type: Character -> Returns 'WIN' if the final selection door
is a car door, or returns 'LOSE' if the final selection door is a goat door.
}
\details{
This function determines if the contestant has won or lost. If their final selection
was a Car door, they have won. If the final selection was a goat, they have lost.
They do not get to keep the goat.
}
\examples{
  determine_winner(final.pick = finalDoor, game = newGame)
}
