% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine whether the contestant won or not.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{...}{the arguments include the final door the contestant picked and the game from
'create_game()'}
}
\value{
The function returns a message on the game's result
}
\description{
\code{determine_winner()} tells the contestant if they won if they choose the door with the car
}
\details{
The contestant is notified if they win the game. If the
contestants pick matches the door with the car, they are
declared the winner. If they choose the door with the goat,
The contestant is informed that they lost the game.
}
\examples{
  determine_winner()

}
