% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{DETERMINE IF CONTESTANT HAS WON}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{Depending}{on the value choosen, this will depend the if argument that will be selected}
}
\value{
the final pick the contestant chose and if its linked  with car, they will WIN or links with the goat, they will LOSE
}
\description{
return function will allow us to see if the contestant won the car or chose the goat door.
}
\details{
If the contestant chooses the door that is linked with car, they will WIN. If the contestant get the door that 
links with the goat, they will LOSE
}
\examples{
return( "LOSE" ) or return( "WIN" )
}
