% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{determine_winner}
\alias{determine_winner}
\title{determine the winner}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{...}{no arguments are used by the function}
}
\value{
Return value WIN/LOSE.
}
\description{
determine whether the participants win the game.
}
\details{
build a function to determine whether the participants picked
a door with goat behind or car behind.
}
