% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine if contestant has won}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
If the final.pick is a car then the return is "WIN"
If the final.pick is a door then the return is "LOSE"
}
\description{
Contestant wins if they chose a door with a car.
}
\details{
After the contestant decides to stay or switch doors,
the final pick will either be a goat or car. If a car
is behind the final chosen door, then the contestant is
a winner.
}
\examples{
determine_winner <- function( final.pick, game )

}
