% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine the Game Winner}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{A numeric value indicating the contestant's final
door selection. The value should be 1, 2, or 3.}

\item{game}{A length 3 character vector representing the location
of two goats and one car behind the three doors.}
}
\value{
A character value of \code{"WIN"} if the final selected door
contains the car or \code{"LOSE"} if it contains a goat.
}
\description{
\code{determine_winner()} determines whether the contestant wins or
loses the Monty Hall game based on the final door selection.
}
\details{
The function checks the prize located behind the contestant's
final selected door. If the selected door contains the car, the
contestant wins. If the selected door contains a goat, the
contestant loses.
}
\examples{
  game <- c("goat", "car", "goat")
  determine_winner(final.pick = 2, game = game)
  determine_winner(final.pick = 1, game = game)

}
