% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{\preformatted{ Determine game outcome
}}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick:}{the integer returned from change_doors\preformatted{ game: the initial set up of the game some
 combination of "goat","goat","car"
}}
}
\value{
\preformatted{ [1] wIN
 [1] LOSE
}
}
\description{
\preformatted{ Determine game outcome
}
}
\details{
\preformatted{ the function compares the value to final.door from
 change_doors against the values "car" and "goat"

 if there is a match with the door and "car" then the
 function returns "WIN"

 if there is a match with the door and "goat" then
 the function returns "LOSE"
}
}
