% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{Game Result}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick:}{an integer between 1 and 3, the final contestant's choice
game: a 3x1 vector with two "goat" and one "car" elements, in random order}
}
\value{
... a string of characters
"WIN" if the contestant pick corresponds to "car"
"LOSE" if the contestant pick corresponds to "goat"
}
\description{
determine_winer( final.pick, game ) reveals whether the contestant final
pick is a car or a goat.
}
\details{
The element of the vector "game" corresponding to the final contestant pick
is revealed; if this element is "car", the contestant wins; otherwise, they
loose.
}
\examples{
  game
  [1] "car"  "goat" "goat"
  a.pick <- select_door()
  a.pick
  [1] 2
  opened.door <- open_goat_door( game, a.pick )
  opened.door
  [1] 3
  final.pick <- change_door(TRUE, opened.door, a.pick )
  final.pick
  [1] 2
  result <- determine_winner( final.pick, game )
  result
  [1] "LOSE"

}
