% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem-Gonzalez.R
\name{determine_winner}
\alias{determine_winner}
\title{The contestant's door is opened to reveal what is behind.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{The}{variables 'final.pick' and 'game' are inserted
into the 'determine_winner' function.}
}
\value{
The function returns a string that either states "WIN"
or "LOSE".
}
\description{
The game ends with a reveal. Behind the final chosen door
will either be a goat or a car. A car is a "WIN" and a
goat is a "LOSE".
}
\details{
All of the previous functions performed the game
logic. The final 'determine_winner' function
matches the 'final.pick' to the 'game' variable. The 'game'
parameter was randomly assigned at the beginning of the
game at the 'create_game' function. Because 'game' is
an index of 3 strings, the 'final.pick' number 1-3 will
be the chosen index character of 'game'. If the index
matches to a "car" string, then the contestant earns a
"WIN". If the index  matches to a "goat" string, then
the contestant earns a "LOSE".
}
\examples{
determine_winner( final.pick= 2, game= '"car", "goat", "goat"')
determine_winner( 1, '"goat", "goat", "cat"')
determine_winner ( 3, '"goat", "car", "goat"')

}
