% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine if the contestant has won.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{...}{the arguments final.pick and game are used by the function.}
}
\value{
The function returns either WIN or LOSE based upon the final pick.
}
\description{
\code{determine_winner()} determines whether or not the contestant has won based on
the final pick.
}
\details{
If the final pick is a car, then the contestant has won.  If the final pick is
a goat, then the contestant has lost.
}
\examples{
  determine_winner(final.pick = 2, game)

}
