% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine if the contestant has won.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{An integer representing the contestant's final pick (a value
between 1 and 3).}

\item{game}{A vector of length 3 representing the contents of the doors (either
"car" or "goat").}
}
\value{
The function returns either WIN or LOSE based upon the final pick.
}
\description{
\code{determine_winner()} determines whether or not the contestant has won based on
the final pick.
}
\details{
If the final pick is a car, then the contestant has won. If the final pick is
a goat, then the contestant has lost.
}
\examples{
  determine_winner(final.pick = 2, game)

}
