% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine whether partipant wins or loses}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick:}{the numerical value from the change_door() function. No default.
game: the vector from the create_game() function. No default.}
}
\value{
The function returns the string "WIN" if the final pick was a "car" door or "LOSE" if the final
pick was a "goat" door.
}
\description{
Now that a final door selection has been made, the function determines whether the final
selection is a car or goat door, and states "WIN" or "LOSE" accordingly.
}
\details{
The function uses the create_game() and change_door() values to determine whether the final
selection was a car or a goat door. A goat door results in a "LOSE" and a car door results
in a "WIN"
}
\examples{
  determine_winner( 3, c( "goat","goat","car" ))
  determine_winner( 1, c( "goat","car","goat" ))

}
