% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.r
\name{determine_winner}
\alias{determine_winner}
\title{Determine if contestant has won}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick:}{the numerical value from the change_door() function. No default.
game: the vector from the create_game() function. No default.}
}
\value{
The function returns the string "WIN" if the final pick was a "car" door or "LOSE" if the final
pick was a "goat" door.
}
\description{
Determine if contestant has won
}
\details{
The function reveals whether the final selection is a car or a goat door.
}
\examples{
  determine_winner( 3, c( "goat","goat","car" ))
  determine_winner( 1, c( "goat","car","goat" ))

}
