% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine if the contestant has won.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{a length 1 numeric vector between 1 and 3, the
door returned by \code{change_door()}.}

\item{game}{a length 3 character vector of "goat" and "car"
returned by \code{create_game()}.}
}
\value{
The function returns a length 1 character vector, either
"WIN" or "LOSE".
}
\description{
\code{determine_winner()} reports whether the contestant's final door
held the car or a goat.
}
\details{
The final pick is used as a position in the game vector to find
out what is behind that door. A car returns "WIN" and a goat
returns "LOSE". Every door holds one or the other, so one of the
two returns always fires.
}
\examples{
  this.game <- create_game()
  my.initial.pick <- select_door()
  opened.door <- open_goat_door( this.game, my.initial.pick )
  my.final.pick <- change_door( stay=F, opened.door, my.initial.pick )

  determine_winner( my.final.pick, this.game )

}
