% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem (4).R
\name{determine_winner}
\alias{determine_winner}
\title{Determine if the contestant has won}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{...}{Based on the game playing strategy if the final.pick equals " the
return is a "WIN" if the final.pick equals "goat" the return is "LOSE"
assigned.}
}
\value{
The function returns "WIN" if final.pick is equal to "car", the
functions returns "LOSE" if final.pick is equal to "goat".
}
\description{
\code{determine_winner( final.pick, game )} generates the outcome of winning
granted the contestant selects the door containing a car in his/her final
pick. It also generates the outcome of losing if the contestant selects the
door containing a goat in his/her final pick.
}
\details{
Selecting the final pick as either the door that has a car behind it, or a
door that has a goat behind yields the results of winning or losing at the
end of the game.
}
\examples{
  determine_winner( final.pick, game )

}
