% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine Game Outcome}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{An integer representing the contestant's final door choice.}

\item{game}{A character vector generated by \code{create_game()}, representing the positions of the car and goats behind the doors.}
}
\value{
The function returns a string: "WIN" if the contestant chose the car, or "LOSE" if they chose a goat.
}
\description{
This function determines whether the contestant has won or lost the game based on their final door selection.
}
\details{
If the contestant's final choice is the door with the car behind it, they win. If their final choice is a door with a goat behind it, they lose.
}
\examples{
game <- create_game()
final.pick <- change_door(stay = TRUE, opened.door = open_goat_door(game, a.pick = select_door()), a.pick = 1)
determine_winner(final.pick, game)

}
