% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{Did the contestant win?}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns "WIN" or "LOSE" depending on the contestant's final door selection.
}
\description{
\code{determine_winner()} provides the conditions to determine if the contestant won from staying with their intial pick or switching doors.
}
\details{
If the contestant's final selected door is a car, they win. If the contestant's final selected door is a goat, they lose.
}
\examples{
  determine_winner()

}
