% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine whether the contestant won.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{An integer (1, 2, or 3) representing the contestant’s final door.}

\item{game}{A character vector created by `create_game()`.}
}
\value{
A character string: `"WIN"` or `"LOSE"`.
}
\description{
`determine_winner()` checks whether the final pick corresponds
   to the car or a goat.
}
\details{
The function compares the contestant’s final pick to the
  hidden prize vector. Returns `"WIN"` if the final door
  hides the car and `"LOSE"` otherwise.
}
\examples{
  gg <- create_game()
  choice <- select_door()
  o <- open_goat_door(g, p)
  final <- change_door(FALSE, o, p)
  determine_winner(final, gg)

}
