% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem(bh).R
\name{determine_winner}
\alias{determine_winner}
\title{Determine if the contestant is a winner.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
If the character vector returned for 'final.pick' is a
"car" door, then the function returns "WIN". If the
character vector for 'final.pick' is a "goat" door,
then the function returns "LOSE".
}
\description{
'determine_winner()' determines whether the contestant is
a winner if their final pick is a "car" door.
}
\details{
The function uses the inputs 'final.pick' and 'game' to
determine if the contestant is a winner based on  whether
the contestant chooses a "goat" or "car" door in the
final selection.
}
\examples{
  determine_winner(final.pick, game)

}
