% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{Reveals whether the final door selection was a winner or loser.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{Integer value representing the final door choice}

\item{game}{Vector of length 3 containing 2 "goat" and 1 "car"}
}
\value{
a character string. "WIN" or "LOSE"
}
\description{
After the contestant makes their final door selection, the
host unveils whether the door held a goat (lose) or a car
(win).
}
\details{
\code{determine_winner()} compares the final door selection with
the original game vector. If the final door's value as an
index in the game vector is "car" the contestant wins and
will return "WIN". Otherwise, if the final door's value as an
index in the game vector is a "goat" the contestant loses.
Requires the final door selection and the initial game vector.
}
\examples{
  determine_winner( 2 , c("goat", "car", "goat"))  #Returns WIN
  determine_winner( 1 , c("goat", "car", "goat"))  #Returns LOSE

}
