% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine the Winner.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{Uses}{2 arguments: final.pick (either switch or stay) and "game" which is the
current game.}
}
\value{
"WIN" or "LOSE" based on the game's outcome.
}
\description{
Function determines the winner of the game.
}
\details{
If the final pick is "car", then the player wins. If the final pick is "goat"
the player loses.
}
\examples{
 game.outcome.stay <- determine_winner( final.pick=final.pick.stay, game=this.game )
 game.outcome.switch <- determine_winner( final.pick=final.pick.switch, game=this.game )
}
