% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine if te contestant won}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{A number}

\item{game}{A array of strings representing the current game.}
}
\value{
A string of either WIN or LOSE
}
\description{
\code{determine_winner()} determines the winner of the game
}
\details{
This function takes the final state of teh game and determines if the
contestant won the car or not
}
\examples{
  determine_winner()
}
