% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{The simulation indicates whether the contestant's final pick results in a
win or a loss of the Monty Hall Problem game.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{The}{arguments 'final.pick' and 'game' are used by the function.}
}
\value{
The function returns a string indicating whether the door that the
contestant ultimately chose results in a win or loss of the game.
}
\description{
\code{change_door} allows the contestant continue on through the game using the
door they originally selected, or to switch their door to the remaining
unopened door.
}
\details{
This is the portion of the game where the contestant decides whether to
continue by using the door they originally selected, or to change their
door choice to the remaining unopened door, knowing that the host has
opened a door with a goat behind it, so one of the remaining two doors has
the other goat behind it, while the other has a car behind it.
}
\examples{
  determine_winner(final.pick, game)
  
}
