% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{Goat or No Goat}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{no}{params}
}
\value{
If the final selection is a car, then 'WIN' is received; if not, then 'LOSE' is
received.
}
\description{
\code{determine_winner()} The final step in the game is for the host to reveal
whether the contestant's final choice is a goat or car.
}
\details{
The contestant has a slightly increased chance to pick a car if the contestant
switched in the previous step. This step reveals if switching doors wins a car or
a goat.
}
\examples{
change_door()=3; select_game()='goat','goat','car'; and determine_winner()=
'WIN'; else the contestant does not switch and change_door()=1...
determine_winner()='LOSE'
}
