% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{\preformatted{Determine Winner.
}}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{...}{\code{final.pick} and \code{game} are the arguments}
}
\value{
\preformatted{the function returns a "WIN" or "LOSE" based on whether or not the game was won.
}
}
\description{
\preformatted{`determine_winner()` returns the item behind the final door chosen, and assigns 
choosing the car as winning, and choosing the goat as losing.
}
}
\details{
\preformatted{Provides final door and determines whether they won or lost.
}
}
\examples{
	determine_winner(final.pick = finalDoor, game = newGame)

}
