% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine whether player won or lost}
\usage{
determine_winner(final.pick, game)
}
\description{
This function determines whether or not the player won a car or a goat
}
