% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{This function determines if the player won or lost the Monty Hall game.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{This}{function takes two parameters, final.pick the output of change_door() and the value of the door that was the player's final pick and game which is the output of create_game() and defines what is behind each door.}
}
\value{
This function returns a string "WIN" if the car door was chosen and "LOSE" if a goat door was chosen.
}
\description{
This function evaluates if the player's final pick corresponds to the door with a car behind it. If so the player wins otherwise they lose.
}
\details{
This function evaluates if the player's final pick corresponds to the door with a car behind it. If so the player wins otherwise they lose.
}
\examples{
determine_winner(my.final.pick, my.game) 

}
