% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine if Contestant has Won}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{number between 1 and 3}

\item{game}{establishes the random scenario that create_game() created: sample( x = c("goat","goat","car"), size=3, replace=F )}
}
\value{
characters "WIN" or "LOSE"
}
\description{
determine_winner() puts the previous arguments together to see if the contestant wins or loses depending on the starting scenario set by create_game() and the contestants choice to stay or switch doors
}
\details{
this function pulls from previous arguments 'final.pick' and 'game' to establish original scenario and contestant choice. the function itself is comprised of a simple logical argument: if 'final.pick' in this 'game' resulted in a car, the contestant wins, if not, they lose. the function will return "WIN" or "LOSE" depending on whether or not the contestant wins.
}
\examples{
  this.game <- c("goat","car","goat")
  determine_winner( final.pick=1, game=this.game ) # should be "LOSE" 
  
  this.game <- c("goat","car","goat")
  determine_winner( final.pick=2, game=this.game ) # should be "WIN" 
  
  this.game <- c("goat","car","goat")
  determine_winner( final.pick=3, game=this.game ) # should be "LOSE" 
  
}
