% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine if the contestant has won.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick:}{a number (between 1 to 3) for the door selected
by the contestant finally;
game:       a vector of 3 characters(2 "goats" and 1 "car")
created by 'create_game()';}
}
\value{
The function returns the result ("WIN" or "LOSE").
}
\description{
\code{determine_winner( final.pick, game )}decides if you win
the car or not
}
\details{
The function is to test if the contestant's final pick is
the car or not. If it is the car, display "WIN", otherwise,
display "LOSE".
}
\examples{
this.game
my.initial.pick
my.final.pick <- change_door( stay=T,
                             opened.door=opened.door,
                             a.pick=my.initial.pick )
determine_winner( final.pick=my.final.pick,
                 game=this.game )
my.final.pick <- change_door( stay=F,
                             opened.door=opened.door,
                             a.pick=my.initial.pick )
determine_winner( final.pick=my.final.pick,
                 game=this.game )
}
