% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine whether game is won or lost}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{Arguments}{are one numeric vector (final.pick) and one 3 character vector (game)}
}
\value{
Returns game outcome - "Win" or "Lose"
}
\description{
This function looks at the final door the contestant chooses to see if it is a goat or a car
}
\details{
If the final door is a car, the game is won. If the final door is a goat, the game is lost.
}
