% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine if the Contestant has Won or Lost the game}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{Parameter}{final.pick is vector length 1 of integer 1, 2, or 3, representing
the final door chosen by the contestant via 'change_door()'. This is compared
to the current game created by 'create_game()'.}
}
\value{
Returns character string being one of "Win" or "Lose" if the final door
chosen is the car door (Win) established in the current game or not (Lose).
}
\description{
\code{determine_winner()} judges the contestant as having won the car or lost the
game and then returns the game results.
}
\details{
If a car is behind the final contestant chosen door, the player won.
Otherwise, the payer has lost.
}
\examples{
determine_winner ()

}
