% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine Winner.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{...}{A Character value}
}
\value{
A single character vector indicating "WIN" or "LOSE"
}
\description{
\code{determine_winner()} determines if the contestant
won the game by making a final selection of a
door with a car behind it.
}
\details{
If the contestant's final selection is a door with
a car behind it, then the contestant won the game.
If the contestant's final selection is a door with
a goat behind it, then the contestant lost the
game.
}
\examples{

`determine_winner()`
[LOSE]

`determine_winner()``
[WIN]

}
