% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine the Winner}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{A single integer between 1 and 3 representing the player's final door selection.}

\item{game}{A vector of length 3 representing the contents behind the doors (either "car" or "goat").}
}
\value{
A character string of either "WIN" or "LOSE".
}
\description{
determine_winner function checks the final selection and returns if the player wins or loses.
}
\details{
The function determines the winner of the game by checking the final selection of the player against the contents of the doors.
}
\examples{
  determine_winner()

}
