% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine if contestant has won.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{and game ... these are default arguments from the previous
function `change_door` to run the function correctly.}
}
\value{
The function will return either "WIN" or "LOSE"
(character vector) that indicating the winning status.
}
\description{
`determine_winner()` this function is to determine
the winning status of the contestant.
}
\details{
After the final discussion of the contestant,
either to change his/her door selection or to stay
with the first selection, the function will determine
his/her winning.
}
\examples{
this.game <- c("goat","car","goat")
determine_winner( final.pick=1, game=this.game )
determine_winner( final.pick=2, game=this.game )
determine_winner( final.pick=3, game=this.game )


}
