% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monty Hall Script.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine whether the final door is a winner}
\usage{
determine_winner(final.pick, this.game)
}
\arguments{
\item{final.pick}{A numeric vector of length one.}

\item{this.game}{A character vector of length three.}
}
\value{
This function returns whether or not the player won or lost.
}
\description{
This function returns whether or not the player's final door choice contains the car or a goat. If
the door was concealing the car, the player wins. If the door was concealing a goat, the player loses.
}
\details{
This function checks whether final door number is equal to the position of the car in the game. If yes,
then "WIN" is returned. If no, then "LOSE" is returned.
}
\examples{
determine_winner(final.pick = 1, this.game = c("car", "goat", "goat"))

determine_winner(final.pick = 2, this.game = c("goat", "car", "goat"))

}
