% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{Monty Hall Prize Reveal}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{...}{It requires the game array from create_game() and the final.pick
from change_door(stay = T, opened.door, a.pick).}
}
\value{
The return is the text stored in game\link{final.pick}.
}
\description{
This function represents the point in the game where the host reveals
what the participant has won.
}
\details{
This function reveals what is stored in the final.pick position of the
game array.
}
\examples{
determine_winner( final.pick, game )
  
}
