% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{Determines whether the contestant wins or loses.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{An integer between 1 and 3}

\item{game}{A length three character vector indicating the
position of goats and car.}
}
\value{
the function returns a string.
}
\description{
\code{determine_winner} determines if the contestant's final pick
is a car door or a goat door. If it is a car door, the
contestant wins, if it is a goat door, they lose.
}
\details{
The game setup replicates the game on the TV show "Let's
Make a Deal" where there are three doors for a contestant
to choose from, one of which has a car behind it and two
have goats. The contestant selects a door, then the host
opens a door to reveal a goat, and then the contestant is
given an opportunity to stay with their original selection
or switch to the other unopened door. There was a famous
debate about whether it was optimal to stay or switch when
given the option to switch, so this simulation was created
to test both strategies.
}
\examples{
  determine_winner(final.pick = 1, game = c("car", "goat", "goat"))
  determine_winner(final.pick = 2, game = c("car", "goat", "goat"))
  
}
