% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine winner}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{keep}{final pick from previous funtion}
}
\value{
return win or lose
}
\description{
Determine if the end user won the game or not
}
\details{
In the final pick function, we figure out if the door picked was a winner by selecting the door with the car or if the door was a goat door, losing the game.
}
\examples{
object <- function return (WIN)
}
