% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine if the contestant is a winner.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
If the character vector for 'final.pick' is a "car" door,
the function returns "WIN". If 'final.pick' is "goat",
the function returns "LOSE".
}
\description{
'determine_winner()' determines the contestant is a winner
if their final pick is a "car" door.
}
\details{
The function uses inputs 'final.pick' and 'game' to determine
if the contestant is a winner.
}
\examples{
  determine_winner(final.pick, game)

}
