% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhallproblem.R
\name{determine_winner}
\alias{determine_winner}
\title{Determines winner of the Monty Hall Problem game.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{is the numeric value of the final door selection.}

\item{Game}{is a character vector. Refer to the function 'create_game'}
}
\value{
The output will be WIN if the final door contains the car.
The output will be LOSE if the final door contains the goat.
}
\description{
\code{determine_winner()} is a function that indicates whether the contestant
won or lost the game.
}
\details{
This function works using an if statement. If the contestant's final
door selection contains the car, the contestant wins. If the contestant's
final door selection contains the goat, the contestant loses.
}
\examples{
  this.game
  my.initial.pick
  my.final.pick <- change_door( stay=T, 
  opened.door=opened.door, 
  a.pick=my.initial.pick )
  determine_winner( final.pick=my.final.pick, 
  game=this.game )
  my.final.pick <- change_door( stay=F, 
  opened.door=opened.door, 
  a.pick=my.initial.pick )
  determine_winner( final.pick=my.final.pick, 
  game=this.game )
  
}
