% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine if contestant has won}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{If}{statement, if the string of the final door selection equals goat, then
return the string, lose.}
}
\value{
The function returns string, "WIN" if contestant's final door selection has
a car. The function returns string, "LOSE" if contestant's final door
selection has a goat behind it.
}
\description{
\code{determine_winner()} This function was created in order to determine whether
the contestant chose the correct door with the car behind it and win the
game.
}
\details{
Using the control "If" statement, if the contestant's final selection is
the door with the car behind it then they have won the game. If the
contestant's final selection is the door with the goat behind it then they
lose the game.
}
\examples{
  determine_winner()
}
