% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine the winner of the game.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{A numeric value between 1 and 3 indicating the contestant's final door selection.}

\item{game}{A length 3 character vector indicating the positions of the goats and the car.}
}
\value{
The function returns a character value of "WIN" if the selected door contains the car or "LOSE" if it contains a goat.
}
\description{
'determine_winner()' determines whether the contestant wins or loses based on their final door selection.
}
\details{
The function checks the contestant's final selected door. If the car is behind the selected door, the outcome is a win. If a goat is behind the selected door, the outcome is a loss.
}
\examples{
game <- c("goat", "car", "goat")
determine_winner(2, game)
determine_winner(1, game)
}
