% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine if the contestant has won.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{a number between 1 and 3}

\item{game}{length 3 character vector}
}
\value{
This function returns a length one character vector
}
\description{
\code{determine_winner()} reveals if the final pick results in a win or lose.
}
\details{
If the final pick is a car door, then the contestant wins. If the final pick
is a goat door, then the contestant loses.
}
\examples{
determine_winner()

}
