% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhallfunctions.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine whether the contestant has won.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{...}{no arguments are used by this function.}
}
\value{
If 'final_pick' is equal to "car", then the function
returns "WIN". If 'final_pick' is equal to "goat", then
the function returns "Lose".
}
\description{
If the contests ends the game with a "car" door,
determine_winner returns their winning status.
}
\details{
'determine_winner()' uses the status of 'final.pick' and 'game'
to determine if the contestant has won.
}
\examples{
  determine_winner(final.pick, game)
}
