% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine whether the contestant wins}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{Integer. The contestant's final selected door.}

\item{game}{Character vector. The game layout.}
}
\value{
A character string \code{"WIN"} if the selected door has the car, otherwise \code{"LOSE"}.
}
\description{
Checks whether the contestant's final pick corresponds to the car.
}
\examples{
g <- create_game()
p <- select_door()
o <- open_goat_door(g, p)
f <- change_door(FALSE, o, p)
determine_winner(f, g)

}
