% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine Winner}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{A number between 1 and 3 representing the contestant's final door selection.}

\item{game}{A character vector representing the game, with two goats and one car.}
}
\value{
The function returns either "WIN" or "LOSE".
}
\description{
Determines whether the contestant wins or loses the Monty Hall game.
}
\details{
The function compares the contestant's final door selection with the game setup.
The contestant wins if the final selection contains the car and loses if it contains a goat.
}
\examples{
game <- create_game()
pick <- select_door()
opened <- open_goat_door(game, pick)
final.pick <- change_door(stay = FALSE, opened.door = opened, a.pick = pick)
determine_winner(final.pick, game)

}
