% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{\itemize{
\item this is the determine a winner function
}}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{-}{in terms of the parameters here, our first argument articulates that if a contestant chooses the car door, they win. The second argument articulates that if they choose a goat door as their final pick, they lose,}
}
\value{
\itemize{
\item we are returned win or lose as a result of two options.
}
}
\description{
\itemize{
\item the function will enable us to decide whether or not a contestant has won a car.
}
}
\details{
\itemize{
\item not much else to share here.
}
}
\examples{
# test code

this.game
my.initial.pick

my.final.pick <- change_door( stay=T, 
                             opened.door=opened.door, 
                             a.pick=my.initial.pick )

determine_winner( final.pick=my.final.pick, 
                 game=this.game )

my.final.pick <- change_door( stay=F, 
                             opened.door=opened.door, 
                            a.pick=my.initial.pick )

determine_winner( final.pick=my.final.pick, 
                 game=this.game )
}
