% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine whether the contestant is a winner.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{This}{function uses the final pick and game functions.}
}
\value{
This function returns 'WIN' or 'LOSE'
}
\description{
\code{determine_winner()} indicates whether the contestant won or loss.
}
\details{
This function indicates whether the final selected door has a car behind it.
}
