% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem (1).R
\name{determine_winner}
\alias{determine_winner}
\title{Determine Winner}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{...}{arguments are (final.pick,game)}
}
\value{
The function prints if its a win or a lost
}
\description{
The function determines if the contestant has won or lost based on the
final pick. The contestant could have either stayed on the initial selection
or switched to a new one.
}
\details{
If the final pick by the contestant is a car door then the contestant wins.
However, if the final pick is a goat door, then the contestant looses.
}
\examples{
determine_winner (final.pick,game)

}
