% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine if the contestant wins or loses the game.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{The door number that the player chooses
as their final door, a number between 1 and 3}

\item{game}{A length 3 character vector represents the label
of the doors, where one door is labeled "car" and the other
two are labeled "goat".}
}
\value{
The function returns a character vector indicating
if the contest wins or loses the game. The contestant
wins the game if there is a car behind the final door.
}
\description{
\code{determine_winner()} generates a win or loss based on if
the contestants final door is a car door or not.
}
\details{
After the contestant chooses to stay or switch their door,
their final door is chosen. If the final door has a goat
behind it the contestant loses, and if the final door has
a car behind it, the contestant wins.
}
\examples{
  determine_winner(final.pick = 2, game = c("goat", "car", "goat"))
    # "WIN"
  
}
