% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine the game outcome.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{The final chosen door.}

\item{game}{The current game setup.}
}
\value{
"WIN" if the final pick is the car, "LOSE" if it's a goat.
}
\description{
Determine the game outcome.
}
\examples{
determine_winner(final.pick = 2, game = c("goat", "car", "goat"))

}
