% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{The function that determines if the participant
won or lost based on what is behind the chosen
door.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{...}{final.pick is an argument that holds the
choice of the participant as to whether or not to
stay or switch doors. game is an argument that holds
information about what is behind each door (e.g., a
car or goat). the game argument tells us what is
behind the door of the final.pick argument (which is
what the contestant chose) which is then used in the
function  determine_winner() to see if the contestant
won or lost. final.pick is numeric and requires numeric
input whereas game is a character vector that requires
input of the same type.}
}
\value{
... this will print either "WIN" or "LOSE"
depending on the if statements in the function.
}
\description{
"determine winner()" allows us to see if the
contestant won based on whether the final door
picked by the contestant has a goat or car.
}
\details{
This function uses the arguments of final.pick and
game. final.pick represents the final pick of the
participant: whether they chose to stay or switch.
"Game" provides information about the ordering
of the goats and cars. If the final.pick door has
the car behind it, then the function will return win.
If the final.pick door has a goat behind it, then
the function will return lose.
}
\examples{

this.game <- c("goat","car","goat")
determine_winner( final.pick=1, game=this.game )
determine_winner( final.pick=2, game=this.game )
determine_winner( final.pick=3, game=this.game )

}
