% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{determine_winner}
\alias{determine_winner}
\title{The game decides if the final door has a car (Win) or a goat (Lose).}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{The}{parameters for this function include the \code{final.pick} from the
previous step and the \code{game} that was created using the \code{create_game()}
function. Creating objects for both scenarios whether
or not the contestant stayed with the door or switched makes it
more useful:
my.final.pick.stay <- \code{change_door( stay = T, opened.doors, my.initial.pick )}
my.final.pick.switch <- \code{change_door( stay = F, opened.doors, my.initial.pick )}}
}
\value{
This function returns "WIN" or "LOSE" depending on the final door
that is selected.
}
\description{
\code{determine_winner()} takes the final selected door and passes it
through to determine if the door has a goat or a car behind it.
}
\details{
This function pretty simply looks at the final door from the
\code{final.pick} and matches it to the \code{game} that was created in the
beginning. If the \code{final.pick} contains a car, it is considered
a win while inversely, if \code{final.pick} contains a goat, it is
considered a loss.
}
\examples{
 `determine_winner( my.final.pick.stay, this.game )`

 `determine_winner( my.final.pick.switch, this.game )`

}
