% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine if the contestant won or lost.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{This is an integer representing the contestant's final
door selection after deciding to stay or switch. Must be between
1 and the number of doors.}

\item{game}{This is a character vector containing "car" and "goat" values
representing the current game setup.}
}
\value{
The function returns a character string: "WIN" if the
contestant's final pick was a car, or "LOSE" if the contestant's
final pick was a goat.
}
\description{
\code{determine_winner()} evaluates the contestant's final door
selection and determines whether they won the car or got a goat.
}
\details{
This function checks what prize is behind the contestant's
final door selection. If the door contains a car, the contestant
wins ("WIN").
If the door contains a goat, the contestant's a \bold{"LOSE"}.
}
\examples{
  game <- c("goat", "car", "goat")
  determine_winner(final.pick = 2, game)
  determine_winner(final.pick = 1, game)

}
