% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_winner.R
\name{determine_winner}
\alias{determine_winner}
\title{Monty Hall Game Winner Test}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{An integer greatr than zero and not larger than the length of the \code{game} vector.}

\item{game}{A vector of any class and length.}
}
\value{
If the \code{final.pick}th element oof the game vector is "car", it returns "WINNER", and if
it equals "goat" it returns "LOSER".
}
\description{
\code{determine_winner( final.pick, game )} determines if the selected door is
a car door and the contestant is a winner.
}
\details{
Given a final door selection and a game vector, this function determines if that door
selection leads to the contestant being the winner. It checks if the door number is
associated with a "car" in the game vector.
}
\examples{
determine_winner( 2, c("goat", "goat", "car" ))
determine_winner( 1, create_game(6), 3 )

\dontrun{
determine_winner( 4, c("goat", "car", "car" ))
}

}
