% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg (2).R
\name{determine_winner}
\alias{determine_winner}
\title{determine winner}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{#no}{arguments are used by the function.}
}
\value{
win" or "lose"
}
\description{
determine_winner() determines if the contestant won or lost the game
}
\details{
determine_winner() determines if the contestant won or lost the game
}
\examples{
determine_winner()
}
