% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine the winner.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{A number between 1 and 3 indicating the
contestant's final selected door.}

\item{game}{A length 3 character vector indicating the
positions of goats and the car.}
}
\value{
The function returns "WIN" if the selected door
contains the car and "LOSE" if it contains a goat.
}
\description{
\code{determine_winner()} determines whether the contestant
wins or loses the game.
}
\details{
The function checks whether the contestant's final selected
door has a car or a goat behind it.
}
\examples{
  game <- create_game()
  final.pick <- select_door()
  determine_winner( final.pick, game )

}
