% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine whether the contestant has won or lost the game.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{The}{arguments for this function are as follows: final.pick, which
contains the numeric value corresponding with the door selection. The
other argument, game, contains the character vector the contains the
current game being played.}
}
\value{
The function returns a character value that tells you if the contestant
wins or loses, those values being "WIN" and "LOSE".
}
\description{
\code{determine_winner()} function takes the character value of "final" pick,
and if the value is "car", the function returns "WIN". And if the value
is "goat", the function returns "LOSE".
}
\details{
"final.pick" contains the numeric value for the door selection. Using the
[] operator on the vector "game", if the value of "final.pick" corresponds
to "car" in the "game" vector, than the function returns the character
value "WIN". If the value of "final.pick" corresponds to "goat" in the
"game" vector, the the function returns the character value "LOSE".
}
\examples{
   determine_winner( final.pick, game )

}
