% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{TITLE}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{details}{}
}
\value{
details
}
\description{
determine_winner
}
\details{
details
}
\examples{
details
}
