% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine Winner}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{if(}{game\link{ final.pick } == "car" ), return( "WIN" )
if( game\link{ final.pick } == "goat" ),return( "LOSE" )}
}
\value{
A message of "WIN" or "LOSE" is returned.
}
\description{
If the player selected the door with the car behind it, they win.
If the player selected a door with a goat behind it, they lose.
}
\details{
If final.pick is equal to "car", the player wins.
If fina.pick is equal to "goat", the player losers.
}
\examples{
 "WIN" or "LOSE"
}
