% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{Pick that Door}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{...}{similar to the other parameters there will be an
if functions where if the pick is car then win if the car is
goat then you loose}
}
\value{
Win or lose will be presented based on the final selection
}
\description{
determine_winner()` determines the winner of the game after
the option to change the door
}
\details{
The game setup replicates the game on the TV show "Let's
Make a Deal" where there are three doors for a contestant
to choose from, one of which has a car behind it and two
have goats. The contestant selects a door, then the host
opens a door to reveal a goat, and then the contestant is
given an opportunity to stay with their original selection
or switch to the other unopened door. There was a famous
debate about whether it was optimal to stay or switch when
given the option to switch, so this simulation was created
to test both strategies.
}
\examples{
determine_winner()
}
