% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{Determines if the contestant is a winner or not.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{`final.pick`}{\itemize{
\item The contestant's final door choice - number 1 through 3.
\code{game} - Vector that holds the door position of the goats
and car.
}}
}
\value{
The function returns either the string \code{WIN} or \code{LOSE} depending
on whether the contestant chose the car door or the
goat door.
}
\description{
\code{determine_winner()} uses the contestant's final door choice
and combines it the knowledge of where the car is to generate
whether the contestant won or lost.
}
\details{
The function checks if the contestant's final choice was the car or the
goat. It returns the string \code{WIN} if it was the car
If it was the goat, it returns the string \code{LOSE}.
}
\examples{
  determine_winner()

}
