% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mh.pack.R
\name{determine_winner}
\alias{determine_winner}
\title{Determining If The Player Won}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{The}{arguments within this function are final.pick and game}
}
\value{
The function returns "WIN" if the player's final pick had a car
behind the door, and "LOSE" if the player's final pick had a goat behind
the door.
}
\description{
\code{determine_winner()} determines if the player won or lost the game based on
their final pick being either a goat or a car
}
\details{
Once the player makes their final choice the door will be opened to reveal
whether the door they chose had a goat or a car behind it. If the player
chose the door with the goat behind it then the player lost the game. If
the player chose the door with the car behind it then the player won the
game.
}
\examples{
  determine_winner()

}
