% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_winner.R
\name{determine_winner}
\alias{determine_winner}
\title{Did I Win?}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{...}{final.pick, game}
}
\value{
returns win or lose
}
\description{
This function determines if you win or not
}
\details{
This function will return a WIN or LOSE depending on the outcome
}
\examples{
  determine_winner()
}
