% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine the winner}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{represents the contestant's final door choice.
`game` represents the outcome of the game.}
}
\value{
The function returns WIN if the door that represents final.pick has a car and 
returns LOSE if the final.pick has a goat
}
\description{
The winner of the game is determined.
}
\details{
If the contestant's final pick is a car, the function will return WIN. If the
contestant's final pick is a goat, the function will return LOSE.
}
\examples{
determine_winner(final.pick = 2, game = c("goat", "car", "goat"))
}
