% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine if Contestant Has Won}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{The}{parameters to this function include the previous steps being completed to provide
the game's results.}
}
\value{
The function returns a character string of "LOSE" or "WIN".
}
\description{
'determine_winner' function returns whether or not the final door choice is
the position of the car.
}
\details{
This step in the game is the final reveal of the doors and whether the contestant
has won the car or not.
}
\examples{
this.game <- c("goat","car","goat")
determine_winner( final.pick=1, game=this.game )

}
