% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine if contestant has won.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{A number, representing the final door pick the
contestant made (the first argument).}

\item{game}{a character vector of length 3, 2 goats and 1 car
(the second argument).}
}
\value{
The function returns a character string of "win" or "lose"
to indicate if the contestant has chosen the door with the car.
}
\description{
\code{determine_winner()} generates a character string that represents
if contestant has won game.
}
\details{
Contest playing the famous TV show game, "Let's Make a Deal," is
revealed whether they have won or lost the game!
}
\examples{
  determine_winner(final.pick=3, game=c('car','goat','goat'))

}
