% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine the Game Outcome}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{The contestant's final chosen door number.}

\item{game}{A character vector representing the game setup.}
}
\value{
"WIN" if the contestant wins, "LOSE" if they lose.
}
\description{
Determines if the contestant wins or loses the Monty Hall Problem game based on their final choice.
}
\details{
The \code{determine_winner()} function evaluates the game's result in the context of the Monty Hall Problem.
It verifies whether the contestant's ultimate selection matches the car's position within the game's configuration.
If the contestant wins, it returns "WIN"; otherwise, it returns "LOSE."
}
\examples{
  determine_winner(final.pick = 2, game = create_game())

}
