% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem functions Hauer.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine Winner or Loser.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{An integer representing the door that the
contestant has finally selected (a value between 1 and 3).}

\item{game}{A vector of length 3 representing the contents of the doors
(either "car" or "goat").}
}
\value{
The results of the game, Win or Lose, is returned.
}
\description{
This function reveals if the contestant has correctly selected the door with
the car behind it, winning the game. Or if the contestant looses.
}
\details{
For this function, in order to determine if the player has won we use
information about the door they have selected and the game set-up
to evaluate if they win the car (win) or the goat (lose).
}
\examples{
determine_winner()

}
