% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine winner}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{The final choice}

\item{game}{the game array containing randomly
spread 'cars' and 'goats'}
}
\value{
The function returns WIN or LOSE
}
\description{
`determine_winner()` determines who is winner on basis of the final
   pick . if its a car then its a win or else loss
}
\details{
If the final pick is car then the result is WIN
  If the final pick is goat then the result is LOSE
}
\examples{
  determine_winner(7,game)
}
