% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{Checks to see if the user won the game.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{...}{The function takes in the parameters: final.pick (integer value)
and game (character values).}
}
\value{
If the final choice corresponds to the position of the car door in game, the
function will return "WIN" indicating the user won. Otherwise, the function
will return "LOSE" if the final choice is not the car door.
}
\description{
\code{determine_winner(final.pick,game)} returns a "WIN" or "LOSE" depending on
the final choice.
}
\details{
The function validates if the user's final choice is equal to the door number
which contains the car behind it.
}
\examples{
  game <- create_game()
  final.pick <- change_door(stay = T, opened.door, a.pick)
  determine_winner(final.pick,game)
  
}
