% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{Did you win?}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a string of win or lose.
}
\description{
Monty now shows you what is behind the door you selected, showing you if
you won a beautiful new car, or a goat!
}
\details{
If the contestant selected the car door, they will see "WIN", if they 
selected the goat door they will see, "LOSE"
}
\examples{
determine_winner 

}
