% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine the winner in a Monty Hall Problem game.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{`final.pick`}{is an integer (1, 2, or 3) representing the door number
that the contestant ends up with after deciding to keep or switch. \code{game}
is a length 3 character vector of 3 game items that show two goats and
one car in any order.}
}
\value{
A character string ("WIN" or "LOSE") determining whether the
contestant's final selection resulted in a win or loss.
}
\description{
\code{determine_winner()} determines if the contestant wins or loses
based on their final selection in the Monty Hall Problem game.
}
\details{
The function compares the contestant's final selection and game setup to
the win/loss criteria for the game. If the contestant's final selection
is a car, they win. If the contestant's final selection is a goat, they
lose.
}
\examples{
  determine_winner()
}
