% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{Did you win?}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{...}{This function takes two arguments. \code{final.pick} which is an integer and the door
number behind which the contestant hopes the car to be. \code{game} is the length 3 character
vector of two "goat" strings and one "car".}
}
\value{
The function returns a character string, either "WIN" or "LOSE" based on which character is
returned for the position \code{final.pick} in the \code{game} character vector.
}
\description{
\code{determine_winner()} will return the outcome string corresponding to whether
the final door picked by the contestant did or did not have a car behind it.
}
\details{
After the contestant chooses to stay with their initial pick or switch to the
remaining door, the outcome is declared. The declaration is through a character
string, either "WIN" if the car was indeed behind the final door chosen by the contestant
and "LOSE" otherwise.
}
\examples{
determine_winner(final.pick=3,game=a.game)
}
