% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine if the Contestant Won}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{This}{is based on the contestant's final pick and the particulr game's doors.}
}
\value{
It will return a value of either "WIN" or "LOSE"
}
\description{
'determine_winner()' will display whether or not the contestant won or lost,
based on their final door pick.
}
\details{
If contestant's final choice was the 'car' door, this will show WIN. If the
contestant's final choice was a 'goat' door, it will show LOSE.
}
\examples{
Sticking with my example above, if I chose to "STAY" with Door 1 = GOAT, my 
result would show "LOSE". 
}
