% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem (1).R
\name{determine_winner}
\alias{determine_winner}
\title{Determine Win or Lose}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{The}{choices made by the plaer and the position of the goat doors}
}
\value{
WIN or LOSE
}
\description{
Dependent upon the choices made previously
}
\details{
if final choice = car, WIN
if final choice  = goat, LOSE
}
\examples{
determine_winner(1, (car, goat, goat)) --> win
determine_winner(2, (car, goat, goat)) --> lose
}
