% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{Pick Winner}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{}
}
\value{
WIN or LOSE
}
\description{
Outcome of the game
}
\details{
The results are shown to the contestant if they win a car
}
\examples{
  game <- create_game()
  final.pick <- change_door(TRUE, open_goat_door(game, select_door()), select_door())
  determine_winner(final.pick, game)
}
