% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine if the contestant wins.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns win or lose.
}
\description{
`determine_winner()` returns if the contestant wins or loses.
}
\details{
Once the contestant makes their final selection, the door
  that they chose is opened. Upon seeing the results, the
  show can determine if they win the car or win the goat.
}
\examples{
  determine_winner()

}
