% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine if the contest won or lost the game}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{The final pick of the contestant}

\item{game}{the order of the contents of the doors.}
}
\value{
Results of the game - "WIN" or "LOST"
}
\description{
Determines the results of the game
}
\details{
If the final pick of the contestant contains a "car" behind it, the contest won. If there a=is a "goat", then the contestant lost.
}
\examples{
determine_winner(2, c("car","goat","goat"))
}
