% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine the outcome of the Monty Hall game.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{game}{A character vector of length 3, representing the positions of the car and the goats. The vector consists of two "goat" values and one "car" value.}

\item{`final.pick`}{An integer (1, 2, or 3) representing the contestant's final door selection.}
}
\value{
The function returns a character string: \code{WIN} if the contestant selected the door with the car, or \code{LOSE} if they selected a door with a goat.
}
\description{
\code{determine_winner()} checks the contestant's final door selection to determine if they have won the car or received a goat.
}
\details{
After the contestant makes their final door selection, this function compares the selection to the game setup to determine whether the contestant wins (by selecting the door with the car) or loses (by selecting a door with a goat).
}
\examples{
`determine_winner()`

}
