% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine if Contestant Has Won}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
'win' or 'lose'
}
\description{
'if(final.pick, game' will be used to create a
text return based on the outcome of the game.
}
\details{
The contestant will either win or lose, based on
their choice whether or not to swap. This will be
created using a if function with a return of either
'win' or 'lose'
}
\examples{
determine_winner
}
