% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty_hall_pkg1.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine if the contestant has won.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{a number between 1 and 3 indicating the contestant's
final door selection after their stay or switch decision.}

\item{game}{a length 3 character vector indicating the positions
of goats and the car.}
}
\value{
The function returns a character vector of either Win or Lose
indicating the results of the game.
}
\description{
\code{determine_winner()} determine if the contestant has won the game
based on their final door selection.
}
\details{
\code{determine_winner()} determines that the contestant has won if
their final door selection contains a car; otherwise,
\code{determine_winner()} determines that the contestant has lost the game
if their final door selection contains a goat.
}
\examples{
  game <- create_game()
  first.pick <- select_door()
  host.door <- open_goat_door(game, first.pick)
  final.pick <- change_door(stay=T, host.door, first.pick )
  determine_winner(final.pick, game)

}
