% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{``Determine if contestant has won}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{'final.pick'}{\itemize{
\item a number between 1 and 3 that
corresponds to the door the contestant ended up with at the end
'game' - the game in there are 3 doors consisting of 2 goats and 1 car
"WIN" - occurs when the contestant's final pick has a car behind it
"LOSE" - occurs when the contestant's final pick has a goat behind it
}}
}
\value{
The function returns the word "WIN" if final.pick equals car or the word
"LOSE" of final.pick equals goat.
}
\description{
The function 'determine_winner()' returns 'WIN' if the contestant's door
is the door with the car behind it or 'LOSE' if it was a door with a goat
behind it.
}
\details{
After deciding whether to stay or switch, the contestant's final choice
of door must be opened. If the host opens the contestant's door and the
car is behind it, they win. If the host opens the contestant's door and
a goat is behind it, they lose.
}
\examples{
  determine_winner()

}
