% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHallKRB.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine game outcome based on guest's choice of doors (to stay or not).}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick, }{a.game}
}
\value{
returns a character vector of length 3 or 4, which represents whether the guest won or lost the game
by choosing a door with or without a car.
}
\description{
Determines if guest wins or loses evaluating the game value (goat or car) final pick of door, whether they change
doors or not.
}
\details{
If the guest's choice is to stay with the first choice of doors and that door corresponds to a
value of "car" in a.game, the guest 'wins'. If the guest decides to change doors and the new door
of their choice is a 'car' door, the guest wins. If the guest stays with the first pick and that
door is a 'goat' door, the gues loses. If the guest changes doors and the new chosen door is a
'goat' door, the guest loses.
}
\examples{
"WIN" or "LOSE"
}
