% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine the Winner}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{Integer indicating the contestant's
final door choice.}

\item{game}{Character vector created by create_game().}
}
\value{
Returns either "WIN" or "LOSE".
}
\description{
Determines whether the contestant wins or loses.
}
\details{
The contestant wins if the final door selection
contains the car.
}
\examples{
game <- create_game()
determine_winner(1, game)

}
