% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg-Wickerham.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine outcome of the game.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{This}{function relies on the arguments final.pick (the door number the
contestant ultimately chose) and game, which contains what is behind
each door.}
}
\value{
This function will return a character value, either WIN or LOSE.
}
\description{
'determine_winner' let's us know if the contestant won or loss. To win,
the final pick must be the car; otherwise the contestant loss.
}
\details{
The game concludes by revealing what was behind the door of the final pick.
}
\examples{
    this.game <- c("goat","car","goat")
    determine_winner( final.pick=1, game=this.game )
}
