% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{Will provide the results of the game as either a win or loss of the game for
the contestant.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{`final.pick`}{Numeric integer}
}
\value{
A character string or "WIN" or "LOSE"
}
\description{
\code{determine_winner} will indicate what the contestants final door choice
resulted in - a win or a loss.
}
\details{
Checks whether the final value of the contestants choice is "car".
}
\examples{
determine_winner("car")
determine_winner("goat")
}
