% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem Script.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine winner}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{Numeric, Player's final pick door between 1&3}

\item{game}{Character, Car or a Goat}
}
\value{
Return Character, "WIN" or "Lose"
}
\description{
This function determine if the player wins or lose.
}
\details{
After the player decided his final selection, whether he wants to stay on
his first selection or choose to change, this function determine if he win
a car or lose it.
}
\examples{
final.pick = 1
game <- create_game()
determine_winner <- function( final.pick, game )

}
