% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine Winner}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{Two}{parameters are required: final.pick and game}
}
\value{
This function returns a string indicating if a contestant has won or lost.
}
\description{
This function evaluates if the final door picked is a goat or a car.
}
\details{
The value of final.pick is evaluated against a string value of "car"
  or "goat". If final.picl matches "car" the contestant has won.
}
\examples{
  determine_winner( final.pick, game )

}
