% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine if the contestant has won.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{Numeric vectors, length 1, between 1 and 3.}

\item{game}{character vector, length 3, any combination of one "car" and
two "goat".}
}
\value{
The function returns a length 1 character vector
indicating the game result.
}
\description{
\code{determine_winner()} returns the "WIN" or "LOSE" game result.
}
\details{
If the contestant's final selection contains the car, they win.
The function will return "WIN"; otherwise they lost/ if a goat
behind their final door selection. The function will return "LOSE"
for this scenario.
}
\examples{
  determine_winner( 1, c("goat","goat","car") )
  determine_winner( 2, c("goat","goat","car") )
  determine_winner( 3, c("goat","goat","car") )
  determine_winner( 1, c("car","goat","goat") )
  determine_winner( 2, c("car","goat","goat") )
  determine_winner( 3, c("car","goat","goat") )
  determine_winner( 1, c("goat","car","goat") )
  determine_winner( 2, c("goat","car","goat") )
  determine_winner( 3, c("goat","car","goat") )

}
