% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine If Contestant Won the Game}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{...}{This function has two arguments: final.pick and game
The argument final.pick is the argument that indicates the
ultimate selection of door by the contestant. The argument game
is the argument that indicates the current game that is being
played}
}
\value{
The function returns a character vector of "lost" if the
\code{final.pick} reflects a door with a "goat", or a character
vector of "won" if the \code{final.pick} reflects a door with
the "car".
}
\description{
\code{determine_winner()} generates the results of whether
the contestant has won or lost the game based on the
'final.pick'
}
\details{
The contestant will be determined a winner of the game if
the final door selected by the contestant shows to have
the car behind it. If the contestant makes a final selection
of a door that has a goat behind it, the contestant
will have lost the game.
}
\examples{
  determine_winner( final.pick, game )

}
