% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine winner of the Monty Hall Game.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{An integer that represents the contestant's final door
choice (1, 2, or 3).}

\item{game}{A vector of length 3 is passed into the function that represents
the placement of the car and goats behind the 3 doors.}
}
\value{
A string indicating the result: \code{"WIN"} if the contestant picks the car,
or \code{"LOSE"} if they pick a goat.
}
\description{
\code{determine_winner()} checks if the contestant's final pick results in a win
(car) or a loss (goat) in the Monty Hall game.
}
\details{
This function takes the final door choice of the contestant and checks
whether the prize behind that door is a car or a goat. If the contestant's
final pick is a car, they win; otherwise, they lose.
}
\examples{
  game <- c("car", "goat", "goat")
  final.pick <- 1
  determine_winner(final.pick, game)
}
