% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine if the contestant won.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick:}{A numeric value (1, 2, or 3) representing the contestant’s final chosen door.
game: A character vector of length 3 created by create_game(), containing two "goat" values and one "car".}
}
\value{
A character string, either "WIN" if the contestant selected the car
or "LOSE" if they selected a goat.
}
\description{
determine_winner() checks whether the contestant’s final door choice
contains the car or a goat.
}
\details{
The function evaluates the contestant's final door selection against
the hidden game setup.
}
\examples{
  game <- create_game()
  final.pick <- 2
  determine_winner(final.pick, game)

}
