% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{DETERMINE IF CONTESTANT HAS WON}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{'final.pick'}{is the final door that the contestant chose (1,2, or 3)
'game' shows the status of each door whether it is car or goat.}
}
\value{
This indicates the result of the game. "WIN" if contestant chooses the car
and "LOSE" if contestant chooses the goat.
}
\description{
This will determine if the contestant won or lost based on the
contestant's choice.
}
\details{
This function checks if the final door choice in the game has the
contestant has won (chose the car) or lost (chose a goat).
}
\examples{
determine_winner(final.pick = 1, game = c("car", "goat", "goat"))
}
