% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Montyhall.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine if the contestant has won the game or not}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{The}{game[final.pick] is a character vector because it contains character values, as the contestant
only has to either choose a "car" door or a "goat" door.}
}
\value{
The function returns a length of one character vector indicating whether the contestant has won or not.
}
\description{
determine_winner()is used along with the if()to determine whether the contestant has won the game or not
     based on his final pick
}
\details{
the determine_winner()uses the if() because based on the result we can tell whether the contestant has won
       or not. It uses the game argument, which is a character vector. It states that if the contestant's final
       choice was the car door, so the returned result will be "win". If the the contestant's final choice was a
       a goat door, the returned result will be "lose".
}
\examples{
        this.game <- c("goat","car","goat")
        determine_winner( final.pick=1, game=this.game )
        determine_winner( final.pick=2, game=this.game )
        determine_winner( final.pick=3, game=this.game )
}
