% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_monty-hall-problem (2).R
\name{determine_winner}
\alias{determine_winner}
\title{Determine if the contestant is a winner}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{the}{parameters are, if final pick is goat=loss and if car = win.}
}
\value{
The return is either a loss or a win.
}
\description{
determine_winner function evaluates whether the final pick from the
contestant is a winner by determining if the door held a car or a goat.
}
\details{
This function determines if the final pick was a car or a goat. If it were
a car, then the game is marked as a win. If its a goat, then it returns
as a loss.
}
\examples{
determine_winner ()

}
