% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine Winner}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{...}{arguments used are (final.pick,game)}
}
\value{
The function returns a WIN or LOSE
}
\description{
The function determines if contestant has won or lost depending on whether the contestant
changed his pick or stayed with original pick.
}
\details{
The if function is used to further demonstrate that if contestant's final pick
whether he changes his pick or not is car , he will win and
if contestant's final pick was 'goat' he will lose.
}
\examples{
determine_winner(final.pick, game )


}
