% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem-brentlinger.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine whether the player won the game.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{Arguments}{used by the function are the player's final pick which was
determined in the 'change_door()' function, and the setup of the game
determined in the 'create_game()' function.}
}
\value{
The function returns a character string, either "Win" or "Lose"
depending on the outcome of the game.
}
\description{
The function will determine whether the player wins or loses the game
and returns the results of the game.
}
\details{
If the player's final pick is a car, the player wins the game.  If the
player's final pick is a goat, the player loses the game.
}
\examples{
determine_winner( final.pick, game )

}
