% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine whether the contestant wins}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{The contestant's final door selection.}

\item{game}{The game configuration vector.}
}
\value{
A character string: "WIN" if the car is selected, otherwise "LOSE".
}
\description{
Determines if the contestant wins the game by selecting the car.
}
\details{
A win occurs when the final door choice contains the car.
}
\examples{
game <- c("goat","car","goat")
determine_winner(2, game)

}
