% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{It pays to win}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{Arguments}{related to the determine winner functions are as follows.
First addressing the if statements which have two primary results such
as 'Win' and 'Lose'. Within the Win function game is being called initially
since this game is being associated with this.game in particular and final.
pick that are being called from 'car' which the various combination have been
assigned from. Hence final.pick will either have the option of being a car
or not a car. which is what the second if statement indicates.
variable is the variable that is being called due to the ties that are being}
}
\value{
'Win' or 'Lose'
}
\description{
Details regarding this piece is specifically catered towards the individuals
that play this game. Of course after all the events take place for the
game play, there must be a winner at some point.
}
\details{
Details regarding this piece is exactly the same as the description, but only
this code is specifically looking probabilities of user that are going to
win. Of course, the analysis of such intensity must be evacuate and
considered for the facts may not appear as they seem. It is obvious that the
winner determination is not directly solid as to weather switching or staying
would benefit the individual. But we can assess the likelihood that the
individuals is to win based on probability and statistics.Hence how this
program was born, by being able to numerically explain that probabilities of
success.
}
\examples{
determine_winner()
}
