% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{Determines whether the contestant wins or loses the game.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{Numeric integer}
}
\value{
Character string, "WIN" or "LOSE"
}
\description{
'determine_winner()' identifies whether the contestant's final door
choice reveals a car door and wins or a goat door and loses.
}
\details{
The function checks to see if the final door selection is a car or a goat.
}
\examples{
determine_winner(final.pick = 1, game = c("goat", "car", "goat"))
}
