% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problemb.R
\name{determine_winner}
\alias{determine_winner}
\title{Determines winner.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{...}{two arguments are used by the function.}
}
\value{
The function returns a character vector
indicating the if the contestant won or lost.
}
\description{
\code{determine_winner()} generates an outcome if contestant won a car
or chose a goat.
}
\details{
This function determines whether or not a contestant won the game.
The contestant chooses a car results in win, goat results in loss.
}
\examples{
  determine_winner()
}
