% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen comments.R
\name{determine_winner}
\alias{determine_winner}
\title{Win or Lose Determination}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{The}{parameter final.pick is an integer 1, 2, or 3, that is the final door
number choice the player chooses at the end.
The parameter game is a vector that factors in the current game.}
}
\value{
This function returns a WIN or LOSE determination.
}
\description{
\code{determine_winner()} tells us if the contestant won the car or lost the game.
}
\details{
This step determines whether the player has won or lost based on whether
the player chose to stay with their original pick or switch to the last
remaining door. If a car is behind the final chosen door, the player won. If
there is a goat behind the final chosen door, the player lost the game.
}
\examples{
determine_winner()

}
