% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{determine_winner}
\alias{determine_winner}
\title{Determines if the user has won the game.}
\usage{
determine_winner(final.pick, a.game)
}
\arguments{
\item{This}{function utilizes the arguements 'final.pick' from the 'change_doors()'
function and 'a.game' from the 'create_game()' function.}
}
\value{
This function will return a character vector of either 'WIN' or 'LOSE'
depending on the result of the game.
}
\description{
'determine_winner' will decide if the user has won
the game.
}
\details{
This function decides if the user has won the game
based on their final selected door.  The final selection
must correspond to the door that has a car behind it as indicated
by the 'create_game' function.
}
\examples{
 this.game
 my.initial.pick
 my.final.pick <- change_door( stay=T,
                           opened.door=opened.door,
                           a.pick=my.initial.pick )
 determine_winner( final.pick=my.final.pick,
                 game=this.game )

 my.final.pick <- change_door( stay=F,
                             opened.door=opened.door,
                             a.pick=my.initial.pick )
 determine_winner( final.pick=my.final.pick,
                 game=this.game )
}
