% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOCUMENTATION.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine winner}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{If}{the final picked door has the car, then the contestant wins.
If the final picked door has a goat, then the contestant loses.}
}
\value{
The function returns four vectors. The first is a length 3
character vector indicating the positions of goats and the car.
The second is a length 1 numerical vector indicating the initial
door selected by the contestant. The third is a length 1 character
vector indicating the game outcome if the contestant choose to
stay with their initial selected door. The third is a length 1
character vector indicating the game outcome if the contestant
choose to exchange their initial selected door for the
final unopened door.
}
\description{
\code{determine_winner()}determines the game outcome.
}
\details{
After the contestant has made their final selection, the game outcome is
determined.
}
\examples{
  this.game
  my.initial.pick
  my.final.pick <- change_door( stay=T, opened.door=opened.door, a.pick=my.initial.pick )
  determine_winner( final.pick=my.final.pick, game=this.game )
  my.final.pick <- change_door( stay=F, opened.door=opened.door, a.pick=my.initial.pick )
  determine_winner( final.pick=my.final.pick, game=this.game )

}
