% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{Function to check if the contestant has won.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{Numeric value.}

\item{game}{length 3 character vector indicating the positions of goats and the car.}
}
\value{
a string  of either \strong{"WIN"} in case the contestant has won  or \strong{"LOSE"} in case the contestant has lost.
}
\description{
\code{determine_winner()} Revealing function to check if the contestant has won or lost.
}
\examples{
determine_winner(final.pick, game)

}
