% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem-cable.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine if the contestant has won the game.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{This}{function uses two arguments: The final door that the contestant picked,
and the current game (as represented by the specific arrangements of cars
and goats behind the three doors).}
}
\value{
The function returns a string of characters. Those characters are either
"WIN" or "LOSE".
}
\description{
\code{determine_winner()} looks at the final door that the contestant chose
within the game. If that door is a car door, then it returns the word
"WIN" and the contestant has won the car. If that door is a goat dore,
then it returns the word "LOSE" and the contestant has not won the car.
}
\details{
Now we reveal if the contestant has chosen a car door and won the car
or if they've chosen a goat door instead and lost the game.
}
\examples{
  determine_winner()
  
}
