% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.r
\name{determine_winner}
\alias{determine_winner}
\title{Determine win or loss}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{Integer 1, 2, or 3. The integer represents the contestant's final door choice.}

\item{game}{Character vector of length 3 with two "goat" and one "car".}
}
\value{
A string: "WIN", if the contestant won. A string "LOSE", if the contestant lost.
}
\description{
Check if the contestant's final door choice is the winning car door.
}
\examples{
  g <- create_game()
  p <- select_door()
  opened.door <- open_goat_door(g, p)
  final.pick <- change_door(stay = TRUE, opened.door= opened.door, a.pick = p )
  determine_winner(final.pick, g)
}
