% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{\preformatted{	Winner?
}}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
the function returns a string of either "WIN" or "LOSE"
}
\description{
\preformatted{Determine is contestant chose the door with the car.
}
}
\details{
\preformatted{	If the contestant has the door with the car then they win,
	if they have a door with a goal then they lose. 
}
}
\examples{
	determine_winner()

}
