% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{Did the contestant win?}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{The final door chosen; integer between 1 and 3.}

\item{game}{A vector containing one "car" and two "goat"s.}
}
\value{
The function returns either "WIN" or "LOSE".
}
\description{
\code{determine_winner()} determines whether the contestant won.
}
\details{
The function returns "WIN" if the contestant's final pick
reveals a car and "LOSE" if the contestant's final pick reveals
a goat.
}
\examples{
  x=c("goat","goat","car")
  determine_winner( final.pick=2, game=x )

  y=c("car","goat","goat")
  open_goat_door( 3, y )

}
