% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_winner.R
\name{determine_winner}
\alias{determine_winner}
\title{Determines if the player won based on their final chosen door.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{The final door chosen.}

\item{game}{Random selection of goats and cars.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
States if the player has won or lost based on if they landed on the goat or car in the end.
}
\description{
Determines if the player won based on their final chosen door.
}
\examples{
## Win based on the original problem.
determine_winner(3, c("goat","goat","car"))

## Determines if the random choices of doors with goats and cars has won or lost.
determine_winner(final.pick = 4, create_game(goats = 5, cars = 1))

}
