% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{Simulate the revealing of contestants final door pick
and determine if they win or lose.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{is a numeric integer.}

\item{game}{is a character vector.}
}
\value{
This function returns a character vector.
}
\description{
\code{determine_winner} is uses if functions to determine if
the door the contestant chooses as their final pick is a
winning or losing door.
}
\details{
This funtion performs the fourth and final step in the game,
where the contestants results are determined using if functions
that show win for car and lose for goat.
}
\examples{
determine_winner()

}
