% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem-R-script.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine if the player won or lost.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{...}{The arguments are \code{final.pick} (the door number
of the final door selected) and \code{game} (character vector
of length 3 indicating the locations of goats and car).}
}
\value{
The function returns a character representing the
outcome of the game, i.e., "WIN" or "LOSE", depending on
whether the final door chosen was a car door or not. The
player wins if the final door had a car behind it.
Example output: "WIN"
}
\description{
\code{determine_winner()} Type: Character -> will return "WIN"
or "LOSE" depending on whether the final door chosen
is a car door or not.
}
\details{
This function determines if the contestant has won
or lost, given their final door choice and locations
of the goats and the car.
}
\examples{
  determine_winner(final.pick = finalDoor, game = newGame)
  determine_winner(final.pick = 3, game = c("goat","goat","car"))

}
