% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{Determines if the contestant won the game}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{ a numeric value between 1 and 3 representing the
contestant's final door choice}

\item{game}{ a vector of length three representing the game structure - see
create_game()}
}
\value{
character object "WIN" or "LOSE"
}
\description{
This function compares the contestant's final selection against the original
game structure and returns the final game result
}
\details{
If the contestant's final selection is the door number with the car behind
it, the function will return "WIN", if not the function will return "LOSE"
}
\examples{
determine_winner(final.pick = 1, game = ["goat", "car", "goat"])
[1] "LOSE"
determine_winner(final.pick = 2, game = ["goat", "car", "goat"])
[1] "WIN"
}
