% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg-3.R
\name{determine_winner}
\alias{determine_winner}
\title{\if{html}{\out{<div class="sourceCode">}}\preformatted{This function determines whether or not the player won a car
}\if{html}{\out{</div>}}}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{This}{function takes in the arguments for final.pick, returned in the previous function, and our game variable}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{This function returns results with the string WIN or LOSE
}\if{html}{\out{</div>}}
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{This function checks the final selection that the player makes and identifies if it matches the integer that represented the door behind which sits a car.
}\if{html}{\out{</div>}}
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{This function takes in the final selection variable and the game variable and compares the selection to the variable representing that integer. If and only if it's a car, the player wins. 
}\if{html}{\out{</div>}}
}
\examples{
	determine_winner(final.pick, this.game)
}
