% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine the winner.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{...}{two arguments (final.pick and game(a vector)) are used.}
}
\value{
It returns the string "WIN" or "LOSE".
}
\description{
It etermines if the final pick is the "car" to declare a win, or if it's a "goat" to signify a loss, 
   then return the selection accordingly.
}
\details{
The function checks if the final pick corresponds to "car" to determine if the contestant has won; 
   if so, return the selection. Similarly, check if the final pick matches "goat" to ascertain if 
   the contestant has lost; then, return the selection.
}
\examples{
   this.game
   my.initial.pick
   my.final.pick <- change_door( stay=T, opened.door=opened.door, a.pick=my.initial.pick )
   determine_winner( final.pick=my.final.pick, game=this.game )
   my.final.pick <- change_door( stay=F, opened.door=opened.door, a.pick=my.initial.pick )
   determine_winner( final.pick=my.final.pick, game=this.game )
}
