% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg copy 2.R
\name{determine_winner}
\alias{determine_winner}
\title{This function determines whether or not the player won a car}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{This}{function takes the arguments \code{final.pick} (returned in the previous
function) and \code{game}.}
}
\value{
The function returns the string "WIN" or "LOSE".
}
\description{
This function checks the player's final selection and identifies whether it
matches the door that has the car.
}
\details{
The function takes the final selection and the \code{game} variable and compares
the selection to the prize behind that door. If and only if it's a car, the
player wins.
}
\examples{
  selection <- change_door(stay=F, goat.door, this.pick)
  determine_winner(selection, this.game)

}
