% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem (1).R
\name{determine_winner}
\alias{determine_winner}
\title{Determine if Contestant has Won}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{:}{… choosing a door with a goat behind it will result in a lost for
the contestant. If there is a car behind the chosen door than the contestant
has won the game.}
}
\value{
The result should either be a “Win” or a “Lose”.
}
\description{
If the contestant picked the car door, they win the game.
}
\details{
The winner of the game is determined by whether the contestant chose the
correct door. If the final pick from the contestant is the car door, then
they have won the game. If the contestant’s final pick resulted with a goat
behind the door, then they lost the game and will receive the message “LOSE”
as a result.
}
\examples{
determine_winner(final.pick) = 1, game = this.game)

}
