% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine the outcome (WIN/LOSE)}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{Integer (1:3) final door selection.}

\item{game}{Character vector from \code{create_game()}.}
}
\value{
"WIN" if car; otherwise "LOSE".
}
\description{
Compares the final pick against the game setup.
}
\examples{
  determine_winner(1, c("car","goat","goat"))
}
