% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine if the participant won}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{...}{final pick a vector that will indicate the door that was last selected (final.pick,game)}
}
\value{
The function returns a WIN or LOSE
}
\description{
The function determines if contestant has won or lost depending on the situation of the final selection
.
}
\details{
This function will show the user if he won or lost depending on his final pick.
If door behind the final pick was a car that the player win and if not the player will lose
}
\examples{
determine_winner( final.pick, game )


}
