% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall_script_rj.R
\name{determine_winner}
\alias{determine_winner}
\title{:
Win/Lose Revealed}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{:}{Text is returned as characters.}
}
\value{
:
Depending on whether a car or a goat was selected, the text "WIN" or "LOSE" will be returned.
}
\description{
:
This final step reveals whether the contestant selected the door with the car behind it.
}
\details{
:
The "determine_winner" function is based on the "final.pick" determined in the previous step. It also calls on the objects assigned to each door in the beginning with the "game" assignment. If the door number selected for the final pick is associated with the car, the word "WIN" is returned. If the final pick is associated with a door that has a goat behind it, the word "LOSE" is returned.
}
\examples{
determine_winner(final.pick=2)
}
