% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty_hall_scripts.R
\name{determine_winner}
\alias{determine_winner}
\title{Game outcome is determined.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{game}{A vector containing one "car" and two "goat" strings in any configuration.}
}
\value{
This function returns a string indicating contestant's outcome in the game: "WIN" or "LOSE".
}
\description{
\code{determine_winner()} simulates whether the contestant has won or lost the game.
}
\details{
In the Monty Hall Problem, once the contestant has opened a door and the host has opened
one of the doors with a goat behind it, the contestant must choose to either stay with their
original door selection or switch to the unopened door. If the final selection has a "car"
behind it then the contestant has won. If the final selection has a "goat" behind it, then the
contestant has lost.

#' @param final.pick
The position of the contestant's final choice after either staying or switching, a number 1-3.
}
\examples{
game <- create_game()
a.pick <- select_door()
opened.door <- open_goat_door(game, a.pick)
final.pick <- change_door( stay = T, opened.door, a.pick)
determine_winner(final.pick, game)

}
