% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine Outcome}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{The}{function uses two inputs, final.pick, which is the contestant's final decision on door, and game, from the create_game()function
that will allow it to determine which door has a car behind it.}
}
\value{
It will return either WIN or LOSE.
}
\description{
This function reveals if the contestant won or lost the game.
}
\details{
The function uses two inputs to determine whether the final pick is associated with the car or not. If it is, the function
declares them the winner (WIN), and if not, declares that they have lost (LOSE).
}
\examples{
determine_winner(final.pick, game)
}
