% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine the winner of the Monty Hall game.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{A numeric value representing the contestant's
final door selection.}

\item{game}{A character vector representing the game, with two
goats and one car.}
}
\value{
A character value of either "WIN" or "LOSE".
}
\description{
\code{determine_winner()} determines whether the contestant wins or
loses based on the final door selection and the game setup.
}
\details{
The function compares the contestant's final door selection with
the game. If the selected door contains the car, the function
returns "WIN". If the selected door contains a goat, it returns
"LOSE".
}
\examples{
  game <- create_game()
  a.pick <- select_door()
  opened.door <- open_goat_door(game, a.pick)
  final.pick <- change_door(TRUE, opened.door, a.pick)
  determine_winner(final.pick, game)

}
