% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{Evaluates the result of the game based on the final door selection}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{The final door selection made by the contestant.}

\item{game}{The game vector containing the prize locations.}
}
\value{
A character value indicating the outcome of the game.  The function returns "WIN" if the
contestant selected the car and "LOSE" if the contestant selected a goat.
}
\description{
The function determines the outcome of the game based on the final door selection.
}
\details{
The accepts the final.pick (from change_door) and the game vector (from create_game).
If the contestant correctly selected the car, the function returns "WIN".  If the
contestant selected a goat, the function returns "LOSE".
}
\examples{
 determine_winner( 1, c("goat", "car", "goat"))
Output: "LOSE"

}
