% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine Winner Function}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{This}{function contains two arguments as described in the details. If
\code{final.pick} equals "car", then the player wins. If \code{final.pick}
equals "goat", thenn the player loses.}
}
\value{
This function returns a length one logical character vector. The choices
are "WIN" or "LOSE".
}
\description{
The function \code{determine_winner()}, determines whether the player's final
pick is a goat door or a car door.
}
\details{
If the player's \code{final.pick} is equal to "car" then they win.  If their
\code{final.pick} is equal to "goat" then they lose.
}
\examples{
  determine_winner()

}
