% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{determine_winner}
\alias{determine_winner}
\title{DETERMINE IF CONTESTANT HAS WON.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{if}{arguments are used by the function.}
}
\value{
The function returns if the contestant WIN or LOSE.
}
\description{
\code{determine_winner()} DETERMINE IF CONTESTANT HAS WON.
}
\details{
After the contestant choose his/her final pick whether to Stay
or Switch doors, the final picked door is open and the results
of determining if the contestant Wins or Loses.
}
\examples{
determine_winner()

}
