% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine if contestant has won - Function}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{The}{function "if" will help to run the code depending on the choice of
the contestant. If the contestant chose the door with a car. then the "if"
function will print in the screen "WIN" if the contestant choice was
a door with a goat, the "if" function will run the code printing the "LOSE"
message.}
}
\value{
Depending on the final pick of the contestant, the word "WIN" or "LOSE"
will be displayed.
}
\description{
This function will help to determine if the contestant has chosen the
"car" door, or if the contestant did not. A message will be displayed the
"WIn" or "LOSE" words.
}
\details{
The function will be utilizing the variables "final.pick" and "game"
previously used in another function.
}
\examples{
 determine_winner( )
}
