% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{Did They Win?}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
If the contestant's final.pick is a car, the game returns "WIN". If they
lost, the game returns "LOSE."
}
\description{
The function deciphers whether the contestant's final pick was a win or
loss.
}
\details{
Feeding off the 'change_door()' function, the contestant's final selection
(final.pick), being either doors 1-3 is determined as a win or loss
}
\examples{
determine_winner(final.pick,game)
}
