% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_game.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine the winner of a Monty Hall game}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{The contestant's final door selection.}

\item{game}{A length-three character vector containing
two goats and one car.}
}
\value{
`"WIN"` if the final selection contains the car,
or `"LOSE"` if it contains a goat.
}
\description{
`determine_winner()` determines whether the contestant's
final door contains the car or a goat.
}
\examples{
game <- create_game()
determine_winner(1, game)

}
