% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine If Contestant Has Won}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{The}{determine_winner function calls upon the final.pick and game arguments.
The IF statements evaluate the game function based upon final.pick function.}
}
\value{
The return is an atomic vector
}
\description{
If statements are used to compare results of previous step to determine if the contestant has won
or lost the game.
}
\details{
The final.pick is called from a previous step and put through if statements to return the game
results. If the final pick was a goat, the result is LOSE. If the contestant picks a car, then the
result is a WIN.
}
\examples{
  WIN  LOSE

}
