% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg-finished.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine if the contestant has won or not.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
This function will return 'WIN' if the contestant has chosen a door with a
car behind it and LOSE' if the contestant has chosen a door with a goat
behind it.
}
\description{
\code{determine_winner()} function reveals if the contestant's choice has
resulted in winning or losing the game.
}
\details{
The contestant will win if they have chosen a door with a
car behind it and lose if they have chosen a door with a goat
behind it.
}
\examples{
  determine_winner()

}
