% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{Detemines if the contestant has won or lost the game.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{The}{function takes two arguments.
1. final.pick- contestant final pick (a number between 1-3)
2. game- character vector of length three containing position
of car and goats.}
}
\value{
The function returns a character vector of length one with either
 "Win" or "lose"
}
\description{
The function decides if the contestant has won or lost the
 game depending upon his/her final pick of door.
}
\details{
The function determines the result of the game. If the contestant's
 final pick is car the s/he has won the game. If the final pick is
 goat the s/he has lost the game.
}
\examples{

 Determine_winner( final.pick, game )

}
