% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine if contestant picked the door with the car (win) or the goat (lose)}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{The}{function uses final.pick and game as arguments where final.pick is the end result of the stay or switch function
and game is the originally produced set and order of doors with corresponding prizes (goat or car)}
}
\value{
Returns a string of characters stating "WIN" if the contestant chose the door with the car or or "LOSE" if any
other door was chosen
}
\description{
The function checks if the selected door contained the car, any other situation is a loss.
}
\details{
determine_winner() checks the contestants final selection against the create_game() arrangement to determine
if the selected door contains a goat or a car. If the door with the car was chosen, the contestant wins, if not, they lose.
}
\examples{
game <- create_game()
a.pick <- select_door()
opened.door <- open_goat_door(game, a.pick)
final.pick <- change_door(stay = TRUE, opened.door, a.pick)
determine_winner(final.pick, game)
}
