% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine Win or Lose}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{`final.pick`}{Integer. The number of the player's final choice
of door.}

\item{`game`}{Character vector. The position of each goat and the car.}
}
\value{
Character vector. This will print the word "win" or "lose".
}
\description{
This function determines whether the door chosen is the winning door
or not.
}
\details{
Win if the final choice of door is the car, lose if it is a goat.
}
\examples{
    determine_winner()

}
