% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine if the contestant won.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{a number between 1 and 3 from \code{change_door()}
giving the contestant's final door.}

\item{game}{a length 3 character vector from \code{create_game()} with
"goat", "goat", and "car" in a random order.}
}
\value{
The function returns a length 1 character vector, "WIN" if
the car was behind the final door and "LOSE" if a goat was.
}
\description{
\code{determine_winner()} checks the contestant's final door and
reports whether they won the car or got a goat.
}
\details{
The contestant wins if the car is behind their final door and
loses if a goat is. The same function scores the stay pick and the switch pick.
}
\examples{
  this.game <- create_game()
  my.initial.pick <- select_door()
  opened.door <- open_goat_door( this.game, my.initial.pick )
  my.final.pick <- change_door( stay=T, opened.door, my.initial.pick )

  determine_winner( my.final.pick, this.game )

}
