% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monty Hall.package.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine the game outcome.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{An integer (1-3), the contestant's final door choice.}

\item{game}{A character vector( length 3) with "goat"/"car" values.}
}
\value{
A string: "WIN" if the final pick is the car, "LOSE" otherwise.
}
\description{
\code{detremine_winner()} determines the contestant wins a car or loses.
}
\details{
Compares the contestant's final pick with the game setup.
}
\examples{
g <- create_game()
p <- select_door()
o <-open_goat_door(g, p)
f <- change_door(FALSE, o, p)
determine_winner(f, g)

}
