% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine the Winner}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{The player's final door choice.}

\item{game}{A vector showing what is behind each door.}
}
\value{
"WIN" if the player selected the car or "LOSE" if
the player selected a goat.
}
\description{
Determines whether the player wins or loses.
}
\details{
The player wins if the final door has the car and loses if
the final door has a goat.
}
\examples{
game <- c("goat", "goat", "car")
determine_winner(3, game)

}
