% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{#'   The contestant is either a winner or a loser!}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{...}{final.pick is returned from the change_door function.
game pulls from the initial arrangement of two goats and a car.}
}
\value{
The function returns a length 1 character vector
indicating the final door and game pairing.
}
\description{
#'   \code{determine_winner()} takes all previous return statements
and reveals whether the contestant is a winner or a loser.
}
\details{
#'   At this last stage of the game, the contestant has already
been given the choice to stay or change doors. Their final
pick is then measured up against what the code knows to be
the winner or the loser. If the guest's door is paired with
car, then they are a winner of the Monty Hall game.
}
\examples{
  determine_winner(1,c("goat","goat","car"))
  
}
