% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine if the final choice the user picked is a goat or car door.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{\itemize{
\item output from change_door()
game - output from create_game()
}}
}
\value{
Return text to show a win or a loss
}
\description{
This step determines the final choice door and declares the game a win or a loss.
}
\details{
x
}
\examples{
 x
}
