% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem edited.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine Whether the Contesant has Won}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{This}{function has two arguments, \code{final.pick} and \code{game}.
\code{final.pick} is the return of \code{change_door}, a numeric value that is usually stored as an object.
\code{game} is the return of \code{create_game}, a vector of character values, also stored as an object.}
}
\value{
The return is a character, either \code{WIN} or \code{LOSE}.
}
\description{
This function checks whether the contestant's final door choice hides the car.
}
\details{
#' The function reveals whether the contestant chose the door with the car
or a goat. It will match the doors 1, 2, 3 to the \code{game} vector
and then see if \code{final.pick} is the car, the contestant
will be determined a winner. If it is a goat, they will be a loser.
}
\examples{
  determine_winner( my.final.pick, this.game )

}
