% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine a winner}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{`final.pick`}{\itemize{
\item A number representing the contestants final pick.
\code{game} - A vector of strings representing the order of goats and cars.
}}
}
\value{
Returns a string of either "Win" or "lose".
}
\description{
\code{determine_winner} will show if the contestant will "WIN" or "LOSE"
}
\details{
In any game show the contestant will either "WIN" or "LOSE". If the final pick contains a "goat" the contestant will Lose, if it is a "car" the contestant is a winner.
}
\examples{
determine_winner(my.final.pick, new.game)
}
