% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{Test}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{...}{no arguments are Test.}
}
\value{
The function returns a Test
}
\description{
Test
}
\details{
Test
  Test
  Test
}
\examples{
  Test()

}
