% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{Reveal}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{The}{initial door saetup reveals which door the contestant has chosen to pick and either stick with or change to the remaining mysterious door}
}
\value{
: WIN/LOSE
}
\description{
The host reveals the door the contestant has picked if it is a car or a goat
}
\details{
The car choice is the winning choice, and the goat is the loosing choice
}
\examples{
: WIN  
}
