% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine if player win or lose}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{the}{first set created by the first function creat_game()}
}
\value{
return win or lose
}
\description{
depends on the player final selection
}
\details{
If the final selection is the car --> "win"
if final selection goat --> "lose"
}
\examples{
  determine_winner(1, (car, goat, goat)) --> win
  determine_winner(2, (car, goat, goat)) --> lose
  
}
