% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.r
\name{determine_winner}
\alias{determine_winner}
\title{Decides whether the contestant has won or lost.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{is the numeric value from the output of `change_door()`
game is the 3 character vector from `create_game()`}
}
\value{
The function returns a string outcome "WIN" or "LOSE"
}
\description{
`determine_winner()` checks if the final pick was a car
  or a goat.
}
\details{
The contestant is given a "WIN" outcome if final pick is
  a car. The contestant is given a "LOSE" outcome if final pick is
  a goat.
}
\examples{
  my.final.pick <- change_door(stay=F, opened.door=opened.door,
  a.pick=my.initial.pick )
  determine_winner( final.pick=my.final.pick, game=this.game )

}
