% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{Game reveals whether player won or lost.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{This}{function uses the previously named
objects 'final.pick'. The 'final.pick'
pulls the result from the 'change_door()' function.}
}
\value{
The function returns a three or four character
word: "WIN" or "LOSE".
}
\description{
'determine_winner()' reveals whether the
player's final choice has a car or goat
behind the door.
}
\details{
The function recalls the final pick from the
previous function and checks whether the door
has a car or goat behind it. If there is a car
then the game reveals that the player won. If
there is a car, then the game reveals that the
player lost.
}
\examples{
determine_winner()
}
