% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-script.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine Winner}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{None}{}
}
\value{
This is the final step to find out if the contestant is a winner.
}
\description{
Once the contestant sticks with their final choice the computer
will expose which door has the car behind it and if the contestant
is a winner or loser.
}
\details{
The final step of the game.
}
\examples{
determine_winner()

}
