% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{Step 5: Determine if contestant has won}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{...}{two positional arguments}
}
\value{
WIN OR LOSE
}
\description{
determines whether the contestant won or lost
}
\details{
this function takes two positional arguments,
final.pick and game. final.pick is the return value of the
change door call game is the return value of
the create_game call
}
