% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{determine_winner}
\alias{determine_winner}
\title{The status of the game is shown whether the contestant win or lose}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{There}{are arguments used by the function.}
}
\value{
The function returns a character vector.
}
\description{
\code{determine_winner()} There two things involve in this case. Either the
contestant win or lose.
}
\details{
The contestant makes the final picks of the doors. The final pick
determines whether the contestant win the or lose the game.
}
\examples{
  determine_winner()

}
