% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine game result}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{final.pick}{An integer of the door picked by the player}

\item{game}{A character vector of the game setup}
}
\value{
A string: "WIN" if car, "LOSE" if goat
}
\description{
Reveals if the player has won a car or a goat
}
\details{
Checks if the final door picked by the player contains the car
}
\examples{
determine_winner (2, c("goat", "car", "goat"))
}
