% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newscriptmontyhall.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine If Contestant Has Won.}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{...}{the door picked by the contestant.}
}
\value{
the result wich is winner or loser.
}
\description{
determine_winner() generates the result which is whether the contestant has won or lost.
}
\details{
If the contestant final choice was a door with car then he is a winner and if his choice was a door with a goat behind it then he is a loser.
}
\examples{
determine_winner()

}
