% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{getlinedat}
\alias{getlinedat}
\title{Get results data formatted for ggplot2 line plots.}
\usage{
getlinedat(ld)
}
\arguments{
\item{ld}{List of win fractions.}
}
\value{
Formatted dataset.
}
\description{
Takes a list of win fractions and results the data in ggplot2 line plot function-ready format.
}
