% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{getlineplot}
\alias{getlineplot}
\title{Generate a ggplot2 line plot.}
\usage{
getlineplot(ld)
}
\arguments{
\item{ld}{List of win fractions.}
}
\value{
Line plot object.
}
\description{
Generates ggplot2 line plot object from the list of win fraction results.
}
