\name{MontyHall-package}
\alias{MontyHall-package}
\alias{MontyHall}
\docType{package}
\title{
MontyHall problem
}
\description{
This package contains functions for simulation of the Monty Hall problem for different number of doors and games and different strategies.
Plots are included as optional logical argument in functions.
}
\details{
\tabular{ll}{
Package: \tab MontyHall\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-08-22\cr
License: \tab GPL (v3 or later)\cr
}

}
\author{
Stephen Downing

Maintainer: Stephen Downing <sdowning.bm02g@nctu.edu.tw>
}
\references{
See for example the Harvard Stat110 lecture at: \code{\link{https://www.youtube.com/watch?v=fDcjhAKuhqQ}}
}
\keyword{ package }
\seealso{
}
\examples{
MontyHallStrat(games = 1000, doors = 3, plot = T)
}
