% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem (1).R
\name{open_goat_door}
\alias{open_goat_door}
\title{The host opens one of the two goat doors, always revealing a goat.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{...}{Arguments: \code{game} (the name of the game)
and \code{a.pick} (its value is the initial selection door's number)}
}
\value{
The function returns an integer representing the number
of the goat door that the host opened.
}
\description{
\code{open_goat_door()} Type: Integer -> Selects one of the goat doors,
generating a number between 1 and 3.
}
\details{
The host will always open a goat door. This will never be the contestant's
initially selected door, or the car door.
}
\examples{
  open_goat_door(game = newGame, a.pick = firstDoor)
}
