% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{open_goat_door}
\alias{open_goat_door}
\title{HOST OPENS GOAT DOOR}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{Arguments}{used that if else statement, depending on the door the contestant opened}
}
\value{
One door should be revealed, this should be the one with a goat behind it.
}
\description{
The Host will want to open one door, not the contestant door selected. Out of the two doors left
the Host will have to choose a GOAT door to open. If the contestant selected the car door, then host
can open either other door. If the contestant selected a goat door, host has to open the other goat door.
}
\details{
open_goat_door <- function( game, a.pick )
}
\examples{
return( opened.door ) 
}
