% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{open_goat_door}
\alias{open_goat_door}
\title{The host open a door with goat.}
\usage{
open_goat_door(game, a.pick)
}
\value{
Return a door apart from the picked door and definitely
with goat behind
}
\description{
After the participants choosed a door, the host open another door
with goat to help the participants.
}
\details{
If the participants picked the door with car behind, then host
opened another door with goat. If the participants picked the
door with goat behind, then host opened the other goat door.
}
