% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Open a Goat Door}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{A length 3 character vector representing the location
of two goats and one car behind the three doors.}

\item{a.pick}{A numeric value indicating the contestant's initial
door selection. The value should be 1, 2, or 3.}
}
\value{
A numeric value indicating the goat door opened by the
host. The returned value will be 1, 2, or 3.
}
\description{
\code{open_goat_door()} selects a door containing a goat for the
host to open after the contestant makes an initial selection.
}
\details{
If the contestant initially selects the car, the function
randomly selects one of the two goat doors. If the contestant
initially selects a goat, the function opens the other goat
door so that neither the car nor the contestant's selected
door is opened.
}
\examples{
  game <- c("goat", "car", "goat")
  open_goat_door(game, 1)

}
