% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{open_goat_door}
\alias{open_goat_door}
\title{\preformatted{ Open a Goat Door
}}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{arguments}{in this function are:
game: randomly assigns goat, goat, and car to doors 1,2,&3
a.pick: randomly assigned integer of 1,2, or 3 as
the contestant's first pick}
}
\value{
\preformatted{ goat_door will return a number that is not assigned
 to the car or to the contestant. the number will reveal
 a goat.

 the number will be placed in goat.door
}
}
\description{
\preformatted{ opens a door not selected by the contestant
 or a car to reveal a goat. This leaves 2 doors
 unopened
}
}
\details{
\preformatted{ this function first checks if the contestant's
 pick is a car by using an IF statement.
 If the contestant has selected the car then the
 function will randomly select one of the remaining
 doors to reveal a goat.

 If the contestant has selected one of the two doors
 with a goat, the function places a number (1,2,or3)
 that does not contain car or the contestant's pick
 into opened.door
}
}
\examples{
     returns one of the following:
     [1] 1
     [1] 2
     [1] 3
}
