% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Step 3: The Host Opens a Goat Door}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game:}{a vector with the elements "goat", "goat", "car" in random order;
the return value of \code{select_door}
a.pick: an integer between 1 and 3; the return value of \code{select_door}}
}
\value{
... an integer between 1 and 3
}
\description{
\code{open_goat_door(game, a.pick)} always reveals a goat behind
one of the doors that the contestant has not picked.
}
\details{
If the contestant picked a door with the car,
\code{open_goat_door(game, a.pick)} (the host) selects
any of the two remaining doors;
otherwise, the host selects the only remaining door with a goat.
}
\examples{
  game
  [1] "car"  "goat" "goat"
  a.pick <- select_door()
  a.pick
  [1] 2
  opened.door <- open_goat_door( game, a.pick )
  opened.door
  [1] 3

}
