% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem-Gonzalez.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Open a door and expose a goat.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{The}{variables 'game' and 'a.pick' are inserted into the
'open_goat_door' function.}
}
\value{
The function returns a length 1 numeric vector to indicate
a door with a goat behind it. It will either be 1, 2, or 3.
The chosen door will not be the same door from 'select_door' function.
}
\description{
'open_goat_door' function reveals that a goat is behind one of
the doors, but not the previous door chosen from 'select_door' function.
}
\details{
After the contestant chooses a door, now the host of the
game will expose a goat behind a different door. The 3
length character variable 'game' holds the randomized
order of 2 goats and 1 car. 'open_goat_door' function needs 2 inputs
to know which doors not to open and which one to reveal. The function
takes in the first choice 'a.pick' and makes sure to
not expose that door, even if there is a goat behind the
contestant's choice. Now, with the goat door exposed, the
contestant has knowledge about what is behind 1 of the doors.
}
\examples{
open_goat_door( game, a.pick)
open_goat_door( game= '"goat", "car", "goat"', a.pick = 1)
open_goat_door( '"car", "goat", "goat"' , 2)
open_goat_door( '"goat", "goat", "car"' , 3)

}
