% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Open a goat door.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{...}{the arguments game and a.pick are used by the function.}
}
\value{
The function returns a number between 1 and 3 that is not the car door
and not the first door that was selected.
}
\description{
\code{open_goat_door()} opens one of the 2 goat doors.
}
\details{
If the contestant first chose a goat door, then there is only on door
available to open.  If the contestant first chose the car door, then any of
the 2 goat doors can be opened.
}
\examples{
  open_goat_door(game, a.pick = 2)

}
