% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Open a goat door given the contestant's pick.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{A character vector returned by \code{create_game()}.}

\item{a.pick}{An integer (1–3) indicating the contestant’s first pick.}
}
\value{
An integer (1–3) corresponding to the door that is opened.
}
\description{
\code{open_goat_door()} reveals one door with a goat, according
to the rules of the Monty Hall problem.
}
\details{
If the contestant initially picked the car, the function
randomly selects one of the two goat doors to open.
If the contestant initially picked a goat, the function
opens the other goat door (not the contestant’s pick).
}
\examples{
  game <- create_game()
  pick <- select_door()
  open_goat_door(game, pick)

}
