% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem (4).R
\name{open_goat_door}
\alias{open_goat_door}
\title{Host opens Goat Door}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{...}{If the contestant selected car, randomly select one of the two
goats, if contestant selected goat, select other door containing a goat.}
}
\value{
The function returns a length 3 character vector
indicating a number between 1 and 3, that will indicate the door opened by
the host revealing a goat.
}
\description{
\code{open_goat_door( game, a.pick )} generates a selection made from the host,
will always open a door containing a goat behind it. This selection cannot
be the door the contestant initially selected.
}
\details{
When the host opens a door revealing a goat, it cannot be a door the
contestant selected, nor can it be the door containing the car. If the
contestant selected the car in his/her initial selection, the host can open
either of the other doors containing goats.
}
\examples{
  open_goat_door( game, a.pick )

}
