% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Open a Goat Door}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{A character vector of length 3 generated by \code{create_game()}, representing the positions of the car and goats behind the doors.}

\item{a.pick}{An integer (1, 2, or 3) representing the contestant's selected door.}
}
\value{
The function returns an integer (1, 2, or 3) representing the door that the host opens to reveal a goat.
}
\description{
This function simulates the host opening one of the doors with a goat behind it, after the contestant has made their initial choice.
}
\details{
If the contestant has chosen the car, the function randomly selects one of the two remaining doors that hide a goat. If the contestant chooses a goat, the host opens the other door that hides a goat.
}
\examples{
game <- create_game()
a.pick <- select_door()
open_goat_door(game, a.pick)

}
