% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{open_goat_door}
\alias{open_goat_door}
\title{The host opens a goat door.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns the number of the goat door opened by the host.
}
\description{
\code{open_goat_door()} simulates the game show host opening a goat door after the contestant makes their initial pick.
}
\details{
The principal of the Monty Hall game is knowing the host will select a goat door which influences whether the contestant should stay with their initial pick or switch doors.
}
\examples{
  open_goat_door()

}
