% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Reveal a goat door.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{A character vector created by `create_game()` representing
the hidden prizes.}

\item{a.pick}{An integer (1, 2, or 3) indicating the contestant’s initial choice.}
}
\value{
An integer 1, 2, or 3 indicating the opened goat door.
}
\description{
`open_goat_door()` returns a door that the host opens
   to reveal a goat.
}
\details{
If the contestant initially selected the car, then the host chooses
   at random one of the two goat doors. If the contestant initially
   selected a goat, the host opens the other goat door.
}
\examples{
  gg <- create_game()
  choice <- select_door()
  open_goat_door(gg, choice)

}
