% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem(bh).R
\name{open_goat_door}
\alias{open_goat_door}
\title{Host opens goat door.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a numeric vector (1, 2, or, 3) that
is different from the one returned by 'select_door()' (a.pick),
and is also not a "car" door. The return is saved as 'opened.door'.
}
\description{
'open_goat_door()' selects a "goat" door not selected by the
contestant
}
\details{
If the contestant's first selection was a "car" door, then the
'open_goat_door()' function randomly selects one of the two
remaining "goat" doors to represent the hosts selection. If the
contestant's first selection was a "goat" door, the function selects
the last remaining "goat" door. The host caanot select a "car" door
or a door that has already been selected by the contestant.
}
\examples{
  open_goat_door( game, a.pick)

}
