% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Opens one of the two un-chosen doors that has a goat behind it.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{vector of length three containing strings "goat" and "car"}

\item{a.pick}{a positive integer between 1 and 3}
}
\value{
an integer between 1 and 3
}
\description{
\code{open_goat_door()} After selecting one of the three doors in the
game, the host reveals a goat behind one of the two un-chosen doors.
}
\details{
This function generates the index value of the un-chosen door
behind which lies a goat. It will not reveal the door with
the car behind it. In the case the initial door value was the car,
the function randomly chooses between the other two doors to
reveal. If the initial door had a goat behind it, then the
function automatically chooses the only remaining door with
a goat.
}
\examples{
  open_goat_door( c("goat", "car", "goat") , 3 )

}
