% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Open 1 of 3 doors.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{Uses}{2 arguments. 'game' where "car" indicates a car and "goat" indicates a
goat and 'a.pick' which is the index of the door that a player picks.}
}
\value{
Function returns the index of the opened door, which is an integer between 1
and 3.
}
\description{
open_goat_door() selects one of the three available doors.
}
\details{
The function will reveal if there is a goat or car behind the door.
}
\examples{
 opened.door <- open_goat_door(this.game, my.initial.pick)
}
