% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Host opens unselected door with goat behind it}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{A array of strings representing the current game.}

\item{a.pick}{A number representing the picked door.}
}
\value{
A number representing a door with a goat behind it that is not the contestant picked door
}
\description{
\code{open_goat_door()} host opens door and reveals a goat.
}
\details{
Host opens one of the two unselected doors revealing a goat.
}
\examples{
  open_goat_door( game, a.pick )
}
