% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Host opens a door with a goat behind it in the Monty Hall Problem Game.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{The}{arguments 'game' and 'a.pick' are used by the function.}
}
\value{
The function returns a number indicating the door that the host
opens. What is behind the door is not indicated, although the person
running the game can determine that it is a goat from the character vector
returned by the first function.
}
\description{
\code{open_goal_door} allows the host of the game to open one of the remaining
doors, not previously selected by the contestant, to reveal a goat.
}
\details{
This is the portion of the game where the host opens one of the remaining
doors, which will have a goat behind it, which sets the contestant up to
make the choice between staying with the door they chose in the previous
step, or changing to the other remaining unopened door.
}
\examples{
  open_goat_door(game, a.pick)
  
}
