% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{open_goat_door}
\alias{open_goat_door}
\title{The host receives the door selection from contestant and opens a different door.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{no}{param}
}
\value{
Returns a numeric value between 1 and 3 that correlates with a door with a goat behind it.
}
\description{
\code{open_goat_door()} Makes it so the host will always open a door different from the door
selected by the contestant and the door the host opens will always have a goat behind it.
}
\details{
The host opens one of the doors that the contestant did not select and the host opens will
always have a goat behind it. The contestant will be asked to stick with the original pick
or change the door the contestant selected.
}
\examples{
open_goat_door() selection()=1 therefore, host selects a door with a goat behind it from
the doors 2 or 3.
}
