% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{open_goat_door}
\alias{open_goat_door}
\title{\preformatted{Open the Goat Door.
}}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{...}{the arguments are \code{game} and \code{a.pick}}
}
\value{
\preformatted{This returns a number representing the number of 
the goat door that the host opened.
}
}
\description{
\preformatted{`open_goat_door()` opens a door with a goat behind
it and that is also not the same door the contestant 
chose.
}
}
\details{
\preformatted{The opened door must have a goat behind it, 
meaning that the door must not have a car behind it.
}
}
