% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Host opens a goat door}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{A character vector representing the contents of the doors in the game}

\item{a.pick}{An integer representing the door selected by the constestant}
}
\value{
A numeric vector representing the doors that have been opened
}
\description{
This function opens one goat door in a Monty Hall game
}
