% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Host opens a door with goat behind it.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{This}{function takes parameters game- the output of create_game()-and a.pick the output of select_door()}
}
\value{
This function returns a number corresponding to the door that was opened.
}
\description{
This function randomly selects from the doors with goats behind them as determined by the create_game() function
}
\details{
only "goat" doors can be opened.
}
\examples{
opened_goat_door(my.game, first.pick)

}
