% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Host Opens Goat Door}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{establishes the random scenario that create_game() created: sample( x = c("goat","goat","car"), size=3, replace=F )}

\item{a.pick}{integer (1,2,3) that aligns with the door that the contestant chose first}
}
\value{
number between 1 and 3
}
\description{
\code{open_goat_door()} randomly returns a value for one of the unopened, unpicked doors that has a goat behind it
}
\details{
starts by pulling data from arguments: \code{game} and \code{a.pick}. creates a doors vector with 3 doors. creates a logical argument where if \code{a.pick} argument was car, it will set \code{goat.doors} as the two remaining unopened doors and randomly choose one of them. if \code{a.pick} was a goat, then it sets the opened door as the remaining unopened door that does not have a car behind it. lastly, it returns the value for \code{opened.door} (1-3) as the result of the logical argument.
}
\examples{
  this.game <- c("goat","car","goat")
  my.initial.pick <- 1
  open_goat_door( this.game, my.initial.pick ) # should be 3
  
  #'   this.game <- c("goat","car","goat")
  my.initial.pick <- 2
  open_goat_door( this.game, my.initial.pick ) # should be 1 or 3
  
  #'   this.game <- c("goat","car","goat")
  my.initial.pick <- 3
  open_goat_door( this.game, my.initial.pick ) # should be 1
  
}
