% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Host selects one of the goat doors.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{Vector with the doors made by create_game().}

\item{a.pick}{is the door selected by the contestant.}
}
\value{
The function returns a single door value, 1-3
}
\description{
\code{open_goat_door()} selects one of the other goat doors as
the host. This door is always a goat door, but is never the same
door that the contestant selects.
}
\details{
The game setup replicates the game on the TV show "Let's
Make a Deal" where there are three doors for a contestant
to choose from, one of which has a car behind it and two
have goats. The contestant selects a door, then the host
opens a door to reveal a goat, and then the contestant is
given an opportunity to stay with their original selection
or switch to the other unopened door. There was a famous
debate about whether it was optimal to stay or switch when
given the option to switch, so this simulation was created
to test both strategies.
}
\examples{
  newgame <- create_game()
  my.initial.pick <- select_door()
  open_goat_door(newgame,my.initial.pick)
}
