% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Host open goat door}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{one}{character vector, one numeric vector
args = current game, chosen door}
}
\value{
number of door that host opened
}
\description{
This function opens a door that the contestant did NOT pick and that has a goat behind it
}
\examples{
open_goat_door(game, a.pick)
}
