% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Game Host opens a confirmed Goat Door.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{Parameter}{game is the current game vector selected by 'create_game()'.
Parameter a.pick is integer that is the contestant's original choice of a
door from 'select_door()'.}
}
\value{
Returns length 1 vector of a number 1, 2, or 3 representing a door with a
goat that is not a.pick.
}
\description{
\code{open_goat_door()} picks a second door known to have a goat behind it
from the remaining two doors not selected previously from \code{select_door()}.
}
\details{
This function has the game Host reveal a door with a goat behind it that the
contestant has not chosen.This door is a game loosing door and provokes the
player to next make a decision of the winning door with this new knowledge of
knowing a game loosing door choice. The game Contestant now has a chosen
mysterious door, a known loosing door, and a third mysterious door that they
may soon choose instead.
}
\examples{
open_goat_door()

}
