% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Open goat door.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{A vector of length 3 representing the contents behind the doors (either "car" or "goat").}

\item{a.pick}{A single integer between 1 and 3 representing the player's initial door selection.}
}
\value{
A single integer between 1 and 3 representing the door number that is opened.
}
\description{
open_goat_door function opens a door with a goat behind it.
}
\details{
The function simulates the game and returns the door number of the door that is opened based on the player's selection and the door with the car behind it.
}
\examples{
  open_goat_door(game, 1)

}
