% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{open_goat_door}
\alias{open_goat_door}
\title{open goat door}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{if}{contestant selected car,
randomly select one of two goats}
}
\value{
The function returns a length 3 character vector
indicating the positions of goats and the car.
}
\description{
The contestant opens a goat door which is one of the 2 available doors
}
\details{
The contestant selects a door, then the host
opens a door to reveal a goat, and then the contestant is
given an opportunity to stay with their original selection
or switch to the other unopened door. There was a famous
debate about whether it was optimal to stay or switch when
given the option to switch, so this simulation was created
to test both strategies.
}
\examples{

}
