% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Step 03: Host opens 'Goat' door.}
\usage{
open_goat_door(a.game, a.pick)
}
\arguments{
\item{Character, }{Numeric}
}
\value{
The function returns a number length one numeric between 1 and 3.
}
\description{
'open_goat_door()' selects one of the remaining two doors that has a goat behind it.
}
\details{
This function simulates how the host will open a door with a goat behind it once the contestant makes the initial selection.
The contestant, however, does not get to open the door of the initial selection at this point.
}
\examples{
Try running 'open_goat_door()' at least three times to ensure that return is random.
open_goat_door()
##CODE CHUNK 3 UNIT TEST

this.game <- create_game()

this.game

my.initial.pick <- select_door()

my.initial.pick

open_goat_door( this.game, my.initial.pick)

open_goat_door( this.game, my.initial.pick)

open_goat_door( this.game, my.initial.pick)
}
