% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HolfordMeetsMonty.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Open a Goat Door}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{This}{function requires two arguments: "game" (a 3 length character vector)
and "a.pick" (a single integer 1:3). These arguments come from the two previous functions and define
\enumerate{
\item Which doors have goats behind them and can be opened
\item Which door the contestant has picked and cannot be opened
}}
}
\value{
This function returns a single integer that indicates the position of a goat
not selected by the contestant.
}
\description{
\code{open_goat_door()} gathers information from the two previous function and
identifies one "goat door" to reveal to the contestant, reducing the number
of doors the car could be behind to two.
}
\details{
Now that the contestant has picked a door, the host reveals one of the doors
that does not hold the car. This reduces the number of doors that the
car could be behind to two: either the door the contestant already picked
or the last remaining door. This will set up the next choice presented to the
contestant in the game and represented in the next function.

This function uses an if function allowing the game to to know which door
to open based on both the location of the car, and the location of the first pick.
}
\examples{
open_goat_door(game = c("goat", "goat", "car"), a.pick = 3)
open_goat_door(game = c("goat", "car", "goat"), a.pick= sample(c(1:3), size = 1))
open_goat_door(game, a.pick)

}
