% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Host opens goat door.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{and a.pick ... these are default arguments from the previous function
`select_door` to run the function correctly.}
}
\value{
The function returns a number (numeric vector) between 1 and 3
indicating the position of the opened door by the host.
}
\description{
`open_goat_door()` this funcction for the host to open
a door with a goat behind it.
}
\details{
After randomly selecting one of the three doors, this function gives
the host the ability to open a door with a goat behind it.
Under the condition that it can㤼㸲t be the door that contestant
has already selected from the previous function.
So this door must not be a car AND the current selection of the
contestant. However, if the contestant selects the car on his/her
first selection the host now can open either door and
if the contestant selects a door of goat,
the host only has one door to open.
}
\examples{
open_goat_door( game = T, a.pick = T )
}
