% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monty Hall Script.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Open a door that is not the player's door or car door}
\usage{
open_goat_door(a.game, a.pick)
}
\arguments{
\item{a.game}{A character vector of length three}

\item{a.pick}{A numeric vector of length one}
}
\value{
This function returns a numeric vector that is the position of the remaining door
}
\description{
This function first determines which goat door should be opened. The door is neither the car door or
the player's door.
}
\details{
This function first determines whether the player chose the car door. If this is true, then a random
goat door is selected. if the player did not chose the car door, then the door that is not the player's
initial choice and the door that does not contain the car are excluded from the remaining doors. The remaining
door is then opened.
}
\examples{
open_goat_door(a.game = c("goat", "car", "goat"), a.pick = 1)

open_goat_door(a.game = c("goat", "car", "goat"), a.pick = 2)

}
