% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Monty Hall Goat Door Reveal}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{...}{Requires the game (array) from create_game() and the initial
door selection (number, a.pick) from select_door() to be included.}
}
\value{
The function returns A number between 1 and 3 representing the
placement in the game array that the opened door object was stored.
}
\description{
\code{open_goat_door()} represents the stage of the game where the game show host
reveals one of the doors the participant did not choose. This door is
guaranteed to not hold the grand prize.
}
\details{
Randomly opens a door provided that the
door doesn't hold the car or was initially selected in select_door().
}
\examples{
open_goat_door(game, a.pick)

}
