% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH.R
\name{open_goat_door}
\alias{open_goat_door}
\title{A confirmed goat door is opened.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{The}{parameter game is a vector that factors in the current game.
The parameter a.pick is an integer that represents the door number that
the player picked first.}
}
\value{
This function returns a number between 1 and 3 which coincides with
a goat behind the door.
}
\description{
\code{open_goat_door()} picks a second door that definitely has a goat behind it
from the remaining two that were not selected from the \code{select_door()}
function.
}
\details{
This function reveals a goat for the contestant. The game host opens one of
the remaining two doors that has a goat behind it.
}
\examples{
open_goat_door()

}
