% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPackageProject_Roach.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Open a goat door.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{Requires}{two specified arguments, one for the game and one for the initial door selection.}
}
\value{
The function returns a length of 1 numeric vector indicating the goat door opened.
}
\description{
'open_goat_door' selects a losing door hiding a goat to be opened after the
contestant's initial door selection.
}
\details{
Creates a variable storing the control flow function with two arguments for
the first part of the game. In this part, the initial door selection has been
made, and a "losing" door hiding a goat will be opened. Therefore, if the
initial door selection hides a goat, the only other goat door must be opened.
However, if the initial door selection hides a car, than either of the remaining
doors may be opened since they hide goats. This function determines which other
door to open based on the initial selection and returns the number of the door
to be opened.
}
\examples{
open_goat_door( game, a.pick)
}
