% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Open Goat Door}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{include}{the game and pick from prior function}
}
\value{
The function will return the number of the first opened door.# if contestant selected car,# number between 1 and 3
}
\description{
the administrator opens up the door that is a goat door.
}
\details{
After the select door function, the administrator will need to open the next door that is not (!==) a car.
}
\examples{
object <- function (parameters)
}
