% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Host opens a goat door.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a numeric vector (1, 2, or, 3) that
is different from the one returned by 'select_door()' (a.pick).
The return is saved as 'opened.door'.
}
\description{
'open_goat_door()' selects one of the doors the contestant did
not pick.
}
\details{
If the contestant selected a "car" door, the function randomly
selects one of the two remaining doors. Both of these are "goat"
doors. If the contestant selected a "goat" door, the function
selects the single remaining "goat" door. The "car" door cannot
be selected.
}
\examples{
  open_goat_door( game, a.pick)

}
