% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhallproblem.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Host opens goat door in the Monty Hall Problem.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{a.pick}{is the value of the door selected by the contestant}

\item{Doors}{is a numeric vector. Refer to the function 'select_door'}

\item{Game}{is a character vector. Refer to the function 'create_game'}

\item{available.doors}{provides the available goat door options,
eliminating the door with the car behind it and eliminating the door
that the contestant selected.}
}
\value{
The output returns the number of one of the goat doors.
This door is eliminated as an option for the contestant to choose
for the rest of the game.
}
\description{
\code{open_goat_door()} indicates the host's selection of goat door
within the game.
}
\details{
This function returns the number of one of the possible goat doors.
The function works by eliminating the possibility of the host selecting
either the contestant's selected door or the door with a car behind it.
}
\examples{
  this.game <- create_game()
  this.game
  my.initial.pick <- select_door()
  my.initial.pick
  open_goat_door( this.game, my.initial.pick )
   
}
