% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Host opens goat door}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{"Doors"}{A numeric vector with values 1, 2,and 3, representing the three doors.}

\item{"If}{statement" If contestant selected door with a goat, then the opened door is the door that
does not contain the car and is not the contestant's selected door.}
}
\value{
The function returns a number between 1 and 3 indicating which door the host
opens.
}
\description{
\code{open_goat_door()} this function was created to select a door that is not a
car and is not the contestant's current door selection.
}
\details{
First, a vector of three doors numbered 1,2,3 are created. Then two control
"If" statements are utilized in order to make sure that the host will choose
a door with a goat behind it taking into account which door the contestant
has already randomly selected.
Case 1: If the contestant picks the door with the car, the host picks either
of the 2 doors with goats. Case 2: If the contestant pick the door with a
goat, the host picks the door that does not contain the car and is not the
picked door. return which door the host opens.
}
\examples{
open_goat_door()
}
