% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Open a door with a goat.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{A length 3 character vector indicating the positions of the goats and the car.}

\item{a.pick}{A numeric value between 1 and 3 indicating the contestant's initial door selection.}
}
\value{
The function returns a numeric value between 1 and 3 indicating the door that was opened.
}
\description{
'open_goat_door()' selects a door with a goat behind it to open after the contestant makes an initial selection.
}
\details{
If the contestant initially selects the car, the function randomly selects one of the two doors containing goats. If the contestant initially selects a goat, the function opens the other door containing a goat.
}
\examples{
game <- c("goat", "car", "goat")
open_goat_door(game, 1)
}
