% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Host opens a goat door.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{length 3 character vector}

\item{a.pick}{number between 1 and 3}
}
\value{
The function returns a number between 1 and 3
}
\description{
\code{select_door()} selects one of the the goat doors that hasn't been
chosen by the contestant.
}
\details{
This function chooses one of the remaining goat doors. As the contestant has
already chosen one door, this function chooses between 2 goat doors if the
initial pick was the car door, or it chooses the only goat door available if
the contestant originally picks a goat door.
}
\examples{
open_goat_door()

}
