% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhallfunctions.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Host opens a goat door.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a numeric vector of 1, 2, or 3 that
represents the door the host picks and is different from
the one chosen by 'select_door()' (a.pick). The return is
then saved as 'opened.door'.
}
\description{
The function 'open_goat_door()' selects one of the two remaining
doors.
}
\details{
'open_goat_door()' functions in one of two ways. If the contestant
selects one of the two "goat" doors, then the function opens the
remaining "goat" door. But if the contestant selects the "car"
door, the function randomly picks one of two unopened "goat" doors.
}
\examples{
  open_goat_door( game, a.pick )
}
