% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Open a Goat Door.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{A character vector of length 3, representing the placement of the "car" and "goat" behind the doors.
a.pick A numeric value (between 1 and 3) representing the door chosen by the contestant.}
}
\value{
A numeric value representing the door number (between 1 and 3) that the host opens, which reveals a goat.
}
\description{
This function simulates the host opening a door that reveals a goat after the contestant has made their initial door selection in the Monty Hall problem.
}
\details{
Depending on whether the contestant's initial pick is a car or a goat, the function selects a door to open. If the contestant picks a car, one of the two doors hiding goats is opened. If the contestant picks a goat, the other goat door is opened.
}
\examples{
game <- c("car", "goat", "goat") 
random game setup
a.pick <- 1 
contestant selects door 1
open_goat_door(game, a.pick)
}
