% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{open_goat_door}
\alias{open_goat_door}
\title{\itemize{
\item Open Goat Door
}}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{-}{the parametrs we setup consist of if statement whuich capture the possibility of the contestant selecting a door with a car or goat behind it, which influences the hosts decision of opening a certain door. Either way, the opened door vector uses the sample function to ensure we open a goat door.}
}
\value{
\itemize{
\item we are returned with an opened door, which contains a goat
}
}
\description{
\itemize{
\item The Game Host Opens a Goat Door, a door with a Goat behind it.
}
}
\details{
The host will always open a door with a goat behind it. But it can’t be a door the contestant has already selected. So it must be a door that is not a car and not a current contestant selection.Note that if the contestant selects the car on the first guess the host can open either door, but if the contestant selects a goat the host only has one option.
}
\examples{
this.game <- create_game()
this.game
my.initial.pick <- select_door()
my.initial.pick
open_goat_door( this.game, my.initial.pick )
}
