% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Select the goat door.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{This}{function has the details of the created game and the original door selection as paramaters.}
}
\value{
This function returns a number from 1-3 that represents the door with the goat behind it.
}
\description{
\code{open_goat_door()} selects one of the doors with a goat behind it.
}
\details{
Function to select one of the goat doors.
}
\examples{
 open_goat_door()

}
