% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem (1).R
\name{open_goat_door}
\alias{open_goat_door}
\title{Open a Goat Door}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{...arguments}{are (game, a.pick)
game is the one launched
a.pick is the initial pick of the contestant}
}
\value{
The function returns a character vector of length 1
indicating the position of opened door.
}
\description{
\code{open_goat_door()}This function opens one goat door only. This door should be different from the
goat door chosen initially.
}
\details{
The host will always open a goat door. However, it should be different
than the goat door chosen by contestant.
}
\examples{
open_goat_door(game,a.pick)

}
