% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Host opens a door with a goat behind it.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{A length 3 character vector represents the label
of the doors, where one door is labeled "car" and the other
two are labeled "goat".}

\item{a.pick}{An integer that represents the initially chosen door,
a number between 1 and 3.}
}
\value{
The function returns a length 1 numeric vector
indicating the opened door of the host. It will be a
different number than the door from select_door()
}
\description{
\code{open_goat_door()} generates a door with a goat behind it
to be opened.
}
\details{
After a door is chosen, the host must open a door that is
different than the chosen door. Regardless if the player
opened a goat door or a car door, the host must open a door
with a goat behind it, because they cannot open a car door.
}
\examples{
  open_goat_door()
  
}
