% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{open_goat_door}
\alias{open_goat_door}
\title{The host opens a goat door.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{...}{uses argument "game" and "a.pick". "game"
represents what is behind the door, per se. a.pick is
the original door that was picked by the contestant.
the function needs this information for the later logical
statements wrapped within the function. game is used to
ensure there is no "car" behind the door that the host
will open, and a.pick is used to ensure that the pick
originally made by the participant will not be chosen
to be opened either. game is a character vector and
requires character input, whereas a.pick is numeric and
requires numeric input.}
}
\value{
... Returns one numeric output/vector. For example, 1.
}
\description{
This function decides which door the host is allowed
to open. Meaning that it determines which is the car
door and which is the door that the contestant has
already opened and then decides the goat door the host
ought to open based on this determination.
}
\details{
This function (open_goat_door()) represents the host's opening
of a door post-contestant opening a door. The open_goat_door()
function needs information from the "game" (created  first
function) and "a.pick" created in the second function. First,
if the contestant originally picked a car, then the available
doors that can be chosen from cannot be a car. Thus, randomly
sample (using the sample() function) one goat door from the two
remaining goat doors and open it (using the opened.door in
the return()). If the contestant originally selected a goat,
then the opened door cannot be a car and it cannot be the
contestant's original pick (a.pick), thus return the only
remaining door (using the opened.door in the return()).
}
\examples{

this.game <- c("goat","car","goat")
the.initial.pick <- 1
open_goat_door( this.game, the.initial.pick )

the.initial.pick <- 2
open_goat_door( this.game, the.initial.pick )

the.initial.pick <- 3
open_goat_door( this.game, the.initial.pick )

}
