% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Open a door that contains a goat.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{This}{function only passes the \code{create_game()} vector (some
combination of 1 car and 2 goats) as \code{game} and the \code{a.pick}
return from the \code{select_doors()} function. A good way to do
this would be to pass the current game and the selected
door into objects:
this.game <- \code{create_game()}
my.initial.pick <- \code{select_door()}}
}
\value{
This function returns a the door number that isn't the one
picked in \code{a.pick} and does not correspond to a car.
}
\description{
\code{open_goat_door()} takes the door number that is not the one
selected from the \code{select_door()} function and a door that
has a goat as determined by the \code{create_game()} function.
}
\details{
\code{open_goat_door()} is a function that returns the number of a
door that contains a goat behind it. This is part of the game
where the host opens a door that has a goat behind it. The
door that he opens is not the selected door. So the contestant
has a choice of staying with the selected door or switching to
the only other closed door.
}
\examples{
 `open_goat_door(this.game, my.inital.pick)`

 `open_goat_door(c("goat", "goat", "car"), 2)`

}
