% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg (1).R
\name{open_goat_door}
\alias{open_goat_door}
\title{Open a Goat Door}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{A character vector representing the game setup (e.g., the output of \code{create_game()}).}

\item{a.pick}{An integer representing the contestant's initial door selection.}
}
\value{
An integer between 1 and 3 representing the opened door.
}
\description{
Opens a door that has a goat behind it and is not the contestant's initial pick.
}
\details{
The host opens a door that is not the contestant's initial pick and does not contain the car.
}
\examples{
game <- create_game()
pick <- select_door()
open_goat_door(game, pick)
}
