% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_door.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Monty Hall Game Goat Door Reveal}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{A vector of any class and length.}

\item{a.pick}{An integer that is not larger than the length of the \code{game} vector.}
}
\value{
The output is an integer number between, and including, 1 and \code{length(game)}
which is different from \code{a.pick}. There should a "goat" element in the \code{game} vector in
an element different from \code{a.pick}th element.
}
\description{
\code{open_goat_door( game, a.pick )} randomly chooses a goat door
that was not picked by the contestant.
}
\details{
This function does the second step of a Monty Hall game where the host reveals a
door with a goat behind it. Given a game vector and a number for the door selectd by
the contestant, this function returns a door number.
}
\examples{
open_goat_door( c("goat", "goat", "car" ), 3 )
open_goat_door( create_game(5), 11 )

\dontrun{
open_goat_door( c("goat", "car", "car" ), 1 )
}

}
