% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg (2).R
\name{open_goat_door}
\alias{open_goat_door}
\title{open goat door}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{#no}{arguments are used by the function.}
}
\value{
number between 1 and 3
}
\description{
open_goat_door() opens the door
}
\details{
the host opens a door to reveal a goat, and then the contestant is
given an opportunity to stay with their original selection
or switch to the other unopened door
}
\examples{
open_goat_door
}
