% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Open a goat door.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{A length 3 character vector indicating the
positions of goats and the car.}

\item{a.pick}{A number between 1 and 3 indicating the
contestant's selected door.}
}
\value{
The function returns a number between 1 and 3
indicating the opened door.
}
\description{
\code{open_goat_door()} selects a door to open that has a goat
behind it.
}
\details{
The function opens a door with a goat behind it while
avoiding the door originally selected by the contestant.
}
\examples{
  game <- create_game()
  a.pick <- select_door()
  open_goat_door( game, a.pick )

}
