% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall-pkg.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Host opens a goat door.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{Arguments}{for this function are game, which is the current game being
played, as in the current order of the goats and car doors. The other
argument is a.pick, which is the contestants original pick.}
}
\value{
The function returns a numeric variable that correlates to the goat door
opened by the host.
}
\description{
In the game, after the contestant makes a pick, the host opens a door
with a goat behind it. If the contestant originally picked a goat door,
the host only has one option to open. However, if the contestant picked
the door with a car behind it, then the host has to pick one of 2 available
doors that have a goat behind it.
}
\details{
For this step, the function will pick a door that is neither the "car"
door or the door picked by the \code{select_door()} function.
This will always be a goat door.
}
\examples{
  open_goat_door( game, a.pick )

}
