% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Open a door.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{The}{function requires objects \code{game} and \code{a.pick}.}
}
\value{
The function returns an object that is the door which is opened.
}
\description{
\code{open_goat_door} opens the door that is a goat and was not selected by the contestant.
}
\details{
The function returns an object that is a number between 1 and 3 that represents the door that was opened.
}
\examples{
  open_goat_door( game, a.pick )
  
}
