% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Open Goat Door}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{if(}{game\link{ a.pick } == "car" )
goat.doors <- doors\link{ game != "car" }
opened.door <- sample( goat.doors, size=1 )
if( game\link{ a.pick } == "goat" )}
}
\value{
A door 1, 2, or 3. Cannot be the same as the door selected above.
}
\description{
The host opens one of the remaining two doors.
}
\details{
The host opens one of the remaining two doors not selected by
the player in the previous step.The door selected has a goat
behind it.
}
\examples{
  1, 2, or 3. 
}
