% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Host opens a door with a goat behind it.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{`'game`}{and \code{a.pick} are passed into the function.
The \code{game} vector stores the door number and whether there is a
goat or car associated with it. '
\code{a.pick} is a numeric vector (length=1) that stores the contestant's
original door choice (number between 1-3)}
}
\value{
The function returns numeric value representing an
open door with a goat behind it.
}
\description{
\code{open_goat_door()} generates a number between 1 and 3
representing the door which the host opens with a goat behind it
and was not the door originally selected by the contestant.
}
\details{
This function uses two \code{if} statements to first determine
if the contestant chose a door with a car behind it. If the original
selection is a car, then the function randomly select one of the goat doors.
If the contestant's original choice was a goat door, then the other goat door
is returned.
}
\examples{
  open_goat_door( game, a.pick )
  open_goat_door( c("car", "goat", "goat"), 1 )

}
