% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mh.pack.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Opening a Goat Door}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{the}{arguments used in this function are game & a.pick}
}
\value{
The function always returns a goat door but alters based on the
scenario depending on if the player chose a car the first time or a goat.
}
\description{
\code{open_goat_door()} a goat door will always be opened, but it cannot be a
door that the player has already chosen, and it cannot be a door
with a car behind it if the player chose a door with a goat behind it
}
\details{
The host within the game will always open a goat door. If the player chooses
the car on the first guess then the host can open one of the two doors
being either the goat door or the car door. However, if the player chooses
the goat door then the host has no other choice but to open a goat door.
}
\examples{
open_goat_door()

}
