% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_goat_door.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Open Goat Door}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{...}{game, a.pick}
}
\value{
The function returns a length 3 character vector
indicating the positions of goats and the car.
}
\description{
\code{open_goat_door}generates a goat door that is not behind your inital pick.
}
\details{
\preformatted{This opens a goat door but it will be a goat door not already seected and it will never open the car door.
}
}
\examples{
  open_goat_door()
}
