% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Open a door with a goat}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{None, }{no arguments required.}
}
\value{
The function returns an object (door) containing a goat behind it.
}
\description{
A door is opened with a goat behind it.
}
\details{
the function opens a door from 1 of the 2 doors that the contestant did not choose. 
If the initial contestant choice is a car the function will return a door with a 
goat behind it. If the initial contestant choice is a goat the the function will still return a goat.
}
\examples{
open_goat_door()
}
