% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Host Opens Goat Door}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{The}{parameters to this function include the amount of doors and the position
of the cars and goats in the game.}
}
\value{
The function returns a length 1 character vector indicating the position of one
of the goats in the game.
}
\description{
'open_goat_door' function provides the position of one of the doors with a goat.
}
\details{
This step in the game is when the host opens one of the doors that is not a
winning door, reducing the remaining choice of doors for the contestant to
choose from.
}
\examples{
this.game <- c("goat","car","goat")
my.initial.pick <- 1
open_goat_door( this.game, my.initial.pick )

}
