% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Open a goat door.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{A length 3 character vector representing the
locations of the two goats and one car.}

\item{a.pick}{A numeric value between 1 and 3 representing
the contestant's initial door selection.}
}
\value{
The function returns a single numeric value between
  1 and 3 representing the door opened by the host.
}
\description{
`open_goat_door()` selects a door containing a goat
  for the host to open after the contestant makes an
  initial selection.
}
\details{
If the contestant initially selects the car, the host
  randomly opens one of the two doors containing goats.
  If the contestant initially selects a goat, the host
  opens the only remaining goat door that was not selected.
}
\examples{
  game <- create_game()
  pick <- select_door()
  open_goat_door(game, pick)

}
