% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Host opens a Goat door.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{a character vector of length 3, 2 goats and 1 car
(the first argument).}

\item{a.pick}{A number, represents contestant's first pick
(the second argument).}
}
\value{
The function returns an integer indicating the door
position the host chooses to reveal one goat.
}
\description{
\code{open_goat_door()} generates a number that represents the
door with the goat the host opens.
}
\details{
From the TV show, "Let's Make a Deal," the contestant has made
their pick, and the host now has to open a door to reveal a goat.
The contestant is then given the opportunity to stay with their
original selection or switch to the other unopened door.
}
\examples{
  open_goat_door(game = c('car','goat','goat'), a.pick = 3)

}
