% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Open a goat door}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{The game with two goats and one car.}

\item{a.pick}{The contestants first door choice.}
}
\value{
Returns the number of the door that was opened.
}
\description{
Opens one of the doors that has a goat.
}
\details{
The host cannot open the contestants door or the door
with the car. If the contestant picked the car one of
the two goat doors is randomly selected.
}
\examples{
game <- c("car","goat","goat")
open_goat_door(game, 2)

}
