% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Open a Goat Door}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{A character vector representing the game setup.}

\item{a.pick}{The contestant's chosen door number.}
}
\value{
The door number that was opened, revealing a goat.
}
\description{
Simulates the host opening a door to reveal a goat, based on the contestant's choice.
}
\details{
The \code{open_goat_door()} function emulates the host's behavior
in the Monty Hall Problem. When the contestant selects a door
containing a car, the host randomly reveals one of the two
unchosen doors with goats by opening it. Conversely, if the
contestant initially chooses a door with a goat, the host
exposes the other unselected door containing a goat.
}
\examples{
  open_goat_door(game = create_game(), a.pick = 1)

}
