% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem functions Hauer.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Host Opens Goat Door.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{A vector of length 3 representing the contents of the doors
(either "car" or "goat").}

\item{a.pick}{An integer representing the door that the contestant has
selected (a value between 1 and 3).}
}
\value{
This function returns an element (door number) as the output.
}
\description{
The host reveals (opens) a new door with a goat behind it to the contestant.
}
\details{
The host must reveal a door that is not a car and not a current contestant
selection.if the contestant selects the car on the first guess the host can
open either door, but if the contestant selects a goat the host only has one
option.
}
\examples{
open_goat_door()

}
