% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{open_goat_door}
\alias{open_goat_door}
\title{opens a goat door}
\usage{
open_goat_door(nDoor, game, a.pick, case_1 = F)
}
\arguments{
\item{nDoor}{number of doors}

\item{game}{the game array containing randomly
spread 'cars' and 'goats'}

\item{a.pick}{The choosen door}

\item{case_1}{True of False depending on
whether program is running for question 1 or not (default = F)}
}
\value{
the vector of gates opened by the host
}
\description{
'open_goat_door()' simulates host opening a goat door after player
     picks his first door
}
\details{
This function returns the doors opened by the host
     after the player picks a door in first step.
     if the first picked door is a goat door, then choose a goat door
     other than the picked one.
     if the first picked door is a car door, then pick any goat door available
     as well as a goat door from the game vector
}
\examples{
     open_goat_door(5, game, 2, T)
     open_goat_door(7, game, 3, F)
}
