% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{open_goat_door}
\alias{open_goat_door}
\title{The host opens up a door containing a goat after the user picks a door number.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{...}{The function takes in the arguments game and a.pick. Game refers
to any object/variable which stores the values generated from create_game().
The object/variable, a.pick, refers to any object/variable that stores the
value from select_door().}
}
\value{
The function returns a door number containing a goat behind it after it
referencing the remaining doors, the user's choice, and the position of the
doors with their corresponding value of goat or car.
}
\description{
\code{open_goat_door(game, a.pick)} selects a door number, containing a goat, out
of the remaining door numbers available after the user chose a door number.
}
\details{
After the user picks a door number, the host will choose one of the
remaining doors only if the door contains a goat behind it. For example, if
the user chooses a door with a goat behind it, the host only has one door
option to open with a goat behind it. If the user chooses a door with a car
behind it, the host has two options left and can open either door.
}
\examples{
  game <- create_game()
  a.pick <- select_door()
  open_goat_door(game, a.pick)
  
}
