% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Host opens a goat door}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{parameters}{are set around the hosts door selection based on the contestant's first pick.}
}
\value{
'opened.door' selects a goat door based on the conditions of the game.
}
\description{
After the contestant selects a door, the host then opens one door, revealing a goat.
This sets the stage for the next step where the contestant can keep his first pick or change his pick.
}
\details{
#' the function 'open_goat_door' is created, combining the a.game and a.pick functions
if, else statements are created select a goat door depending on the contestant's first pick
if the contestant picks the car door, the host can open one of two remaining doors.
if the contestant picks a goat door, the host must open the one remaining goat door.
}
\examples{
open_goat_door -> 3
}
