% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{open_goat_door}
\alias{open_goat_door}
\title{The goat door.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a length 3 character vector
  indicating the door with one of goats.  If the contestant's first door hides the car,
  Monty can show either of the other doors, however, if the first door selected was a goat,
  Monty can only show the contestant the other goat door.
}
\description{
Monty will now open a door with a goat behind it.
}
\details{
Ever the showman, Monty will show the lucky contestant one of the two goats.
}
\examples{
open_goat_door

}
