% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Host opens a goat door in Monty Hall Problem game.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{`game`}{is a length 3 character vector of 3 game items that show two
goats and one car in any order. These represent what's behind each of the
doors. \code{a.pick} is an integer (1, 2, or 3) representing the contestant's
initial door selection from \code{select_door()}.}
}
\value{
An integer (1, 2, or 3) representing the door number that the host
opens, which will always contain a goat.
}
\description{
\code{open_goat_door()} simulates the action of a game host opening a door
with a goat after the contestant makes a pick in the Monty Hall problem.
}
\details{
This function simulates the action of the host in the Monty Hall Problem
game. The host opens one of the remaining goat doors after the contestant
makes their initial pick. Based on the contestant's door selection from
\code{select_door()}, the remaining doors the host can open will either have
two goats or one goat and one car. This function considers the
contestant's initial choice and returns a door number that the host opens
(which will always have a goat).
}
\examples{
  open_goat_door()
}
