% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Host opens the door with a goat behind it.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{...}{This function takes two arguments, \code{game} which is
a character vector of length 3 has two "goat" strings and 1
"car". The other argument is \code{a.pick}, a whole number, which
represents the initial door picked by the contestant.}
}
\value{
The function returns an integer between 1 to 3, associated
with the position of the "goat" in the vector \code{game} sans the
the position already taken by \code{a.pick}. This is stored in
\code{opened.door}.
}
\description{
\code{open_goat_door()} will return the integer (1 or 2 or 3)
representing the door number behind which there
is a goat and is not the initial pick.
}
\details{
The host, as per the original show, picks a door
where he knows there is a goat behind and was not
the initial pick of the contestant. This will setup
our Monty Hall problem as it will influence the decision
to "stay" or "switch" for the contestant.
}
\examples{
open_goat_door(game=a.game,a.pick=1)
}
