% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem (1).R
\name{open_goat_door}
\alias{open_goat_door}
\title{The hosts opens the first door}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{The}{randomized three doors created in create_game &
The first selection made by the player}
}
\value{
The goat door number that corresponds to what the host will open
}
\description{
The host opens another door - NOT the one chosen by the contestant, and NOT the car door
}
\details{
If the car door was chosen previously, host opens either goat. If a goat was chosen, the host opens the OTHER goat.
}
\examples{
open_goat_door -> 2
}
