% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Goat Door Opens}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{a.pick}
}
\value{
Value of opened door
}
\description{
Host shows one door has a goat
}
\details{
If they chose a car, then host opens either. IF they chose a goat the host opens other goat door.
}
\examples{
  game <- create_game()
  a.pick <- select_door()
  open_goat_door(game, a.pick)
}
