% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Open a goat door.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a goat door that was not selected
  by the contestant.
}
\description{
`open_goat_door()` selects a goat door to open.
}
\details{
Once the contestant picks a door, then one of the doors
  that is not the contestant's selection must be opened.
  The host will open a door that has a goat behind it, so
  the contestant must choose if they want to stick with
  their original door or choose the other unopened door.
}
\examples{
  open_goat_door()

}
