% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Host opens a door with a goat behind it}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{The contents behind of the three doors corresponding to the numbers of the doors in order.}
}
\value{
the number of the door that was opened by the host (goat behind the door)
}
\description{
After contestant selects a door, the host reveals a door with a goat behind it.
}
\details{
If the contestant's initial pick has a goat behind it, the host opens the only remaining door with a goat. If the contestant's initial pick has a car behind it, the host randomly selects one of the goat doors to open.
}
\examples{
open_goat_door(c("car","goat","goat"),2)
}
