% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Host Opens Goat Door}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function will return the door selected by the host
}
\description{
The function 'if()' is used to create a if
statement to filter the doors
}
\details{
The host will randomly selected one of the doors
remaining that is NOT a car door. If the contestant
has selected the car door, the host will randomly
select one of the remaining two goat doors.
}
\examples{
open_goat_door
}
