% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty_hall_pkg1.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Host opens a goat door.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{a length 3 character vector indicating the positions of
goats and the car.}

\item{a.pick}{a number between 1 and 3 indicating the contestant's selection.}
}
\value{
The function returns a number between 1 and 3 indicating the
host's selection.
}
\description{
\code{open_goat_door()} allows the host to open one of the remaining two doors
that were not selected by the contestant to reveal a goat.
}
\details{
\code{open_goat_door()} randomly selects one of the two goat doors in the game
if the contestant selected a car door; otherwise, \code{open_goat_door()}
select the remaining goat door in the game that were not selected by the
contestant.
}
\examples{
  game <- create_game()
  first.pick <- select_door()
  open_goat_door(game, first.pick)

}
