% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Host opens a door with a goat behind it}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{'doors'}{\itemize{
\item the numbers 1, 2, 3 that correspond to the the three doors in
the game
'goat.doors' - a door that has a goat behind it, aka does not have a car
behind it
'a.pick' - one door numbered either 1, 2, or 3, which is the contestant's
previously selected door
'opened.door' - the door that the host opens that is not the door the
contestant selected nor has a car behind it
}}
}
\value{
The function returns 'opened.door', which is a number from 1 to 3 that
corresponds to the door the host opens, which must be a door that has a
goat behind it but is not the door the contest previously chose
}
\description{
The 'open_goat_door()' function generates a number between one and three
that corresponds to the door the host opens, which cannot be the door the
contestant previously chose nor the door with the car behind it.
}
\details{
After the contestant chooses a door, the host continues the game by
eliminating one of the other doors. The host does this by choosing and
opening one of the two doors left that were not selected by the
contestant.The host also cannot choose the door with the car behind it as
this would spoil the game. Thus, the host, who has insider information on
which doors have goats and which has the car, opens an unselected door
with a goat behind it in order to keep the contestant stumped and the
game going.
}
\examples{
    open_goat_door()

}
