% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHallKRB.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Open a door that is not a.pick or corresponds to the value of 'car' in a.game.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{a.pick, }{a.game}
}
\value{
returns a length one numeric vector, 1,2, or 3, which corresponds to the host-opened door
that is neither the car nor the guest's first choice.
}
\description{
Opens one of three doors that is neither the guest's first choice (a.pick, return of select_door)
nor has a car behind it.
}
\details{
Because it cannot be a door with a car behind it, it must be a door with a
goat behind it. This represents the host opening another door, thus increasing the odds that the
remaining door will have the car behind it.
}
\examples{
opened.door
}
