% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Open a Goat Door}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{A character vector created by create_game().}

\item{a.pick}{An integer indicating the contestant's
initial door selection.}
}
\value{
An integer indicating which door is opened.
}
\description{
Determines which goat door the host can open after
the contestant makes an initial selection.
}
\details{
The host cannot open the contestant's selected door.
If multiple goat doors are available, one is selected
at random.
}
\examples{
game <- create_game()
pick <- select_door()
open_goat_door(game, pick)

}
