% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Simulate the host's selection of one other door containing a goat.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{Character vector}

\item{a.pick}{Numeric integer}
}
\value{
The function returns a numeric vector
}
\description{
\code{open_goat_door} utilizes other R functions to randomly select 1 of the
remaining two doors not selected by the contestant to reflect the host's
choice of door.
}
\details{
This function is used to represent the host's choice of 1 door
out of the 2 remaining doors presented, ensuring only 'goat' doors are
eliminated by host.
}
\examples{
open_goat_door(1:3, 2)
}
