% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem Script.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Host opens a goat door}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{Character, car or one of the goats}

\item{a.pick}{Numeric, door numbers}
}
\value{
a number between 1 and 3 that contain a goat behind it
}
\description{
\code{open_goat_door(game,a.pick)} return a door that contain a goat behind it.
}
\details{
Before the selected door is opened however, someone who knows what's
behind the doors (Monty Hall) opens one of the other two doors, but it
can't be a door the contestant has already selected . So it must be a door
that is not a car and not a current contestant selection.
}
\examples{
game <- create_game()
a.pick <- select_door()
open_goat_door ( game, a.pick )

}
