% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Host Opens Goat Door}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{...}{This function has two arguments: game and a.pick:
game is inputted by the create_game function
and a.pick is inputted by the select_door function.}
}
\value{
The function returns a length of one numeric vector
that is not \code{a.pick}, nor a numeric value that corresponds
to "car" by the \code{create_game()} function. This will be
saved as 'opened.door'.
}
\description{
\code{open_goat_door()} generates a selection of a door by the
host through a numeric vector of 1, 2, or 3, but will not
be \code{a.pick} nor a numeric vector value that corresponds to
the character vector value of "car" by \code{create_game()}.
}
\details{
The host will make a selection of one of the two remaining
doors after the contestant has made their initial selection.
If the contestant initially selected a door with a goat
behind it, the host will choose the remaining door with a
goat. If the contestant initially selected a door with the
car, the host will select either one of the two remaining
doors with a goat behind it.
}
\examples{
  open_goat_door( game, a.pick )

}
