% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{open_goat_door}
\alias{open_goat_door}
\title{HOST OPENS GOAT DOOR}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{The}{game vector indicates the status of each door ("car" or "goat")
The a.pick vector is the door number that the contestant initially chose (1,2, or 3)}
}
\value{
This will return the door number that the host opened to reveal the goat (1,2, or 3)
}
\description{
The host will open a door that reveals a goat based on the contestant's first choice.
}
\details{
This function will determine which door the host will open to reveal a goat.
In the case that the contestant chose the door with a car, the host will randomly
select one of the two doors with a goat. In the case that the contestant chose the
door with a goat, the host will select and open the only remaining goat door.
}
\examples{
open_goat_door(game = c("car", "goat", "goat"), a.pick = 1)
}
