% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Host opens a door with a goat from the remaining doors.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{: the game with two doors with goats behind them and one with a car.}

\item{a.pick:}{the door selected by the contestant.}
}
\value{
The function returns a number between 1 and 3 corresponding to a door with a goat behind it.
}
\description{
\code{open_goat_door()} selects one of the remaining two doors with a goat behind it.
}
\details{
This function requires two arguments.
}
\examples{
open_goat_door(game=new_game, a.pick=contestant_choice)
}
