% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Montyhall.R
\name{open_goat_door}
\alias{open_goat_door}
\title{The host opens a goat door
 @description
            The host has to open a door that is not initially selected by the contestant or a door with a car behind it.
            In case that the contestant initially selected the Car door, so the host can choose any of the other goat
            doors.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{The}{doors object is a numeric vector containing numeric values while the game object includes character
vectors. The result returned after running the function is a numeric value between 1 & 3}
}
\value{
The function return a length of one numeric vector indicating the goat door opened by the host.
}
\description{
The host opens a goat door
 @description
            The host has to open a door that is not initially selected by the contestant or a door with a car behind it.
            In case that the contestant initially selected the Car door, so the host can choose any of the other goat
            doors.
}
\details{
The open_goat_door()was used here for the host to open a goat door. Inside this function, we referred back
       to two of the previously created objects, game & a.pick. The game object includes character vector "goat,
       goat, car". The function uses the "if ()". It states that in case the contestant initially selected
       the car door, so the host will open any of the other 2 doors. The second "if ()" states that if the
       contestant initially selected one of the two goat doors, so the host will open a door that was not initially
       selected by the host and that is not a car door.
}
\examples{
        this.game <- c("goat","car","goat")
        my.initial.pick <- 1
        open_goat_door( this.game, my.initial.pick )
        my.initial.pick <- 2
        open_goat_door( this.game, my.initial.pick )
        my.initial.pick <- 3
        open_goat_door( this.game, my.initial.pick )
}
