% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_monty-hall-problem (2).R
\name{open_goat_door}
\alias{open_goat_door}
\title{The host reveals a goat.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{If}{contestant selected car, randomly select one of the two goats.
The parameter for this function is to open a door that is not the current
pick and does not contain a car.}
}
\value{
This function will remove one door from the selection process by revealing
a goat from door 1,2, or 3.
}
\description{
'open_goat_door' allows the host to reveal one goat by opening one
of the three doors. This leaves two doors remaining.
}
\details{
The host opens one of the three doors to reveal a goat. The contestant
is given an opportunity to switch to the unopened door or stay with
their original selection. The door reveals will always have a goat behind
it. If the contestant selected a car, then either door can be opened.
However, if a goat is selected, then there is only one door that can be
revealed.
}
\examples{
open_goat_door()


}
