% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Host opens goat door}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{...}{arguments used are ( game, a.pick )}
}
\value{
The function returns a  numeric vector of length one
indicating the opened door by host.
}
\description{
\code{open_goat_door()} This function will open one goat door only
and it should be the door that is goat door and the door that the contestant did not pick.
}
\details{
The host will always open a door with a goat behind it.
But it can't be a door the contestant has already selected.
It must be a door that is not a car and not a current contestant selection.
If the contestant selects the car on the first guess
the host can open either door, but if the contestant
selects a goat the host only has one option.
}
\examples{
open_goat_door(game, a.pick)


}
