% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem-brentlinger.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Host opens a goat door.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{arguments}{used by the function are the 'game', which should be the
setup of the game created by the 'create_game()' function, or a vector of
three prizes, for example game<-c("goat", "goat", "car") and the 'a.pick'
which should be the player selection from the 'select_door()' function,
or a number of which door was selected}
}
\value{
The function returns a length 1 number vector, indicating the door
position of the opened goat door
}
\description{
The host of the game will open one of the remaining doors that were not
selected by the player.  The opened door must be a losing goat door, which
excites the game.  The player must next pick whether they want to switch
doors or keep their selection.
}
\details{
The opened door will be a door that is not equal to the player selection,
and not equal to a car.
}
\examples{
open_goat_door(game, a.pick)
x<-c("goat", "goat", "car")
y<-3
open_goat_door(x, y)

}
