% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{open_goat_door}
\alias{open_goat_door}
\title{\preformatted{Open one of the Goat Doors.
}}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
\preformatted{The function returns a numeric vector 1-3.
}
}
\description{
\preformatted{ `open_goat_door()` will return a numeric vector of 
 one of the door numbers, the door that is opened. 
}
}
\details{
\preformatted{ In the original "Let's Make a Deal" game, after
 the contestant selects one of the three doors as 
 their intial pick, one of the remaining doors is
 opened to reveal one of the two goats. In this
 game, the if the contestant selected door one, 
 either door 2 or 3 would be opened to reveal one 
 of the goats behind one of the doors. This will 
 bring the contestant to the point of having to 
 decide between staying or switching doors. This 
 function  will randomly select one of 
 two doors that the contestant did not select. The 
 randomization is set for both scenarios in which 
 the contestant's intial pick was the car or one of
 the goats. 
}
}
\examples{
     open_goat_door()
     
}
