% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Simulates the host opening the remaining goat door.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{Character vector}

\item{a.pick}{Numeric integer}
}
\value{
The function returns a numeric vector
}
\description{
'open_goat_door' will select a goat door using 'sample',
that was not selected by the contestant using 'which'.
}
\details{
This function is the second step in the game. It eliminates one
of the goat doors, leaving only 1 goat door and only 1 car door.
}
\examples{
open_goat_door(1:3, 2)
}
