% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen comments.R
\name{open_goat_door}
\alias{open_goat_door}
\title{A goat door is opened.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{The}{parameter game is a vector that factors in the current game.
The parameter a.pick is an integer that represents the door number that
the player picked first.}
}
\value{
This function returns a number between 1 and 3 which coincides with
a goat behind the door.
}
\description{
\code{open_goat_door()} picks a second door that has a goat behind it
and is not the door the contestant selected \code{select_door()}
function.
}
\details{
This function reveals a goat for the contestant. A door that is not the
car and not the participants door is opened.
}
\examples{
open_goat_door()

}
