% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOCUMENTATION.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Open goat door}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{The}{second door in play can not be the initial door the contestant
selected nor can it be the door with the car behind it.}
}
\value{
The function returns three vectors. The first is a length 3
character vector indicating the positions of goats and the car.
The second is a length 1 numerical vector indicating the initial
door selected by the contestant. The third is a length
1 numerical vector indicating the door selected by the host.
}
\description{
\code{open_goat_door()} selects the second door to open in the game
}
\details{
This is the door the game host opens.The host will always open a door
with a goat behind it.
}
\examples{
  this.game <- create_game()
  this.game
  my.initial.pick <- select_door()
  my.initial.pick
  open_goat_door( this.game, my.initial.pick )

}
