% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Door opening function by the host}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{length 3 character vector indicating the positions of goats and the car.}

\item{a.pick}{Numeric value.}
}
\value{
Numeric value between 1 and 3.
}
\description{
\code{open_goat_door()} The host chooses a door with a goat behind it,
which can't be a door the contestant had already selected.
}
\examples{
open_goat_door()

}
