% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem-cable.R
\name{open_goat_door}
\alias{open_goat_door}
\title{The host opens a goat door.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{This}{function uses two arguments: the game that has been created
(represented by the specific arrangements of goats and cars behind
the three doors), and a.pick (the door that the contestant originally
chose).}
}
\value{
The function returns a length 1 numeric vector indicating the door
that the host opened.
}
\description{
\code{open_goat_door()} sets three doors labeled "1", "2", and "3". Then,
If within this game the contestant originally chose a car door,
it sets the non-car doors as goat doors and
takes a random sample of one of the goat doors.
If the contestant originally chose a goat door,
it opens the door that is not the car door and also not the chosen door.
Whichever situation occurs, it then returns the value of the goat door
that the host opened.
}
\details{
Now that the contestant has chosen their door, the host reveals a goat
door. This reduces the contestant's final options down to two doors:
a car door and a goat door. If the contestant chose a goat door originally
then the host has to open the other goat door. If the contestant chose the
car door originally then the host randomly opens one of the goat doors.
}
\examples{
  open_goat_door()
  
}
