% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Open a door with a goat behind it.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{A vector containing one "car" and two "goat"s.}

\item{a.pick}{The initial door chosen; integer between 1 and 3.}
}
\value{
The function returns an integer between 1 and 3, representing
the door the host opens (which has a goat behind it).
}
\description{
\code{open_goat_door()} opens an unopened door with a goat
behind it.
}
\details{
The function selects a goat door from among the unopened doors.
}
\examples{
  x=c("goat","goat","car")
  open_goat_door( game=x, a.pick=1)

  y=c("car","goat","goat")
  open_goat_door( y, 3)

}
