% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem-R-script.R
\name{open_goat_door}
\alias{open_goat_door}
\title{The host opens a goat door (door with a goat behind it).}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{...}{The arguments are \code{game} (its value is the name of
the game) and \code{a.pick} (its value is the initial door's number)}
}
\value{
The function returns an integer representing the number
of the goat door that the host opened. Returns a value which
can be stored in a variable \code{openedDoor}.
Example output: 2
}
\description{
\code{open_goat_door()} Type: Integer -> generates a random door number
from 1 to 3, where the opened goat door generated by this function
will neither be a door that was initially selected nor will it be
a door with a car behind it (car door).
}
\details{
The opened door must have a goat behind it. This
implies that the door must not have a car behind it.
}
\examples{
  open_goat_door(game = newGame, a.pick = firstDoor)
  open_goat_door(game = c("goat", "goat", "car"), a.pick = 1)

}
