% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Opens a door not selected by the contestant that has a goat behind it}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{ a vector of length three representing the game structure - see
create_game()}

\item{a.select}{ a numeric value between 1 and 3 representing the contestant's
initial door selection - see select_door()}
}
\value{
Generates a numeric value between 1 and 3
}
\description{
Randomly generates a number between 1 and 3, taking inputs of the game
structure and initial contestant pick into account
}
\details{
If the contestant has selected the car door, the function will return one of
the two goat door numbers at random. If the contestant has selected a goat
door, the function will return the only remaining goat door number.
}
\examples{

open_goat_door(game = ["goat", "car", "goat"], a.pick = 1)
open_goat_door(game = ["goat", "car", "goat"], a.pick = 2)
open_goat_door(game = ["goat", "car", "goat"], a.pick = 3)

}
