% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg-3.R
\name{open_goat_door}
\alias{open_goat_door}
\title{\if{html}{\out{<div class="sourceCode">}}\preformatted{	Have the host open a door not picked by the player. 
}\if{html}{\out{</div>}}}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{This}{function takes the game and the a.pick variables as arguments to ensure that the result is not the player selected door.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{This function returns an opened.door variable which contains an integer between 1 and 3 inclusive. 
}\if{html}{\out{</div>}}
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{open_goat_door() represents the host opening a door behind which sits a goat. This door must not be the same door that the player picked. 
}\if{html}{\out{</div>}}
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{The function generates another number between 1 and 3 inclusive, but this time, it must not be the number the player selected AND it must be the number that represents a goat. 
}\if{html}{\out{</div>}}
}
\examples{
	open_goat_door(this.game, this.pick)
}
