% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{open_goat_door}
\alias{open_goat_door}
\title{The host makes a selection.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{...}{two arguments (game and a.pick) are used.}
}
\value{
The function returns "opened.door" variable.
}
\description{
'open_goat_door' creates a vector of doors "doors".
    When selecting a door to open, ensure it is not "a.pick" 
    chosen by the contestant and does not conceal the "car."
}
\details{
The host is required to choose a door containing a goat that 
   hasn't been selected by the contestant, ensuring it's neither 
   the car "car" nor the door "a.pick" initially chosen by the 
   contestant. When the contestant initially picks the car, the  
   host can open either door, but if the contestant selects a goat, 
   the host's only option is determined.
}
\examples{
   this.game <- create_game()
   this.game
   my.initial.pick <- select_door()
   my.initial.pick
   open_goat_door( this.game, my.initial.pick )
   
}
