% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg copy 2.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Have the host open a door not picked by the player}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{This}{function takes the \code{game} and \code{a.pick} variables as arguments to ensure
that the result is not the player's selected door.}
}
\value{
This function returns an \code{opened.door} variable containing an integer between 1 and 3.
}
\description{
open_goat_door() represents the host opening a door behind which sits a goat.
This door must not be the same door that the player picked.
}
\details{
The function generates another number between 1 and 3, but this time it must
not be the number the player selected AND it must correspond to a goat.
}
\examples{
  this.game <- create_game()
  this.pick <- select_door()
  open_goat_door(this.game, this.pick)

}
