% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Host opens a goat door}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{The}{function uses the arguments of "game" and "a.pick",
typically taken from the previous two functions.}
}
\value{
the function returns a single numeric, either 1, 2, or 3.
}
\description{
\code{open_goat_door()} first determines if the participant
selected a door with a car or a goat. Then, the function
chooses a different door with a goat behind it to
"reveal" the goat.
}
\details{
This function acts as the second part to the game, where
the host reveals a door with a goat behind it. The host
must choose a door with a goat behind it. If the
participant initially chose a goat door, the host will
reveal the other. If the participant initially chose
the door with the car, then the host can choose
either of the two un-selected doors.
}
\examples{
  open_goat_door(c("goat","car","goat"), 1)
}
