% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Host opens one goat door}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{...}{this function will take two parameters the first parameter game is a vector indicating the positions of goats and the car.
( game, a.pick )}
}
\value{
The function returns a  numeric vector of length one
the second parameter will define the door previously selected from the participant indicating the opened door by host.
}
\description{
\code{open_goat_door()} This function will generate a number referring to a door with one goat behind
different from the one previously selected
}
\details{
The host will make the game with more suspense will choose a door with one goat behind it and
will leave the participant with two doors that one contains another goat and the other contain the car
}
\examples{
open_goat_door(game, a.pick)


}
