% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall_script_rj.R
\name{open_goat_door}
\alias{open_goat_door}
\title{:
Game show host reveals what is behind a different door.}
\usage{
open_goat_door(a.game, a.pick)
}
\arguments{
\item{:}{The door number returned is a numeric value.}
}
\value{
:
The function returns the number of the new "opened.door".
}
\description{
:
This function builds on which door was selected by the contestant in the previous function. The host selects a different door to reveal what is behind it. Only a non-winning door with a goat behind it shoudl be selected.
}
\details{
:
The first row sets open_goat_door as a function based on what was returned for a.pick and the correlated "goat" or "car" as set in the "game" assignment. The second row uses an if argument to prompt the host to reveal what is behind a different door. Inside the "if" function, "goat.doors" is set to include only doors that are not the door with a car behind it. Then "opened.door" selects a sample, this time it is picking from "goat.doors" which we just set to be the two doors that do not have cars behind them. The size is set to one so it selects one of those goat doors. The next "if" function runs if the contestant picked a goat door. The host will then reveal a door that is not the car and is not the original door they selected. At the end of the function, it returns the new opened door.
}
\examples{
:
open_goat_door()
}
