% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty_hall_scripts.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Host opens one of the goat doors.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{A vector containing one "car" and two "goat" strings in any configuration.}

\item{a.pick}{The position of the selected door, a number between 1 and 3.}
}
\value{
This function returns a single object containing the position of the opened goat door: 1, 2, or 3.
}
\description{
\code{open_goat_door()} simulates the game show host opening a door that the contestant has not
opened and that has a goat behind it.
}
\details{
In the Monty Hall Problem, once the contestant has opened a door, the game show host opens
one of the closed doors that has a goat behind it.
}
\examples{
game <- create_game()
a.pick <- select_door()
open_goat_door(game, a.pick)

}
