% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Open a Door with a Goat Behind It}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{This}{function takes in two inputs: game, which will be linked to create_game(), and a.pick() which will be linked to select_door().}
}
\value{
This function will return a number, either 1, 2, or 3. It will not be the same number as from select_door().
}
\description{
This function opens a door that both has a goat behind it and is not the door the contestant initially picked.
}
\details{
This function uses the information from the previous two functions to eliminate one of the doors the contestant can choose.
It will open a door in which there is a goat behind it, but it will not be the door the contestant initially picked.
}
\examples{
open_goat_door(game, a.pick)
}
