% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Open a goat door in the Monty Hall game.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{A character vector representing the game, with two
goats and one car.}

\item{a.pick}{A numeric value representing the contestant's
initial door selection.}
}
\value{
A numeric value representing the door opened by the host.
The value will be 1, 2, or 3.
}
\description{
\code{open_goat_door()} determines which door the host should open
after the contestant makes an initial selection.
}
\details{
If the contestant selects the car, the function randomly selects
one of the two goat doors. If the contestant selects a goat, the
function opens the remaining door that contains a goat.
}
\examples{
  game <- create_game()
  a.pick <- select_door()
  open_goat_door(game, a.pick)

}
