% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Open a goat door in the Monty Hall game}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{A character vector of length 3 representing the current game state,
as created by the `create_game()` function.}

\item{a.pick}{An integer between 1 and 3 representing the contestant's initial choice,
as returned by the `select_door()` function.}
}
\value{
An integer between 1 and 3, representing the door number opened by the host.
}
\description{
`open_goat_door()` simulates the host opening a door to reveal a goat in the Monty Hall Problem.
}
\details{
This function mimics the action of the game show host, who always opens a door
that contains a goat and is not the door chosen by the contestant. If the
contestant initially chose a door with a goat, the host will open the other
door with a goat. If the contestant chose the door with the car, the host
randomly selects one of the two goat doors to open.
}
\examples{
# Create a game and make a selection
game <- create_game()
pick <- select_door()

# Open a goat door
open_goat_door(game, pick)

}
