% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Host opens a door that displays a goat}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{A character vector of length 3 representing the
distribution of prizes behind the doors (e.g., c("goat", "car", "goat")).}

\item{a.pick}{An integer 1, 2, or 3 which represents the choice selected (or
in this case, simulated) by the player.}
}
\value{
This function returns the number (1 through 3) of the door that the host opens.
}
\description{
\code{open_goat_door()} chooses one of the three doors not selected by participants
to open and show a goat.
}
\details{
This function simulates the host's action of opening a door to reveal a goat in
the Monty Hall game. After the contestant's initial selection, the host must open
one of the two remaining doors, adhering to two crucial rules: the opened door
must not be the contestant's choice, and it must not contain the prize (typically
a car). The function takes into account the contestant's selection and the actual
location of the prize, then determines which of the remaining doors can be safely
opened to reveal a goat. This step is critical in the game as it provides
additional information to the contestant before they make their final decision to
stay with their original choice or switch doors. The function's behavior mirrors
the host's knowledge of the prize's location and their deliberate action to
maintain the game's suspense while adhering to the established rules.
}
\examples{
  open_goat_door( my.game, 2 )

}
