% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Reveal a 'Zonk' (goat)}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{The returned value from 'create_game()' containing the vector of doors}

\item{a.pick}{The returned value from the 'select_door()' function containing the contestant selection}
}
\value{
A numeric value representing the revealed door is returned as the vector index of 'create_game()'
}
\description{
The host requests that one of the models/assistants reveal a losing
door for the contestant.  The revealed door cannot reveal the car
or include the contestants selection (a.pick).
}
\details{
The function evaluates the constestant selection.  If the contestant selected
the car, the function will randomly reveal one of the other doors.  If the
contestant selected a goat, the function will always reveal the other goat door.
The reveal uses the 'create_game()' vector as the baseline and excludes the car and
contestant selection.
}
\examples{
 open_goat_door( game, a.pick)
 Output: 2
 open_goat_door( c("goat", "car", "goat"), 2)
 Output: 1

}
