% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{open_goat_door}
\alias{open_goat_door}
\title{HOST OPENS GOAT DOOR.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{if}{arguments are used by the function.}
}
\value{
The function returns the opened door.
}
\description{
\code{open_goat_door()} HOST OPENS GOAT DOOR.
}
\details{
The host will always open a door with a goat behind it,
but it can’t be a door the contestant has already selected.
So it must be a door that is not a car and not a current
contestant selection.
Note that if the contestant selects the car on the first guess
the host can open either door, but if the contestant selects
a goat the host only has one option.
}
\examples{
open_goat_door()

}
