% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Host Opens a Goat Door}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
opened.door
}
\description{
'open_goat_door()' is when the host will open 1 of the 2 remaining
doors to reveal a goat
}
\details{
The function uses the 'select_door()' function as we cannot choose one of
doors the contestant already chose( 1, 2 or 3). The host also cannot choose
the door with a car behind it, hence why the function cannot equal the win.
}
\examples{
open_goat_door(game, a.pick )
}
