% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_game.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Open a goat door

`open_goat_door()` determines which door the host opens
after the contestant makes their initial selection.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{A length-three character vector containing
two goats and one car, as created by `create_game()`.}

\item{a.pick}{The contestant's initial door selection,
represented by a number between 1 and 3.}
}
\value{
The number of the door opened by the host.
}
\description{
Open a goat door

`open_goat_door()` determines which door the host opens
after the contestant makes their initial selection.
}
\examples{
game <- create_game()
pick <- select_door()
open_goat_door(game, pick)

}
