% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg-finished.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Create a function that opens a goat door}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a number between 1 and 3 that corresponds with a goat door
}
\description{
\code{open_goat_door()} generates a new function that opens one of the two
remaining doors that is a losing goat door
}
\details{
After the contestant selects their chosen door, the host
will open one of the two remaining doors. However, the
door that is opened MUST be one of the losing goat doors.
}
\examples{
  open_goat_door()

}
