% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Open a goat door.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{`game`}{Character vector. This argument inputs the order of the
doors (car and goats) to ensure a goat door is selected.}

\item{`a.pick`}{Integer. This is the initial door the player picked.}
}
\value{
Integer. This returns the number of the opened door.
}
\description{
\code{open_goat_door} returns the number of a door with a goat behind it
from the two remaining doors after the initial player choice.
}
\details{
Monty always opens a door with a goat behind it from the remaining two.
}
\examples{
    open_goat_door()

}
