% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Host opens a goat door.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{a length 3 character vector from \code{create_game()} with
"goat", "goat", and "car" in a random order.}

\item{a.pick}{a number between 1 and 3 from \code{select_door()}
giving the door the contestant picked.}
}
\value{
The function returns a number between 1 and 3 indicating
the door the host opened.
}
\description{
\code{open_goat_door()} opens one of the doors the contestant did not
pick to reveal a goat.
}
\details{
The host never opens the car door or the contestant's door. If
the contestant picked the car he opens either goat door, and if
they picked a goat there is only one left to open.
}
\examples{
  this.game <- create_game()
  my.initial.pick <- select_door()
  open_goat_door( this.game, my.initial.pick )

}
