% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monty Hall.package.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Hosts opens a goat door}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{A character vector (length 3) with "goat"/"car" values, from \code{create_game ()}
a.pick  An integer (1-3) indicating the contestant's first choice.}
}
\value{
An integer (1-3) indicating the door that was opened by the host.
}
\description{
\code{open_goat_door()} selects one of the doors with a goat that was not picked by the contestant.
}
\details{
If the contestant initially picks the car, the host randomly chooses one of the two goat doors.
If the contestant initially picks a goat, the host must open the other goat door.
}
\examples{
g <- create_game()
p <- select_door()
open_goat_door (g,p)

}
