% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Open a Goat Door}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{A vector showing what is behind each door.}

\item{a.pick}{The door originally selected by the player.}
}
\value{
The number of the goat door that is opened.
}
\description{
Opens one of the doors that has a goat behind it.
}
\details{
The function chooses a goat door that is different from the
door the player selected.
}
\examples{
game <- c("goat", "goat", "car")
open_goat_door(game, 1)

}
