% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{open_goat_door}
\alias{open_goat_door}
\title{#'   One of the two goat doors is opened by the host.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{...}{game argument is from create_game function and a.pick is from the
select_door function.}
}
\value{
The function returns a length 1 numerical vector
indicating the position of a goat that has not been chosen by the guest.
}
\description{
#'   \code{open_goat_door()} looks behind the scenes at which doors have goats
and opens one goat door that has not been chosen by the contestant.
}
\details{
#'   The game requires a goat door to be opened after the contestant
makes the initial selection. This will lead into the contestant
choosing to stay or change doors in the following stage of the game.
}
\examples{
  open_goat_door(c("goat","goat","car"), 1)
  
}
