% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Open one of the goat doors}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{a.pick}{\itemize{
\item the output from select_door(), a number between 1-3
game - output from create_game(), a length 3 character vector that detail positions of goats and cars
}}
}
\value{
the function returns a number between one and 3 representing the opened goat door
}
\description{
This function represents the next step of the game, where the host opens one of the doors that has a goat behind it.
}
\details{
This step samples between the remaining doors depending on whether the initial pick is a goat or car.
}
\examples{
 open_goat_door( create_game(), select_door())
}
