% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem edited.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Host Opens Another Door.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{This}{function has two argments. The first is \code{game} which is a character vector
created in the \code{create_game} function, and the second is \code{a.pick} which is
a single numeric value that results from the \code{select_door} function.}
}
\value{
This function returns a single numeric value between 1 and 3, from the doors vector (1, 2, 3).
}
\description{
\code{open_goat_door} matches the vector created in create_game to a vector
of doors (1, 2, 3), and chooses a door that contains a goat
but was not chosen in \code{select_door}.
}
\details{
At this point in the game, the host opens of the two doors
which were not chosen by the contestant. The host only opens
a door with a goat behind it. The contestat still does not know
what is behind the other two doors.
}
\examples{
  open_goat_door( this.game, my.initial.pick )

}
