% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Test}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{...}{no arguments are Test.}
}
\value{
The function returns a Test
}
\description{
Test
}
\details{
Test
  Test
  Test
}
\examples{
  Test()

}
