% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Opening the goat door}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{if}{the contestant has selected the car door, the host can open either of the two remaining doors, if the contestant has initially picked a goat door, the doors will revel the remaining goat}
}
\value{
return a number 1, 2, or 3.
}
\description{
Here the host of the game reveals a door which was not picked by the contestant
}
\details{
the door reveals a goat
}
\examples{
: 3
}
