% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\name{open_goat_door}
\alias{open_goat_door}
\title{The host open the first door}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{The}{random set of the 3 doors created at crest_game() &
The first selection by player}
}
\value{
The "goat" door number that the host will open
}
\description{
The host open a door different from the picked door & the car door
}
\details{
If the picked door is "car", the host can open any of the other two.
If the picked door is "goat" the host has only one option to open
}
\examples{
  open_game_door(game, a.pick) --> 2
}
