% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.r
\name{open_goat_door}
\alias{open_goat_door}
\title{Open Goat Door.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{is the 3 character vector from `create_game()`
and a.pick is the single numeric output of `select_door()`}
}
\value{
The single numeric value is returned, which represents
  the opened door of the goat.
}
\description{
`open_goat_door()` reveals a unopened door with a goat
  behind it.
}
\details{
The removes the newly opened door, as one of the choices for the
  contestant.
}
\examples{
  this.game <- create_game()
  this.game
  my.initial.pick <- select_door()
  my.initial.pick
  open_goat_door( this.game, my.initial.pick )

}
