% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Host reveals a door with a goat.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{This}{function uses the doors object, which
is a three character numeric vector. It also uses
an if, then statement to establish that if the player
initially selected a goat door, then the other goat door
should be opened and not the car door. If the player
initially selected the car door, then either of the goat
doors can be opened. Both scenarios are built so that
the game reveals a door that was not initially selected.}
}
\value{
The function returns one character to show which
door (1, 2, or 3) has a goat behind it.
}
\description{
'open_goat_door()' reveals one of the doors with
a goat behind it.
}
\details{
Once the player selects a door, then the host
will reveal one of the doors that contains a goat.
The door will not be the number that has already been
selected. This means that if the player selected the
car on the first round, then either of the other
two doors can be opened. If the player selected
a goat door initially, then the other goat door
will be revealed.
}
\examples{
open_goat_door()
}
