% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Opens an unselected door with a goat behind it.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{`'game`}{and `a.pick` are passed into the function.
`game` holds the info on which door hold either a
goat or car. `a.pick` is the chosen door by the
contestant.}
}
\value{
The function returns numerical value representing an
  open door with a goat behind it.
}
\description{
`open_goat_door()` generates a number between 1 and 3
  representing a door with a goat behind it and not the
  door that was selected by the contestant.
}
\details{
This function uses `if` statements to first determine
  if the contestant selected the car. If he did select
  the car, then a random goat door is opened. If the
  contestant chose a goat door, then the other goat door
  is opened.
}
\examples{
  open_goat_door( new.game, first.pick )
  open_goat_door( c("goat", "car", "goat"), 2 )

}
