% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-script.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Open Goat Door}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{None}{}
}
\value{
Host will open a door with a goat behind it, even if the
contestant chose the door with the car behind it.
}
\description{
Host will open a door that will have a goat behind it.
}
\details{
If contestant selected car, randomly select one of two goats
}
\examples{
open_goat_door()

}
