% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Step 3 : Host opens goat door}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{...}{game, the return value of create_game call
a.pick is the return value of select_door call}
}
\value{
an integer representing the opened door
}
\description{
opens a door
}
\details{
This function receives two positional
arguments, game and a.pick
}
