% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{open_goat_door}
\alias{open_goat_door}
\title{The contestant to open one of the door.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{There}{are arguments used by the function.}
}
\value{
The function returns a numeric vector.
}
\description{
\code{open_goat_door()} Contestant opened one of the doors with goat.
}
\details{
The contestant selects a door, then the host opens a door to reveal a goat
and the contestant was given a chance to select again.
}
\examples{
  open_goat_door()

}
