% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Host opens a door revealing a goat}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{A character vector representing the game setup}

\item{a.pick}{An integer indicating the door selected by the player.}
}
\value{
An integer representing the door number opened by the host.
}
\description{
Determines which door the host opens to reveal a goat
}
\details{
If the player's pick is the car, the host randomly pick one of the  remaining goat doors. If it's a goat, the host opens the only remaining goat door.
}
\examples{
open_goat_door(c("goat", "car", "goat"), 2)

}
