% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Host opens a goat door}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{game}{A length-3 character vector (from \code{\link{create_game}}),
containing "goat"/"car" in a random order.}

\item{a.pick}{An integer 1, 2, or 3, representing the door
the contestant initially chose (from \code{\link{select_door}}).}
}
\value{
An integer \code{1}, \code{2}, or \code{3} for the door opened by the host.
}
\description{
\code{open_goat_door()} simulates the host revealing exactly one goat
door, following Monty Hall constraints.
}
\details{
If the contestant's chosen door has the car, the host randomly picks
one of the two goat doors to open. If the contestant has a goat,
the host must open the only other goat door.
}
\examples{
  current_game <- create_game()
  my_pick <- select_door()
  open_goat_door(current_game, my_pick)

}
