% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newscriptmontyhall.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Host Opens Goat Door.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{...}{initial door selected by the contestant}
}
\value{
The function returns a one numerical number indicating which door is slected by the host.
}
\description{
open_goat_door() generates door with goat behind it to be opened by the host.
}
\details{
The host will open a goat door and a door different than the one that was chosen by the contestant.So if the contestant has chose initially a car door then the host can now choose randomly any other door, but if the contestant has initially chose a door where there is a goat behind it, then now the host must choose the other goat door so the host will have in this case one option door to be choosed.
}
\examples{
open_goat_door()

}
