% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem (1).R
\name{play_game}
\alias{play_game}
\title{Plays an entire game, from door setup to
determinining if the contestant has won a car or not.}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
This function returns a dataframe with two columns: "strategy" and "outcome".
"Strategy" contains the options "stay" and "switch", and "outcome" contains the
options"WIN" and "LOSE". This dataframe reports the outcome of the game
as well as the contestant's strategy.
}
\description{
play_game() Combines all above functions to play an entire game at once.
}
\details{
This function reports outcomes based on whether the contestant stayed or switched
doors. The intent is to determine if staying or switching is the better strategy.
}
\examples{
  play_game()
}
