% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{Play the Monty Hall Problem game.}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns if the game was won or lost.
}
\description{
\code{play_game()} generates an entire game played by the computer.
}
\details{
All functions are called in the function to result in a full
game play of the Monty Hall Problem.
}
\examples{
  play_game()

}
