% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{PLAY GAME IN ORDER}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
To see if the contestant won or lost
}
\description{
play_game() function which executes each step of a single game in order
}
\details{
Using the functions you created last week or those provided in the solutions, package them all 
together into a single play_game() function which executes each step of a single game in order.
}
\examples{
return( game.results ) to see answer
}
