% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{play_game}
\alias{play_game}
\title{\if{html}{\out{<div class="sourceCode">}}\preformatted{Testing code for the whole game.
}\if{html}{\out{</div>}}}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Run every part of the game and return game result.
}\if{html}{\out{</div>}}
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Run every part of the game.
}\if{html}{\out{</div>}}
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Run the whole game from crate game to report result.
}\if{html}{\out{</div>}}
}
