% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{Let's Play the Monty Hall game!}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
A list of the resuls of each function will be returned.
}
\description{
This game plays the entire Monty Hall game
from start to finish.
}
\details{
Within this game, all functions: create_game(),
select_door(), open_goat_door(), change_door(),
and determine_winner() will run all together.
}
\examples{
play_game()
}
