% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{\preformatted{ Play a game
}}
\usage{
play_game()
}
\value{
\preformatted{ [1] WIN
 [1] LOSE
}
}
\description{
\preformatted{ goes through all of the steps (functions) that are
 in the monty hall game
}
}
\details{
\preformatted{ one iteration of the entire game and will return the
 outcome of the game
}
}
\examples{
     play_game
}
