% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_game}
\alias{play_game}
\title{Full Set Monty Hall Game Functions}
\usage{
play_game()
}
\arguments{
\item{...}{no parameters for this function}
}
\value{
... a data frame with two columns: "strategy" (Stay or Switch)
and "outcome" ("win" or "loose")
}
\description{
\code{play_game()} contains all the functions for the Monty Hall Game
nested under a single function
}
\details{
This function calls the Monty Hall game functions and returns the result of the game
for both possible contestant strategies.
}
\examples{
 P <- play_game()
    strategy outcome
  1     stay     WIN
  2   switch    LOSE

}
