% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem-Gonzalez.R
\name{play_game}
\alias{play_game}
\title{The entire Goat Game in a single function from start
to finish.}
\usage{
play_game()
}
\arguments{
\item{There}{are no paramaters for the 'play_game' function.
The parameters are all within the main function and the
inner functions were declared and defined above.}
}
\value{
The function returns a dataframe named 'game.results'.
This data frame is made up of two columns, 'strategy'
and 'outcome'. The 'strategy' vector is a character string
and will either be "stay" or "switch". The 'outcome'
vector is also a character string that can be "WIN" or
"LOSE".
}
\description{
This combined function incorporates both choices shown
as results- to either stay with the original choice of
door or to change doors.
}
\details{
This comprehensive function incorporates all of the
functions that were previously stated above. The
variables are all held within and carry through from
beginning to end. An added feature is the inclusion
of 2 different outcomes. 'outcome.stay' and 'outcome.switch'
provides the results from both staying with the original
'final.pick.stay' door and the 'final.pick.switch' door.
After the two different results are gathered, a data
frame is created with 'strategy' in one column and 'outcome'
in the other. The very end of this 'play_game' function
returns the 'game.results' which includes both outcomes of
staying and switching doors.
}
\examples{
play_game()

}
