% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{Play one game.}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns either WIN or LOSE based upon the final pick.
}
\description{
\code{play_game()} simulates one entire playthrough of the game.
}
\details{
The function will run one course of the game and return the results.
}
\examples{
  play_game()

}
