% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.r
\name{play_game}
\alias{play_game}
\title{Game wrapper function}
\usage{
play_game()
}
\arguments{
\item{No}{arguments are used by this function.}
}
\value{
The function returns a data frame indicating whether each strategy wins
or loses.
}
\description{
Wraps all previous functions into one function.
}
\details{
The function performs each function necessary to return win or lose results. Each game is played
with both a stay and a switch result, so the user can determine which strategy has a higher
proportion of wins and losses.
}
\examples{
play_game()

}
