% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem (4).R
\name{play_game}
\alias{play_game}
\title{Play the game}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns different outcomes when it is run. It will
determine if staying with initial selections wins or loses, same goes for
switching.
}
\description{
\code{play_game()} generates a scenario incorporating all of what we have built
putting it all together to run a simulation of whether staying with your
initial selection, or switching will yield a better outcome of winning.
}
\details{
Going through the entire game, step by step, will allow for a better
understanding of the game in an attempt to determine if staying with, or
switching the initial selection is best to increase chances of winning at
the game.
}
\examples{
  play_game()

}
