% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{Simulate a Single Monty Hall Game}
\usage{
play_game()
}
\value{
The function returns a data frame with two rows: one for the "stay" strategy and one for the "switch" strategy. The outcome of each strategy is recorded as either "WIN" or "LOSE".
}
\description{
This function simulates a single instance of the Monty Hall game, testing both strategies: staying with the initial choice and switching.
}
\details{
The function runs through the entire Monty Hall game process, from selecting a door, opening a goat door, to determining whether staying or switching results in a win or loss.
}
\examples{
play_game()

}
