% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{The whole Monty Hall Problem game}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns "WIN" or "LOSE" as the game.results
}
\description{
\code{pkay_game()} combines all of the game functions to quickly and easily generate game conditions and the subsequent win/lost outcome.
}
\details{
In one function, a game is set-up, an initial door selected, a goat door opened, the stay/switch decision made, and the final outcome associated with the final selected door is revealed.
}
\examples{
  play_game()

}
