% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_game}
\alias{play_game}
\title{Play one round of the Monty Hall game.}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
A data frame with two columns: `strategy` and `outcome`.
}
\description{
`play_game()` simulates a full round of the Monty Hall game
  including the contestant’s initial choice, the host opening a goat door,
  and the outcome for both stay and switch strategies.
}
\details{
Returns a data frame with two rows, one for the "stay" strategy and one
  for the "switch" strategy, and a column indicating whether that strategy
  resulted in a win or loss.
}
\examples{
  play_game()

}
