% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_game}
\alias{play_game}
\title{Runs one full iteration of the Monty Hall Problem.}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function}
}
\value{
data.frame of dimentions 2 by 2. The first column displays
the strategies "stay" and "switch". The second column reports
the outcome of each strategy "win" or "lose".
}
\description{
The Monty Hall Problem is a famous mathematical probability
problem that replicates a game show. A contestant is shown
three doors, two of which veil a goat and one of which holds
a car. The contestant first chooses one door, after which the
host unveils one of the other two doors which has a goat behind
it. The contestant then chooses to stay with their first door
or switch to the other. This function evaluates the outcomes
of both strategies in a single, random round of the Monty Hall
Problem game.
}
\details{
\code{play_game()} requires no inputs. It first creates a random vector
of three items: 2 goats and 1 car. A random value between 1 and 3
is generated as the contestants first choice and one of the two
remaining values is revealed to hold a goat. The function then
reports the results of both strategies the contestant might
choose, "stay" or "switch"
}
\examples{
  play_game()

}
