% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{Play a Game}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used in this function.}
}
\value{
Returns a 2x2 dataframe that shows if the strategy of "stay" or "switch" is
a win or a loss.
}
\description{
Create and play a game of Monty Hall.
}
\details{
This function runs the entire game and then catalogs the outcomes into a
dataframe. The outcomes will show some percentage of wins and losses based on
the strategy of staying and switching.
}
\examples{
 game.outcome <- play_game()
}
