% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{Play the game}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
A string with the game result
}
\description{
\code{play_game()} starts the monty hall game
}
\details{
This function plays the game and calls all the necessary functions.
At the end it will report if the player won or lost
}
\examples{
  play_game()
}
