% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_game}
\alias{play_game}
\title{A wrapper function is created to allow all of the simulation's previously
created functions to run in a single step.}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function, although the previously
indicated arguments are used by the previously defined functions within
this function.}
}
\value{
The function returns a table with two rows and two columns,
indicating the outcomes of the game (win/lose) if the contestant decides to
stay with the door they originally chose or decides to switch to the door
remaining after the host opens the goat door.
}
\description{
\code{play_game} allows the entire simulation, including all previously defined
functions, to be run in a single step.
}
\details{
All of the functions defined in previous portions of the simulation are
combined (wrapped) into a single function, allowing the game-runner to play
the game in a single step.
}
\examples{
  play_game()
  
}
