% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{Plays a full monty hall game and returns the results for one game.}
\usage{
play_game()
}
\arguments{
\item{none}{}
}
\value{
Return indicates whether staying or switching won for this one game.
}
\description{
\code{play_game()} returns the results (i.e., win or lose) for one full monty hall
game.
}
\details{
This function allows the user to test the results of one randomized monty hall
game; whether staying or switching won or lost.
}
\examples{
play_game(): Outcome: Stay: 1; Switch: 0.
}
