% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{\preformatted{Play the Game
}}
\usage{
play_game()
}
\arguments{
\item{...}{No parameters}
}
\value{
\preformatted{This returns a dataframe with columns `strategy` and `outcome`
strategy can be either stay or switch
outcome can be either win or lose
}
}
\description{
\preformatted{`play_game()` combines the five functions above into this one function.
}
}
\details{
\preformatted{This function lists the corresponding outcomes for each strategy, which will vary each time.
}
}
\examples{
	play_game()

}
