% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{New game}
\usage{
play_game()
}
\value{
This funtion returns the final results of the game
}
\description{
This is a new game created based on the original codes created above for iteration
}
