% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{This function plays an entire Monty Hall game all the way through.}
\usage{
play_game()
}
\arguments{
\item{...}{this function has no parameters}
}
\value{
\code{play_game()} returns two strings one with the strategy selected either "switch" or "stay" and one with the outcome of the game either "WIN" or "LOSE".
}
\description{
This function uses all the previously defined functions to play a single Monty Hall game all the way through.
}
\details{
This function uses all the previously defined functions to play a single Monty Hall game all the way through.
}
\examples{
play_game()

}
