% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_game}
\alias{play_game}
\title{Play the Monty Hall Game}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
a dataframe table with two columns: strategy and outcome, and two rows: the 'stay' and 'switch'. the two rows will have the outcome "WIN" or "LOSE" depending on the scenario and the contestant choices
}
\description{
play_game() puts all the previous functions together so you can easily play a full Monty Hall game
}
\details{
this function creates multiple arguments using the previous functions in order to nest them all together under one function. it also creates 'final.pick.stay' and 'final.pick.switch' for the two outcomes of 'change_door', as well as 'outcome.stay' and 'outcome.switch' for the two outcomes of the contestant's final pick depending on the 'create_game' scenario. the final part of this function creates a vector called 'strategy' that has either "stay" or "switch", allowing the following code line to create a data frame that has the strategy vectors as row names, making it easier to vizualize the game results. The last part of this function calls the 'game.results' which returns the data frame that makes it easy to read all the game results in one spot.
}
\examples{
  play_game() 
  
}
