% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{Play single round of Monty Hall Problem game}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns the result ("WIN" or "LOSE") of both
strategies (stay or switch) for a single round of play.
}
\description{
\code{play_game()}generates a Monty Hall Problem game, play it
randomly and display the result of the game.
}
\details{
The function integrates all the functions of playing a
Monty Hall Problem game, and plays a single game in order.
}
\examples{
  "play_game()"
}
