% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_game}
\alias{play_game}
\title{Create wrapper function}
\usage{
play_game()
}
\arguments{
\item{no}{arguments needed for this function}
}
\value{
Returns game results
}
\description{
This function wraps all the previously created variables up into one function
}
\details{
This function assigns previous functions to objects and also stores the results
of switching doors and the results of keeping the same door. It saves the game
results as a data frame
}
\examples{
play_game()
}
