% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{Play a round of the Monty Hall Problem Game}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function; but the other functions are used:
create_game(), select_door(), open_goat_door(), change_door(), and
determine_winner().}
}
\value{
Returns the final outcome of the game instance being a "Win" or a "Lose".
}
\description{
\code{play_game()} executes an entire round of the game, executing the steps in
order as one function made of the functions combined.
}
\details{
Game is created. Three doors: one with a car and two with goats behind them.
The Contestant select an initial door choice without knowing its occupant
goat or car. The Host discloses a goat door to the contestant that is neither
a door with a car or the previously chosen door. Contestant now decides to
stay with their original choice of door or to switch o the reaming door. If
this final choice of door by the contestant has a car behind it they win;
if there is a goat behind it, they lose. This game is played once.
}
\examples{
play_game()

}
