% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_game}
\alias{play_game}
\title{Complete Monty Hall Problem Game.}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
2 x 2 data frame of character values: strategy
(i.e., "stay" or "switch") and outcome (i.e, "WIN" or "LOSE")
}
\description{
\code{play_game()} allows the user to play the entire Monty
Hall Problem Game.
}
\details{
A wrapped version of all the applicable Monty Hall Problem
Game functions: \code{create_game()}, \code{select_door()}, \code{open_goat_door()},
\code{change_door()}, and \code{determine_winner()}. Finally, all
of these game results are stored as a data frame.
}
\examples{
`play_game()`
}
