% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{Play the Monty Hall Game}
\usage{
play_game()
}
\arguments{
\item{None}{}
}
\value{
A data frame with two rows and two columns. The first column is the strategy used, either "stay" or "switch", and the second column is the outcome, either "WIN" or "LOSE".
}
\description{
The function simulates the Monty Hall game and returns the outcome of the game for both the "stay" and "switch" strategies.
}
\details{
The function generates a new game, selects a door, opens a goat door, and calculates the final outcome for both "stay" and "switch" strategies.
}
\examples{
play_game()

}
