% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{Play_game for creating anew game with 3 doors}
\usage{
play_game()
}
\arguments{
\item{if}{final pick is a car is a winner
if final pick is a goat loses}
}
\value{
switch or stay
}
\description{
New game is created with 3 doors , 2 goats and one car
}
\details{
The game shows the 3 options of 2 goats and one car doors and the contestants have to choose one of thesr doors
}
\examples{

}
