% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_game}
\alias{play_game}
\title{Step 06: Run full game.}
\usage{
play_game()
}
\arguments{
\item{Character, }{Numeric}
}
\value{
'play_game()' returns a 2-column, 2-row table that provides both contestant strategy options in Column 1 (STAY, SWITCH), and
the result (WIN, LOSE) based on the game created and contestant choice.
}
\description{
This function runs the game from beginning to end.
}
\details{
'play_game()' simluates a contestant choosing to STAY and SWITCH their initial door selection after starting a new randomized game.
}
\examples{
play_game()
}
