% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{Test or play one round of the game.}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function will return a dataframe containing the winning
and losing outcome that related to the two strategies
(switching and staying).
}
\description{
`play_game()` this function is to test the two strategies of the game.
}
\details{
This function combines all the other functions of Monty Hall Problem
game. For one round only, It tests the two strategies of the
game (changing or staying with the first selection).
}
\examples{
play_game()

}
