% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monty Hall Script.R
\name{play_game}
\alias{play_game}
\title{Play game with desired number of cars and goats}
\usage{
play_game()
}
\arguments{
\item{...}{This function does not take any arguments.}
}
\value{
This function returns a character vector that states the results of one game.
}
\description{
\code{play_game()} returns the results of the game given the number of cars and goats inputted by
the player.
}
\details{
This function is a wrapper function that runs through the game. If used independently, this
function would only run through the game once as opposed to the full simulation.
}
\examples{
play_game()

}
