% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_game}
\alias{play_game}
\title{Monty Hall Master Function}
\usage{
play_game()
}
\arguments{
\item{...}{This function does not require any inputs.}
}
\value{
This function returns a 2 object array containing the strategy
(whether the participant switched during that run or stayed) and the outcome
(what prize the participant won).
}
\description{
This function plays the entire game from start to finish.
}
\details{
This function runs all functions needed to play the Monty Hall game.
}
\examples{
play_game()
    
}
