% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{Play the Game}
\usage{
play_game()
}
\arguments{
\item{there}{are no parameters}
}
\value{
return game results
}
\description{
Play the game using create game() function, select foor, open door functions.
}
\details{
Play the game, the first door opened will be the goat door, the next is the pick from the user, does the user want to switch or stay, and did the user win or lose.
}
\examples{
return data frame of switch or sstay and what the outcome was in each case.
}
