% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{Play a full Monty Hall Problem game.}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
A table with the game outcome (WIN or LOSE) for each
strategy (stay or switch).
}
\description{
'play_game()' returns the game outcomes of a new Monty Hall
game each time it is run.
}
\details{
The function uses results generated by 'create_game()',
'select_door()', 'open)goat_door()', 'change_door()', and
'determine_winner()'. These results are printed in a table
describing the outcome of two game strategies.
}
\examples{
  play_game()

}
