% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhallproblem.R
\name{play_game}
\alias{play_game}
\title{Plays the entire Monty Hall Problem game.}
\usage{
play_game()
}
\arguments{
\item{Strategy}{is a character vector that indicates whether the contestant
chose to switch doors or stay with the same door.}

\item{Outcome}{is a character vector that shows the outcome of the game
if the contestant stays and the outcome of the game if the contestant
switches.}
}
\value{
The output is a data frame called 'Game Results' that shows
the outcomes of the game if the contestant switches and if the
contestant stays.
}
\description{
\code{play_game()} is a function that combines all of the functions needed
to create the Monty Hall game for ease of testing with a large
sample size.
}
\details{
This function works by combining the create_game, select_door,
open_goat_door, change_door, and determine_winner functions
so they run as one.
}
\examples{
  play_game()
  
}
