% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_game}
\alias{play_game}
\title{Simulation Set-Up}
\usage{
play_game()
}
\arguments{
\item{When}{the function is called, a new game is created and a door representing
the contestant's pick is randomly selected with the select_door() function.
Then, the function open_goat_door selects the host's door that is neither
a door with the car or the contestant's door.}

\item{final.pick.stay}{and final.pick.switch is based on whether the contestant
chooses to stay with their original picked door "True" or switch to another
door "False".}

\item{outcome.stay}{and outcome.switch record whether the contestant "won" the car
or "lost" if the door had a goat behind it.}

\item{The}{game strategy and win/lost results from this game are then input into
a dataframe.}
}
\value{
Returns a dataframe containing 2 columns labeled strategy and outcome. The first
row returns the game results "Win" or "lose" if the contestant stayed with
their selection. The second row returns the game results "win" or lose" if
the contestant switched their door selection.
}
\description{
\code{play_game()} this function was created so that all of the functions that were
created earlier would be all executed together to simulate a game from start
to finish.
}
\details{
First, a new game is created and the contestant's first door selection is
recorded in "first.pick". Then the host opens another door with a goat behind
it. Then both choices are recorded if the contestant chose to stay or switch.
The door is revealed and if it is a car they win, if it is a goat then they
lose. The results of this game are recorded and returned in a table.
}
\examples{
 play_game()
}
