% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_game}
\alias{play_game}
\title{Play one Monty Hall game.}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the funcction.}
}
\value{
The function returns a data frame containing the stay and switch strategies and the outcome of each strategy.
}
\description{
'play_game()' simulates one complete Monty Hall game using both the stay and switch strategies.
}
\details{
The function creates a new game, randomly selects an initial door, opens a door containing a goat, and determines the outcome for both staying with the original selection and switching to the remaining unopened door.
}
\examples{
play_game()
}
