% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{play_game}
\alias{play_game}
\title{Play the game.}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
This function returns a vector of the game set up, numbers between 1 and 3
for the initial pick, opened door, and final door, "Stay"/"Switch", and
"Win"/"Lose"
}
\description{
\code{play_game()} combines each step of the game to run as one function and show
the conditions and results of a game
}
\details{
This function includes every step of the game to show the door selection,
the first pick, the opened door, if the contestant stays or switches, the
final pick, and the result of the game.
}
\examples{
play_game()

}
