% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhallfunctions.R
\name{play_game}
\alias{play_game}
\title{Play the Monty Hall Problem game.}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by this function.}
}
\value{
This function returns a table with the outcome (WIN or LOSE)
associated with each strategy (stay or switch).
}
\description{
'play_game()' returns the outcome of each game
every time it is run.
}
\details{
Using the results from 'create_game()', 'select_door()',
'open_goat_door()', 'change_door()', and 'determine_winner()',
this function completes a Monty Hall Problem game and records its
outcomes in a table.
}
\examples{
  play_game()
}
