% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{Monty Hall game play function.}
\usage{
play_game()
}
\arguments{
\item{This}{function does not have any parameters.}
}
\value{
This function returns the results of a game play: 'WIN' or 'LOSE'.
}
\description{
\code{play_game()} compiles all of the functions to complete an entire game of Monty Hall.
}
\details{
This function provides the steps for game play of the Monty Hall game.
}
\examples{
 play_game()

}
