% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem (1).R
\name{play_game}
\alias{play_game}
\title{Play Game}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments used}
}
\value{
the function returns a data frame that includes the strategy used by
contestant and the outcome if stayed or switched (lose or win)
}
\description{
This function includes game demonstration
}
\details{
The contestant and host start a game round. Each one of them take their turn
of selection.
}
\examples{

play_game ()

}
