% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_game}
\alias{play_game}
\title{Play a simulation game of the montyhall scenario.}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a data frame in the form of
a table with columns of "stay" and "switch" and results
of WIN" or "LOSE" indicating if the contest wins or loses
the game. The contestant wins the game if there is a car
behind the final door.
}
\description{
\code{play_game()} uses the previous code to construct a
comprehensive simulation of the entire game.
}
\details{
The simulation begins by the game being created. Next, the
contestant chooses the first door. After the door is selected,
the host opens a door with a goat behind it. Next, the
contestant decides if they want to stay with their chosen door
or choose the last unopened door. Finally, the door they
chose is revealed, and it is determined if they win or lose
the game.
}
\examples{
play_game()

}
