% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{Play a single Monty Hall game.}
\usage{
play_game()
}
\value{
A data frame with two rows, one for staying and one for switching strategies,
each containing the strategy and outcome.
}
\description{
Play a single Monty Hall game.
}
\examples{
play_game()

}
