% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_game}
\alias{play_game}
\title{Creates a data frame of the outcome of one game if
a contestant stays or if they switch.}
\usage{
play_game()
}
\arguments{
\item{...}{No arguments are used.}
}
\value{
... A data frame that details the outcomes of
each strategy (stay/switch).
}
\description{
Condenses the all functions and wraps them in the
play_game function. It produces the outcome winning
for both staying and switching based on prior information
from created functions, namely, create_game(),
select_door(), opened_goat_door(), etc. which are stored in
new_game, first.pick, and opened.door, respectively.
}
\details{
play_game does not need any arguments inside the
function as all of the necessary pieces of information
are provided later. create_game(), select_door(), and
open_goat_door() are assigned to an object, namely,
new_game, first.pick, and opened.door, respectively.
Next are the strategies, which is either to stay or
switch: final.pick.stay and final.pick.switch.
This is represented by change_door, stay=TRUE
and stay=FALSE, respectively. This function needs
information from opened.door and first.pick. The
determine_winner() function is then stored in either
outcome.stay or outcome.switch, corresponding to the
aforementioned stored objects of final.pick.stay and
final.pick.switch. determine_winner requires information
from new_game() in order to determine what is behind the
doors and know whether or not the contestant won. The
final objects in the function create the data frame.
"strategy" represents one of the columns in the data frame,
which will be either stay or switch. outcome represents the
other column, which will be either outcome.stay or
outcome.switch. Finally, game.results stores the strategy
and outcome objects stored prior, and transforms them into
a data frame using data.frame(), ensuring that it does not
become a factor (through stringsAsFactors=F).
return() returns the data frame.
}
\examples{
play_game()

premeditated.game <- c("goat","car","goat")
contestant_first_pick <- 1
play_game <- function( )
{
 new.game <- premeditated.game 
 first.pick <- contestant_first_pick
 opened.door <- open_goat_door( new.game, first.pick )
 final.pick.stay <- change_door( stay=T, 
 opened.door = opened.door, 
 a.pick = first.pick )
 final.pick.switch <- change_door( stay=F, 
 opened.door = opened.door, 
 a.pick = first.pick )
outcome.stay <- determine_winner( final.pick.stay, new.game  )
outcome.switch <- determine_winner( final.pick.switch, new.game )
strategy <- c("stay","switch")
outcome <- c(outcome.stay,outcome.switch)
game.results <- data.frame( strategy, outcome,
                             stringsAsFactors=F )
 return( game.results )
}
play_game()
premeditated.game <- c("goat","car","goat")
contestant_first_pick <- 2
play_game <- function( )
{
 new.game <- premeditated.game 
 first.pick <- contestant_first_pick
 opened.door <- open_goat_door( new.game, first.pick )
 final.pick.stay <- change_door( stay=T, 
 opened.door = opened.door, 
 a.pick = first.pick )
 final.pick.switch <- change_door( stay=F, 
 opened.door = opened.door, 
 a.pick = first.pick )
outcome.stay <- determine_winner( final.pick.stay, new.game  )
outcome.switch <- determine_winner( final.pick.switch, new.game )
strategy <- c("stay","switch")
outcome <- c(outcome.stay,outcome.switch)
game.results <- data.frame( strategy, outcome,
                             stringsAsFactors=F )
 return( game.results )
}
play_game()
premeditated.game <- c("goat","car","goat")
contestant_first_pick <- 3
play_game <- function( )
{
 new.game <- premeditated.game 
 first.pick <- contestant_first_pick
 opened.door <- open_goat_door( new.game, first.pick )
final.pick.stay <- change_door( stay=T, 
opened.door = opened.door, 
a.pick = first.pick)
final.pick.switch <- change_door( stay=F, 
opened.door = opened.door, 
a.pick = first.pick )
outcome.stay <- determine_winner( final.pick.stay, new.game  )
outcome.switch <- determine_winner( final.pick.switch, new.game )
strategy <- c("stay","switch")
outcome <- c(outcome.stay,outcome.switch)
game.results <- data.frame( strategy, outcome,
                            stringsAsFactors=F )
return( game.results )
}
play_game()


}
