% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{play_game}
\alias{play_game}
\title{Goes through all of the game functions to run a full game simulation to
return the game results if you either stay with your door or switch.}
\usage{
play_game()
}
\arguments{
\item{Because}{this function uses all the previous functions, it is
unnecessary for a parameter to be used. Simply run the function
\code{play_game()} and it will return which strategy was successful
in winning the game.}
}
\value{
\code{play_game()} returns a dataframe of the strategy by the outcome.
For example, a dataframe that will be returned would have the stay and
the switch strategies and next to it whether or not that strategy won
or lost.
}
\description{
\code{play_game()} is the function that will run through all of the previous
functions together and return a dataframe of the game result if I stay
and if I switch doors.
}
\details{
This function uses all of the other functions together and runs through
every one so one can easily run the game over and over to see which
method is best to win the game.
}
\examples{
 `play_game()`

}
