% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_game}
\alias{play_game}
\title{Play a complete Monty Hall game.}
\usage{
play_game(Num_Doors = 3, Num_Cars = 1, Reveal_Goats = 1, Reveal_Cars = 0)
}
\arguments{
\item{Num_Doors}{This is an Integer value. The number of doors in the game.
The default is 3.}

\item{Num_Cars}{This is an Integer value. The number of cars in the game.
The default is 1.}

\item{Reveal_Goats}{This is an Integer value. The number of goat doors
Monty opens. The default is 1.}

\item{Reveal_Cars}{This is an Integer value. The number of car doors
Monty opens. The default is 0.}
}
\value{
The function returns a data frame with two rows and two columns:
\itemize{
\item strategy: A character column with values "stay" or "switch"
\item outcome: A character column with values "WIN" or "LOSE"
}
}
\description{
\code{play_game()} simulates one complete round of the Monty Hall
problem, returning the outcomes for both stay and switch strategies.
}
\details{
This function runs a complete simulation of the Monty Hall game
by executing all steps in sequence: creating a new game with the
specified number of doors and cars, having the contestant select
a door, having Monty open the specified number of doors
(goats or cars), and then determining the outcome for both possible
strategies (staying with the original pick vs. switching to another
door). The function returns a data frame, which shows  the outcome for
each strategy, facilitating comparisons of the two approaches.
}
\examples{
  # Play a standard 3-door game
  play_game()

  # Play a 5-door game with 2 cars
  play_game(Num_Doors = 5, Num_Cars = 2)

  # Play a game where host reveals both a goat and a car (Part I)
  play_game(Num_Doors = 5, Num_Cars = 2, Reveal_Goats = 1, Reveal_Cars = 1)

}
