% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg (2).R
\name{play_game}
\alias{play_game}
\title{play game}
\usage{
play_game()
}
\arguments{
\item{#no}{arguments are used by the function.}
}
\value{
game results
}
\description{
play_game() plays the full monty hall game
}
\details{
full test of montyhall game
}
\examples{
play_game()
}
