% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_game}
\alias{play_game}
\title{Play a Monty Hall game.}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a data frame containing the
strategy and outcome for staying and switching.
}
\description{
\code{play_game()} plays one Monty Hall game using both the
stay and switch strategies.
}
\details{
The function creates a new game, selects a door, opens a
goat door, and determines the outcome for both staying
and switching.
}
\examples{
  play_game()

}
