% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall-pkg.R
\name{play_game}
\alias{play_game}
\title{Combine steps into single play_game() function}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
Function returns a table with that shows if staying or swithing wins or
loses.
}
\description{
\code{play_game()} combines all previous step into one single function so you
can now complete a game for the Monty Hall problem.
}
\details{
This function uses the results generated from previous steps and will
result in a table that shows how each strategy (stay vs switch) works
out (win vs loss).
}
\examples{
   play_game()

}
