% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{Game Strategy}
\usage{
play_game()
}
\arguments{
\item{new.game}{<- create_game()
first.pick <- select_door()
opened.door <- open_goat_door( new.game, first.pick )
final.pick.stay <- change_door( stay=T, opened.door, first.pick )
final.pick.switch <- change_door( stay=F, opened.door, first.pick )
outcome.stay <- determine_winner( final.pick.stay, new.game  )
outcome.switch <- determine_winner( final.pick.switch, new.game )
strategy <- c("stay","switch
outcome <- c(outcome.stay,outcome.switch)
game.results <- data.frame( strategy, outcome,stringsAsFactors=F )}
}
\value{
game.results are returned, "WIN" or "LOSE".
}
\description{
The game is assigned to a function, play_game.
}
\details{
Each step to playing the game is list and assigned to vectors.
}
\examples{
 "WIN" or "LOSE"
}
