% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{Pick that Door}
\usage{
play_game()
}
\arguments{
\item{...}{are combining all function above to create a game}
}
\value{
will return if you win the game or loose the game. This is
random with no choice to switch or stay
}
\description{
play_game()` will start the game into effect as a contestant
}
\details{
The game setup replicates the game on the TV show "Let's
Make a Deal" where there are three doors for a contestant
to choose from, one of which has a car behind it and two
have goats. The contestant selects a door, then the host
opens a door to reveal a goat, and then the contestant is
given an opportunity to stay with their original selection
or switch to the other unopened door. There was a famous
debate about whether it was optimal to stay or switch when
given the option to switch, so this simulation was created
to test both strategies.
}
\examples{
play_game()
}
