% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{Plays a single game}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a data frame showing the results of
a single game with the winning and losing outcomes from
staying or not staying.
}
\description{
\code{play_game()} generates a full game with both outcomes from either
staying with the initial door, or switching doors.
}
\details{
The function first creates a new game with the
contestant's first door pick. It then opens a goat door for the h.
A final pick for staying vs not staying is created.
Then, it determines which the two final picks was the winner and loser. Finally,
the results are returned as a data frame.
}
\examples{
  play_game()

}
