% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mh.pack.R
\name{play_game}
\alias{play_game}
\title{Playing the Game}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a data frame with the best strategy for the game which
is either the player switching or staying with their initial pick.
}
\description{
\code{play_game()} generates the entire game play by play in order and works as a
simulation set up along with returning a data frame if the player would have
lost or won the game if they would have stayed or switched from their initial
choice
}
\details{
To play the game it begins with setting up the game with 3 doors where
2 of the doors have a goat behind them and 1 door has a car. Then the player
begins by choosing one of the three doors at random not knowing which door
has the goat or car. Then one goat door is opened every time at the start
whether a goat or car door is chosen initially. If a car door is chosen the
player has the the choice to open either the goat or car door, not knowing
which one is behind each of the doors, but if a goat door is chosen the
player has no choice but to open a goat door. However, the player
has the choice to change doors from the initial pick or to stay with their
initial choice. Finally, at the end based on the player's final pick the
game will indicate whether the player won or lost based on if they chose
a door with a goat or car behind it. Based on the game a data
frame will be created that indicates what the best strategy is whether
that is staying or switching from the player's initial pick and this will be
stored as game results.
}
\examples{
  play_game()

}
