% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_game.R
\name{play_game}
\alias{play_game}
\title{Lets Win a Car!}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
all the functions to run the game and splits the results
}
\description{
Lets Win a Car!
}
\details{
Tells you if won by staing or if you won by switching . The victory is the car
}
\examples{
  play_game()
}
