% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_game}
\alias{play_game}
\title{Play the Game}
\usage{
play_game()
}
\arguments{
\item{None, }{no arguments required.}
}
\value{
The function returns a data frame of the game with the outcome representing what
would happen if they stayed or what would happen if they switched. The results
are then revealed.
}
\description{
The contestant plays the game.
}
\details{
A contestant plays the game once the doors have been chosen so the results can then
be revealed.
}
\examples{
play_game()
}
