% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_game}
\alias{play_game}
\title{Full Game Test}
\usage{
play_game()
}
\arguments{
\item{No}{arguments used in this function.}
}
\value{
The function returns a table of the strategies and the results of either "LOSE" or "WIN".
}
\description{
'play_game' function combines all of the previous steps into one function to play
a full game and provide the resulting outcomes.
}
\details{
This would be a full game played without individual steps, providing a table of the strategy
and the resulting outcome based on that choice.
}
\examples{
play_game( )

}
