% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_game}
\alias{play_game}
\title{Game Results}
\usage{
play_game()
}
\arguments{
\item{logical}{}
}
\value{
The function will be true or false, which is defined as win or lose.
}
\description{
'play_game()' function shows us the game results with the two strategies
of stay and switch.
}
\details{
There are two strategies of the game, which results in different outcomes.
}
\examples{
   play_game()
}
