% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_game}
\alias{play_game}
\title{Stay or Switch: Results of "Let's Make a Deal!"}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
A data frame with two rows for strategy, switch/stay, and
a column for outcome, LOSE/WIN.
}
\description{
\code{play_game()} generates a data frame representing if
contestant wins or loses based on strategy: Switch, Stay.
}
\details{
The game setup replicates the game on the TV show "Let's
Make a Deal" where there are three doors for a contestant
to choose from, one of which has a car behind it and two
have goats. The contestant selects a door, then the host
opens a door to reveal a goat, and then the contestant is
given an opportunity to stay with their original selection
or switch to the other unopened door. There was a famous
debate about whether it was optimal to stay or switch when
given the option to switch, so this simulation was created
to test both strategies.
}
\examples{
  play_game()

}
