% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{Play a Monty Hall Problem Game}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
"WIN" if the contestant wins, "LOSE" if they lose.
}
\description{
Simulates the entire Monty Hall Problem game, including the contestant's choice, the host's action, and the game outcome.
}
\details{
The \code{play_game()} function offers a comprehensive simulation of the Monty Hall Problem game.
It encompasses the contestant's initial door selection, the host's action of revealing a goat door,
and the ultimate game outcome, determined by the contestant's decision to either stay or switch.
}
\examples{
  play_game()

}
