% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem functions Hauer.R
\name{play_game}
\alias{play_game}
\title{Play a Simulation of a Single, Entire Game.}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the simulation.}
}
\value{
This simulation will return the results of a single game, Win or Lose.
}
\description{
Package the game functions together into a single play_game() function which
executes each step of a single game in order.
}
\details{
We can use the play_game () to run a full simulation to examine the
effectiveness of both strategies Stay and switch) by playing the game over
and over and looking at which strategy yields higher returns.
}
\examples{
play_game()

}
