% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{play one round of Monty Hall Problem game.}
\usage{
play_game(goats, cars, question)
}
\arguments{
\item{goats}{number of goats}

\item{cars}{number of cars}

\item{question}{true if the program is running for the 1st Question
false if the program is running for 2nd question}
}
\value{
returns whether the player wins or looses the game
}
\description{
'play_game()' execute the game steps and return the
     result of the game
}
\details{
This function first creates a new game by calling 'create_game()'
     first.pick stores the door picked by the user, returned by 'select_door()'
     The host opens new Door(s) and that is returned by 'open_goat_door()'
     and stored in opened.door
     Then 'determine_winner()' gets the winner for both the scenarios of staying
     and switching the picked door stratery
     Finally the verdict is retured from the function in 'game.results'
}
\examples{
      play_game(3, 2, T)
      play_game(5, 7, F)
}
