% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_game}
\alias{play_game}
\title{Plays the Monty Hall game.}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
Returns a data frame since it tests the strategy of staying and switching.
}
\description{
\code{play_game()} returns a data frame containing either "WIN" or "LOSE".
}
\details{
The function takes in all the previous functions generated to functionally
play the Monty Hall game. It will simulate each function and output a data
frame.
}
\examples{
  outcome <- play_game()
  
}
