% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{Play the game from start to finish to test the code}
\usage{
play_game()
}
\arguments{
\item{parameters}{are set around the play_game function.

the stay command is incorporated into the final code}
}
\value{
return(game.results)
}
\description{
Now that the code is created,
the game can be play from beginning to end using one function.
All of the previous functions are packaged into play_game.
}
\details{
the 'play_game' function encompasses all of the steps of the game
}
\examples{
play_game ->
Game setup: GOAT GOAT CAR
Initial Selection: 1,
The opened goat door: 2
Final door selection: 3
Game Outcome: WIN

}
