% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{Play both opportunities to see which strategy wins.}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
returns win lose values.
}
\description{
In this section you will play both the stay and switch strategy to identify which 
method gives you a greater chance at winning.
}
\details{
This section will capture all wins and losses playing both stay and switch strategies.
}
\examples{
play_game

}
