% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_game}
\alias{play_game}
\title{Monty Hall Problem game simulator.}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
A data frame with the outcomes for both the "stay" and "switch"
strategies. The data frame has two columns where \code{strategy} represents the
possible choices from \code{change_door()} ("stay" or "switch") and \code{outcome}
represents the outcomes for each strategy ("win" or "lose").
}
\description{
\code{play_game} simulates a single Monty Hall Problem game from start to
finish.
}
\details{
This function simulates a single Monty Hall Problem game from start to
finish. Using other package functions, it sets up the game, selects the
contestant's initial door, reveals a host-selected door with a goat
behind it, simulates what happens when the contestant keeps their door and
switches doors, determines the game outcome in both scenarios, and
returns the outcome.
}
\examples{
  play_game()
}
