% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{Let us play the entire Monty Hall Game for you!}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used in this function}
}
\value{
This function returns a 2x2 dataframe, with columns
strategy and outcome based on the random Monty Hall game played.
}
\description{
\code{play_game()} will run the entire one-game simulation of
the Monty Hall Problem for you.
}
\details{
This function goes through all the steps in the Monty Hall
problem, randomly creating the game with two doors hiding
a goat and one a car. It will then select a door at random,
open a door hiding a goat that is different from the initial pick
and declare a winner before presenting the final result for both
"stay" and "switch" strategies and their respective outcomes.
}
\examples{
play_game()
}
