% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_game}
\alias{play_game}
\title{Simulation Set-up}
\usage{
play_game()
}
\arguments{
\item{No}{parameters in this function.}
}
\value{
It will return the results of the game based on the results of the
particular game values.
}
\description{
This code chunk will execute all of the previous functions of the game in
order.
}
\details{
By having a "game recipe", we will not need to individually run each function
every time we want to run a new game.
}
\examples{
If for this game, door 3 is the 'car' door and I chose door 2 = goat. The
host will open door 1 = goat. From here, I chose to "SWITCH", indicating that
my final choice is door 1. I will receive the value of "LOSE". 
}
