% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem (1).R
\name{play_game}
\alias{play_game}
\title{Playing the game through}
\usage{
play_game()
}
\arguments{
\item{none}{}
}
\value{
the steps of the game from start to resulting win or lose
}
\description{
call all functions in the game
}
\details{
create a game, chose an initial door, open a goat door, chose stay or switch, determine win or lose
}
\examples{
play_game()
}
