% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_game}
\alias{play_game}
\title{Play game one round}
\usage{
play_game()
}
\arguments{
\item{Stay}{and or swtich result and their win}
}
\description{
Monty Hall simulation in one go
}
\details{
Runs the game as one instance
}
\examples{
  play_game()
}
