% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{play_game}
\alias{play_game}
\title{Play Monty Hall problem game.}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function runs the game from start to finish.
}
\description{
`play_game()` runs the Monty Hall problem steps.
}
\details{
The game sets up and replicates the game on the TV show
  "Let's Make a Deal". It begins by creating the game,
  setting up three doors, one of which has a car behind it and two
  have goats. The contestant selects a door, then the host
  opens a door to reveal a goat, and then the contestant is
  given an opportunity to stay with their original selection
  or switch to the other unopened door. Once the contestant
  selects their answer, the game determines if the contestant
  wins or loses.
}
\examples{
  play_game()

}
