% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{play_game}
\alias{play_game}
\title{Play the whole game}
\usage{
play_game()
}
\arguments{
\item{.}{. . no arguements}
}
\value{
Results of the game based on the strategy used.
}
\description{
Plays the whole game from the start to the end
}
\details{
Generates the contents of the three doors, has the contestant select an initial pick out of the three doors. Then has the host reveal a goat door. After that, the contestnat is given an option to use a "switch" or a "stay" strategy. This function explores both possibilities. Then collects thte results  into a data frame.
}
\examples{
play_game()
}
