% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_game}
\alias{play_game}
\title{Simulate the Monty Hall game.}
\usage{
play_game()
}
\arguments{
\item{...}{No arguments are used by the function.}
}
\value{
A data frame with two columns (stategy and outcome)
}
\description{
\code{play_game()} simulates a single round of the Monty Hall game, including selecting an initial door, opening a door with a goat, and determining the outcome for both staying with the initial pick or switching to the other door.
}
\details{
This function orchestrates a full Monty Hall game round which includes: creating a new game with a random number distribution of the car and goats, having the contestant make an initial door selection, having the host open one of the other doors to reveal a goat. Then, the function determines the outcome for both strategies: staying with the initial selection or switching to the other unopened door. It then return a data frame showing the outcomes for both strategies.
}
\examples{
`play_game()`

}
