% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_game}
\alias{play_game}
\title{Simplify to a Single Function}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
'win' or 'lose' will be returned
}
\description{
Simplify to a single function that can be
used to run the game
}
\details{
By storing the game in a single 'play game' function,
the process of running the game can be simplified.
}
\examples{
play_game
}
