% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty_hall_pkg1.R
\name{play_game}
\alias{play_game}
\title{Play a round of the Monty Hall Problem game.}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a dataframe with two rows and two columns. The
first column indicates the strategy used by the contestant, and the
second column indicates the outcome of the game for each startegy.
}
\description{
\code{play_game()} simulates a single round of the Monty Hall Problem game
and returns the final outcome for both scenarios in which the
contestant engages in the "stay" and "switch" strategies.
}
\details{
\code{play_game()} is a wrapper function that uses the \code{create_game()},
\code{select_door()}, \code{open_goat_door()}, \code{change_door()}, and determine_winner()`
to play through all the stages of a single round of the Monty Hall Problem game.
This functions show the outcomes for both scenarios in which the contestant
engages in the "stay" and "switch" strategies.
}
\examples{
  play_game()

}
