% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_game}
\alias{play_game}
\title{Play a Complete Monty Hall Game}
\usage{
play_game()
}
\value{
A data frame containing strategies and outcomes.
}
\description{
Simulates a complete Monty Hall game and reports
outcomes for both stay and switch strategies.
}
\details{
Creates a game, allows a contestant to select a
door, opens a goat door, and evaluates both
possible strategies.
}
\examples{
play_game()

}
