% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg-Wickerham.R
\name{play_game}
\alias{play_game}
\title{Game simulation}
\usage{
play_game()
}
\arguments{
\item{...no}{arguments used for this function}
}
\value{
The return is a table showing 'stay' and 'switch' in the first
column, and the results in the second column.
}
\description{
The play_game() function runs a simulation of a game. It goes
through each step in order and returns the results.
}
\details{
The simulation runs through a Monty Hall scenario, returning
the result of the game in a scenario where the contestant
stayed and a scenario where the contestant switched.
}
\examples{
    play_game()
}
