% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_game}
\alias{play_game}
\title{Runs through game twice, once where contestant keeps original door choice
and another game where contestant switches their door choice.}
\usage{
play_game()
}
\arguments{
\item{...no}{arguments are used by the function.}
}
\value{
List with results of keeping or switching door.
}
\description{
\code{play_game()} generates all game elements to determine if switching or
keeping the door results in a win.
}
\details{
Function imports game outputs and extracts the outcome variable from the
resulting list to indicate whether contestant won or lost.
}
\examples{
play_game()
}
