% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem Script.R
\name{play_game}
\alias{play_game}
\title{Monty Hall Game}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
Strategy of the CONTESTANT : Stay or Switch and in each case the
outcome: WIN or Lose
}
\description{
\code{play_game()} is a function that tests the Monty Hall game and tells what
will happen if the CONTESTANT choose to stay or change his selection.
}
\details{
THe function will play the  game will be played by running the functions
in order in a script. The game script will be in run a functions in
the following order:
\enumerate{
\item One function that sets up the game (three doors, one car, two goats).
\item One function that selects a door for your first guess.
\item One function that reveals a goat for the host.
\item One function that makes your final pick
(depending upon if you intend to stay or switch).
5.One function that decides if you win the car or not.
}
}
\examples{
play_game <- function( )

}
