% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{Play Game}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
This function returns a string indicating the results of the game.
}
\description{
play_game pulls together many of the other functions in this package to
  ultimately run a simulated Monty Hall game problem.
}
\details{
create_game, a vector or strings, is stored in new.game. select_door, an
   integer, is stored in first.pick. open_goat_door with arguments new.game
   and first.pick are stored in opened.door.

   Two scenarios are then run, one in which the contestant stays and the
   other in which the contestant switches doors. This is done by setting
   their values to change_door with on taking the argument stay=T and the
   other taking the argument stay=F.

   Outcomes for each scenario are stored in outcome.stay and outcome.switch
   by passing final.pick.stay or final.pick.switch as arguments to
   determine_winner.
}
