% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{Play the Monty Hall Problem game.}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a length 1 character vector
indicating the result of the game.
}
\description{
\code{play_game()} packages all above functions together and plays
the game from the beginning to the end.
}
\details{
The function will execute all functions by order, generate a random
game setting, play around, and returns the result.
}
\examples{
  play_game()

}
