% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{Play a Full Game of Monty Hall}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a character vector table of two
columns, "strategy" and "outcome", and two rows under
the strategy column, "stay" or "switch". The outcome
column will show the character vector results of the
game, "win" or "lose" based on the strategy used.
}
\description{
\code{play_game()} generates all the steps of one full game
including \code{create_game()}, \verb{select_door(), }open_goat_door()\verb{, }change_door()\verb{, and }determine_winner()`. This function
shows the results of the game in a table.
}
\details{
Playing a full game of Monty Hall will include having
the contestant making an initial selection of one of
three doors. Two of the doors have a goat behind it, and
one door has the car behind it, which is the prize.
The host will then open a door with a goat behind it.
The contestant will be given the choice to either stay
with their door selection or switch it to the other unopened
door. Once the contestant's final selection is made, the
game will then determine the winner of the game. If the
contestant ultimately chose the door with the car behind
it, the contestant will win the game. If the contestant
ultimately chose the door with a goat behind it, the
contestant will have lost the game.
}
\examples{
  play_game()

}
