% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_game}
\alias{play_game}
\title{Play a round of the Monty Hall Game.}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
A data frame with two rows and two columns:
\code{strategy}: The strategies used ("stay" or "switch").
\code{outcome}: The outcome of each strategy ("WIN" or "LOSE").
}
\description{
\code{play_game()} simulates a round of the Monty Hall game, including making
a door selection, opening a door with a goat, and determining the outcome
for both strategies (staying or switching).
}
\details{
This function simlulates a complete round by:
\enumerate{
\item Creating a new game with one car and two goats.
\item Allowing the contestant to select an initial door.
\item Opening a door with a goat.
\item Determining the final pick if the contestant stays with the initial door
or switches to the remaining unopened door.
\item Returning the outcomes of both strategies ("stay" and "switch") in
a data frame.
}
}
\examples{
  play_game()
}
