% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_game}
\alias{play_game}
\title{Play Game}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
A data frame with the strategy ("stay" or "switch")
and the corresponding outcome ("WIN" or "LOSE")
}
\description{
This will simulate a Monty Hall game.
}
\details{
This function will simulate a Monty Hall game. This includes all the
steps of contestant choosing an initial door, the host opening a goat
door,and if the contestant chooses to stay with their initial pick or
switch.
}
\examples{
play_game()
}
