% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_game}
\alias{play_game}
\title{Simulate a Monty Hall game and return if contestant wins when they keep their door or change door.}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function}
}
\value{
returns a 2x2 data frame with a strategy column with 'stay' or 'switch' options and a result column showing whether the strategy won or lost the game.
}
\description{
\code{play_game()} function simulates one Monty Hall game and returns a 2x2 data frame with a strategy column with 'stay'
or 'switch' options and a result column showing whether the strategy
won or lost the game.
}
\examples{
play_game()
}
\seealso{
Other montyhall: 
\code{\link{change_door}()},
\code{\link{create_game}()},
\code{\link{determine_winner}()},
\code{\link{open_goat_door}()},
\code{\link{play_n_games}()}
}
\concept{montyhall}
