% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Montyhall.R
\name{play_game}
\alias{play_game}
\title{Wrapping all the previous 5 game steps into single play_game()}
\usage{
play_game()
}
\arguments{
\item{The}{stay argument is a logical vector, as it includes values of either "True" or "False"}
}
\value{
The returned result from the function is a length of one character vector indicating whether the contestant
      has won or not
}
\description{
The play_game()wraps up all the previously stated game steps into one function and returns the data frame
           that contains the results from one game.
}
\details{
The play_game() includes all the codes used in the previous 5 steps with the aim of returning the data frame
       that includes all the results related to one game, which means means no repetition. Therefore, we need to
       use the loop() to check which is the more dominant strategy, the stay or the switch strategy.The returned
       result after running the function is a length of one character vector indicating whether the contestant
       has won or not
}
\examples{
        play_game()
}
