% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_monty-hall-problem (2).R
\name{play_game}
\alias{play_game}
\title{create a new Monty Hall Problem game using the previous functions}
\usage{
play_game()
}
\arguments{
\item{no}{arguments are used in this function}
}
\value{
play_game will run an entire single game of Monty Hall and produce a win or
loss determination depending on the strategy.
}
\description{
play_game combines all the functions established above to play one round
of the Monty Hall game.
}
\details{
This function will execute each step of a single game in order
}
\examples{
play_game ()


}
