% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{Play game}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The play game function returns game results in both scenarios (stay & switch)
}
\description{
The play_game() executes each step
of a single game in order and see which strategy wins/loses.
}
\details{
The objects assigned to each of the
functions to  help determine the final game results if contestant switches his pick or
stays with his same pick.All the functions
are packaged into this single play_game () function which executes
each step of a single game in order.
}
\examples{
play_game()

}
