% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem-brentlinger.R
\name{play_game}
\alias{play_game}
\title{Play the Monty Hall game}
\usage{
play_game()
}
\arguments{
\item{no}{arguments are used by the function, but it relies on all other functions:
create_game(), select_door(), open_goat_door(), change_door(),determine_winner()}
}
\value{
The function will return the results or outcome of the instance of the game played.
}
\description{
This function combines all game functions into one function to be able to
play the entire game with one function.
}
\details{
By running this function, the entire game is played once.
}
\examples{
play_game()

}
