% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{Play the Entire Monty Hall Game with One Function}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
\preformatted{play game will return a data frame stating if
STAY or SWITCH was the winning or losing option.
}
}
\description{
\preformatted{`play_game` runs the entire game simulation of the 
functions in this package. 
}
}
\details{
\preformatted{This funtion will run all the Monty Hall Game functions
as a loop. It will play thorugh an entire game of the 
"Let's Make a Deal" process. 
}
}
\examples{
    play_game()
    
}
