% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_game}
\alias{play_game}
\title{Executes the game and stores the outcome values.}
\usage{
play_game()
}
\arguments{
\item{...}{No arguments are used by the function.}
}
\value{
Data frame containing 2 columns: strategy and outcome.
}
\description{
'play_game' plays through the game and returns the game outcome values.
}
\details{
This function gives the output values for randomly placing the car and goats,
the contestant's initial choice, a goat door revealed by the host, the contestant's
choice to switch or stay, the final door selected by the contestant, and whether
or not the contestant wins or loses if they stay, and whether or not the contestant
wins or loses if they switch
}
\examples{
play_game()
}
