% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problemb.R
\name{play_game}
\alias{play_game}
\title{Plays the game.}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a character vector
indicating the overall game result.
}
\description{
\code{play_game()} generates a the function to play the overall game
}
\details{
This function wraps all other functions into playing the game.
}
\examples{
  play_game()
}
