% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen comments.R
\name{play_game}
\alias{play_game}
\title{An entire round of Monty Hall problem game}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
This function returns a win or lose determination at the end of the
game.
}
\description{
\code{play_game()} runs through one whole round of the door game, executing each
step in order.
}
\details{
A game is created where there are three doors; one with a car behind it and
two with goats behind them. The player selects a door number between 1 and 3.
The game host then opens a second door with a goat behind it. The player
must then decide to stay with their original pick or switch their pick to the
remaining door. Whether the player switches or stays, if the door number they
ultimately choose has a car behind it, they won. If it has a goat behind it,
they lose.
}
\examples{
play_game()

}
