% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{play_game}
\alias{play_game}
\title{Packages all functions into a single funtion to
play the Monty Hall Problem game.}
\usage{
play_game()
}
\arguments{
\item{This}{function does not rely on any arguements.}
}
\value{
This function will return a data frame that will show the
final results (win or lose) based on the strategy chosen
(to stay or switch).
}
\description{
'play_game' will package all required functions
to allow for the completion of the Monty Hall Problem.
}
\details{
This function allows a user to play the Monty Hall
Problem game without the need to complete each required
function individually, as everything required is bundled
into this function. It will track results based on outcome
(win or lose) as well as based on strategy (stay or switch)
to determine the best possible strategy.
}
\examples{
  return(game.results)
}
