% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOCUMENTATION.R
\name{play_game}
\alias{play_game}
\title{Play game}
\usage{
play_game()
}
\arguments{
\item{fill}{in}
}
\value{
fill in
}
\description{
\code{play_game()}plays a full round of the game and returns the game outcome.
}
\details{
This function plays a full round of the game and returns the game outcome.
}
\examples{
  fill in

}
