% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{Function to encapsulate all the Monty Hall Problem game functions into one function.}
\usage{
play_game()
}
\arguments{
\item{...}{the function uses no arguments.}
}
\value{
data.frame containing the strategy of staying or switching and the outcome.
}
\description{
\code{play_game()} encapsulates \code{create_game()} , \code{select_door()} , \code{open_goat_door()} , \code{change_door()} , \code{determine_winner()} to determine the game results.
}
\examples{
play_game()

}
