% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem-cable.R
\name{play_game}
\alias{play_game}
\title{Play the game in a single function.}
\usage{
play_game()
}
\arguments{
\item{No}{arguments are used by the function.}
}
\value{
This function returns a data frame of 2x2.
}
\description{
\code{play_game()} sets a specific arrangement of cars and goats as this
particular new game, then the contestant selects their first door,
and then the host opens a goat door based on the arrangement of
this new game and the door that the contestant originally opened.
It then sets what the final door will be if the contestant stays
and what the final door will be if the contestant switches.
Then it determines the winner within this game when the contestant stays
and when the contestant switches. Finally, it prepares to make a data frame
showing the game results of staying vs switching. It does this by
making the strategy our column headers, the outcome of staying and
switching our next row, and then making a data frame of those objects.
}
\details{
This function plays the entire game. The doors are set, the contestant
chooses their original door, the host opens a goat door, the contestant
decides if they want to stay with their original choice or switch to the
other unopened door, and then the winner is determined.
}
\examples{
  play_game()

}
