% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.r
\name{play_game}
\alias{play_game}
\title{Play a single game of the Monty Hall Problem}
\usage{
play_game()
}
\description{
This function simulates one game of the Monty Hall Problem.
}
\details{
A data frame with a strategy row ("stay" or "switch") and an outcome row ("WIN" or "LOSE")
}
\examples{
play_game()
}
