% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{\preformatted{Let's Play!
}}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
the function returns a dataframe of "WIN" and "LOSE" and "Switch" and "Stay"
}
\description{
\preformatted{Run through the whole game from start to finish.
}
}
\details{
\preformatted{Create a new game, select a door, open a goat door, 
choose whether the contestant wants to stay or switch,
determine if the contestant wins. 
}
}
\examples{
	play_game()

}
