% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{Play one round of the Monty Hall game.}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a one-vector data frame that
includes whether the contestant stayed with their initial
door choice or switched and whether they won or lost.
}
\description{
\code{play_game()} calls a series of functions to play one
complete round of the Monty Hall game.
}
\details{
The function goes through the steps of the Monty Hall game:
The contestant chooses a door, the host opens a different door
(which has a goat behind it), the contestant is then given the
option to stay with their initial choice or switch to the other
unopened door.  Their final choice is then opened, revealing
a car if they won and a goat if they lost.
}
\examples{
  play_game()

}
