% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{placeholder}
\usage{
play_game()
}
\arguments{
\item{placeholder}{}
}
\value{
placeholder
}
\description{
placeholder
}
\details{
placeholder
}
\examples{
placeholder
}
