% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_game.R
\name{play_game}
\alias{play_game}
\title{Monty Hall game where user defines number of goats and cars}
\usage{
play_game(goats, cars)
}
\arguments{
\item{goats}{Numer of goats that will be hidden behind doors.}

\item{cars}{Number of cars that will be hidden behind doors.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The number of goats added to the number of cars makes the total number of doors to choose from.
}
\description{
Monty Hall game where user defines number of goats and cars
}
\examples{
## Runs a single observation of the Monty Hall problem with 2 goats and 1 car.
play_game(goats = 2, cars = 1)

## Runs a single observation of the Monty Hall problem with 10 goats and 5 cars.
play_game(10, 5)

}
