% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_game}
\alias{play_game}
\title{Play game function to run the game twice, one time the contestant
stays with their original door and the other they swtich doors.}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
Data frame with the results of staying with original door and switching doors.
}
\description{
'play_game()' puts together all of the game elements to determine
if staying with the original door or switching doors is more
likely to end in a win.
}
\details{
The function imports all parts of the game and returns the results
of staying with the original door and the results of switching doors.
}
\examples{
play_game()

}
