% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem-R-script.R
\name{play_game}
\alias{play_game}
\title{The Monty Hall Problem game is played.}
\usage{
play_game()
}
\arguments{
\item{...}{This function does not take any parameters.}
}
\value{
This function returns a dataframe with
columns labelled "strategy" and "outcome".
"strategy" can be either "stay" or "switch".
"outcome"  can be either "WIN" or "LOSE".
The dataframe states corresponding outcomes
had the player stayed or switched.
Example output:play_game()| strategy  | outcome |1 |     stay  |     WIN |2 |   switch  |    LOSE |
}
\description{
The five functions described above are
packaged into one single function, play_game().
}
\details{
play_game() lists the corresponding outcomes
for each strategy, stay and switch. The outcome
of staying and switching may not always be same.
Every time play_game() is run, results might
be different.
}
\examples{
  play_game()


}
