% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{A wrapper function that allows a user to simulate a full Monty Hall game}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function}
}
\value{
A data.frame object containing stay/switch strategy and final game
outcome
}
\description{
Call play_game() to simulate a monty hall game outcome and output the final
result in the form of a data frame containing the strategy used by the
contestant and the final game outcome if that strategy was used
}
\details{
This function calls all of the helper functions defined in this package
to create a new game, select a door, open a goat door, and simulate what
the contestant's final selection would be if they keep their original
choice or switch. The final results are written into a data frame that
shows both stay and switch strategies and what the outcome of each strategy
would have been in actual game play.
}
\examples{
play_game()
}
