% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg-3.R
\name{play_game}
\alias{play_game}
\title{\if{html}{\out{<div class="sourceCode">}}\preformatted{The full play_game function runs through each previous function to simulate an entire Monty Hall game scenario.
}\if{html}{\out{</div>}}}
\usage{
play_game()
}
\arguments{
\item{This}{function takes no arguments}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{	This function returns a game.results variable which shows either the player swapped or stayed and whether they won or lost that game. 
}\if{html}{\out{</div>}}
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{This function serves as a way for us to simulate a game to check results where the player both stays and swaps. 
}\if{html}{\out{</div>}}
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{	This function creates our game variable with the prizes and doors, simulates the player selections, runs the scenario where the player swaps and stays, then determines whether they won or lost their game, storing that result in a results variable. 
}\if{html}{\out{</div>}}
}
\examples{
		play_game()
}
