% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{Play the Monty Hall game.}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
It returns "game.results" which is "WON" or "LOSE".
}
\description{
It enables the entire game to be played with one function call.
}
\details{
This function puts together all game functions into a single function, 
   making it possible to play the entire game with just one function call.
}
\examples{
   play_game()
}
