% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg copy 2.R
\name{play_game}
\alias{play_game}
\title{The full play_game function runs through each previous function to simulate an entire Monty Hall game scenario.}
\usage{
play_game()
}
\arguments{
\item{This}{function takes no arguments.}
}
\value{
The function returns \code{game.results}, which shows whether the player stayed
or switched and whether they won or lost that game.
}
\description{
This function simulates a full game so we can check results for both the
stay and switch strategies.
}
\details{
The function creates the \code{game} variable with the prizes and doors,
simulates the player's selection, runs the scenarios where the player
swaps and stays, then determines whether they won or lost, storing the
results in a \code{results} variable.
}
\examples{
  play_game()

}
