% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem (1).R
\name{play_game}
\alias{play_game}
\title{Test the Game}
\usage{
play_game()
}
\arguments{
\item{:}{… The commands that are chosen are a result of the contestant
decides to either “Stay” or “Switch”.}
}
\value{
The result should be 4 separate answers that shows whether the choices made
wins or loses.
}
\description{
The game must be tested to ensure that it works.
}
\details{
The commands must work together as in a single play. The commands that need
to be tested are the commands that generate results. Each result must rely on
each other and cannot produce conflicting results. The first step that need
to be tested is the game setup which is the location of the goats and the car.
The initial pick of the contestant, the opened goat door, and the final pick
must work together. For example, the contestant’s door cannot be the same as
the goat door. The final result has to be either a “Win” or a “Lose”. The
“Win” should not appear if the final door chosen is a goat door.
}
\examples{
  this.game <- c("goat", "car", "goat")
      determine_winner (final.pick = 1, game = this.game)

}
