% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_game}
\alias{play_game}
\title{Runs a complete Monty Hall Game simulation}
\usage{
play_game()
}
\arguments{
\item{No}{arguments are used by this function}
}
\value{
The function returns a 2 x 2 data frame, showing the two
strategies used in the game (stay = TRUE or stay = FALSE)
and the results of each strategy (WIN or LOSE).
}
\description{
\code{play_game()} takes all previous functions in the package
and wraps them into a single function.
}
\details{
This function runs two complete Monty Hall Game
simulations from the previously used functions. This package
simulates one game where the contestant keeps their initial
door choice, and another where the contestant changes to
the other un-opened door.
}
\examples{
  play_game()
}
