% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{Play the game one time}
\usage{
play_game()
}
\arguments{
\item{N/A}{}
}
\value{
will return a small table that indicates the winngin and losing strategy
}
\description{
play_game general function that contains all the previous functions will return
WIN  or Lose
}
\details{
The game consist of choosing one door from three doors, one of which has a car
behind it and two have goats The contestant selects a door, then the host
opens a door to reveal a goat, and then the contestant is
given an opportunity to stay with their original selection
or switch to the other unopened door. If the last pick was a car the participants
considered a winner otherwise a looser
}
\examples{
play_game()

}
