% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall_script_rj.R
\name{play_game}
\alias{play_game}
\title{:
Putting it all together and playing the game.}
\usage{
play_game()
}
\arguments{
\item{:}{The outcome is returned as characters.}
}
\value{
:
A data frame is returned displaying "WIN" or "LOSE" and "STAY" or "SWITCH".
}
\description{
:
All of the steps are run sequentially so that a round of the game is played.
}
\details{
:
The play_game function runs all of the above functions in order. This step adds a "game.results" function that builds on outcomes based on whether the contestant stayed or switched doors. It then builds a data frame to display the final results.
}
\examples{
play_game()
}
