% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{Play the Monty Hall Problem}
\usage{
play_game()
}
\arguments{
\item{This}{function takes no inputs.}
}
\value{
The function returns a 2x2 data frame, with each strategy getting a WIN or LOSE judgement.
}
\description{
This executes all the functions of the Monty Hall package in the correct order to play one game.
}
\details{
The code assigns functions to specific objects to allow for function input in the functions that require it. The code
the result of both strategies in the game. It then prints this information in an easily interpretable data frame.
}
\examples{
play_game()
}
