% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_game}
\alias{play_game}
\title{Play a complete Monty Hall game.}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
A data frame containing two rows showing the strategy
("stay" or "switch") and the resulting outcome ("WIN" or "LOSE").
}
\description{
\code{play_game()} simulates a complete Monty Hall game and compares
the outcome of staying with the original door selection to the
outcome of switching doors.
}
\details{
The function creates a new game, randomly selects an initial
door, opens a goat door, determines the final selection for both
the stay and switch strategies, and determines the outcome of
each strategy.
}
\examples{
  play_game()

}
