% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_game}
\alias{play_game}
\title{Simulate the outcome of the game}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
A dataframe containing the result of each strategy (stay or switch) and the outcome (win or lose) of each.
}
\description{
The function determines the outcome of the game based on the final door selection.
}
\details{
The function calls each of the previous functions to simulate a game passing
}
\examples{
 play_game()
  - create_game() output: c("goat", "car", "goat")
  - select_door() output: 1 
  - open_goat_door( c("goat", "car", "goat"), 1) output: 3
  - change_door( stay=T, 3, 1) output: 1
   - determine_winner( 1, c("goat", "car", "goat")) output: "lose"
  - change_door( stay=F, 3, 1) output: 2
   - determine_winner( 1, c("goat", "car", "goat")) output: "win"
Output:
  game.results("stay", "lose"
              "switch", "win")
}
