% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{play_game}
\alias{play_game}
\title{Simulations of the Game.}
\usage{
play_game()
}
\arguments{
\item{arguments}{are used by the function.}
}
\value{
The function returns game results.
}
\description{
\code{play_game ()} Simulations of the Game.
}
\details{
This is a simulation of the game to go through all previous
steps in one export.
}
\examples{
play_game ()

}
