% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{The Monty Hall Game}
\usage{
play_game()
}
\arguments{
\item{change_door(}{stay=T, opened.door, first.pick )
change_door( stay=F, opened.door, first.pick )}
}
\value{
game.results
}
\description{
Play a full Monty Hall Game
}
\details{
The play_game function combines our previous functions into a single
function and returns the results of the game each time. The results are
printed into a table indicating a loss or win.
}
\examples{
A table indicating either a win or a loss
}
