% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_game.R
\name{play_game}
\alias{play_game}
\title{Play one Monty Hall game}
\usage{
play_game()
}
\value{
A data frame containing the strategy used and
whether that strategy resulted in a win or loss.
}
\description{
`play_game()` simulates one Monty Hall game and calculates
the outcome of both the stay and switch strategies.
}
\examples{
play_game()

}
