% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{Simulation Set-Up}
\usage{
play_game()
}
\value{
The game will return the vector of the strategy function as well as
the vector of the outcome function in a table with the appropriate headers.
}
\description{
Take the packages from the previous steps and combine them into one package.
}
\details{
After completing the process of creating small packages to handle specific functions of the
game, the time has come to combine them all to run together as one complete package.
}
\examples{
  Strategy     Outcome
  stay          WIN
  switch        LOSE

}
