% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg-finished.R
\name{play_game}
\alias{play_game}
\title{Complete game play through}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function will return a data frame based on the contestant's chosen strategy
}
\description{
\code{play_game()} function will do a complete put together play through of the
Monty Hall game.
}
\details{
The contestant will make a choice of door, the host will open a non-winning door,
the contestant the has the choice to switch to the remaining door or keep their original
one. It will then be revealed if they win based on whether or not they have chosen to
stay or switch their original choice of door.
}
\examples{
  play_game()

}
