% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{A simulation of the Monty Hall problem game.}
\usage{
play_game()
}
\arguments{
\item{No}{argumnets are required.}
}
\value{
The function returns a dataframe containing the strategy
 (Stay and Switch) and the result in both the scenarios.
}
\description{
This function simulates all the steps of the game and returns the result.
}
\details{
This is a collection of all the steps of the game. It runs a simulation
 of the whole game. The function returns the game result for both the
 scenarios where the contestant has decided to stay with his/her original
 choice and where he has decided to switch doors.
}
\examples{

 play_game()

}
