% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_game}
\alias{play_game}
\title{Play a single game of "Let's Make a Deal" or more commonly called, the Monty Hall game.}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by this function}
}
\value{
This returns a data frame with 2 rows and 2 columns. THe dataframe includes a column representing
the strategy (stay/switch) and another for the outcomes (win/lose)
}
\description{
The function runs through an entire game of "Let's Make a Deal" where each round is random.
}
\details{
assigned car or goats, the contestant selects a door, the host reveals a goat door, the contestant then chooses to
keep the original door or switch. Finally, win or lose is presented depdening on the outcome of the door selection.
}
\examples{
play_game()
}
