% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_game}
\alias{play_game}
\title{Play Game}
\usage{
play_game()
}
\arguments{
\item{No}{parameters.}
}
\value{
A data frame. The frame contains each strategy (switch or stay) and
the result of each.
}
\description{
This function plays an entire round of the Monty Hall game. It assigns
the positions of the goats and car, has the player's initial pick, opens
a goat door, and then reveals if they won depending on if they switched.
}
\details{
This goes through each step previously documented.
}
\examples{
    play_game()

}
