% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monty Hall.package.R
\name{play_game}
\alias{play_game}
\title{Play one round of the Monty Hall game}
\usage{
play_game()
}
\arguments{
\item{NO}{arguments are used by this function.}
}
\value{
A data frame with two rows:
\code{strategy} : "stay" or "switch"
\code{outcome} : "WIN" or "LOSE"
}
\description{
\code{play_games()} determines the contestant wins a car or loses.
}
\details{
This function wraps all the steps: create a game, select a door, open a goat door #' and decide whether to stay or witch. It returns the outcomes for both strategies.
}
\examples{
play_game()

}
