% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_game}
\alias{play_game}
\title{Play a Game}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
A data frame showing the strategy and whether it
resulted in a win or loss.
}
\description{
Plays one Monty Hall game.
}
\details{
The function plays one game and compares what happens when
the player stays versus switches doors.
}
\examples{
play_game()

}
