% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{#'   Pay the Monty Hall game in its entirety.}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a 2 by 2 data frame
indicating the Win or Loss matching with Stay or
Change doors.
}
\description{
#'   \code{play_game()} uses the entire code already shown to
output the win or loss based on staying or changing doors.
}
\details{
#'   The game setup has already been established by the previous
code, so all that's left is to play the Monty Hall game. By
using this function, all previous functions are called in order
such that the final output is a data frame based on the options
to stay or change doors.
}
\examples{
  play_game()
  
}
