% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_game}
\alias{play_game}
\title{Create and play a Monty Hall problem game}
\usage{
play_game()
}
\arguments{
\item{No}{paremeters}
}
\value{
Returns game results win or lose
}
\description{
Adds up all the functions to play a game
}
\details{
No details
}
\examples{
 play_game()
}
