% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem edited.R
\name{play_game}
\alias{play_game}
\title{Playing the Monty Hall Game}
\usage{
play_game()
}
\arguments{
\item{This}{function does not have any arguements.}
}
\value{
The return is a table showing the contestant's choice and whether they won or lost.
}
\description{
`play is a wrapper function that plays the entire Monty Hall game from start to finish.
}
\details{
This function combines the functions \code{create_game}, \code{select_door},
\code{open_goat_door}, \code{change_door}, and \code{determine_winner} to complete the
entire Monty Hall game from the initial placing of goats and car to the final selection
and result from the contestant.
}
\examples{
  play_game()

}
