% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_game}
\alias{play_game}
\title{Play a new game}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
A data frame detailing the results.
}
\description{
Running all functions to simulate a new game.
}
\details{
This code is combining all functions in order to simulate a new game.
}
\examples{
play_game()
}
