% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{Test}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are Test.}
}
\value{
The function returns a Test
}
\description{
Test
}
\details{
Test
  Test
  Test
}
\examples{
  Test()

}
