% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{play_game}
\alias{play_game}
\title{Play the Monty Hall game}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a data frame of the game results
}
\description{
\code{play_game()} plays a single Monty Hall game and displays the outcome
}
\details{
The function plays a 3-door Monty Hall game with 2 goats and 1 car.
The outcome is displayed for whether the contestant chose to stay
with their initial door selection or switch to the other door.
}
\examples{
  play_game()

}
