% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\name{play_game}
\alias{play_game}
\title{play the game from the start}
\usage{
play_game()
}
\arguments{
\item{no}{parameters}
}
\value{
The strategy & the result
}
\description{
call all the previous functions
}
\details{
creating the game, selecting the first door, open the first door,
choosing to stay or to switch door, finding if win or lose
}
\examples{
  play_game()
  
}
