% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.r
\name{play_game}
\alias{play_game}
\title{Monty Hall Game}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
A table showing the return values, or outcomes, for staying and
  switching, given the inital pick.
}
\description{
`play_game()` runs one full Monty Hall game.
}
\details{
A game is created and the outcomes for staying and switching are
  determined for one initial pick.
}
\examples{
  play_game()

}
