% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{Play game in its entirety.}
\usage{
play_game()
}
\arguments{
\item{this}{function uses several objects and functions.
a 'new.game' is created with the create_game() function.
Then the 'first.pick' is immediately selected using the
select_door() function. Then the game reveals a goat door
using the open_goat_door() function and the newly created
objects new.game and first.pick. Then the final.pick objects
are defined for whether the player stays or switches their pick.
The two outcomes are defined whether the final pick is stay or
switch using the determine_winner() function. And finally, the
game.results object is created using all of the previous information
put into a data frame and the results of the game are revealed
to the user.}
}
\value{
The output shares the results for each step and reveals
whether the player won or lost the round.
}
\description{
'play_game()' will cycle through the entire game.
}
\details{
This function includes game setup all the way through
game results. When this function is run, the user will
find out whether staying with the initial pick or switching
would result in a win or a loss.
}
\examples{
play_game()
}
