% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{Plays a single game}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a data frame showing the results of
  a single game with the winning and losing outcomes from
  staying or not staying.
}
\description{
`play_game` generates a full game with both outcomes from either
  staying with the intial door, or switching doors.
}
\details{
This function first creates a new game along with the
  contestant's first door pick. It then opens a goat door.
  Next it creates a final pick for poss possiblities of
  staying vs not staying. Afterwards, it determines which
  of the two final picks was the winner and loser. Finally,
  it creates the results in a data frame and retuns the results.
}
\examples{
  play_game()

}
