% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-script.R
\name{play_game}
\alias{play_game}
\title{Play Game}
\usage{
play_game()
}
\arguments{
\item{None}{}
}
\value{
A WIN or a LOSS,
}
\description{
All the functions are put together to run a full game once.
}
\details{
Test to see if switching or staying with initial pick will produce
a WIN or a LOSS.
}
\examples{
play_game()

}
