% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_game}
\alias{play_game}
\title{Wrapper function for the game}
\usage{
play_game(num.cars, num.goats)
}
\arguments{
\item{...}{two integer positional arguments (number of cars and number of goats)}
}
\value{
returns the game result
}
\description{
Makes it easier to simulate
}
\details{
The function takes two positional arguments, number of cars,
and number of goats
}
