% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_game}
\alias{play_game}
\title{The strategy used by the contestant in the game.}
\usage{
play_game()
}
\arguments{
\item{No}{arguments used by the function.}
}
\value{
The function return character vector
}
\description{
\code{determine_winner()} The strategy used by the contestant assessed in this
session of the functions.
}
\details{
The pay off of the strategy used by the contestant was determine. Base on
the assessment of the strategy contestant was able to know whether the
strategy used is good.
}
\examples{
  new.game()

}
