% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_game}
\alias{play_game}
\title{Simulate one game}
\usage{
play_game()
}
\arguments{
\item{...}{No argument are required}
}
\value{
A data frame with columns: strategy and outcome
}
\description{
Simulates one round of Monty Hall game using both strategies: stay and switch
}
\details{
Simulates full game process and returns a data frame showing the outcome of both strategies
}
\examples{
play_game()
}
