% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newscriptmontyhall.R
\name{play_game}
\alias{play_game}
\title{Wrap the game steps into a single function.}
\usage{
play_game()
}
\arguments{
\item{...}{numeric and logical vectors of the game steps.}
}
\value{
the result of the game which is whether the contestant has won or lose.
}
\description{
play_game() wraps the game steps into a single function which includes within it all the required steps.
}
\details{
This is a wrap function that includes within it all other functions and by the end we can know depending on whether the contestant has chose to switch or stay whther he is a winner or loser.
}
\examples{
play_game()

}
