% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem (1).R
\name{play_n_games}
\alias{play_n_games}
\title{Play the Monty Hall game any number of times.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{...}{Argument: \code{n} (the amount of times you want to play the game)}
}
\value{
This function returns a dataframe reporting the outcomes and strategy for
each of the games it just played.
}
\description{
Use this to generate a data set to determine the best strategy for the game.
}
\details{
This function plays the game \code{n} numer of times (default n = 100), reporting
the outcome and strategy for all of the games in one data frame.
}
\examples{
  play_n_games(100)
  play_n_games(5000000)
  play_n_games(10)
}
