% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{Plays a new Monty Hall Problem game that repeats 'n' times.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{...}{the argument is the number of times the game will be simulated.}
}
\value{
The function returns a table with the proportions of losing and winning
if they choose to stay or switch their final selection door.
}
\description{
\code{play_n_games()} generates 'n' new games that consists of two doors
with goats behind them, and one with a car.
}
\details{
The game setup replicates the game on the TV show "Let's
Make a Deal" where there are three doors for a contestant
to choose from, one of which has a car behind it and two
have goats. The contestant selects a door, then the host
opens a door to reveal a goat, and then the contestant is
given an opportunity to stay with their original selection
or switch to the other unopened door. There was a famous
debate about whether it was optimal to stay or switch when
given the option to switch, so this simulation was created
to test both strategies. The game will be simulated 'n' times.
}
\examples{
  play_n_games()

}
