% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{Adding the Game to a Loop, Running Multiple Times}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
We will want a statistic outcome layout of strategies of staying versus switching, 
In this we will want to see if we win or lose
}
\description{
Adding a for loop to run the game multiple time to get results
}
\details{
When running simulations for inferential purposes the general rule of thumb is they should run at
 least 10,000 times in order for the simulated statistics to converge 
close to the actual theoretial value (average proportion of wins achieved by each strategy in this case).
In this case we will be running 100 times, n=100.
}
\examples{
  return( results.df )
}
