% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{play_n_games}
\alias{play_n_games}
\title{\if{html}{\out{<div class="sourceCode">}}\preformatted{ Play the game n times and store the result.
}\if{html}{\out{</div>}}}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{...}{no arguments are used by the function}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Return a table of multiple game results.
}\if{html}{\out{</div>}}
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Play the game multiple times.
}\if{html}{\out{</div>}}
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ Play the game multiple times and store the result.
}\if{html}{\out{</div>}}
}
