% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{Play the Monty Hall Many Times}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{"n"}{could be set to whatever number you want.}
}
\value{
A table will be returned with your results after
the game is played 100 times.
}
\description{
This program will play the Monty Hall game
multiple times.
}
\details{
The function is set up to play the Monty Hall
game "n" number of times. And "n" is set at 100
times. So the game will loop the plays and play
the game 100 times.
}
\examples{
play_n_games <- function( n=100 )

}
