% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{\preformatted{ Play a number of games
}}
\usage{
play_n_games(n = 100)

play_n_games(n = 100)
}
\arguments{
\item{the}{function can run for any integer of gameplays.
n=100 initially sets the game at 100 plays}
}
\value{
\preformatted{ the outcome of all games returned as a dataframe
 table
}

\preformatted{ the outcome of all games returned as a dataframe
 table. for example:

         outcome
         strategy LOSE  WIN
         stay   0.76 0.24
         switch 0.24 0.76
}
}
\description{
\preformatted{ plays the full version of the games a number of times
 initially set at 100 games
}

\preformatted{ plays the full version of the games a number of times
 initially set at 100 games
}
}
\details{
\preformatted{ plays the game and stores the results in a list named
 results.list. this is repeated via a loop
 n number of times.

 the results of the games are stored in a dataframe
 called results.df
}

\preformatted{ plays the game and stores the results in a list named
 results.list. this is repeated via a loop
 n number of times.

 the results of the games are stored in a dataframe
 called results.df
}
}
