% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_n_games}
\alias{play_n_games}
\title{Monty Hall Game Simulation}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n:}{number of times the game is played in the simulation}
}
\value{
'play_n_games(n)' returns a data frame with 2*n observations
of two chr variables: strategy ("stay" or "switch" ); and
outcome ("LOSE" or "WIN")
}
\description{
play_n_games(n) will repeat the game n times
}
\details{
It allows to evaluate the probability of winning depending on the chosen
strategy
}
\examples{
  P <- play_n_games(10000)
              outcome
  strategy  LOSE   WIN
      stay  0.66  0.34
    switch  0.34  0.66
  str(P)
 'data.frame':	20000 obs. of  2 variables:
  $ strategy: chr  "stay" "switch" "stay" "switch" ...
  $ outcome : chr  "LOSE" "WIN" "LOSE" "WIN" ...

}
