% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{Simulating the Game}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{...}{This function has one argument which is n = amount of games.}
}
\value{
This function returns a probability table using the dplyr package.
It shows the win rates for each strategy.
}
\description{
\code{play_n_games} is a function made to simulate the strategies
used in the game and to table the results. This function is designed
to test which strategy is better at winning the game by playing the
game many times and tallying up the wins.  For the most accurate
results, set n = 10000.
}
\details{
This function shows the results for each strategy for a specified
amount of games. It can simulate the game up to n amount of times.
}
\examples{
 plan_n_games(n=10000)
 
 To test this game you must insert an integer for n.
 
 Test example:
 play_n_games(n=10)
 play_n_games(n=100)
 play_n_games(n=1000)
 
 This should result in a proportions table that displays the 
 win rate of each strategy for the specified amount of games. 
 
}
