% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{Play multiple games.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n}{An integer representing the number of games to be played. Default value
is 100.}
}
\value{
The function returns the number of times WIN or LOSE was the result
based upon the final pick in each game.
}
\description{
\code{play_n_games()} simulates n number of entire playthroughs of the game.
}
\details{
The function will run n number of courses of the game and return the results
of all games.
}
\examples{
  play_n_games(n=100)

}
