% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{Play games simulator}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n:}{a number determined by the user. Defaults to 100.}
}
\value{
The function returns a table with the proportion of wins and losses for each strategy and
a dataframe showing the results for each game played in the loop.
}
\description{
Function performs the play_game() function n times
}
\details{
In order to determine which strategy is more effective, a statistically
significant sample size is necessary. In order to create such a sample, this function
simulates the game n times, returning the proportion of wins and
losses for both strategies (stay, switch).
}
\examples{
  play_n_games( 10000 )
  play_n_games( 500 )

}
