% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.r
\name{play_n_games}
\alias{play_n_games}
\title{Simulate the game}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n:}{the contestant can determine the number of times they would like to play the game.
Defaults to 100.}
}
\value{
The function returns a table with the wins and losses for each selection.
}
\description{
\code{play_n_games} is the function that allows the contestant to simulate the game
as many times \code{n} as they want.
}
\details{
This function is useful to determine which strategy is better to choose.
}
\examples{
play_n_games(100)
play_n_games(10000)
}
