% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem (4).R
\name{play_n_games}
\alias{play_n_games}
\title{Building a Simulation of 100 games}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{...}{Because the aim is to find the differences between staying and
switching, we are working with a sample size of at least 100 games played at
random to see the different outcomes of simulated games.}
}
\value{
The function returns the different outcomes, showing how many games
out of 100 were "WIN" and how many were "LOSE" after running the simulation.
}
\description{
\code{play_n_games( n=100 )} generates a simulation where 100 games are played
at random to determine the outcome of winning or losing.
}
\details{
Going through a simulation where a series of games are played at random
helps identify which strategy is best, when attempting to determine if
switching or staying is the better move to make to increase chances of
winning.
}
\examples{
  play_n_games( n=100 )

}
