% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{100 rounds of the Monty Hall Problem game}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns the number of wins and losses in a data frame.
}
\description{
\code{play_n_games()} runs the game 100 times and reports the results of wins and losses.
}
\details{
The function uses a loop to run the game 100 times and reports the number of wins and losses.
}
\examples{
  play_n_games()

}
