% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_n_games}
\alias{play_n_games}
\title{Run n iterations of the Monty Hall Problem and return
winning proportions of both strategies.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n}{A positive integer. Number of simulations to run.
Defaults to 100}
}
\value{
a proportion table of the observed winning and losing
percentages of each strategy.
}
\description{
The Monty Hall Problem is a famous mathematical probability
problem. This function allows one to simulate the problem n
times in order to explore the empirical probability of both
strategies for the game "stay" and "switch".
}
\details{
\code{play_n_games()} runs the \code{play_game()} function from this
package n times and stores the results of each strategy in
a data frame. This results are then summarized and reported
in a proportion table showing the winning and losing percentage
of each strategy.
}
\examples{
  play_n_games(1000)

}
