% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{Play the game 100x.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{uses}{the function( n=100 ) to iterate a simulation of 100x.}
}
\value{
Returns a dataframe of game results based on strategy of "stay" or "switch" along
with the corresponding statistical percentages.
}
\description{
Continually simulates playing the game 100 times.
}
\details{
This function will continually simulate the play_game() function 100 times. It
then will store the results into results.list and then bind them to a new data
frame, results.df. This will then run a statistical analysis of the game results
and generate a prop.table detailing the results.
}
\examples{
 play_n_games()
}
