% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{Play the game multiple times}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n}{A number}
}
\value{
A Dataframe with the results of the n number of games
}
\description{
\code{play_n_games()} Plays the game multiple times
}
\details{
This plays the game n number of specified times and stores the
results in a dataframe for later analysis
}
\examples{
  play_n_games(n=50)
}
