% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_n_games}
\alias{play_n_games}
\title{The game is run on a loop.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{...}{no arguments are used by the function, although the previously
indicated arguments are used by the previously defined functions within
this function.}
}
\value{
The function returns a table with two rows and three columns,
indicating the proportion of times each game strategy (stay/switch) results
in a given outcomes of the game (win/lose) depending on whether the
contestant decides to stay with the door they originally chose or decides
to switch to the door remaining after the host opens the goat door.
}
\description{
\code{play_n_games} allows the entire simulation, including all previously
defined functions, to be run repeatedly, 'n' number of times, in a single
step.
}
\details{
The simulation is run 'n' number of times to determine the proportion of
times that the contestant will win and lose the Monty Hall Problem game
based on their decision to stay with the door that they originally chose or
switch to the remaining door after the host opens a door with a goat behind
it.
}
\examples{
  play_n_games()
  
}
