% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{Runs 100 games providing cumulative results.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{none.}{}
}
\value{
Indicates the percentage of wins or loses for staying or switching over 100
games.
}
\description{
\code{play_n_games()} runs the monty hall game 100 times, compiles the results,
and provides a table that shows the percentages won or lost for each strategy.
}
\details{
The results show the win or lose rate for each strategy given the parameters
of the game. This win determine which strategy is better over a large sample
of attempts.
}
\examples{
Outcome: Staying--Won 30\% vs Lose 70\%. Switching--Won 41\% vs. Lose 59\%
}
