% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{\preformatted{Play the Game n times (by default, n=1000).
}}
\usage{
play_n_games(n = 1000)
}
\arguments{
\item{...}{"n", or the number of times to run the loop}
}
\value{
\preformatted{This function creates a dataframe that combines the outcomes of total games and lists then out of 1
}
}
\description{
\preformatted{creates the game structure on a loop to examine the outcomes 
of running the game a multitude of times. Gives probability by 
playing the game as many times as needed using `play_n_games()`.
}
}
\details{
\preformatted{The default for the game is 1000 times, but it can be played n time
Corresponding outcomes are listed for each game.
}
}
\examples{
	play_n_games(n=1000)

}
