% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{This function plays n Monty Hall games and returns the results as a table.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{This}{function takes one parameter n which dictates the number of Monty Hall games played with default 100.}
}
\value{
Returns a table of counts of wins vs. counts of losses by strategy.
}
\description{
This function plays n Monty Hall games and returns the results along side the strategy as a table.This function will allow you to evaluate which strategy "switch" or "stay" yields more wins on average.
}
\details{
This function plays n Monty Hall games and returns the results along side the strategy as a table.This function will allow you to evaluate which strategy "switch" or "stay" yields more wins on average.
}
\examples{
play_n_games(n=1000)

}
