% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_n_games}
\alias{play_n_games}
\title{Play 100 Monty Hall Games}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n, }{integer which is set to 100 here as default}
}
\value{
returns a data frame with n rows that have the game outcome.
}
\description{
play_n_games() utilizes play_game() and loops to cycle through play_game() n times. by default, n is set to 100
}
\details{
this function starts by setting n to equal 100, then loads the dplyr package, sets up a 'results.list' collector, and loop count of 1. Then the loop portion iterates one loop n times, which is by default set to be 100, it will do the following: create the argument game.outcome to hold the play_game() function, set that with the collector, and add an additional +1 to the loop count. next it sets the collector to become 'results.df', creates a table with a pipe operator for row proportions, number rounding, and prints it. finally, this function returns the results.df table
}
\examples{
  play_n_games( n=100 )
  
}
