% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{Play multiple rounds of Monty Hall Problem game}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n:}{a positive integer indicating number of times the
game played.}
}
\value{
a list of each game's results of both strategies (stay or switch) .
}
\description{
\code{play_n_game()}generates the results of multiple rounds
of random Monty Hall Problem game.
}
\details{
The function plays n rounds of Monty Hall Problem games.
It reports the proportion of WIN and LOSE for both strategies
(stay or switch) and lists the results of each game.
}
\examples{
 "play_n_game( n = 1000)"
}
