% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_n_games}
\alias{play_n_games}
\title{Create a simulation}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n}{= number of times you'd like the simulation to run}
}
\value{
Returns data frame outputted as a 4x4 table (switch/stay by win/lost)
}
\description{
This function runs the previous wrapper function 100 times and
stores the outcome each time.
}
\details{
This simulation provides the outcome in a table comparing wins and losses for
each of the two game strategies (switch and stay).
}
