% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{Play 100 Games to Determine Win Probability, given Stay or Switch Strategy.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{Parameter}{n=100 is the input for the number of games to run. The Rule of
Large Numbers implies that the greater n is, the nearer to the actual
probability will the outcome approach.}
}
\value{
Creates a dataframe of the number of Win and Lose game.outcome in a combined
results.list, based on contestant's strategy of Stay or Switch.
Returns a table showing the proportion of Win and Lose based on the
contestant's choice of stay with the original door choice or to switch to the
remaining mystery door.
}
\description{
\code{play_n_games()} plays n complete game iterations, n being 100, and create a
random sample of game results from staying and switching strategies in order
to determine the probability of each Win and Lose, given having either
switched or stayed in a given game.
}
\details{
Loops the game n=100 times to accumulate a sample from which the average
proportion of Win if the contestant chooses to stay compared to switch.
}
\examples{
play_n_games()

}
