% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_n_games}
\alias{play_n_games}
\title{Simulation of Multiple Monty Hall Problem Games}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
A 2 x 2 table displaying the proportion of lost and
won outcomes based on the strategy of "switch" or the strategy
of "stay."
}
\description{
\code{play_n_games()} simulates n number of Monty Hall Problem Games.
}
\details{
The user can choose the number of times they would like the
Monty Hall Problem Game to be simulated and this loop will run
the game n number of times.
}
\examples{
`play_n_games <- function (n = 50)`
`play_n_games <- function (n = 125)`

}
