% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{Play n Monty Hall games}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n}{An integer representing the number of games to play. Default is 100.}
}
\value{
A data frame containing the outcome of each game.
}
\description{
Play n Monty Hall games and summarize the results.
}
\details{
The function plays the Monty Hall game 'n' times, records the outcome of each game, and returns a table summarizing the proportion of wins and losses for each strategy (i.e., staying with the initial door selection or switching to another door).
}
\examples{
play_n_games(n = 1000)

}
