% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_n_games}
\alias{play_n_games}
\title{Step 07: Multiple Game Simulations}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{Vector}{}
}
\value{
This function returns a 2-column, 200-row table that provides both contestant strategy options in Column 1 (STAY, SWITCH), and
the result (WIN, LOSE) based on the game created and contestant choice.
}
\description{
'play_n_games()' runs 100 simulations of the game in Step 06.
}
\details{
This function produces a table reporting the results of 100 simulations using pairs of results for the STAY/SWITCH strategies.
}
\examples{
play_n_games()
}
