% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{Test or play multiple rounds of the game.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n}{... the numbers of the times for playing or testing the game.}
}
\value{
The function will return a dataframe containing the winning
and losing outcome that related to the two strategies
(switching and staying).
}
\description{
`play_n_games()` this function is to test the two
strategies of the game for multiple times.
}
\details{
This function combines all the other functions of Monty Hall Problem
game. For mulitple 'n' rounds, It tests the two strategies of the
game (changing or staying with the first selection).
}
\examples{
play_game( n = 10 )



}
