% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_n_games}
\alias{play_n_games}
\title{Monty Hall Game Loop}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{...}{The default value for n is 100, however, you can call the function with
your own value.}
}
\value{
It returns the results data frame.
}
\description{
This function loops through the game n amount of times in order to
determine the probability of winning or loosing based on whether or not
the participant stays with their original choice.
}
\details{
For every iteration of n, the function calls play_game() and stores the
results in an array using rbind. The function then displays the results
in a 2x2 table.
}
\examples{
play_n_games()
play_n_games(200)
play_n_games(n)
  
}
