% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{Play Games}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{run}{the loop 100 times}
}
\value{
return results in a table.
}
\description{
Play the game on a loop
}
\details{
Walk through the game multiple times return the dataframe of the game outcome
}
\examples{
for (i in 1:n) loop
}
