% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{Play game loops.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n}{Set equal to any number to change how many times the game
loop is played.}
}
\value{
A table with proportions of LOSE and WIN outcomes
from stay and switch strategies.
}
\description{
'play_n_games()' will play the Monty Hall game n amount of times
and return a table of game outcomes.
}
\details{
The function will automatically run the game 100 times (n = 100).
It will return a table reporting proportions of game outcomes
based on two strategies. These proportions should be rounded to
2 decimal places.
}
\examples{
  play_n_games(n = 1000)

}
