% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhallproblem.R
\name{play_n_games}
\alias{play_n_games}
\title{Plays the Monty Hall Problem n number of times.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n}{the number of times the game is played within the loop}

\item{results.list}{collects the results of the game loop}

\item{i}{iterator that plays the game n number of times}

\item{game.outcome}{places the play_game function into the loop}

\item{bind_rows}{binds the results and game outcome into the results list}
}
\value{
The output is a table that displays how many times the
contestant won or lost the game for each strategy. The result
is rounded to two decimal points for ease of viewing.
}
\description{
\code{play_n_games()} plays the Monty Hall game a set number of times
to see the average number of times a contestant will win depending
on their strategy.
}
\details{
This function works by using a for loop to play the game a certain
number of times. The results are collected into a list and analyzed
to determine the average number of wins and losses per strategy. The
package dplyr is required to perform this analysis.
}
\examples{
  play_n_games()

}
