% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_n_games}
\alias{play_n_games}
\title{Play the game 100 times}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{When}{the function is called a results.list is created which will store all 100
game results. Then the game will run through 100 iterations and the results
will be stored in a table which contains the proportions of wins and losses
by game strategy.}
}
\value{
returns a table which contains the proportion of each game-strategy's outcome.
}
\description{
\code{play_game()} this function was created in order to automatically run
100 games and receive the results of those 100 games. The results are
recorded and returned in a table to understand which game strategy produces
more "Wins".
}
\details{
With inferential simulations, best practice is to simulate the game
numerous times and record each outcome. This ensures that the simulated
statistical outcomes converge closely to the theoretical values.
The average proportion of wins achieved by each strategy in this scenario
are returned.
}
\examples{
 play_n_games()
}
