% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_n_games}
\alias{play_n_games}
\title{Play multiple Monty Hall games.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n}{A numeric value indicating the number of games to simulate. The default is 100.}
}
\value{
The function returns a data frame containing the strategy and outcome for each simulated game.
}
\description{
'play_n_games()' simulates multiple Monty Hall games and compares the outcomes of the stay and switch strategies.
}
\details{
The function repeatedly calls 'play_game()' for the specified number of games and combines the results into a single data frame. It also calculates and prints the proportion of wins and losses for the stay and switch strategies.
}
\examples{
play_n_games(n=10)
}
