% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{play_n_games}
\alias{play_n_games}
\title{Play the game 100 times}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n}{a value indicating the number of games}
}
\value{
This function returns a table of the probabilities of winning or losing
depending on if the player stays or switches doors.
}
\description{
\code{play_n_games()} replays the game "n" times
}
\details{
This function replays the game with every step mentioned previously and runs
through it "n" number of times
}
\examples{
play_n_games()

}
