% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhallfunctions.R
\name{play_n_games}
\alias{play_n_games}
\title{Looping play game.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n}{equals the number of games played.}
}
\value{
The function returns a table with proportions of LOSE and WIN
outcomes associated with each strategy.
}
\description{
The function 'play_n_games()' will simulate n number of Monty
Hall Problem games and return the table of outcomes.
}
\details{
'play_n_games()' will automatically run the game as many n times
as selected. In addition, the function will return the results
associated with each strategy (stay or switch) as a proportion
to two decimal places.
}
\examples{
  play_n_games(n = 100)
}
