% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_n_games}
\alias{play_n_games}
\title{Play game for 100 doors and store outcomes.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n}{The number of games to be played (default is 100).}
}
\value{
A data frame containing the outcomes for each game strategy (stay or switch) and whether the contestant won or lost.
}
\description{
This function simulates multiple rounds of the Monty Hall game, for a specified number of games, and returns the list of outcomes for each game. By default, it plays 100 rounds.
}
\details{
The function simulates multiple Monty Hall games, each with a contestant's initial choice, the host revealing a door, and the contestant making a final choice. It then stores the results of both strategies (stay and switch) and computes the proportion of wins for each strategy.
}
\examples{
Play 100 Monty Hall games and store the results
play_n_games(100)
}
