% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_n_games}
\alias{play_n_games}
\title{Play Multiple Monty Hall Games}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n}{The number of games to simulate. The default is 100.}
}
\value{
The function returns a data frame containing the strategy and outcome for each game.
}
\description{
Simulates multiple Monty Hall Problem games and summarizes the outcomes of the stay and switch strategies.
}
\details{
The function runs the Monty Hall game the specified number of times and returns
the results for each game. It also prints the proportion of wins and losses
for each strategy.
}
\examples{
play_n_games()

}
