% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{\itemize{
\item the game loop
}}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{-}{here, we set n to 100, which means we will play the game 100 times. In addition, we add a collector calld results.list, as well as an iterator which will run the actual simulation. Enclosed inside our simulator are game.outcome, results.list, as well as loop.count. The game.outcome stores the play_game function, while results.list stores loop count and feeds it the game.outcome. the loop count is then incremented by one.}
}
\value{
\itemize{
\item we are returned a results dataframe which binds rows from the results list collector. We then execute a table command using the results.df and convert our results into proportions.
}
}
\description{
\itemize{
\item we construct a loop in order to simulate the goat game as manu times as we want.
}
}
\details{
When running simulations for inferential purposes the general rule of thumb is they should run at least 10,000 times in order for the simulated statistics to converge close to the actual theoretial value (average proportion of wins achieved by each strategy in this case).
}
\examples{
##         outcome
}
