% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{Monty Hall game results.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{This}{function has the number of game rounds as a parameter.}
}
\value{
This function returns the results of multiple rounds of Monty Hall.
}
\description{
\code{play_n_games()} is a function that produces results of multiple rounds of Monty Hall.
}
\details{
This function lists the results of multiple rounds of the Monty Hall game.
}
\examples{
play_n_games()

}
