% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_n_games}
\alias{play_n_games}
\title{Play Multiple Games}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n}{An integer specifying the number of games to simulate (default is 100).}
}
\value{
A data frame containing the outcomes of all games played.
}
\description{
Simulates multiple games of the Monty Hall problem and summarizes the results.
}
\details{
Repeats the Monty Hall game \code{n} times, calculates proportions of wins for each strategy, and returns a summary.
}
\examples{
play_n_games(100)
}
