% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem (1).R
\name{play_n_games}
\alias{play_n_games}
\title{Play N Games}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{...argument}{is the number of games and here it is n=100}
}
\value{
This function returns a data frame of results
}
\description{
Play 100 games, bind the results in a data frame and then tabulate
them
}
\details{
Create a list to collect the results of every round (100 rounds in this case)
Iterate through the list until we fill all the results
Bind rows of results in data frame
Create a table of results
}
\examples{
play_n_games (n=100)

}
