% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_n_games}
\alias{play_n_games}
\title{Play multiple simulations of the Monty Hall game}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n}{represents the number of games that are simulated.}
}
\value{
A data frame that displays the results of each game in a
table format.
}
\description{
\code{play_n_games()} simulates the Monty Hall game being played
multiple times
}
\details{
The simulation runs the game 100 times by default, but can be
changed by changing "n". For each n, a new game is created, and
the Monty Hall game is played based off of the functions above.
This function calculates the amounts of wins and losses for all
the games played.
}
\examples{
  results <- play_n_games(10000)
    # shows results for 10,000 games for both staying and switching
  
}
