% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{Play multiple Monty Hall games.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n}{The number of games to play.}
}
\value{
A data frame summarizing the outcomes of all games played.
}
\description{
Play multiple Monty Hall games.
}
\examples{
play_n_games(100)

}
