% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_n_games}
\alias{play_n_games}
\title{Repeatedly run monty hall games for a specified
number of times.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{...}{n=100 or whatever desired number is the
argument in this function. It's used to specify the
number of times you want to repeatedly run the
function and then runs it. numeric and requires numeric
input.}
}
\value{
... A table of the probability of winning
or losing according to strategy (switching or staying).
}
\description{
play_n_games runs the monty hall games for a given
number of times and then then produces an table
displaying the proportion of the time that each
result would occur given the strategy taken (stay
or switch).
}
\details{
play_n_games requires the specification of how many
times one wants to run the function, provided by n=100
dplyr is then called. the function list() is stored
in results.list in order to create a tally of the
outcomes. 1 is then stored in the in loop.count
for( i in 1:n ) specifies the number of times the
games will be run. game.outcome stores the play_game()
function that holds all of the necessary components
of the game like create_game(), select_door(),
open_goat_door() and more. game.outcome is then stored
in results.listloop.count which applies the results
of the game to the list. loop.count + 1 is then assigned
to the object loop.count, binding to it the number of
loops.  dplyr::bind_rows( results.list ) is assigned
to results.df which works to organize the list into rows.
Then, table(results.df) is called to create the table
and piped into prop.table(margin = 1) to provide the
probabilities of winning and losing for each strategy,
rounding to two decimal places. print() prints this table
and return() provides the table. dplyr used for piping.
}
\examples{

my.premeditated.game <- c("car","goat","goat")
my_contestant_first_pick <- 1
my_play_game <- function( )
{
 new.game <- my.premeditated.game
 first.pick <- my_contestant_first_pick
 opened.door <- open_goat_door( new.game, first.pick )
 final.pick.stay <- change_door( stay=T, 
 opened.door = opened.door, 
 a.pick = first.pick )
final.pick.switch <- change_door( stay=F, 
opened.door = opened.door, 
a.pick = first.pick )
outcome.stay <- determine_winner( final.pick.stay, new.game  )
outcome.switch <- determine_winner( final.pick.switch, new.game )
strategy <- c("stay","switch")
outcome <- c(outcome.stay,outcome.switch)
game.results <- data.frame( strategy, outcome,
                            stringsAsFactors=F )
return( game.results )
}
play_n_games <- function( n=100 )
{
 library( dplyr )
results.list <- list()   # collector
loop.count <- 1
for( i in 1:n )  # iterator
{
game.outcome <- my_play_game()
results.list[[ loop.count ]] <- game.outcome 
loop.count <- loop.count + 1
}
results.df <- dplyr::bind_rows( results.list )
table( results.df ) \%>\% 
prop.table( margin=1 ) \%>\%  # row proportions
round( 2 ) \%>\% 
print()
return( results.df )
}
play_n_games(n=100)
my.premeditated.game <- c("car","goat","goat")
my_contestant_first_pick <- 2
my_play_game <- function( )
{
new.game <- my.premeditated.game
first.pick <- my_contestant_first_pick
opened.door <- open_goat_door( new.game, first.pick )
final.pick.stay <- change_door( stay=T, 
opened.door = opened.door, 
a.pick = first.pick )
final.pick.switch <- change_door( stay=F, 
opened.door = opened.door, 
a.pick = first.pick )
outcome.stay <- determine_winner( final.pick.stay, new.game  )
outcome.switch <- determine_winner( final.pick.switch, new.game )
strategy <- c("stay","switch")
outcome <- c(outcome.stay,outcome.switch)
game.results <- data.frame( strategy, outcome,
                             stringsAsFactors=F )
return( game.results )
}
play_n_games <- function( n=100 )
{
library( dplyr )
results.list <- list()   # collector
loop.count <- 1
for( i in 1:n )  # iterator
{
game.outcome <- my_play_game()
results.list[[ loop.count ]] <- game.outcome 
loop.count <- loop.count + 1
}
results.df <- dplyr::bind_rows( results.list )
table( results.df ) \%>\% 
prop.table( margin=1 ) \%>\%  # row proportions
round( 2 ) \%>\% 
print()
return( results.df )
}
play_n_games(n=100)
my.premeditated.game <- c("car","goat","goat")
my_contestant_first_pick <- 3
my_play_game <- function( )
{
new.game <- my.premeditated.game
first.pick <- my_contestant_first_pick
opened.door <- open_goat_door( new.game, first.pick )
final.pick.stay <- change_door( stay=T, 
opened.door = opened.door, 
a.pick = first.pick )
final.pick.switch <- change_door( stay=F, 
opened.door = opened.door, 
a.pick = first.pick )
outcome.stay <- determine_winner( final.pick.stay, new.game  )
outcome.switch <- determine_winner( final.pick.switch, new.game )
strategy <- c("stay","switch")
outcome <- c(outcome.stay,outcome.switch)
game.results <- data.frame( strategy, outcome,
                              stringsAsFactors=F )
return( game.results )
}
play_n_games <- function( n=100 )
{
library( dplyr )
results.list <- list()   # collector
loop.count <- 1
for( i in 1:n )  # iterator
{
game.outcome <- my_play_game()
results.list[[ loop.count ]] <- game.outcome 
loop.count <- loop.count + 1
}
results.df <- dplyr::bind_rows( results.list )
table( results.df ) \%>\% 
prop.table( margin=1 ) \%>\%  # row proportions
round( 2 ) \%>\% 
print()
return( results.df )
}
play_n_games(n=100)

my.premeditated.game <- c("car","goat","goat")
my_play_game <- function( )
{
new.game <- my.premeditated.game
first.pick <- select_door()
opened.door <- open_goat_door( new.game, first.pick )
final.pick.stay <- change_door( stay=T, 
opened.door = opened.door, 
                              a.pick = first.pick )
final.pick.switch <- change_door( stay=F, 
opened.door = opened.door, 
                               a.pick = first.pick )
outcome.stay <- determine_winner( final.pick.stay, new.game  )
outcome.switch <- determine_winner( final.pick.switch, new.game )
strategy <- c("stay","switch")
outcome <- c(outcome.stay,outcome.switch)
game.results <- data.frame( strategy, outcome,
                         stringsAsFactors=F )
return( game.results )
}
play_n_games <- function( n=100 )
{
library( dplyr )
results.list <- list()   # collector
loop.count <- 1
for( i in 1:n )  # iterator
{
game.outcome <- my_play_game()
results.list[[ loop.count ]] <- game.outcome 
loop.count <- loop.count + 1
}
results.df <- dplyr::bind_rows( results.list )
table( results.df ) \%>\% 
prop.table( margin=1 ) \%>\%  # row proportions
round( 2 ) \%>\% 
print()
return( results.df )
}
play_n_games(n=100)
my_contestant_first_pick <- 3
my_play_game <- function( )
{
new.game <- create_game()
first.pick <- my_contestant_first_pick
opened.door <- open_goat_door( new.game, first.pick )
final.pick.stay <- change_door( stay=T, 
opened.door = opened.door, 
                          a.pick = first.pick )
final.pick.switch <- change_door( stay=F, 
opened.door = opened.door, 
                        a.pick = first.pick )
outcome.stay <- determine_winner( final.pick.stay, new.game  )
outcome.switch <- determine_winner( final.pick.switch, new.game )
strategy <- c("stay","switch")
outcome <- c(outcome.stay,outcome.switch)
game.results <- data.frame( strategy, outcome,
                              stringsAsFactors=F )
                              return( game.results )
}
play_n_games <- function( n=100 )
{
library( dplyr )
results.list <- list()   # collector
loop.count <- 1
for( i in 1:n )  # iterator
{
game.outcome <- my_play_game()
results.list[[ loop.count ]] <- game.outcome 
loop.count <- loop.count + 1
}
results.df <- dplyr::bind_rows( results.list )
table( results.df ) \%>\% 
prop.table( margin=1 ) \%>\%  # row proportions
round( 2 ) \%>\% 
print()
return( results.df )
}
play_n_games(n=100)
play_n_games(n=100)
play_n_games(n=100)
play_n_games(n=100)


}
