% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{play_n_games}
\alias{play_n_games}
\title{Allows one to simulate the game however many times they want.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{The}{only paramter that is passed into this function is an integer
that will tell the function how many times to run the \code{play_game()}
function.}
}
\value{
This function returns a dataframe of however many games that is
set in the parameter. Telling the contestant for each case whether
or not each strategy was a Win or Lose.
}
\description{
\code{play_n_games} is the function that will allow the contestant to
simulate the game an \code{n} amount of times.
}
\details{
This function is useful to run if one wants to play the game over
and over to try and find out which strategy was dominant over a
set number of times.
}
\examples{
 `play_n_games( n = 100 )`
 `play_n_games( 100 )`
 `play_n_games( 10000 )`

}
