% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg (1).R
\name{play_n_games}
\alias{play_n_games}
\title{Play Multiple Monty Hall Games}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n}{An integer specifying the number of games to simulate (default is 100).}
}
\value{
A data frame summarizing the results of the simulations.
}
\description{
Simulates multiple Monty Hall games and summarizes the results.
}
\details{
The function simulates \code{n} games and calculates the proportion of wins for both staying and switching strategies.
}
\examples{
play_n_games(n = 100)
}
