% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg (2).R
\name{play_n_games}
\alias{play_n_games}
\title{play_n_games}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{#no}{arguments are used by the function.}
}
\value{
table of results
}
\description{
play_n_games() plays the game in a loop
}
\details{
play_n_games() plays the game in a loop
}
\examples{
play_n_games()
}
