% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_n_games}
\alias{play_n_games}
\title{Play multiple Monty Hall games.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n}{A numeric value indicating the number of games
to play.}
}
\value{
The function returns a data frame containing the
strategy and outcome for each game.
}
\description{
\code{play_n_games()} plays multiple Monty Hall games.
}
\details{
The function repeatedly plays the Monty Hall game and
combines the results for the stay and switch strategies.
}
\examples{
  play_n_games( n=100 )

}
