% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall-pkg.R
\name{play_n_games}
\alias{play_n_games}
\title{Play game simulation/loop}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{The}{argument for this function is n where you can set it n equal to the
number of times you wish to simulate the game.}
}
\value{
Returns a table listing the numeric count of wins and losses for the
strategies.
}
\description{
This function allows you to loop the \code{play_game()} function n times to get
a better idea of which strategy works best.
}
\details{
This function will currently run the game 100 times and print those
results in a table to show how many times each strategy (stay or switch)
wins and losses.
}
\examples{
   play_n_games (n=100)
   play_n_games (n=1000)
   play_n_games (n=10000)

}
