% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{play multiple games}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{details}{}
}
\value{
details
}
\description{
determine_winner
}
\details{
details
}
\examples{
details
}
