% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{Play the Game 100 Times}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{play_n_games}{<- function( n=100 )
results.list <- list()   # collector
loop.count <- 1
game.outcome <- play_game()
results.list[\link{ loop.count }] <- game.outcome
loop.count <- loop.count + 1
results.df <- dplyr::bind_rows( results.list )
table( results.df )}
}
\value{
The results of 100 game plays are returned.
}
\description{
Uses a loop to simluate playing the game 100 times.
}
\details{
A function is created that loops the game until it is played 100 times.
The results of the games are returned.
}
\examples{
  WIN: 70 LOSE: 30
}
