% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{Runs the game a given amount of times and shows results.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{`n`}{\itemize{
\item Input to determine the number of times the game is run.
}}
}
\value{
The function returns a results table showing the proportional
results of \code{n} runs of the game.
}
\description{
\code{play_n_games()} plays the game a number of
times based on the numerical parameter passed to it.
Proportional results are then shown.
}
\details{
The function uses a \code{for} loop that continues based on
the number parameter passed into it. The outcome are recorded in a
results list. A table showing the proportional results is then created
and returned.
}
\examples{
  play_n_games(5)
  play_n_games(90)

}
