% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_n_games.R
\name{play_n_games}
\alias{play_n_games}
\title{Lets Play x 100}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{...}{n=100}
}
\value{
return( results.df )
}
\description{
Plays the game a hundred times
}
\details{
PLays a hundred time and shows your results after those hundred and it determine
what had the better winning percentage: staying or winning
}
\examples{
   play_n_games()
}
