% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_n_games}
\alias{play_n_games}
\title{Play the game (n number of times)}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n=}{a value that represents the number of times the game is played. The default is
100 if not specified.}
}
\value{
The function returns the result of the number of games played in the form of
a data frame and prints the proportions.
}
\description{
A contestant plays the game but you choose how many times.
}
\details{
The contestant plays the game with the chosen amount of times. The proportion
is then shown to reveal the proportion of games that the contestant
won and lost.
}
\examples{
play_n_games()
}
