% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_n_games}
\alias{play_n_games}
\title{Adding the Game to a Loop}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{The}{parameters to this function include the previous steps being completed to provide
the game's probability of each result when ran n times.}
}
\value{
The function returns a table of the strategies and the probability of either "LOSE" or "WIN".
}
\description{
"play_n_games' function plays a certain number of games and provides the
probability of how likely the strategy will result in each outcome.
}
\details{
This would be a full game played without individual steps, providing a table of the strategy
and the resulting probability of each outcome based on that choice.
}
\examples{
play_n_games( )

}
