% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_n_games}
\alias{play_n_games}
\title{Probability Table}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{Numeric}{}
}
\value{
The function will be a numeric value that shows the win and loss rates.
}
\description{
'play_n_games()' function lets us see the probability charts for
the wins and losses of the game.
}
\details{
This shows us the probability when 100 games are played.
}
\examples{
   play_n_games()
}
