% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_n_games}
\alias{play_n_games}
\title{Play multiple Monty Hall games.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n}{A numeric value indicating the number of Monty Hall
games to simulate. The default is 100.}
}
\value{
The function returns a data frame containing the
  stay and switch outcomes for all simulated games.
}
\description{
`play_n_games()` runs the Monty Hall game multiple times
  and records the outcomes for both the stay and switch
  strategies.
}
\details{
The function repeatedly calls `play_game()` for the number
  of games specified by `n`. It combines the results into
  one data frame and prints the proportion of wins and losses
  for each strategy.
}
\examples{
  play_n_games(n=10)

}
