% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_n_games}
\alias{play_n_games}
\title{Simulate "Lets Make A Deal" the number of times you'd like.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n}{a number, represnting the number of times to simulate
game, default is 100 games.}
}
\value{
a data frame, with one row per strategy for each game
simulated.
}
\description{
\code{play_n_games()} simulates the game a number of n times, the
default is 100 games.
}
\details{
The game setup replicates the game on the TV show "Let's
Make a Deal" where there are three doors for a contestant
to choose from, one of which has a car behind it and two
have goats. The contestant selects a door, then the host
opens a door to reveal a goat, and then the contestant is
given an opportunity to stay with their original selection
or switch to the other unopened door. There was a famous
debate about whether it was optimal to stay or switch when
given the option to switch, so this simulation was created
to test both strategies. This simulation repeats the game n
times. Default is 100.
}
\examples{
  play_n_games (n=100)

}
