% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{Play Multiple Monty Hall Problem Games}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
A data frame with the outcomes of each simulated game.
}
\description{
Simulates multiple Monty Hall Problem games and records the outcomes.
}
\details{
The \code{play_n_games()} function enables you to simulate numerous Monty Hall Problem game instances.
It iterates the \code{play_game()} function 'n' times, records the results (WIN or LOSE) for each game,
and presents them in a data frame. The 'n' parameter allows you to specify the number of games to simulate.
}
\examples{
  # Simulate 50 Monty Hall Problem games
  play_n_games(50)

}
