% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem functions Hauer.R
\name{play_n_games}
\alias{play_n_games}
\title{Monty Hall Simulation}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{...}{no arguments are used by the simulation.}
}
\value{
a table of the outcome proportions and
returns the results data frame.
}
\description{
A function that simulates 'n' number of Monty Hall game plays and
returns the outcome of each game.
}
\details{
We use the play_game function to simulate one game, and use the dplyr library
to bind the outcome of each game into a single data frame.
The function then creates outcome proportions and a results data frame.
}
\examples{
play_n_games()

}
