% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{Play games for n times}
\usage{
play_n_games(n = 100, question = 1)
}
\arguments{
\item{n}{number of times game is played}

\item{question}{question number (default = 1)}
}
\value{
The function returns a table containing the results obtained
        by running play_games()  n number of times.
}
\description{
`play_n_games()` runs the game for n times and notes the reults.
}
\details{
The function calls play_game( ) function repeatedly
  The results obtained from play_game() as return value
  are noted and displayed as a table
}
\examples{
  play_n_games(1000,2)
}
