% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_n_games}
\alias{play_n_games}
\title{Plays the Monty Hall game in x amount of times.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{...}{the function takes in the variable/object, n. n represents an
integer value which determines how many times the function will iterate the
Monty Hall game.}
}
\value{
It will return a data frame, after x amount of iterations, containing the
probability of winning and losing if a user decides to switch or stay.
}
\description{
\code{play_n_games(n)} returns a data frame, after x amount of iterations, of the
probability of winning and losing if user decides to switch or stay.
}
\details{
The function iterates for x amount of times, depending on the integer input
(n), and will output a data frame into a table indicating the wins and loses
for staying and switching.
}
\examples{
  play_n_games(n=10)

}
