% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{Simulate the monty hall game}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{parameters}{are set around the number of iterations the game will have}
}
\value{
return(results.df)
}
\description{
Now that the game has been created and packaged into one function,
we will simulate the game using loops to test whether staying
or switching is the dominant strategy for the contestant
}
\details{
the game is played a set number of times using 'n'
the 'for()' argument is made to iterate the number of time the game is played
the results are then displayed in a dataframe 'results.df'
}
\examples{
table(results.df)
outcome ->
strategy -> stay: Lose-60, Win-40; switch: Lose: 40, Win: 60
}
