% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{This is the table that shows you the results of staying and switching.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
wins and losses in a table.
}
\description{
The table will catalog all wins and losses for 100 games to give the viewer 
an understanding of the best strategy for playing Monty Hall.
}
\details{
This simulates 100 games playing a stay strategy, or playing the same 100 games
with a switch strategy.
}
\examples{
play_n_games

}
