% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_n_games}
\alias{play_n_games}
\title{Looped Monty Hall Problem game simulator.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{`n`}{is a numeric value specifying the number of games to play. It
defaults to 100.}
}
\value{
A data frame with the results of each individual game. The data
frame has two columns where \code{strategy} represents the possible choices
from \code{change_door()} ("stay" or "switch") and \code{outcome} represents the
outcomes for each strategy ("win" or "lose"). It also returns a proportion
table summarizing the frequency of wins and losses by strategy.
}
\description{
\code{play_n_games} simulates multiple Monty Hall Problem games from start to
finish.
}
\details{
This function simulates multiple Monty Hall Problem games from start to
finish. It creates a data frame of results for a user-defined number of
games and summarizes the results in a proportion table to show the
likelihood of winning and losing based on strategy.
}
\examples{
  play_n_games()
}
