% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{Let us simulate the Monty Hall problem n times!}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{...}{This function takes only one argument, a whole number n, representing
the number of times the game should be played.}
}
\value{
This function returns a dataframe with labels "stay" and "switch" in column strategy,
and win and lose as the other columns. In these columns, the proportion of outcomes corresponding
to the relevant strategy and outcome are populated.
}
\description{
\code{play_n_games()} will simulate the Monty Hall problem
a specified n number of times. Returning the aggregated results
of the simulations in a combined table.
}
\details{
This function uses iterative looping to run the \code{play_game()} function
a specified n times. It will store the results of each simulation in the same
list, which is then binded together to form a dataframe. This attempts to find
an answer to the Monty Hall problem through iterative runs of the game and
presenting the compiled results.
}
\examples{
play_n_games(n=200)
}
