% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_n_games}
\alias{play_n_games}
\title{Adding a Game Loop}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{In}{the play_n_games() the function indicates that the game will repeat 100
separate times.}
}
\value{
The return value will give a percentage of the wins/losses of those 100 games
and also indicate the number of wins/losses if you choose to "STAY" or to
"SWITCH".
}
\description{
This code chunk indicates running the simulation 100 times to help with
result randomization.
}
\details{
The for() loop indicates that the game will run 100 separate times. From here,
each time that play_game() function is called, it will generate a new
'game_outcome".
}
\examples{
For example, the results could show something similar to this:
}
