% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem (1).R
\name{play_n_games}
\alias{play_n_games}
\title{Putting the game into a Loop}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n}{= number of games played}
}
\value{
table with wins and loses
}
\description{
Creating a loop to play the game enough times to compare the strategies of staying vs switching
}
\details{
The game is played n times the results reported as total wins and total loses
}
\examples{
play_n_games (n=1000)
}
