% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_n_games}
\alias{play_n_games}
\title{Simulate game N times}
\usage{
play_n_games(n = 100)
}
\value{
Frame of wins and losses
}
\description{
Runs loop of function for n rounds
}
\details{
Plays n times and presents results and win/loss rate
}
\examples{
  play_n_games(100)
}
