% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{play_n_games}
\alias{play_n_games}
\title{Repeating the game n times to see results}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n}{amount of times the game will be run}
}
\value{
Data frame of the results and the strategy used
}
\description{
Runs the game n amount of times and collects the results and strategy used.
}
\details{
Runs the game n amount of times, then collects the reults into a data frame based on if the results are "WON" or "LOST" and the strategy used - "stay" or "switch"
}
\examples{
play_n_game(100)
}
