% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_n_games}
\alias{play_n_games}
\title{Simulate multiple Monty Hall games.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{`n`}{An integer specifying the number of games to simulate. Default is 100.}
}
\value{
A data frame with the results of each game, including the strategy used and the outcome for that strategy.
}
\description{
play_n_games()` runs a specified number of Monty Hall game simulations and calculates the proportion of wins and losses for both staying with the initial pick and switching to the other door.
}
\details{
This function simulates a given number of Monty Hall games using \code{play_game()}. It then collects the results of each game and combines all results into a single data frame. Afterwards, it computes and prints the proportion of wins and losses for each strategy. Finally, it returns a data frame with the combined results of all simulations.
}
\examples{
`play_n_games()`

}
