% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_n_games}
\alias{play_n_games}
\title{Create a loop to run the game 100 times}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The return will be a table with the relevant game data
}
\description{
Prove whether it is better to swap or stay
with data by running the game multiple times
}
\details{
.
By creating a loop, the game can be ran several
times which will record data to justify whether
it is better to swap or stay during the game.
}
\examples{
play_n_games
}
