% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty_hall_pkg1.R
\name{play_n_games}
\alias{play_n_games}
\title{Play multiple rounds of the Monty Hall Problem game.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n}{a number indicating the number of rounds of the Monty Hall Problem game.
The default value is 100.}
}
\value{
The function returns a dataframe indicating the win and lose rate
for each strategy. The rows represent the strategy ("stay" or "switch"), and
the columns represent the outcome ("WIN" or "LOSE").
}
\description{
\code{play_game()} simulates multiple rounds of the Monty Hall Problem game
and returns the final outcome for both scenarios in which the
contestant engages in the "stay" and "switch" strategies for all games played.
}
\details{
\code{play_n_games()} repeatedly run the \code{play_game()} function for \code{n} times.
.  This function is defaulted to simulate 100 games. It combines the results
of all games played into a single list or dataframe and creates a summary table
showing the win and lose rate of both the "stay" and "switch" strategies based on
}
\examples{
  play_n_games()

}
