% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{Play the game 100 times}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{'n'}{\itemize{
\item can be set to equal the number of times the game is going to be
played
}}
}
\value{
'results.df' is returned and shows a table with proportions of wins and
loses based on strategy.
}
\description{
'play_n_game()' allows the game to be played many times, 100 in this case
}
\details{
The n can be changed to whatever number a user wants. The game can be
played many times. An important purpose of this is to determine the
probability of a win versus a lost based on switching or staying, so
that it can be determined which strategy is more likely to have a good
outcome.
}
\examples{
    play_n_games()

}
