% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHallKRB.R
\name{play_n_games}
\alias{play_n_games}
\title{Determine the likelihood of winning a car utilising the Monty Hall Problem.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{game.outcome}{}
}
\value{
Returned is a proportion table of wins and losses by guest door choice, of class dataframe.
}
\description{
A loop generates the game results ("Win" or "Lose") for 100 iterations of the Monty Hall Problem.
Results are assigned to a collector vector list and displayed as a table of proportions.
}
\details{
The two game outcomes of play_game, win or lose, are played out 100 times using the conditions of
play_game, or the Monty Hall Problem. The proportion of wins and losses, by stay or change choice,
is displayed in the table.
}
\examples{
results.df
}
