% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_n_games}
\alias{play_n_games}
\title{Game simulation function.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{`n`}{Integer for number of simulations}
}
\value{
Proportion table and data frame of outcomes
}
\description{
\code{play_n_games} performs a simulation to play through the game and switch or
keep the door multiple times to produce an output table to indicate \% of
wins or losses for each strategy.
}
\details{
The function loops through play_game() multiple times and returns data frames
of outcomes and propotion of wins and losses for each strategy
}
\examples{
play_n_games(10)
}
