% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem Script.R
\name{play_n_games}
\alias{play_n_games}
\title{Adding the Game to a Loop}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n=100}{Numeric, 100 times iteration}
}
\value{
a table includes strategy (Switch or Stay), Outcome(WIN or Lose),
and the probability of each outcome according to the chosen stategy.
}
\description{
\code{play_n_games(n)} add the Monty Hall game to a loop consist of 100 iteration
and then tabulate the results of number of wins and loses out of this
iteration
}
\details{
When running simulations, at least we should run the game 100 time to get a
result that we can rely on. In this function we simulate the game 100 time,
then the function will calculate the probability of winning and loosing
according to the contestant choice/strategy (Stay or switch) out of 100
time iteration
}
\examples{
play_n_games()

}
