% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{Simulate the Monty Hall Problem game.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n}{numeric vectors, length 1, default at 100}
}
\value{
The function returns a 2 by 2 data frame
indicating the probability of win or lose for stay
and switch strategies.
}
\description{
\code{play_n_games()} simulates the game by n rounds/ default at 100 and
generates the strategy table with the probability of stay or switch to
the respective chance on win or lose.
}
\details{
The function will create a list to document the game result with a loop
by n rounds, append each additional game result to the list according
to the index position. and transform the list into a table with row
proportions rounded to the second decimal digits.
}
\examples{
  play_n_games()

}
