% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_n_games}
\alias{play_n_games}
\title{Play multiple rounds of the Monty Hall Game.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n}{An integer specifying the number of games to simulate.}
}
\value{
A data frame containing the results of all simulated games, with columns
for \code{strategy} and \code{outcome}. Additionally, the function prints the
proportions of wins for each strategy.
}
\description{
\code{play_n_games()} simulates multiple rounds of the Monty Hall game and calculates
the proportion of wins for each strategy ("stay" or "switch") across all
rounds.
}
\details{
This function performs the following steps:
\enumerate{
\item Runs the \code{play_game()} function \code{n} times, where \code{n} is the number of
games to simulate.
\item Collects the results of each game into a list.
\item Combines the results from all games into a single data frame.
\item Calculates and prints the proportion of wins for each strategy based on
the collected data.
}
}
\examples{
  play_n_games(n = 100)
}
