% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_n_games}
\alias{play_n_games}
\title{Playing Multiple Games}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n}{An integer specifying the number of games to simulate. Default is 100}
}
\value{
A data frame with the results of all games.
The proportions of wins for each strategy are also printed.
}
\description{
Simulates multiple games of Monty Hall
and calculates the proportion of wins for each strategy.
}
\details{
This function runs a specified number of Monty Hall games and computes
the proportion of wins for both the stay and switch strategies.
It returns a data frame with the results and prints the proportions.
}
\examples{
play_n_games(n = 100)
}
