% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_n_games}
\alias{play_n_games}
\title{Simulate user-defined number of Monty Hall games.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n:}{a number corresponding to the number of games to be run.}
}
\value{
returns a data frame with a strategy column with 'stay' or 'switch' options and a result column showing whether the strategy won or lost the game for each simulation.
}
\description{
\code{play_n_games()} function simulates x number of Monty Hall games and returns a data frame
of results with a strategy column with the 'stay' or 'switch' options and a result column showing
whether the strategy won or lost the game.
}
\details{
This function requires one argument.
}
\examples{
# play 100 games
play_n_games(100)
# play 50 games
play_n_games(50)
}
\seealso{
Other montyhall: 
\code{\link{change_door}()},
\code{\link{create_game}()},
\code{\link{determine_winner}()},
\code{\link{open_goat_door}()},
\code{\link{play_game}()}
}
\concept{montyhall}
