% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Montyhall.R
\name{play_n_games}
\alias{play_n_games}
\title{Adding the game to a loop}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n}{& i are numeric vectors.}
}
\value{
The function returns the results as proportion figures indicating the proportion of the wins and losses of
      each strategy
}
\description{
The game will be added to a loop in order to be played 100 times for inferential purposes.
}
\details{
The game will be added to a loop and played 100 times, as the simulated results will help us get closer to
       the actual theoretical value (i.e. the proportion value of wins achieved by each strategy). Our loop consists
       of three parts; the collector, iterator & the binding step. In order to make it visually more appealing,
       by using the table proportions function, prob.table() in order to report results as proportions for each
       strategy.
}
\examples{
        play_n_games()
}
