% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{Monty Hall simulation}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{...}{arguments used are number of simulations user specifies}
}
\value{
number of times where stay and switch strategies show WIN and LOSE outcomes.
}
\description{
Allows user to run the simulation for n times to know the WIN/LOSE probability.
}
\details{
The function helps us understand game outcomes if contestant stays or switches
and it takes into account the number of times the contestant plays.Results are stored in
a data frame and returns probabilities of the two game strategies that are stay and switch.
}
\examples{
play_n_games()


}
