% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_n_games}
\alias{play_n_games}
\title{Play multiple Monty Hall games}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n}{Number of games to simulate.}
}
\value{
A data frame containing the results of all simulated games.
}
\description{
Simulates many Monty Hall games and summarizes the results.
}
\details{
The function repeats the Monty Hall simulation \code{n} times and prints
the proportion of wins and losses for each strategy.
}
\examples{
play_n_games(10)

}
