% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{\preformatted{Run Mulitple Simulations of the Monty Hall Game
}}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
\preformatted{This function will return a data frame with the strategy
and percentage chance of winning or losing for each 
strategy. Strategies include staying or switching. 
}
}
\description{
\preformatted{`play_n_games()` will play the Monty Hall Game several 
times based on the input of n for number of games. 
}
}
\details{
\preformatted{ This funtion will run all the Monty Hall Game functions
 as a loop. It will play thorugh an entire game of the 
 "Let's Make a Deal" process. The loop will run based on 
 the number of games indicated by n. 
}
}
\examples{
    play_n_games()
    
}
