% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{Do it again}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{results.lists}{holds the value of lists which as been mentioned in the comment
is the collector of this information within the iterator. loop count minimum
has to be at least 1.  the for i is the amount specified within the how many
times the loop shall run and 1:n is the range for the data to which be
collected. game.outcome is from play_game which will simulate the game.
results list will then be saved in the game outcome. loop count will always
have to have a minimum of 2 for it to be conisdered technically a loop.
results.df is being saved from the results list and being bound from the
two data frames that are coming from each loop which will technically be it's
own dateframe. Then it is all formated nice
to a cleaner format. then}
}
\value{
results.df which is being bound together from the various loop simulations
}
\description{
this export is specially looking at the repetition that allows the game to
be played as many times are the code requires
}
\details{
The purpose for this piece is to look at the game play for a specified amount
of times that the user decides to input. The purpose for this is that it will
then allow us to simulate to the game as many times as we want in order to
get an accurate assessment on the best option for the game player to either
stay or switch.
}
\examples{
play_n_games(n=1000)
play_n_games(n=10)
play_n_games(n=50)
}
