% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_n_games}
\alias{play_n_games}
\title{Monty Hall Game Simulation}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n}{Integer for number of game simulations}
}
\value{
Data frame and proportion tables
}
\description{
'play_n_games' plays through the Monty Hall game simulation 'n' number of times
to produce an output table to determine the percentage of wins or losses
for staying or switching.
}
\details{
This function loops through the game 100 times and returns a data frame of
game outcomes and proportion of wins and loses for the strategy of staying
and the strategy of switching.
}
\examples{
play_n_games(100)
}
