% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problemb.R
\name{play_n_games}
\alias{play_n_games}
\title{Plays multiple games at once.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{...}{one arguments are used by the function.}
}
\value{
The function returns a data frame with the overall
results of the n = x number of games ran.
}
\description{
\code{play_n_games()} generates an option to play multiple games for analysis.
}
\details{
This function allows for running analysis of win/loss based on
contestant decisions of stay or switch.
}
\examples{
  play_n_games()
}
