% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen comments.R
\name{play_n_games}
\alias{play_n_games}
\title{Play 100 full games to see the probability of winning}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{The}{parameter n=100 represents the number of times the game will be ran
to determine the probabilities of winnign and losing.}
}
\value{
This function returns a table showing the proportion of wins and loses
if the player stays with the original pick or switches to the remaining door.
}
\description{
\code{play_n_games()} runs through 100 iterations of the full game to determine
the probability of winning if the player stays or switches.
}
\details{
This function adds a loop to the game, playing it "n" number of times to tell
us the average proportion of wins the player can achieve with each strategy
(stay or switch).
}
\examples{
play_n_games()

}
