% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{play_n_games}
\alias{play_n_games}
\title{Simulates the Monty Hall Problem game}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{This}{function does not rely on any arguements but
requires the use of the 'dplyr' package.}
}
\value{
This function will return a data frame detailing the
final results (win or lose) based on the strategy
chosen (to stay or switch doors)
}
\description{
'play_n_games' creates a loop to simulate the
Monty Hall Problem game 100 times.
}
\details{
This function creates a loop of the Monty Hall Problem
Game. The function will be simulated 100 times and the
results tabulated in a data frame showing the outcome
based on the chosen strategy.
}
\examples{
 return(results.df)
}
