% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOCUMENTATION.R
\name{play_n_games}
\alias{play_n_games}
\title{Play game loop}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{fill}{in}
}
\value{
fill in
}
\description{
\code{play_n_games( n= )} plays the game on a loop for n amount of times.
}
\details{
fill in
}
\examples{
  fill in

}
