% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{Function to stimulate the Monty Hall Problem game.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n}{is the number of stimulations.}
}
\value{
a list of the results
}
\description{
\code{play_n_games()} Runs stimulation of the Monty Hall Problem game and return the results as a list.
}
\examples{
play_n_games(n=100)

}
