% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem-cable.R
\name{play_n_games}
\alias{play_n_games}
\title{Play the game multiple times.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{This}{function has one argument: n. N refers to the number of games that
the function will play and give statistics on. Defaults to 100 games.}
}
\value{
This function prints a table of the win/lose proportions of the
results data frame, and also returns the results data frame.
}
\description{
\code{play_n_games()} loads the "dplyr" library, then sets the list function
as the results list and runs the game in a single loop. Then for each
instance between 1 and however many games you're playing (defaults to 100),
it plays the game and determines the game outcome, saves that game result
as the game result for its iteration within the loop,
and then sets the new loop count as the next iteration of the game until
it runs out of games to play. It then makes a data frame of the list
of all the results from all the games that were played and outputs
those results in a table that shows what proportion of the time
each game strategy resulted in winning or losing. It also returns
the count of how many winning and losing games there were in this set.
}
\details{
This function plays the entire game several times, records how many times
the strategy of staying or switching led to a winning game or a losing
game, and then exports both the count of those game stats and the
proportions of winning and losing games per strategy.
}
\examples{
  play_n_games()
  play_n_games(n=1000)

}
