% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.r
\name{play_n_games}
\alias{play_n_games}
\title{Play many Monty Hall Problem games}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n}{Integer. Represents the number of games played.}
}
\value{
A data frame with the results of all of the simulated games with strategy ("stay" or "switch")
and outcome ("WIN" or "LOSE").
}
\description{
This simulates multiple rounds of the Monty Hall Problem and summarizes the results.
}
\examples{
  play_n_games()
}
