% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{\preformatted{Loops
}}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
the function returns a dataframe of "WIN" and "LOSE" and "Switch" and "Stay"
}
\description{
\preformatted{Complete n number of loops to simulate playing the game n times.
}
}
\details{
\preformatted{Create loops of playing the game to determine the outcome over
many iterations. Can determine the if its more optimal to stick with 
orginal door or swtich doors. 
}
}
\examples{
	play_n_games()

}
