% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{Play the Monty Hall game multiple times and save results.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n}{The number of times to play the game; integer default n=100.}
}
\value{
The function returns a data frame containing the "stay"
result and the "switch" result for each round of the game.
}
\description{
\code{play_n_games()} plays n complete rounds of the Monty Hall
game and saves all of the results as a data frame.
}
\details{
The function goes through the steps of the Monty Hall game n
times.  For each game, the function saves what the result would
be if the contestant stayed with their intitial pick and what the
result would be if they switched.  Those results are then reported
as a table which shows the win percentage for each strategy.
}
\examples{
  play_n_games( n=500 )
  play_n_games( 100 )

}
