% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_n_games}
\alias{play_n_games}
\title{Game simulation function.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n}{is an integer for number of simulations.}
}
\value{
Data frame and proportion table of results.
}
\description{
'play_n_games()' performs a Monte Carlo style simulation to play through
the game staying and switching the door many times and then produces a results
table that indicates the proportion of wins and losses for each strategy.
}
\details{
The function loops through play_game() a set number of times and returns
data frames of the results and the proportion of wins and losses for each strategy.
}
\examples{
play_n_games()

}
