% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem-R-script.R
\name{play_n_games}
\alias{play_n_games}
\title{Play the Monty Hall Problem game n times (by default, n=100)}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{...}{the parameter for this function is 'n', the number of
times the game is played. By default, it is 100. The number can
be changed if desired by passing another integer into 'n'.}
}
\value{
This function returns a dataframe combining all the
outcomes of staying and switching for each game out of
the total of 'n' games.
For example:play_n_games(n=4)\preformatted{         |     outcome   |
strategy |  LOSE |   WIN |
    stay |  0.75 |  0.25 |
  switch |  0.25 |  0.75 |
}\preformatted{strategy  outcome
}

1     stay      WIN
2   switch     LOSE
3     stay     LOSE
4   switch      WIN
5     stay     LOSE
6   switch      WIN
7     stay     LOSE
8   switch      WIN
}
\description{
Do the results of playing the game 100 times suggest that
staying is a better strategy or that switching is a better strategy?
}
\details{
The game is played n times (default is 100 times).
Corresponding outcomes for each strategy are listed
for each game. All the results are combined into
a single dataframe.
}
\examples{
  play_n_games(n=100)  (default)
  play_n_games(n=50)

}
