% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{A wrapper function that allows a user to simulate a specified number of
Monte Hall games}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n}{ numeric value indicating the number of Monte Hall simulations
to run. The default value of n is 100.}
}
\value{
A proportion table indicating the outcomes of both switch and stay
strategies over n simulations
}
\description{
This function will simulate a user-defined number (n) of Monte Hall games
and store the aggregated results in a data table.
}
\details{
Simulates all steps of the Monte Hall game from the game setup
(assigning objects behind doors), initial contestant door selection, the
host opening a goat door, the contestant strategy of switch door selection
or stay, and the final game result. The results are simulated a user-defined
n-times and stored in a results data frame. The function returns a
proportion table to make the results more easily legible.
}
\examples{
play_n_games(n=1000)
play_n_games()
}
