% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg-3.R
\name{play_n_games}
\alias{play_n_games}
\title{\if{html}{\out{<div class="sourceCode">}}\preformatted{Plays n games of Monty Hall 
}\if{html}{\out{</div>}}}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{This}{function takes in an integer n, representing how many times we want to play the game}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ This function returns results.df, a list of results for each game simulated. 
}\if{html}{\out{</div>}}
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Plays n games of Monty Hall with both strategies (stay or swap doors) and records the results. 
}\if{html}{\out{</div>}}
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{This function lets us run the previous function n times and records all the results by trying both strategies. This allows us to check across many runs which strategy is the best. 
}\if{html}{\out{</div>}}
}
\examples{
	play_n_games(n=100)
}
