% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{Playing the game for n times (default=100)}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{...}{The arugment (n) is used for simulation.}
}
\value{
The function returns a dataframe of analysis of game results.
}
\description{
This function is a simulation of playing the games multiple times
}
\details{
The function will execute the entire game, eeping track of the outcomes from each simulation,
   and then provide a "count"of how many times switching or sticking with the initial choice and 
   the results respectively.
}
\examples{
   play_games(200)
}
