% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg copy 2.R
\name{play_n_games}
\alias{play_n_games}
\title{Plays n games of Monty Hall}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{`n`}{— an integer representing how many times we want to play the game.}
}
\value{
This function returns \code{results.df}, a list of results for each game simulated.
}
\description{
Plays \code{n} games of Monty Hall with both strategies (stay or switch) and records the results.
}
\details{
This function runs the previous function \code{n} times and records all the
results by trying both strategies. This allows us to check across many
runs which strategy is best.
}
\examples{
  play_n_games(n=100)

}
