% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem (1).R
\name{play_n_games}
\alias{play_n_games}
\title{Counting the odds}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{:}{… no arguments are used by the function.}
}
\value{
The result is that there will be 100 samples of results from the
game being played repeatedly.
}
\description{
After the game is created, it needs to be tested multiple times to determine
if the odds increase if the contestant stays or if they switch doors.
}
\details{
The function created for this test needs to be tested over 100 times. The
test must be run on a loop for the count of 100.
}
\examples{
i.	For( i in 1:n)
{
 	game.outcome <- play_game()
 	Results.list[[loop.count]] <- game.outcome
 	Loop.count <- loop.count + 1
 	}

}
