% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_n_games}
\alias{play_n_games}
\title{Simulate the Monty Hall game any number of times}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{The}{function requires an argument of a numeric, with the
default being 100.}
}
\value{
The function returns a 3 x 3 table showing the two
strategies (stay = TRUE or stay = FALSE) and then the
number of WINs and LOSSes for both. The number of
WINs and LOSSes should correspond with the numeric that
was input for the arguments function.
}
\description{
\code{play_n_games()} allows the user to determine the number
of times they want to run the Monty Hall simulation.
}
\details{
This function lets the user choose how many times they
want to run the previous "play_game()" function. This
allows the user to simulate the function many times to
determine which strategy, staying or switching, is the
best strategy for winning.
}
\examples{
  play_n_games(50)
}
