% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{Play the game one hundred time}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n}{which indicates how many times the game is repeated}
}
\value{
will return a small table that indicates the winning and losing strategy
}
\description{
play_n_game is a general function that contains all the previous functions will repeat the game 100 times
and will return  WIN  or Lose
}
\details{
The game consist of choosing one door from three doors, one of which has a car
behind it and two have goats The contestant selects a door, then the host
opens a door to reveal a goat, and then the contestant is
given an opportunity to stay with their original selection
or switch to the other unopened door. If the last pick was a car the participants
considered a winner otherwise a looser
}
\examples{
function( 100 )

}
