% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall_script_rj.R
\name{play_n_games}
\alias{play_n_games}
\title{:
Play multiple rounds}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{:}{Values will be returned as integers.}
}
\value{
:
A data frame is returned displaying a summary of wins, losses, switches, and stays.
}
\description{
:
Use a simulation to run multiple rounds of the game
}
\details{
:
The play_n_games function allows you to set the number of iterations to play. Each time the game is played, the games are stored in the data frame.The dplyr package will be required to bind rows in the results table.
}
\examples{
  play_n_games <- function( n=100 )
}
