% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty_hall_scripts.R
\name{play_n_games}
\alias{play_n_games}
\title{Monty Hall Problem game simulation of "n" games}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n}{An integer representing the number of games to play.}
}
\value{
This function returns a table of the game winning and losing probabilities based on the
number of game plays and the results of each round in a dataframe.
}
\description{
\code{play_n_games()} simulates the Monty Hall Problem game for a series of 'n' games.
}
\details{
This function plays the Monty Hall Problem game for as many iterations as the user enters.
}
\examples{
play_n_games(n = 150)

}
