% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{Game Loop}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{This}{function takes in an input of a specified 'n'.}
}
\value{
This function will return a table containing the proportions of wins/losses for each respective strategy (switch or stay) rounded
to two place.
}
\description{
This function loops the Monty Hall game 'n' times to allow for a summary statistics to be made regarding chance of winning.
}
\details{
This function will play the Monty Hall game a specified 'n' times and collect information regarding both strategy and won/lost
games into a data frame. It will then pipe this information into a proportion table, rounded to two decimal points.
}
\examples{
play_n_games(n=100)
}
