% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{play_n_games}
\alias{play_n_games}
\title{Game Set-Up.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{for, }{table, round and print arguments are used by the function.}
}
\value{
The function returns table to report the results.
}
\description{
\code{play_n_games()} game set-up.
}
\details{
Return a data frame that contains results from one game, then
use a loop to build a simulation that plays the game 100 times,
and create a table to report the results.
}
\examples{
play_n_games()

}
