% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{Play "n" games - Function}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{The}{function will run the game utilizing the function ""loop.count""play_game()",
the library (dplyr), the function "for(in)". Some of the variables used in this
function are loop.count, results.list, and game.outcome.}
}
\value{
The function will return the outcome of the strategy chosen. How many times a
contestant could win if chooses to stay or how may times the contestant could
win if chooses to switch?
}
\description{
This function will be using the play_game() function to simulates and test
the game a hundred times. The game will be using a loop to run the function "n"
times (n=100 in this came).
}
\details{
The function will be using a loop that will help to running the game continually
up to 100 times to see the different results.
}
\examples{
 play_n_games()
}
