% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{Monty Hall Game Loops}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{We}{set n to 100 loops}
}
\value{
results.df
}
\description{
Play the game n amount of times
}
\details{
The 'play_n_games' function loops the 'play_game' function n amount of times.
In this case, n is set to 100. A data frame was created to list the total
count of losses and wins out of the 100 loops.
}
\examples{
play_n_games(n = 10000)
}
