% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_game.R
\name{play_n_games}
\alias{play_n_games}
\title{Play multiple Monty Hall games}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n}{The number of games to simulate. The default is 100.}
}
\value{
A data frame containing the strategy and outcome
for each simulated game.
}
\description{
`play_n_games()` simulates the Monty Hall game multiple
times and returns the results for both the stay and
switch strategies.
}
\examples{
play_n_games(100)

}
