% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{Adding The Game To A Loop}
\usage{
play_n_games(n = 10000)
}
\arguments{
\item{The}{play_n_games function calles upon the argument of a numeric vector.
The library dplyr calls upon the package.
The results.list function will act as a collector for atomic vectors from the loop.
The loop.count will act as a counter
The for loop runs the functions play.game and creates new functions to gather data
and increment the counter when complete.
Results are stored in results.df and put into a table and formatted with the round
function.}
}
\value{
A formatted table with headers and stategy results formatted to show proportion.
}
\description{
We want to run simulations to gather data for inferential statistics.
}
\details{
By creating a loop to run the program 10,000 tiems, we will gather enough
data to make inferences upon the game strategy for each case, stay or switch.
}
\examples{
      outcome
  strategy   LOSE  WIN
   stay      .67   .33
   switch    .33   .67

}
