% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg-finished.R
\name{play_n_games}
\alias{play_n_games}
\title{Simulation of 100 randomized Monty Hall games}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function will return a table of game results based off of a simulation
of 100 different games with randomized player strategies of staying or switching.
}
\description{
\code{play_n_games()} function will run a series of 100 random Monty Hall games
}
\details{
To determine what the best player strategy is, a series of 100 random games
will be played to see if staying with the original door or
switching to the remaining door, is the more effective strategy.
}
\examples{
play_n_games()

}
