% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{Runs a simulation of the Monty Hall Problem game to
  determine the probability of win or lose.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{The}{Function takes a number for which the simulation needs to be
run as an argument.}
}
\value{
The function returns a data frame with probabilities of win and
 lose for both scenarios (stay and switch).
}
\description{
The function is a compliation of all the steps in playing
 the game. It allows the user to run the simluation for any number
 of times to determine probability of win or lose.
}
\details{
This function can run the simulation any number of times. It stores
 the result in a data frame. After running the simulation for supplied
 number of times it calculates the probabilities of win and lose for
 both scenarios (stay and switch).
}
\examples{

 play_n_games ( n=100 )

}
