% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_n_games}
\alias{play_n_games}
\title{Play more than one rounds of the Monty Hall Game}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{The}{function uses n as a parameter which represents the number of games to be simulated.}
}
\value{
Returns a data frame that holds the win/lose for each game as well as the strategy stay/switch.
}
\description{
The function plays multiple rounds of the game and stores the results in a data frame to analyze
}
\details{
This function enables users to simulate numerous rounds of the game in order to analyze the most effective strategies.
By choosing 'n' games, you can review both strategies and compare the results of each in a prop.table..
}
\examples{
play_n_games(n = 1,000)
}
