% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_n_games}
\alias{play_n_games}
\title{Play multiple games.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{`n=`}{Integer. How many times the game will be played.}
}
\value{
Data frame. The frame will contain each strategy and their
respective win rates.
}
\description{
This plays the entire game multiple times depending on the input.
}
\details{
This will calculate the overall win rate of each game strategy.
}
\examples{
    play_n_games()

}
