% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monty Hall.package.R
\name{play_n_games}
\alias{play_n_games}
\title{Play multiple Monty Hall games.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{An}{integer specifying the number of games to simulate.
Defaults to 100.}
}
\value{
A data frame containing the strategy used ("stay" or "switch") and the
corresponding outcome ("WIN" or "LOSE") for each game played.
}
\description{
\code{play_n_games()} simulates a specified number of Monty Hall games
by repeatedly calling the \code{play_game()} function and storing the results.
}
\details{
This function runs `n independent simulations of the Monty Hall Problem.
For each game, both strategies( staying with the initial.pick and switching
to the other door) are evaluated. The results are combined into a single data
frame, and a summary table showing the proportion of wins and losses for each
strategy is printed.
}
\examples{
# Simulates 50 Monty Hall games #' results <- play_n_games(50)
}
