% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_n_games}
\alias{play_n_games}
\title{Play Multiple Games}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n}{The number of games to play. The default is 100.}
}
\value{
A data frame containing the results from all games.
}
\description{
Plays the Monty Hall game multiple times.
}
\details{
The function repeats the game and records the results for
staying and switching.
}
\examples{
play_n_games(n=10)

}
