% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_n_games}
\alias{play_n_games}
\title{Simulate many iterations of the Monty Hall game}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{x}{}
}
\value{
x
}
\description{
x
}
\details{
x
}
\examples{
 x
}
