% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_n_games}
\alias{play_n_games}
\title{Simulate many iterations of the Monty Hall game}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n}{\itemize{
\item number of trials to simulate
}}
}
\value{
a data frame of results
}
\description{
Description
}
\details{
Details
}
\examples{
 play_n_games(5)
}
