% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem edited.R
\name{play_n_games}
\alias{play_n_games}
\title{Loop the Monty Hall Game}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{This}{function has one argument, which is a numeric value for the number of times to recreate
the simulation, such as n=10,000.}
}
\value{
The return is a table showing the decimal percentages of how often
the contestan won or lost.
}
\description{
\code{play_n_games} creates a looped simulation to see what the results
of the Monty Hall Game are over \code{n} simulations, and compare whether
staying with the contestant's original choice or switching to the other
unopened door provides a better chance of winning the car.
}
\details{
This function adds a loop feature into the \code{play_game} function, so that
in the result you can see whether the contestant won the car more often by chooseing
\code{STAY=T} or \code{STAY=F}.
}
\examples{
  play_n_games( n=100 )

}
