% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_n_games}
\alias{play_n_games}
\title{Simulate the game 100}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{'n'}{\itemize{
\item A number representing the number of times to run the game. The default value is 100.
}}
}
\value{
This returns the amount of loses and wins out of n times in a table in percent form.
}
\description{
A loop of the game will run 100 times.
}
\details{
This code is telling the game to run n times in a row with different results and to stop after a number of times.
}
\examples{
play_n_games(100)
}
