% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{Repitition}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{Proportional}{results will be given as to the strategy that the contestant has picked (stay or switch)}
}
\value{
Same result of WIN/LOSE
}
\description{
This function allows the game to be repeated to see the outcome percentages over time
}
\details{
n specifies the number of times the game will be simulated, and a table of the results will be provided
}
\examples{
WIN
}
