% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\name{play_n_games}
\alias{play_n_games}
\title{play multiple times}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{The}{number og games played}
}
\value{
return a table with the numbers of winning & losing for each strategy
}
\description{
to compare the results of stating option & switching the door option
}
\details{
The game played n times, and return the results of each strategy
}
\examples{
  play_n_games(n=1000)
  
}
