% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.r
\name{play_n_games}
\alias{play_n_games}
\title{Play Monty Hall n times.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{An}{integer value n represents the number of times the game will be
played.}
}
\value{
A table of proportions would be displayed with a win rate of staying
  and switching.
}
\description{
This runs the Monty Hall game simulation n times.
}
\details{
The game results would be similar to the theoretical values
  if the number of games  increases. This functions allows for that.
}
\examples{
  play_n_games(100)

}
