% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{Loop multiple games to see the probabilty
of each outcome.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{additional}{libraries are loaded for
this function. R will loop the gameplay for
the amount of times specified by the user.
If the user does not input a number in the
play_n_games() function, then R will default
to looping the game 100 times.}
}
\value{
A table will be provided to show the
probabilities of winning and losing.
}
\description{
'play_n_games()' loops the full game cycle
n times and produces the probability of
winning or losing.
}
\details{
The user can change the
number of times the game is looped by
entering a number in the function. If
nothing is entered in the function, then
R will run the loop 100 times.
}
\examples{
play_n_games()
play_n_games(10000)
}
