% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{play_n_games}
\alias{play_n_games}
\title{Runs the Monty Hall game a given amount of times and shows results.}
\usage{
play_n_games(n)
}
\arguments{
\item{`n`}{- This determines the number of times the game is run.}
}
\value{
The function returns a results table showing the proportional
  results of `n` runs of the game.
}
\description{
`play_n_games` plays the Monty Hall game a number of
  times based on the numerical parameter you pass to it.
  It then shows the proportional results.
}
\details{
The function uses a `for` loop that continues based on
  the parameter passed into the function. It records each
  outcome in a results list. When the loop finishes, it
  creates a table showing the proportional results.
}
\examples{
  play_n_games(30)
  play_n_games(100)

}
