% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-script.R
\name{play_n_games}
\alias{play_n_games}
\title{Run a loop.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{None}{}
}
\value{
Multiple loops run at once.
}
\description{
Running a loop several times can show which option is better.
}
\details{
This gives results of odds of winning if contestant chooses
to switch or stay.
}
\examples{
play_n_games()

}
