% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_n_games}
\alias{play_n_games}
\title{Simulate multiple games}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n}{Number of games to play. Default is 100}
}
\value{
A data frame of all outcomes
}
\description{
Plays the Monty Hall game multiple times and summarizes the results.
}
\details{
Useful for observing the statistical advantage of switching doors over multiple trials
}
\examples{
play_n_games(10)
}
