% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_n_games}
\alias{play_n_games}
\title{Play many Monty Hall games}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{n}{An integer specifying how many Monty Hall games to simulate.}
}
\value{
A data frame with \code{2*n} rows (because each game
has 2 rows: "stay" and "switch"), plus prints a table
of row proportions (WIN vs. LOSE) for each strategy.
}
\description{
\code{play_n_games()} repeats \code{\link{play_game}} \code{n} times,
collecting the outcomes for each strategy,
and prints row proportions (WIN/LOSE).
}
\details{
By default, it loops 100 times, but you can specify
a larger \code{n} for more precise estimates
(e.g., 10,000 or 1e5).
}
\examples{
\dontrun{
  results_100 <- play_n_games(100)
  results_10000 <- play_n_games(10000)
}

}
