% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newscriptmontyhall.R
\name{play_n_games}
\alias{play_n_games}
\title{Build a loop Function.}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{...}{iterator values}
}
\value{
the fuction returns the result of the the decisions taken whether to switch or stay.
}
\description{
play_n_games() generates the result if the game has been palyed for n tims and here for instance n is 100.
}
\details{
This is a loop function designed to build a simulation that plays the game for 100 times.
}
\examples{
play_n_games()

}
