% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem (1).R
\name{select_door}
\alias{select_door}
\title{The contestant makes their initial door selection.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns the number of the randomly selected door, between 1 and 3.
}
\description{
\code{select_door()} Type: Integer -> randomly selects a door number from 1 to 3.
}
\details{
Selects door 1, 2, or 3. The door contains a goat or a car.
}
\examples{
  select_door()
}
