% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{Select a door.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a value corresponding with the door the contestant selected.
}
\description{
\code{select_door()} allows the player to pick a random door out of the three.
}
\details{
The contestant selects a door of which they assume the car is behind.
}
\examples{
  select_door()

}
