% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{CONTESTANT SELECTS A DOOR}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function allows for one door to be selected, numbered 1:3
}
\description{
select_door() function allows the contestant to select a door, number 1:3.
}
\details{
The contestant will be selecting 1 our of 3 doors on the "game show."
}
\examples{
select_door()
}
