% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{select_door}
\alias{select_door}
\title{Select a door from three doors}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a random value in the door vector.
}
\description{
\code{select_door()} selects a door from the three different doors
}
\details{
build a new vector with 1-3 represents 3 doors and randomly
return a number from the vector to represent a picked door.
}
\examples{
    select_door
}
