% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{Contestants selects a door}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a length of 1 numeric vector
indicating that one door has been selected.
}
\description{
Contestant will randomly select a door 1, 2, or 3.
}
\details{
Each door will randomly have a goat or car behind
it with a total of 2 goats and 1 car.
}
\examples{
select_door()

}
