% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Select an initial door in the Monty Hall game.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a single integer (1, 2, or 3) representing
the index of the selected door.
}
\description{
\code{select_door()} simulates a contestant making their initial selection
of one of the three doors at the start of the game.
}
\details{
In the Monty Hall Problem, the contestant must choose one of three
closed doors before any goats are revealed. This choice represents
the initial state of the simulation, which will later be evaluated
against the host's revealed goat door.
}
\examples{
  select_door()

}
