% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Randomly select a door}
\usage{
select_door()
}
\value{
An integer (1–3).
}
\description{
Chooses one of the three doors at random.
}
\details{
Returns an integer 1, 2, or 3 representing the selected door.
}
\examples{
select_door()
}
