% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Select a Door}
\usage{
select_door()
}
\value{
A numeric value indicating the selected door. The returned
value will be 1, 2, or 3.
}
\description{
\code{select_door()} randomly selects one of the three doors in a
Monty Hall game as the contestant's initial choice.
}
\details{
The function creates a numeric vector representing doors 1, 2,
and 3, then randomly samples one door from the vector.
}
\examples{
  select_door()

}
