% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{Select the contestant's first pick of doors}
\usage{
select_door()
}
\arguments{
\item{no}{arguments are used in select_door}
}
\value{
the function returns a randomly generated number
(1-3) indicating the door which the contestant first
"picked"\preformatted{ the function will place the number into a.pick
}
}
\description{
the contestants first pick of a door, numbered 1-3.
randomly assigns a number 1-3 as the contestant's first pick
}
\details{
all three doors are still closed and available
to the contestant
}
\examples{
select_door will return one of the following:
     [1] 1
     [1] 2
     [1] 3

}
