% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Step 2: The Contestant Selects a Door}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
... number between 1 and 3
\code{select_door( )} returns a random integer between 1 and 3
}
\description{
\code{select_door( )} produces a random integer between 1 and 3
}
\details{
The contestant selection points to one of the elements
in the vector \code{game}. The door remains closed (what is behind
it remains unknown to the contestant).
}
\examples{
  game
  [1] "car"  "goat" "goat"
  a.pick <- select_door()
  a.pick
  [1] 2

}
