% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem-Gonzalez.R
\name{select_door}
\alias{select_door}
\title{Choose a door for the Goat Game.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a length 1 numeric vector to indicate
which door was choosen. It will either be 1, 2, or 3.
}
\description{
The 'select_door' function provides the initial door choice for the
Goat Game.
}
\details{
This is a simulation of when a contestant choooses  a door.
In the real Goat Game, a human would be given the choice between
3 doors. The person will choose either 1, 2, or 3. This first
choice is done with no knowledge of what is behind the doors.
}
\examples{
select_door()

}
