% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{Selecting an Initial Door at Random}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used in this function}
}
\value{
The function returns a length of 1 numeric vector
indicating what door the contestant has initially picked.
}
\description{
\code{select_door} randomly samples a vector that represents the possible
door options a contestant could have chosen.
}
\details{
This function replicates the "Let's Make a Deal" game by selecting
one of the three doors a contestant could have chosen.  This function
randomly selects a door which will then be used later to test the
two strategies of switching vs staying with a selected door.
}
\examples{
 select_door()
 
 To test this code simply use it several times.
 select_door()  
 select_door()
 select_door()
 
 This should results in random doors being selected.
  
}
