% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{Select a door.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a number between 1 and 3 that corresponds to the
\code{create_game()} function.
}
\description{
\code{select_door()} randomly selects a door between 1 and 3.
}
\details{
Due to randomization, the first door selected could contain either a goat or a
car.
}
\examples{
  select_door()

}
