% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{Select a door from the three available doors}
\usage{
select_door()
}
\arguments{
\item{No}{arguments are used by the function.}
}
\value{
The function returns a number between 1 and 3.
}
\description{
\code{select_door()} randomly selects a number between 1 and 3 to represent the door the
the participant has selected.
}
\details{
The game strategy begins with the participant selecting a door. In order to replicate this,
a number (representing the door) is randomly selected. A goat door will be revealed next
and the participant will have the option to stay with their selected door or switch doors.
This is a stand-alone function with no information needed from the previous step.
}
\examples{
  select_door()

}
