% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem (1).R
\name{select_door}
\alias{select_door}
\title{Selecting an initial door randomly}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a length 1 numeric vector indicating which door that the contestant initially chose.
}
\description{
select_door randomly samples a vector that shows the possible door options for contestant can choose.
}
\details{
This function replicates the game show, "Let's Make a Deal" by selecting one of the three doors that a contestant can choose.
This function randomly selects a door that will be used to test if switching or changing the door strategy works better.
}
\examples{
select_door()

 To test this code simply use it several times.
 select_door()  
 select_door()
 select_door()
 
}
