% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{To determine participant's door selection.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
La function devuelve un numero entre 1 y n aleatoriamente para representar
la eleccion del participante.
}
\description{
\code{select_door()} devuelve un numero entre 1 y n aleatoriamente para representar
la eleccion del participante.
}
\details{
Una vez seteado el juego, el participante procede a seleccionar una puerta.
}
\examples{
  select_door()
  
}
