% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.r
\name{select_door}
\alias{select_door}
\title{Select a door from the three available doors}
\usage{
select_door()
}
\arguments{
\item{No}{arguments are used by the function.}
}
\value{
The function returns a number between 1 and 3.
}
\description{
\code{select_door()} randomly selects a door number between 1 and 3.
}
\details{
The contestant selects a door. To replicate this, doors are randomly selected.
A goat door will be revealed next and the contestant has the option to stay
with their current selection or switch.
}
\examples{
  select_door()

}
