% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Contestant selects a door.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a length 1 numeric vector, a number
between 1 and 3 indicating the door the contestant selected.
}
\description{
\code{select_door()} randomly selects one of the three doors as the
contestant's initial pick.
}
\details{
Since the contestant will not know the position of the car when
they select a door, no information about the game set-up is
shared before the selection is made. The door is drawn at random
from the three positions.
}
\examples{
  select_door()

}
