% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem (4).R
\name{select_door}
\alias{select_door}
\title{Contestant chooses a door}
\usage{
select_door()
}
\arguments{
\item{...}{Number between 1 and 3}
}
\value{
The function returns a length 3 character vector
indicating a selection of position between 1 and 3.
}
\description{
\code{select_door()} generates a random selection for the first choice the
contestant makes.
}
\details{
In selecting a door, the contestant will later decide if he/she wishes
to stay with the originally selected door, or if they will decide to switch.
}
\examples{
  select_door()

}
