% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{Select a Door}
\usage{
select_door()
}
\arguments{
\item{None.}{The function does not take any arguments.}
}
\value{
The function returns a single integer between 1 and 3, representing the contestant's selected door.
}
\description{
This function randomly selects one of the three doors in the Monty Hall game.
}
\details{
The function generates a random selection from the three available doors (labeled 1, 2, and 3) to simulate the contestant's choice.
}
\examples{
select_door()

}
