% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{Contestant selects a door.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns the number of the door selected.
}
\description{
\code{select_door()} randomly selects one of the three doors as the initial pick of the contestant.
}
\details{
This function creates the initial door selection for the contestant. This selection will influcence the following functions.
}
\examples{
  select_door()

}
