% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Select a door in the Monty Hall Problem game.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
An integer 1, 2, or 3 indicating the chosen door.
}
\description{
`select_door()` returns a random door (1, 2, or 3).
}
\details{
This simulates the contestant’s initial choice of one of the three
  available doors. This is independent of the car and goat locations.
}
\examples{
  select_door()

}
