% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem(bh).R
\name{select_door}
\alias{select_door}
\title{Contestant selects first door.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a single numeric vector: 1, 2, or 3.
}
\description{
'select_door()' randomly generates a numeric vector (1-3) which
represents of the doors created in the create_game() function.
}
\details{
The doors that are generated by the 'select_door()' function each
represent either a "goat" or a "car". Each door is assigned a number
(1-3) and the randomly generated numbers from this function represent
either a "goat" or "car" door created in the create_game() function.
The contestant selects a door one of these doors, and a numeric vector
(1-3) is returned.
}
\examples{
  select_door()

}
