% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Randomly select one of three doors. This is the first door pick in
the Monty Hall Problem game.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns an integer vector of length 1 indicating the
index value of the door first chosen by the contestant.
}
\description{
\code{select_door()} returns a random integer between 1 and 3 and serves
as the first door selected while playing the Monty Hall Problem
game.
}
\details{
This is the first step in playing the Monty Hall Problem. One of
three doors is randomly selected. Since no other information is known
by the contestant at the time of the first door choice, a random
simulation is appropriate.
}
\examples{
  select_door()

}
