% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{Select a random door}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments used by the function.}
}
\value{
The function returns a randomly selected number between 1 and 3.
}
\description{
'select_door()' selects a door at random.
}
\details{
The function works as an initial selection for which door the contestant chooses.
}
\examples{
  select_door()
}
