% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{Select a random door}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a integer of the selected door
}
\description{
\code{select_door()} selects a random door of the three available
}
\details{
Contestant selects one of three available doors. The item behind
the selected door would be their prize.
}
\examples{
  select_door()

}
