% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{Select a door}
\usage{
select_door()
}
\arguments{
\item{A}{number value representing the door chosen by the player (1, 2, or 3)}
}
\value{
The funtion returns a number between 1 and 3
}
\description{
Select a door in Monty Hall Problem game
}
