% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{player selects a door}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments used by the function}
}
\value{
The function returns a single value for the chosen door
}
\description{
\code{select_door()} chooses a door numbered between 1 and 3 randomly
}
\details{
Simulates a players first pick in a monty hall game
}
\examples{
select_door()

}
