% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Contestant selects first door.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
the output will be a vector
}
\description{
\code{select_door()} randomly returns a value for one door (1-3)
}
\details{
this function first creates a vector of doors numbered 1-3, then randomly selects one of the doors, then returns the selection. this mimics a gameshow contestant randomly picking one of three closed doors
}
\examples{
  select_door()   # randomly chooses door
  
}
