% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{Contestant selects a door for the first guess.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a number between 1 and 3
indicating the door initially opened by contestant.
}
\description{
\code{select_door} select a door randomly from doors
numbered 1,2,3.
}
\details{
The function simulates the process when contestant selects
a door randomly.
}
\examples{
  'select_door()'
}
