% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{Select a door to start}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns the door that was chosen in the sample to start the game.
}
\description{
`select_door()` Samples a door from the created game.
}
\details{
The game setup replicates the game on the TV show "Let's
  Make a Deal" where there are three doors for a contestant
  to choose from, one of which has a car behind it and two
  have goats. The contestant selects a door, then the host
  opens a door to reveal a goat, and then the contestant is
  given an opportunity to stay with their original selection
  or switch to the other unopened door. There was a famous
  debate about whether it was optimal to stay or switch when
  given the option to switch, so this simulation was created
  to test both strategies.
}
\examples{
 select_door()
}
