% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Choose original contestant door}
\usage{
select_door()
}
\arguments{
\item{No}{arguments needed for this function}
}
\value{
Will return an integer vector of the contestant's chosen door
}
\description{
This function randomly decides the original contestant choice from the available doors
}
\details{
In this case the function is picking randomly from three doors.
}
\examples{
select_door()
}
