% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{Player randomly selects a door from the previously created game.}
\usage{
select_door()
}
\arguments{
\item{...}{No arguments are used by the function.}
}
\value{
Function returns a  vector of length 1 containing the number 1, 2, or 3 to
indicate the location of the selected door.
}
\description{
\code{select_door()} picks a door selection from those created in the current game
by \code{create_game()}.
}
\details{
\code{select_door()} randomly selects one of a three door selection from the game
previously created by \code{create_game()}.
Neither any Car nor Goat location is known to the Player.
}
\examples{
select_door()

}
