% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{\if{html}{\out{<div class="sourceCode">}}\preformatted{Contestant selects a door in the Monty Hall Problem game.
}\if{html}{\out{</div>}}}
\usage{
select_door()
}
\arguments{
\item{...}{A numeric value}
}
\value{
The function returns a length 1 numeric vector
indicating the position of the door they chose.
}
\description{
\code{select_door()} generates a random contestant selection of
one of the three available game doors.
}
\details{
This is the contestants first selection of one of the three
available game doors and the first door selection in the Monty
Hall problem game.
}
\examples{
  `select_door()`
  [1]
  
  `select_door()`
  [2]
  
  `select_door()`
  [3]
  

}
