% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{select door}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by this function}
}
\value{
The function returns a length 3 character vector
indicating the positions of goats and the car.
}
\description{
'
\code{select_door}opens a door with the possibility of having a goat behind or a car behind
}
\details{
'
The contestant selects a door, then the host
opens a door to reveal a goat, and then the contestant is
given an opportunity to stay with their original selection
or switch to the other unopened door. There was a famous
debate about whether it was optimal to stay or switch when
given the option to switch, so this simulation was created
to test both strategies.
}
\examples{

}
