% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Step 02: Contestant makes an initial door selection.}
\usage{
select_door()
}
\arguments{
\item{Numeric}{}
}
\value{
The function returns a number length one numeric between 1 and 3.
}
\description{
'select_door()' generates an initial door selection in which the contestant chooses a door but does not know what is behind it.
}
\details{
This function should only return one number, representing a blind choice by the contestant.
}
\examples{
Try running 'select_door()' at least three times to ensure that return is random.
##CODE CHUNK 2 UNIT TEST
select_door()
select_door()
select_door()
}
