% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HolfordMeetsMonty.R
\name{select_door}
\alias{select_door}
\title{Choose Your Lucky Door}
\usage{
select_door()
}
\arguments{
\item{No}{parameters are used in this function.}
}
\value{
The function returns a single, random integer between 1 and 3 representing
the door selected by the contestant -> a.pick
}
\description{
\code{select_door()} randomly samples 1 number from the vector (1,2,3) to
represent the door the contestant has chosen.
}
\details{
Once the game has been created, we have 3 potential prizes, goat, goat, car
each positioned in a character vector. \code{select_door()} represents the selection
of the contestant and will be associated with the position of the character vector
with each prize.
}
\examples{
 select_door()
}
