% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{Contestant selects a random door.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a number (numeric vector) between 1 and 3
indicating the position of the selected door by the contestant.
}
\description{
`select_door()` this function to let the contestant makes his/her
first selection.
}
\details{
After creating the game, this second function gives the contestant
a chance to randomly select one of the three doors.
}
\examples{
select_door()

}
