% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monty Hall Script.R
\name{select_door}
\alias{select_door}
\title{Player selects a door}
\usage{
select_door()
}
\arguments{
\item{...}{This function does not take any arguments}
}
\value{
This function returns a numeric vector of length one called a.pick
}
\description{
This function creates a numeric vector of length one. This vector represents the door that the
player selected and is either 1, 2, or 3.
}
\details{
The door is selected at random using \code{sample()} and will represent the either a goat or car based on
the positions in a.game
}
\examples{
select_door()

}
