% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Monty Python Door Selection}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a number between 1 and 3 which represents the
door placement within the game array
}
\description{
\code{select_door()} randomly generates a door selection.
}
\details{
This function replicates the participant of the game show making their
initial door selection.
}
\examples{

select_door()

}
