% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPackageProject_Roach.R
\name{select_door}
\alias{select_door}
\title{Select the initial door.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\description{
'select_door' selects the initial door choice in the Monty Hall game.
}
\details{
The door is randomly selected from the three available doors,
and returns the selection as the contestant's door selection.
It can be any of the three doors, including the door hiding the car.
}
\examples{
select_door()
}
