% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{Select Door}
\usage{
select_door()
}
\arguments{
\item{No}{parameters}
}
\value{
return the first door selection  # number between 1 and 3
}
\description{
Select a Door Function
}
\details{
select a door of 1
}
\examples{
object <- function
}
