% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{Contestant selects a door.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a single numeric vector: 1, 2, or 3.
}
\description{
'select_door()' randomly generates a numeric vector (1-3) which
indicates a door produced from the create_game() function.
}
\details{
The randomly generated numbers from this function represent either
a "goat" or "car" door from the create_game() function.
}
\examples{
  select_door()

}
