% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Contestant selects a door}
\usage{
select_door()
}
\arguments{
\item{"Doors"}{A numeric vector with values 1, 2,and 3, representing the three doors.}

\item{"A.pick"}{is chosen through the sample function which takes a random sample
of 1 of the 3 doors.}
}
\value{
The function returns a number between 1 and 3 indicating which door the
contestant has selected.
}
\description{
\code{select_door()} This function was created to simulate a contestant randomly
choosing one of the three doors.
}
\details{
First, a vector of three doors numbered 1,2,3 are created. Then one of those
doors will be randomly selected using \code{sample()}. Finally, the function will
return the randomized door selection.
}
\examples{
 select_door()
}
