% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Select a door.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function}
}
\value{
The function returns a numeric value returns a numeric value between 1 and 3 indicating the selected door.
}
\description{
'select_door()' randomly selects one of the three doors in the Monty Hall game.
}
\details{
The function randomly selects one door from doors 1, 2, and 3 to represent the contestant's initial choice.
}
\examples{
  select_door()
}
