% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{select_door}
\alias{select_door}
\title{Select an initial door.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a number between 1 and 3.
}
\description{
\code{select_door()} retuns a door selection for the initial pick by the contestant.
}
\details{
This is the first choice of the three doors. The contestant can pick any one.
}
\examples{
  select_door()

}
