% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhallfunctions.R
\name{select_door}
\alias{select_door}
\title{Contestant selects a door.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a numeric vector of 1, 2, or 3.
}
\description{
'select_door()' randomly selects a number from 1-3 and
matches a door generated by the create_game() function.
}
\details{
The randomly generated numbers of 1-3 created by this
function correspond to either a "car" or a "goat" door,
as determined by the create_game() function.
}
\examples{
  select_door()
}
