% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Randomly select one of three doors}
\usage{
select_door()
}
\value{
An integer (1, 2, or 3) representing the contestant's initial pick.
}
\description{
Randomly selects a door number from 1 to 3.
}
\examples{
select_door()

}
