% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Select a Random Door.}
\usage{
select_door()
}
\arguments{
\item{None}{}
}
\value{
A single numeric value representing the selected door (between 1 and 3).
}
\description{
This function randomly selects one of the available doors in the Monty Hall game.
}
\details{
The Monty Hall problem involves three doors, behind one of which is a car and behind the other two are goats. This function simulates the contestant's initial choice of a door.
}
\examples{
Randomly select a door, select_door()
}
