% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{This is a select door function.}
\usage{
select_door()
}
\arguments{
\item{...}{Here, the arguments stored in the a.pick vector,  samples from our doors vector which contains three option choices. Per the offcial documentation," sample takes a sample of the specified size from the elements of x using either with or without replacement". The size argument within represents a non-negative integer specifying the number of items to choose from.}
}
\value{
\itemize{
\item we are returned a.pick, which stores the sample taken from the door options
}
}
\description{
This select door function enables a player to select a door for their first case
}
\details{
The contestant makes their first selection, and this function enables us to choose one door at random.
}
\examples{
select_door()
}
