% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{Select original door.}
\usage{
select_door()
}
\arguments{
\item{This}{function does not have any parameters.}
}
\value{
This function returns a number from 1-3 that represents the door of the original pick.
}
\description{
\code{select_door()} allows contestant to pick their first door.
}
\details{
Function to select original door.
}
\examples{
 select_door()

}
