% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Select a Door}
\usage{
select_door()
}
\value{
A numeric value representing the chosen door (1, 2, or 3).
}
\description{
Randomly selects one of the three doors for the contestant.
}
\details{
This function simulates the contestant's initial random choice of one of the three doors.
}
\examples{
select_door()
}
