% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem (1).R
\name{select_door}
\alias{select_door}
\title{Door Selection}
\usage{
select_door()
}
\arguments{
\item{no}{arguments are used by the function.}
}
\value{
The function returns a numeric vector which indicates the number of the
chosen door
}
\description{
\code{select_door()}The contestant selects a door.
}
\details{
The contestant selects only one door. It can be either a goat door or a car door.
}
\examples{
select_door ()
}
