% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Select a door for the game.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a numeric vector
indicating the number of the chosen door.
}
\description{
\code{select_door()} random generates a chosen door numbered 1-3
}
\details{
One of the three doors will be selected at random. Two
of the doors have goats behind them, and one of the doors
has a car behind it.
}
\examples{
  first pick <- select_door()
  
}
