% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{Select a door.}
\usage{
select_door()
}
\value{
A number between 1 and 3 representing the chosen door.
}
\description{
Select a door.
}
\examples{
select_door()

}
