% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Selects a door or doors from the doors available.}
\usage{
select_door()
}
\arguments{
\item{...}{This function does not input arguments.}
}
\value{
... Returns one numeric vector/output.For example, 1.
}
\description{
'select_door' returns a random door or doors (depending
on what is specified in the sample function) from the
previously determined number of doors.
}
\details{
'select_door' returns a random door from the previously
determined number of doors. this is determined by the
doors <- c(1, 2, 3). a.pick is the randomization process
as it includes the function sample() that uses the
previously created "door" as a input and "size = 1" to
determine how many doors to return. return() uses
"a.pick" as an input to provide the random door number
requested as an output.
}
\examples{

select_door()
select_door()
select_door()

}
