% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{select_door}
\alias{select_door}
\title{Select a random door for the contestant.}
\usage{
select_door()
}
\arguments{
\item{This}{function passes no parameters in it.}
}
\value{
This function returns a single value from the \code{doors} vector (a
number between 1 and 3)
}
\description{
\code{select_door()} generates 3 doors numbered 1-3 and randomly assigns
one to the contestant.
}
\details{
This function specifically grabs 3 doors and numbers them from 1-3
and then returns a random number. This random number is the number
that corresponds to the door. In this case, the order of the doors
is corresponding to the order of the \code{create_game()} function that
determines whether or not the randomly selected door has a goat or
a car behind it.
}
\examples{
 `select_door()`

}
