% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg (1).R
\name{select_door}
\alias{select_door}
\title{Select a Door}
\usage{
select_door()
}
\value{
An integer between 1 and 3 representing the selected door.
}
\description{
Randomly selects one of the three doors.
}
\details{
This function simulates the contestant's initial door selection.
}
\examples{
select_door()
}
