% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_door.R
\name{select_door}
\alias{select_door}
\title{Monty Hall Door Selector}
\usage{
select_door(game)
}
\arguments{
\item{game}{A vector of any class and length.}
}
\value{
The output is a number between, and including, 1 and \code{length(game)}.
}
\description{
\code{select_door(game)} randomly chooses a door among in a Monty Hall door.
}
\details{
This function does the first step of a Monty Hall game where the contestant has to
choose a door. Given a game vector, this function returns a door number.
}
\examples{
select_door(c("car", "car", "goat"))
select_door(c(1, 2, 3, 4, 5))
select_door(create_game(4,1))

}
