% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg (2).R
\name{select_door}
\alias{select_door}
\title{select door}
\usage{
select_door()
}
\arguments{
\item{#no}{arguments are used by the function.}
}
\value{
The function returns a number between 1 and 3
}
\description{
\code{select_door()} allows game to select door
}
\details{
The contestant selects a door, then the host
opens a door to reveal a goat, and then the contestant is
given an opportunity to stay with their original selection
or switch to the other unopened door
}
\examples{
select_door()
}
