% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Select a door.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a number between 1 and 3
indicating the selected door.
}
\description{
\code{select_door()} randomly selects one of three doors.
}
\details{
The function creates three numbered doors and randomly
selects one of them for the contestant.
}
\examples{
  select_door()

}
