% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall-pkg.R
\name{select_door}
\alias{select_door}
\title{Contestant selects a door.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a numeric variable that correlates to the
door number selected.
}
\description{
\code{select_door()} replicates a contestant choosing one of
the 3 available doors.
}
\details{
This step allows us to replicate the randomized selection of a door,
which directly imitates how this action is done during the real game.
}
\examples{
   select_door()

}
