% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{Select a door.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a length 1 numeric vector
indicating the contestants selection.
}
\description{
\code{select_door()} generates a numeric scalar that consists of an integer between 1 and 3.
}
\details{
The function creates an object that serves as the contestants selection.
}
\examples{
  select_door( )

}
