% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{Selecting a Door}
\usage{
select_door()
}
\arguments{
\item{doors}{<- c(1,2,3)
a.pick <- sample( doors, size=1 )
return( a.pick )  # number between 1 and 3}
}
\value{
a. pick, a number between 1 and 3
}
\description{
A door is randomly selected.
}
\details{
A vector is created assigning 1, 2, and 3 to "doors".
A door 1, 2, or 3 is randomly selected.
A selection is returned.
}
\examples{
  1, 2, or 3
}
