% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{Contestant selects a door out of three available.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a vector (length = 1) with numerical value
between 1 and 3 that indicates with door the contestant has chosen.
}
\description{
\code{select_door} allows the user to makes a choice of a door from 1 to 3,
representing the three doors.
}
\details{
First, the variable \code{doors} is created, holding the
numbers 1 through 3. Then one of the numbers is
selected randomly and stored in a variable \code{a.pick}
which is returned when the function is called.
}
\examples{
  select_door()

}
