% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_door.R
\name{select_door}
\alias{select_door}
\title{Door Selection}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns an interger selected which door is your intinital pick
}
\description{
\code{select_door} generates a randome door that will be your inital selection
}
\details{
\preformatted{  Doors that are selected are from 1-3 and are picked at random and that will be your inital pick
}
}
\examples{
  select_door()
}
