% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Select a Door}
\usage{
select_door()
}
\arguments{
\item{None, }{no arguments required.}
}
\value{
The function returns an object representing one of three doors that were randomly selected.
}
\description{
`select_door` randomly selects a door from one of three options
}
\details{
The function creates a vector containing 3 doors (either 1, 2, or 3) and randomly
selects on door.
}
\examples{
select_door()
}
