% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Contestant Selects A Door}
\usage{
select_door()
}
\arguments{
\item{No}{arguments used in this function.}
}
\value{
The function returns a length 1 character vector indicating the position of
the chosen door.
}
\description{
'select_door' function sets up the possible door selections and makes a
selection at random.
}
\details{
This step in the game is when a contestant sees the amount of doors in the
game and chooses a door at random.  The function will return the door selection.
}
\examples{
select_door()

}
