% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Select a door.}
\usage{
select_door()
}
\arguments{
\item{Numeric}{}
}
\value{
The function returns a numeric value that indicates
which door was picked.
}
\description{
\code{select_door()} allows a contestant to choose 1 door.
}
\details{
The contestant chooses one door out of the three.
}
\examples{
  select_door()
}
