% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Contestant selects a door.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a numeric value indicating
the position the contestant is choosing.
}
\description{
\code{select_door()} generates a number that represents the
contestant's first door selection.
}
\details{
Contestant chooses from three closed doors from TV show,
"Let's Make a Deal." The doors have a car behind it and two
have goats. The contestant selects a door, then the host
opens a door to reveal a goat, and then the contestant is
given an opportunity to stay with their original selection
or switch to the other unopened door.
}
\examples{
  select_door()

}
