% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Select a door}
\usage{
select_door()
}
\value{
Returns a number between 1 and 3.
}
\description{
Randomly selects one of the three doors.
}
\details{
This is the contestants first choice before one of the
goat doors is opened.
}
\examples{
select_door()

}
