% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{Select a door.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The chosen door number (1, 2, or 3).
}
\description{
Simulates a contestant selecting a door from three available choices.
}
\details{
The \code{select_door()} function simulates the contestant's initial choice
in the Monty Hall Problem. It randomly selects one of the three available doors,
and the outcome represents the door number chosen by the contestant.
}
\examples{
  select_door()

}
