% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem functions Hauer.R
\name{select_door}
\alias{select_door}
\title{Contestant Selects a Door.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
One of the three elements (a door) is randomly returned as
the output.
}
\description{
\code{select_door()} selects a door for the player's first guess.
}
\details{
The function creates a vector of three elements- door1,
door2, and door3 and uses sample () to select one random
door and store this door in the variable 'a.pick'.
}
\examples{
select_door()

}
