% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{selects door}
\usage{
select_door(nDoor)
}
\arguments{
\item{nDoor}{number of doors is taken as argument.}
}
\value{
The function returns a the door choosen
}
\description{
`select_door()` selects a door out of n doors.
}
\details{
In this function we wrote code such that the function
  will choose a door out of n availavble door
}
\examples{
  select_door(10)
}
