% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{The user selects a door number.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a random number between 1 to 3 which the user
picks for the door choice.
}
\description{
\code{select_door()} generates a random number choice between 1 to 3.
}
\details{
The random choice is what the user picks as the door they believe contains
the car and the host will always show a goat door that is not related to
the door choice the user picked.
}
\examples{
  select_door()
  
}
