% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{Contestant randomly selects one of the three doors}
\usage{
select_door()
}
\arguments{
\item{no}{arguments are used by the function.}
}
\value{
'a.pick' randomly selects a sample of one of the three integers in the vector 'doors'
}
\description{
'doors' divides each of the items into three separate parts.
'sample()' randomly assigns each of the items to a door
and 'a.pick' randomly selects one of the three doors.
}
\details{
doors creates a vector of three integers
a.pick samples one of the three integers (1,2,or 3) as doors
}
\examples{
create(select_door)
}
