% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{Select a door.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a length 3 character vector
  indicating the door the contestant selected.
}
\description{
`select_door()` selects one of the three door options available to the player.
}
\details{
Like the Monty Hall game show the contestant is asked to pick a door.
}
\examples{
 select_door

}
