% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Contestant selects a door in Monty Hall Program game.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
An integer (1, 2, or 3) representing the door selected by the
contestant.
}
\description{
\code{select_door()} randomly selects a door to mimic a contestant making
their first selection in a Monty Hall Problem game.
}
\details{
The function randomly selects one of the three doors (labeled 1, 2, or 3)
that represents the contestant's choice. The selection is made by
generating a random number between 1 and 3. This is the first step in the
Monty Hall Problem game.
}
\examples{
  select_door()
}
