% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{Contestant makes their first pick!}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function}
}
\value{
The function returns a single integer value which is
stored in \code{a.pick}.
}
\description{
\code{select_door()} generates at random an integer between 1 and 3
inclusive, representing a door number in the game.
}
\details{
This follows the game setup again. The chosen door may or may not
have a car behind it. But the contestant has to make their initial
pick as per design.
}
\examples{
select_door()
}
