% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Contestant selects a door}
\usage{
select_door()
}
\arguments{
\item{There}{are no arguments used in this function.}
}
\value{
The contestant has the chose of 3 doors and can only choose 1 of the doors.
}
\description{
'select_door()' creates a function for the contestant to randomly chose one
of the three doors.
}
\details{
The 'create_game()' function generates the game by assigning two doors with
goats and one door with the car. The next step is for the contestant to
randomly select one of the doors - they will not know if they got a goat or
car.
}
\examples{
Per the select_door(), contestant can choose a door from 1-3. 
}
