% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem (1).R
\name{select_door}
\alias{select_door}
\title{Player selects first door}
\usage{
select_door()
}
\arguments{
\item{none}{used}
}
\value{
Return the first chosen door
}
\description{
Choose randomly from 1, 2 or 3
}
\details{
using a sample function, we pick an inital random door
}
\examples{
select_door()
}
