% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Select a Door}
\usage{
select_door()
}
\arguments{
\item{n/a}{}
}
\value{
Numeric value of door chosen
}
\description{
Randomly selects a door from the three available.
}
\details{
Contestants pick from one of three random doors, 2 have goats 1 with car
}
\examples{
  select_door()
}
