% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{select_door}
\alias{select_door}
\title{Select a door in the Monty Hall problem.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns the contestant's selection.
}
\description{
`select_door()` selects one of the three doors as the contestant's door.
}
\details{
Once the "Let's Make a Deal" game is set up with three doors,
  the contestant selects a door.
}
\examples{
  select_door()

}
