% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{select_door}
\alias{select_door}
\title{Contestant selects a door}
\usage{
select_door()
}
\arguments{
\item{.}{. . no arguments}
}
\value{
contestant's initial pick
}
\description{
Randomly selects a door
}
\details{
Randomly selects a door as initial pick for the contestant
}
\examples{
select_door()
}
