% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Select a random door in the Monty Hall game.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a single integer between 1 and 3, representing the door selected by the contestant.
}
\description{
\code{select_door()} randomly selects one of the three doors in the Monty Hall problem.
}
\details{
In the Monty Hall game, a contestant initially selects one of the three doors. This function simulates the contestant's first random selection of either goat or a car.
}
\examples{
`select_door()`
}
