% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Contestant Selects A Door}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The return will be the randomly selected door from the created
}
\description{
The function 'function()' creates a custom function to preform a task.
'sample' is used to select 1 option in the vector, randomly
}
\details{
This section creates a vector of three doors that the contestant
can choose from, and then randomizes the selection
}
\examples{
select_door
}
