% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty_hall_pkg1.R
\name{select_door}
\alias{select_door}
\title{Contestant selects a door.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a number between 1 and 3 indicating the
contestant's inital selection.
}
\description{
\code{select_door()} allows the contestant to select one of
the three doors in the game.
}
\details{
\code{select_door()} randomly selects one of the three doors created at the start
of the game, which represents the contestant's first selection.
}
\examples{
  select_door()

}
