% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{Randomly select a door from 1-3 in the game.}
\usage{
select_door()
}
\arguments{
\item{'a.pick'}{\itemize{
\item one door numbered either 1, 2, or 3, which is the contestant's
chosen door
'doors' - the numbers 1, 2, 3 that correspond to the the three doors in
the game
}}
}
\value{
The function returns 'a.pick', which is a number between 1 to 3 that
represents the door the contestant has chosen.
}
\description{
The 'select_door()' function randomly selects a number between 1 and 3,
which correspond to numbers on the doors in the game.
}
\details{
The contestant must randomly
select a door in order to begin playing the game. One door has a
car behind it, while two doors have goats. The 'select_door' function
serves to allow the "contestant" to choose one
of the three doors created by the 'create_game' function.
}
\examples{
    select_door()

}
