% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Select a Door}
\usage{
select_door()
}
\value{
An integer representing the selected door.
}
\description{
Simulates a contestant selecting one of three doors.
}
\details{
A random door numbered 1 through 3 is selected.
}
\examples{
select_door()

}
