% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg-Wickerham.R
\name{select_door}
\alias{select_door}
\title{The contestant selects one of the three doors.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function}
}
\value{
The function returns a numeric vector representing the door
selected by the contestant.
}
\description{
'select_door()' will return a number between 1 and 3. This number
represents the position of a door with either a goat or car
behind it. It is the contestant's first choice in the game.
}
\details{
This is the first step of the actual game. The contestant will
have the opportunity to switch their door number or keep the original
one after the host makes their choice.
}
\examples{
    select_door()
}
