% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem Script.R
\name{select_door}
\alias{select_door}
\title{Selecting a door}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a number between 1 and 3
}
\description{
\code{select_door()}Randomly select one door out of ( 1,2,3 ) set and return the
selection
}
\details{
Assume that a room is equipped with three doors. Behind two are goats,
and behind the third is a shiny new car. In this function the player is
picking a random door out of the three doors, and the game will continue
as the player will see below till he win or lose.
}
\examples{
  select_door()
 
}
