% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{Select Door}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The functions returns a length 3 integer from 1 to 3.
}
\description{
`select_door()` generates a vector of integers and randomly
  selects a door number for the game.
}
\details{
The select door functions creates a new vector called 'doors'.
  It then randomly selects a door number from 'doors'. It then
  stores the selection from 'doors' in a variable called 'a.pick'.
  Select_door returns the value stored in a.pick.
}
\examples{
  select_door()

}
