% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{The contestant selects a door.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a length 1 numeric vector
indicating the door position or number that one contestant
initially selects.
}
\description{
\code{select_door()} generates a door selection and the contestant
randomly chooses from.
}
\details{
This function creates a vector of doors numbered 1, 2, 3.
Three doors that one contestant can choose from. Then the
contestant randomly selects ONE of the doors. The function
returns the selection that the contestant has made, the door
number between 1 and 3.
}
\examples{
  select_door()

}
