% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{Contestant Selects a Door.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a length of one numeric vector, either
1, 2, or 3 and is saved as 'a.pick'.
}
\description{
\code{select_door()} generates an initial selection of a door by
the contestant through a numeric vector of 1, 2, and 3.
}
\details{
The contestant will make an initial selection of one of the
three doors of the game, not knowing if there is a goat or
car behind the door they select.
}
\examples{
  select_door()

}
