% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Select one of the three doors.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
A single integer is returned that represents the chosen door: 1, 2, or 3.
}
\description{
\code{select_door()} randomly selects one of the three doors in the Monty Hall
game.
}
\details{
This function simulates a contestant randomly choosing one of the three
doors (numbered 1, 2, or 3) in the Monty Hall problem.
}
\examples{
  select_door()
}
