% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Select a random door.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function}
}
\value{
A single numeric value (1, 2, or 3) representing the door chosen.
}
\description{
select_door() simulates the contestant's first action in the Monty Hall
problem, choosing one of the three doors at random.
}
\details{
The contestant does not know what is behind each door and has an equal
chance of selecting door 1, 2, or 3.
}
\examples{
select_door()

}
