% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{CONTESTANT SELECTS A DOOR}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function}
}
\value{
The function returns an integer representing the door number chosen by the contestant.
}
\description{
The contestant makes their first selection at random.
}
\details{
This function will simulate the contestant's initial choice in the Monty Hall game by selecting at random one of the three doors.
}
\examples{
 select_door()
}
