% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Determine contestant's first door choice.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a number between 1 and 3.
}
\description{
\code{select_door()} randomly one of the doors for the contestant.
}
\details{
Generally, you will not need to run this function yourself; it should be run automatically when you run a full simulation with \code{play_game()}.
}
\examples{
select_door()
}
