% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_monty-hall-problem (2).R
\name{select_door}
\alias{select_door}
\title{select a door for your first guess.}
\usage{
select_door()
}
\arguments{
\item{If}{the door selected reveals a car, the game ends.}
}
\value{
The function returns one character vector of either a goat or car
}
\description{
'select_door" randomly selects a door for the contestant
to reveal a goat or a car.
}
\details{
there are three doors for a contestant to choose from, one of
which has a car behind it and two have goats.
The contestant selects a door, then the host opens a door to reveal
 a goat or a car.
}
\examples{
select_door()

}
