% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{Contestant selects a door.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a numeric vector of length one
indicating the chosen door by contestant.
}
\description{
\code{select_door()} allows contestant to choose one  door out of the three available doors at random.
}
\details{
Contestant is asked to pick a door out of the three available doors.
Two of the doors are goat doors and only one of the doors is a car door.
}
\examples{
 select_door()

}
