% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem-brentlinger.R
\name{select_door}
\alias{select_door}
\title{Player randomly selects a door.}
\usage{
select_door()
}
\arguments{
\item{no}{arguments are used by the function.}
}
\value{
The function returns a length 1 number vector, indicating the door position of the one selected
}
\description{
\code{select_door()} generates a door selection from game created in \code{create_game()}
}
\details{
The door is randomly selected from the game created in \code{create_game()}
}
\examples{
select_door()

}
