% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Randomly select a door}
\usage{
select_door()
}
\value{
An integer value of 1, 2, or 3 indicating the selected door.
}
\description{
Randomly selects one of the three doors for the contestant.
}
\details{
The contestant has no information about what is behind each door,
so the selection is made uniformly at random.
}
\examples{
select_door()

}
