% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{Select one of the three doors.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a numeric vector 1-3.
}
\description{
\code{select_door()} generates a randome selection of an intial
door number of one through three.
}
\details{
\preformatted{As part of the original game, "Let's Make a Deal" contestants
pick an initial door that they think/hope the car is behind. 
Contestants pick one door number, 1 through 3. 
}
}
\examples{
    select_door()

}
