% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{Selecting a door}
\usage{
select_door()
}
\arguments{
\item{the}{first argument is the vector of doors to which is numbered from 1-3
in this example.Therefore Identifying the amount of doors that are going to
to be used in the game. Following, the object a.pick must be assigned
based on the logical values that are being pulled from the variable 'doors'
it is first encapsulated by the function 'sample' which is ultimately use
in calling a certain 'sample' from the vector or 'x' that is specific. In
this case, that vector is 'doors' which is then followed by the argument
size of 1 which identifies the amount that is being 'sampled'. This logical
statement is assigned to the a.pick object which is then being returned from
the 'select_door' function.}
}
\value{
return is the door that is randomly assigned to the variable 'a.pick' based
on the logic as describe before which in this case is limited to doors 1-3.
when return function is being called it is essentially asking 'what should
be displayed when 'select_door' function is called?' How this value is
returned is based on the arguments which is stated prior
}
\description{
\code{select_door()} is a function that will generate a door to be randomly
selected to which will then be identified as the users initial pick
}
\details{
details regarding this function is that there are 3 doors to choose from
to which only one must be sampled from the 3 thus becoming "a.pick" and
being returned as such.
}
\examples{
select_door()
}
