% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Simulates the contestant making initial door choice.}
\usage{
select_door()
}
\arguments{
\item{...}{No arguments are used by the function.}
}
\value{
The function returns an integer between 1 and 3.
}
\description{
'select_door()' will randomly select a door number between 1-3 using 'sample'.
}
\details{
This function is the first step after creating the game, with a 33.33\% chance of selecting
the winning door that will reveal the car.
}
\examples{
select_door()
}
