% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problemb.R
\name{select_door}
\alias{select_door}
\title{Select a door by the contestant.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a length 1 vector
indicating the positions of the contestant choice.
}
\description{
\code{select_door()} generates a contestant choice of a door.
}
\details{
The contestant needs to choose and make an initial pick
in the game, this randomizes the choice the contestant
makes.
}
\examples{
  select_door()
}
