% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Player selects an initial door}
\usage{
select_door(pick = NULL)
}
\arguments{
\item{pick}{Integer door index (1, 2, or 3). If \code{NULL}, picks at random.}
}
\value{
pick Integer in (1,2,3).
}
\description{
Player selects an initial door
}
\examples{
select_door()      # random
select_door(2)     # choose door 2
}
