% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{select_door}
\alias{select_door}
\title{Select a door within the Monty Hall Problem Game}
\usage{
select_door()
}
\arguments{
\item{No}{arguements are used by this function.}
}
\value{
The function returns an integer variable that
corresponds to the number of the door that is selected.
}
\description{
'select_door'chooses 1 door
within the Monty Hall Problem game
}
\details{
This function will select 1 door at random from 3
the possible doors available for a contestant's
initial selection.
}
\examples{
 select_door()
}
