% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOCUMENTATION.R
\name{select_door}
\alias{select_door}
\title{Select initial door}
\usage{
select_door()
}
\arguments{
\item{No}{arguments are used by the function}
}
\value{
The function returns a length 1 numerical vector indicating
the first door in play.
}
\description{
\code{select_door()} randomly selects and opens the first door in the
game
}
\details{
This is the first door the contestant chooses in the game.
}
\examples{
  select_door()
  
}
