% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{Door selection function by the contestant.}
\usage{
select_door()
}
\arguments{
\item{...}{the function uses no arguments.}
}
\value{
Numeric value between 1 and 3.
}
\description{
\code{select_door()} The contestant selects one door from 3 selections.
}
\details{
This function selects a random door from 3 selections.
}
\examples{
select_door() 

}
