% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.r
\name{select_door}
\alias{select_door}
\title{\if{html}{\out{<div class="sourceCode">}}\preformatted{Select a random door.
}\if{html}{\out{</div>}}}
\usage{
select_door()
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{The function returns an integer 1, 2, or 3. The integer
represents the door that was randomly selected as the contestant's
first choice. 
}\if{html}{\out{</div>}}
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Randomly selects one of three doors as the contestant's initial choice. 
}\if{html}{\out{</div>}}
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{This function simulates the contestant's first choice in the
Monty Hall problem,	where they choose one of three doors. The 
function chooses door one, two, or three at random. 
}\if{html}{\out{</div>}}
}
\examples{
	select_door()
}
