% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.r
\name{select_door}
\alias{select_door}
\title{Select a random door.}
\usage{
select_door()
}
\value{
The function returns an integer 1, 2, or 3. The integer
represents the door that was randomly selected as the contestant's
first choice.
}
\description{
Randomly selects one of three doors as the contestant's initial choice.
}
\details{
This function simulates the contestant's first choice in the
Monty Hall problem,	where they choose one of three doors. The
function chooses door one, two, or three at random.
}
\examples{
  select_door()
}
