% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{\preformatted{Select a door
}}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
the function returns a number between 1 and 3 to represent one of the three doors
}
\description{
\preformatted{`select_door()` selects one door of the three options
}
}
\details{
\preformatted{only one door can be selected, this will be the first door that is picked, 
}

the contestant will have a chance to change doors if desired later in the game.
}
\examples{
	select_door()

}
