% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{Randomly select one of three doors.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns an integer between 1 and 3.
}
\description{
\code{select_door()} selects one of three doors.
}
\details{
This is the contestant making their first choice from
among doors 1, 2, and 3.
}
\examples{
  select_door()

}
