% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{placeholder}
\usage{
select_door()
}
\arguments{
\item{placeholder}{}
}
\value{
placeholder
}
\description{
placeholder
}
\details{
placeholder
}
\examples{
placeholder
}
