% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Simulate contestant's inital door choice.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns an integer between 1 and 3.
}
\description{
\code{select_door()} uses sample to select a number between 1-3,
and that number corresponds to one of the three doors.
}
\details{
This is the frist step in the game where the contestant chooses one
of three doors, where they have a 1 in 3 chance of choosing the door with the car.
}
\examples{
select_door()

}
