% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem-R-script.R
\name{select_door}
\alias{select_door}
\title{The contestant selects their initial door.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns an integer representing the number
of the door that the participant selects. Returns a value
which can be stored in a variable called \code{firstDoor}.
Example output: 1
}
\description{
\code{select_door()} Type: Integer -> generates a random door number
from 1 to 3.
}
\details{
The selected doors can be 1, 2, or 3. The door
may or may not have a car behind it.
}
\examples{
  select_door()

}
