% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg-Spino.R
\name{select_door}
\alias{select_door}
\title{Select an Initial Door}
\usage{
select_door()
}
\value{
A single integer between 1 and 3 identifying the selected door.
}
\description{
Randomly select one of the three doors.
}
\details{
No arguments are required. Each door has an equal chance of
  selection, independently of the prize locations.
}
\examples{
pick <- select_door()
pick
stopifnot(length(pick) == 1, pick \%in\% 1:3)
}
