% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{Select a door at random in a Monty Hall game}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by this function}
}
\value{
The function returns a numeric value between 1 and 3
}
\description{
Randomly generates a number between 1 and 3, representing the door number
selected by the contestant
}
\details{
The door selected by the contestant will be used later in the game and
presented as a choice, stay or switch and win the prize behind the final
door selected
}
\examples{
select_door()
}
