% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg-3.R
\name{select_door}
\alias{select_door}
\title{\if{html}{\out{<div class="sourceCode">}}\preformatted{Have the game "player" select the door
}\if{html}{\out{</div>}}}
\usage{
select_door()
}
\arguments{
\item{No}{arguments are used by the function.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{The function returns a pick variable which contains either the integer 1, 2, or 3
}\if{html}{\out{</div>}}
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{select_door() represents the player selecting a door numbered from 1 through 3 inclusive. 
}\if{html}{\out{</div>}}
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{The function generates a random pick between 1 and 3. This represents the contestant's choice in initially selecting their door. 
}\if{html}{\out{</div>}}
}
\examples{
		select_door()
}
