% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{The contestant makes their first selection.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a variable named 'a.pick'.
}
\description{
'select_door()' creates a vector of doors numbered 1,2,3. 
   After the contestant selects the one of doors randomly, 
   the function reuturns the variable (a.pick).
}
\details{
This function creates a vector named(doors). The contestant
   selects a number and it creates a variable which is randomly 
   assigned from "doors". The selected door by the contestant
   which is between 1 and 3 is returned.
}
\examples{
    select_door()
    select_door()
    select_door()
}
