% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg copy 2.R
\name{select_door}
\alias{select_door}
\title{Have the game "player" select the door}
\usage{
select_door()
}
\arguments{
\item{No}{arguments are used by the function.}
}
\value{
The function returns a variable \code{a.pick} containing either the integer 1, 2, or 3.
}
\description{
select_door() represents the player selecting a door numbered 1 through 3.
}
\details{
The function generates a random pick between 1 and 3. This represents the
contestant's initial door choice.
}
\examples{
  select_door()

}
