% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem (1).R
\name{select_door}
\alias{select_door}
\title{Contestant Selects a Door}
\usage{
select_door()
}
\arguments{
\item{…}{no arguments are used by the function.}
}
\value{
The return for the function is the randomly selected door by the “contestant”.
}
\description{
The contestant is given the option to choose between three doors.
}
\details{
The select door function consists of the doors and the contestant’s choice.
The doors for the game has to be created and identified as 1, 2, or 3. After
the doors are created, the contestant is able to choose which door they want.
Using the code “sample” allows the computer to randomly select which door is
selected for each game.
}
\examples{
select_door ()

}
