% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{select_door}
\alias{select_door}
\title{Select an initial door}
\usage{
select_door()
}
\value{
Integer 1, 2, or 3.
}
\description{
Randomly select one of doors 1, 2, or 3 as the player's first pick.
}
\examples{
  set.seed(1); select_door()
}
