% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{selecting one of the three door randomly}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a length of 1 character vector
indicating the selected door
}
\description{
\code{select_door()}generate a random door
}
\details{
this function will represent the choice of the participant where he
will have equal chances to select a door
}
\examples{
select_door()

}
