% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall_script_rj.R
\name{select_door}
\alias{select_door}
\title{:
Randomly select on of three doors.}
\usage{
select_door()
}
\arguments{
\item{:}{The door number returned is a numeric value.}
}
\value{
:
"a.pick" is set as the return since this displays which door was selected.
}
\description{
:
This function creates a choice of three doors. It then randomly picks one door and reveals the door number.
}
\details{
:
The first assignment, "doors", creates a vector with three numbers. The next assignment, "a.pick" then chooses a random sample of one of the doors. The first position in the parentheses determines what vector to pick from. The next position, "size=1" determines how many choices to select. Since we only want to open one door, the size is set to 1. Finally, we return the door that was selected.
}
\examples{
:
  select_door()
}
