% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty_hall_scripts.R
\name{select_door}
\alias{select_door}
\title{Contestant selects a door.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
This function returns a single object containing the position of the selected door: 1, 2, or 3.
}
\description{
\code{select_door()} generates a random door position to open.
}
\details{
In the Monty Hall Problem there are three doors. A car is behind one door and goats are behind
the other two doors. The contestant must pick a door to open. This function simulates the
contestant selecting one of the three doors.
}
\examples{
select_door()

}
