% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{Initial Door Selection}
\usage{
select_door()
}
\arguments{
\item{No}{arguments are used by the function.}
}
\value{
The function returns one number corresponding to a door, which will either be 1, 2, or 3.
}
\description{
'select_door' creates a function that chooses the initial pick of the contestant
}
\details{
This function replicates the contestant initially picking a door behind which they believe is a car.
}
\examples{
select_door()
}
