% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Select a door for the Monty Hall game.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
A numeric value representing the door selected by the
contestant. The value will be 1, 2, or 3.
}
\description{
\code{select_door()} randomly selects one of the three doors in the
Monty Hall game.
}
\details{
The function randomly selects one door from the three available
doors. The selected door represents the contestant's initial choice.
}
\examples{
  select_door()

}
