% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Select a door in the Monty Hall game}
\usage{
select_door()
}
\value{
An integer between 1 and 3, representing the selected door number.
}
\description{
`select_door()` simulates the contestant's initial door selection in the Monty Hall Problem.
}
\details{
This function randomly selects one of the three doors, mimicking a contestant's
initial choice in the Monty Hall Problem. The selection is made without any
knowledge of what's behind the doors, simulating the blind choice a real
contestant would make.
}
\examples{
# Select a door
select_door()

# Make multiple selections
replicate(5, select_door())

}
