% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Game player selects a door from the three doors}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
This function returns a number between 1 and 3, representing the selected door.
}
\description{
\code{select_door()} chooses one of the three doors created in \code{create_game()} at random
}
\details{
The function randomly selects one of three doors, mimicking the initial choice
made by a contestant in the Monty Hall game. This selection represents the first
step in the game show scenario, where the contestant must choose a door without
any prior information about what lies behind it. The function uses a random number
generator to ensure an unbiased selection among the three available options,
labeled as doors 1, 2, and 3. This random selection is crucial for simulating the
contestant's initial guess in the Monty Hall problem, allowing for further
exploration of the game's probability dynamics and the effectiveness of different
strategies (staying with the initial choice or switching) when given the opportunity
to change doors.
}
\examples{
  select_door()
  
}
