% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Player selects a door}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by this function.}
}
\value{
A numeric value representing the contestant selection is returned as the vector index of 'create_game()'
}
\description{
'select_door()' randomly selects one of the three doors from 'create_game'.
}
\details{
This step simulates the constesant responding to the offer of three doors
by selecting the one they hope will reveal the car.  The constestant has
no prior knowledge of the prize locations.  The selection uses a random
sample choice.  The sample seed can be set by adding set.seed(seed value)
before the calling the function.
}
\examples{
  select_door()
  Output: 2

}
