% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{select_door}
\alias{select_door}
\title{Select Door Function.}
\usage{
select_door()
}
\arguments{
\item{No}{arguments are used by the function.}
}
\value{
The function returns a length one character vector which is a number
between 1 and 3 and represents the number of the initial door selected.
The returned value is assigned to the variable \code{a.pick}.
}
\description{
\code{select_door()} is the function which simulates the user selecting a
random door.
}
\details{
The initial door selection simulates the player choosing one of three
closed doors.  There are goats behind two of the doors and a car
behind one of the doors.  At this point in the game all three doors are
still closed so the player doesn't know if they picked the car door or
a goat door.
}
\examples{
  select_door()

}
