% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{select_door}
\alias{select_door}
\title{CONTESTANT SELECTS A DOOR.}
\usage{
select_door()
}
\arguments{
\item{no}{arguments are used by the function.}
}
\value{
The function returns a 1/3 chance to choose a Car and 2/3 chance
to choose a goat.
}
\description{
\code{select_door()} The contestant makes their first selection.
}
\details{
The contestant makes their first selection of a random door from
Doors 1-3, with 1/3 chance to choose a Car and 2/3 chance to
choose a goat.
}
\examples{
select_door()

}
