% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{Contestant Selects a door - Function}
\usage{
select_door()
}
\arguments{
\item{Three}{numbers are placed on a vector (representing each door, 1, 2, and 3).}
}
\value{
The game returns the selection (pick) of the contestant.
Door: 1 goat
Door: 2 goat
Door: 3 door
}
\description{
In this function of the program, de contestant choose their first door.
}
\details{
The function will create a vector with 3 doors. ONE door is randomly selected.
}
\examples{
 select_door( )
}
