% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{Contestant's First Selection}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
'a.pick' will return either 1,2 or 3
}
\description{
'select_door()' generates the contestant's first door selection
}
\details{
The function records the contestant's first selection, removing it from
selections in our future functions. The contestant will either choose
doors 1-3 which were generated in our 'create_game()' function.
}
\examples{
'select_door()'
}
