% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_game.R
\name{select_door}
\alias{select_door}
\title{Select a door}
\usage{
select_door()
}
\value{
A number between 1 and 3 representing the
contestant's selected door.
}
\description{
`select_door()` randomly selects one of the three doors
as the contestant's initial choice.
}
\examples{
select_door()

}
