% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{Contestant Selects A Door}
\usage{
select_door()
}
\arguments{
\item{The}{select_door function calls upon a generic function with no arguments.
The doors function uses logical vectors 1-3.
The a.pick function uses the sample function to randomly generate a logical vector
taken from the doors function.}
}
\value{
The function returns a length three character vector that will represent
the door number. The vectors can only be : 1, 2, or 3
}
\description{
This block of code will randomly select ONE of the three doors
and return the result as the selection.
}
\details{
Since the contestant will not know the position of the car when
they select a door, we do not need to share the information about the
game set-up before the selection is made. We will create a vector
containing three integers to represent a door number.
}
\examples{
  1
  2 
  3

}
