% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg-finished.R
\name{select_door}
\alias{select_door}
\title{Create a function to select an initial door}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a number between 1 and 3 that corresponds
with a door number
}
\description{
\code{select_door()} generates a new function that allows a contestant to pick
one of three doors in the game
}
\details{
According to the rules, to get the game started, a contestant
may choose any of the three doors in the game with no limitations
placed on which one is chosen. the chosen door is then kept a secret,
so the contestant does not know if they have chosen a losing goat door
or a winning car door.
}
\examples{
  select_door()
  
}
