% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{Contestant selects a door out of three available.}
\usage{
select_door()
}
\arguments{
\item{No}{arguments are used by the function.}
}
\value{
The function returns a numberic vector of length one, incdicating
 the door number the contestant has chosen.
}
\description{
`select_door()` allows the user to makes a choice of a door
 "1,2, or 3"
}
\details{
The select_door() let users pick a door. In this function, one of the door
 out of three is selected at random and the choice is stored in a vector.
 At this point the contestant does not know what is behind the door s/he
 has selceted.
}
\examples{
 select_door()

}
