% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Contestant selects the door they believe the car is behind.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by this function}
}
\value{
This function returns a single numeric value representing the door randomly selected by the contestant and
ranging between 1 and 3.
}
\description{
In line with the game show flow, the contestant selects the doors they
believe the car is behind. The select_door function allows the contestant to randomly select one of the
available doors
}
\details{
This function serves as the contestants guess as to which door holds the car. It randomly selects a door from 1-3.
It works by calling the 'sample()' on a vector of doors and returns the value of the door number the contestant chose.
}
\examples{
select_door()
}
