% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Pick the player's initial door.}
\usage{
select_door()
}
\arguments{
\item{This}{function uses no arguments.}
}
\value{
Integer. The function returns a single number between 1 and 3.
}
\description{
\code{select_door()} randomly returns one of the three doors.
}
\details{
This door is selected randomly to represent the player's lack of
knowledge surrounding which door is which.
}
\examples{
    select_door()

}
