% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monty Hall.package.R
\name{select_door}
\alias{select_door}
\title{Randomly select a door}
\usage{
select_door()
}
\arguments{
\item{No}{arguments are used by this function.}
}
\value{
An integer between 1 and 3 indicating the door selected.
}
\description{
\code{select_door()} randomly selects one of the three doors.
}
\details{
Simulates the contestant's initial door choice
}
\examples{
select_door ()

}
