% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Select a Door}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
A number between 1 and 3 representing the selected door.
}
\description{
Randomly selects one of the three doors.
}
\details{
The function picks door 1, 2, or 3 at random.
}
\examples{
select_door()

}
