% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{#'   Contestant will select a door for the Monty Hall Problem game.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a length 1 numerical vector
indicating the initial pick of the contestant.
}
\description{
#'   \code{select_door()} randomly chooses a door among the 3 doors.
}
\details{
#'   The game setup has 3 possibilities, so the contestant must choose
1 of 3 doors.
}
\examples{
  select_door()
  
}
