% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Select the initial picked door for the game.}
\usage{
select_door()
}
\arguments{
\item{no}{arguments are used for the function}
}
\value{
the function returns a number between one and three
}
\description{
\code{select_door()} selects a value to represent which door is picked in the Monty Hall game
}
\details{
This function selects the door for the game by sampling from a set of 1-3
}
\examples{
  select_door)
}
