% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem edited.R
\name{select_door}
\alias{select_door}
\title{The Contestant Makes and Initial Choice.}
\usage{
select_door()
}
\arguments{
\item{No}{arguments are used in this function.}
}
\value{
The function returns a single numeric value from the vector (1, 2, 3).
}
\description{
\code{select_door} creates a vector for the 3 doors in the game, (1, 2, 3),
and randomly selects one to represent the contestant's intial choice.
}
\details{
This function first creates a sample of the three doors,
a numeric vector of (1, 2, 3). The items ordered in create_game
are hidden behind these doors. Then it randomly samples one of the
doors as the initial choice of the contestant. At this point, only
the host knows what is behind the door.
}
\examples{
  select_door()

}
