% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Select a door}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns the picked door.
}
\description{
\code{select_door} Randomly chooses a door between 1-3.
}
\details{
Randomly selecting a door is similar to Lets make a deal when the contestant chooses a door of their liking.
}
\examples{
select_door()
}
