% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{Chooses a door.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a number of the door that was selected
}
\description{
`select_door()` presents 3 doors and chooses one at random.
}
\details{
There are three doors for a contestant
  to choose from, one of which has a car behind it and two 
  have goats. The contestant selects a door.
}
\examples{
  select_door()

}
