% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{Contestant door selection}
\usage{
select_door()
}
\arguments{
\item{this}{door is used to determine which other door should be opened by the host to reveal the goat behind it}
}
\value{
return a number 1, 2, or 3.
}
\description{
Here the contestant chooses 1/3 doors
}
\details{
first door selection
}
\examples{
: 2
}
