% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montyhall.R
\name{select_door}
\alias{select_door}
\title{Select a random door between 1 and 3}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a number between 1 and 3
}
\description{
\code{select_door()} generates 3 doors and randomly selects one
}
\details{
The selection serves as the initial door choice for the contestant
in the Monty Hall game.
}
\examples{
  select_door()

}
