% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\name{select_door}
\alias{select_door}
\title{The player select his first door}
\usage{
select_door()
}
\arguments{
\item{No}{parameters used}
}
\value{
Return the first picked door
}
\description{
We choose a random door as a first selection
}
\details{
We use sample function to return a picked door
}
\examples{
  select_door()
  
}
