% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.r
\name{select_door}
\alias{select_door}
\title{Select Door}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a numeric value between 1 and 3,
  representing the initial pick.
}
\description{
`select_door()` represents the door that the contestant
  has chosen as their initial pick
}
\details{
This returns the initial pick of contestant.
}
\examples{
  select_door()

}
