% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{Player selects a door}
\usage{
select_door()
}
\arguments{
\item{The}{function uses the object doors, which is a three character
numeric vector. a.pick randomly selects one of the doors from
the previous vector. The function return() is used to produce
the selected door as one character (1, 2, or 3)}
}
\value{
The function returns 1 character indicating
the door selection
}
\description{
'select_door()' picks which door the player will select
without knowing what is behind the door.
}
\details{
The first step of the game once it is set up is for the
player to select a door. Later in the game, the player
will have the opportunity to stay with this choice or
switch their selection. They will receive more information
about the doors before needing to make a decision.
}
\examples{
select_door()
}
