% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{Selects a door.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a numerical value between 1 and 3.
}
\description{
`select_door` picks a number from 1 to 3,
  representing the three doors.
}
\details{
First, the variable `doors` is created, holding the
  numbers 1 through 3. Then one of the numbers is
  selected randomly and put into the variable `a.pick`
  which is returned when the function is called.
}
\examples{
  select_door()

}
