% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-script.R
\name{select_door}
\alias{select_door}
\title{Select Door}
\usage{
select_door()
}
\arguments{
\item{None}{}
}
\value{
The function will return one door that is chosen by the contestant.
}
\description{
Contest chooses 1 door.
}
\details{
Contest will only be able to choose 1 door no matter how many options
are available.
}
\examples{
select_door()

}
