% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Step 2: Contestant selects a door}
\usage{
select_door()
}
\arguments{
\item{...}{NO arguments}
}
\value{
returns an integer
}
\description{
The contestant makes their first selection
}
\details{
this function receives one positional
argument and it should be the return value of
the create_game function
}
