% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem.R
\name{select_door}
\alias{select_door}
\title{Host selected one of the doors before the game start.}
\usage{
select_door()
}
\arguments{
\item{No}{arguments are used by the function.}
}
\value{
The function returns a numeric vector
}
\description{
\code{select_door()} selected a door for the game to start.
}
\details{
The host selected a door among the three doors that are closed. The
host should select a door without a car.
}
\examples{
  select_door()

}
