% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Player selects a door}
\usage{
select_door(door)
}
\arguments{
\item{door}{Integer in 1:3, the chosen door number.}
}
\value{
Integer in 1:3, the selected door number.
}
\description{
Player makes the initial choice of one of the three doors.
}
\examples{
pick <- select_door(1)
pick
}
