% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Player selects a door}
\usage{
select_door()
}
\arguments{
\item{...}{No arguments are required}
}
\value{
An integer between 1 and 3 representing the selected door
}
\description{
Randomly selects one of three doors
}
\details{
Simulates a player making an initial choice between three doors
}
\examples{
select_door()

}
