% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{Randomly select one door}
\usage{
select_door()
}
\arguments{
\item{...}{No arguments are used by this function.}
}
\value{
An integer \code{1}, \code{2}, or \code{3}, representing
the chosen door.
}
\description{
\code{select_door()} chooses exactly one door among the three
with equal probability.
}
\details{
Since there are 3 doors labeled 1, 2, 3, the function
simply samples one door index with uniform probability.
}
\examples{
  select_door()

}
