% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newscriptmontyhall.R
\name{select_door}
\alias{select_door}
\title{Inidicate which door number was selected by contestant.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a one numerical number indicating the chosen door.
}
\description{
select_door()generates the door number that was chosen by the contestant.
}
\details{
The contestant makes their first selection and they have to choose one of the three doors where there are two goats and one car hiding.
}
\examples{
select_door()

}
