% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\docType{data}
\name{this.game}
\alias{this.game}
\title{UNIT TEST A
can only be door 3}
\format{
An object of class \code{character} of length 3.

An object of class \code{character} of length 3.

An object of class \code{character} of length 3.

An object of class \code{character} of length 3.
}
\usage{
this.game

this.game

this.game

this.game
}
\description{
UNIT TEST A
can only be door 3

can only be door 1

UNIT TEST B

UNIT TEST C
}
\keyword{datasets}
